### 2025-08-04 #8701 @editorialbot
[PRE REVIEW]: PFD Toolkit: Unlocking Prevention of Future Death Reports for Research
**Submitting author:** @Sam-Osian (<a href="http://orcid.org/0009-0000-7644-6218">Samuel Osian Andrews</a>)
**Repository:** https://github.com/Sam-Osian/PFD-toolkit
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.3
