+++ mkdocs.yml
@@ -1,6 +1,6 @@
 site_name: PFD Toolkit
 site_description: "Turn raw PFD reports into actionable insights"
-site_url: https://pfdtoolkit.org
+site_url: https://pfdtoolkit.org/
 repo_url: https://github.com/Sam-Osian/PFD-toolkit/
 edit_uri: edit/main/docs/
 theme:
