+++ mkdocs.yml
@@ -1,5 +1,5 @@
 site_name: PFD Toolkit
-site_description: "Turn raw PFD reports into actionable insights"
+site_description: "AI toolkit for analysing Prevention of Future Deaths (PFD) reports."
 site_url: https://pfdtoolkit.org/
 repo_url: https://github.com/Sam-Osian/PFD-toolkit/
 edit_uri: edit/main/docs/
