+++ pyproject.toml
@@ -60,6 +60,7 @@ dev = [
     "jupyter-core>=5.8.1",
     "ruff>=0.12.0",
     "twine>=6.1.0",
+    "openpyxl>=3.1.5",
 ]
 
 [project.urls]
