+++ tests/model_comparison.py
@@ -1,0 +1,141 @@
+"""Compare Screener performance across multiple LLM models against ONS consensus labels.
+
+This script reads report sections and ground-truth labels directly from
+``ons_replication/ONS_master_spreadsheet.xlsx`` and measures the accuracy,
+sensitivity, specificity, and elapsed time for several LLM models. Results are
+written to ``model_comparison.txt``.
+"""
+
+from __future__ import annotations
+
+import os
+import time
+from pathlib import Path
+
+import pandas as pd
+from dotenv import load_dotenv
+
+from pfd_toolkit import LLM, Screener
+from pfd_toolkit.config import GeneralConfig
+
+# ---------------------------------------------------------------------
+# 1. Load ONS ground-truth data
+# ---------------------------------------------------------------------
+ons_path = (
+    Path(__file__).resolve().parent.parent
+    / "ons_replication"
+    / "ONS_master_spreadsheet.xlsx"
+)
+ons_df = pd.read_excel(ons_path, sheet_name=0)
+
+ons_df = ons_df.rename(
+    columns={
+        "Investigation section": GeneralConfig.COL_INVESTIGATION,
+        "Circumstances of death section": GeneralConfig.COL_CIRCUMSTANCES,
+        "Matters of concern section": GeneralConfig.COL_CONCERNS,
+        "Consensus": "consensus",
+        "Ref": GeneralConfig.COL_ID,
+    }
+)
+
+keep_cols = [
+    GeneralConfig.COL_ID,
+    GeneralConfig.COL_INVESTIGATION,
+    GeneralConfig.COL_CIRCUMSTANCES,
+    GeneralConfig.COL_CONCERNS,
+    "consensus",
+]
+reports = ons_df[keep_cols].copy()
+
+reports["consensus"] = (
+    reports["consensus"]
+    .astype(str)
+    .str.strip()
+    .str.lower()
+    .map({"yes": True, "no": False, "1": True, "0": False})
+)
+
+print(f"Loaded {len(reports)} reports from ONS spreadsheet.")
+
+# ---------------------------------------------------------------------
+# 2. Initialise models and query
+# ---------------------------------------------------------------------
+load_dotenv(Path(__file__).resolve().parent.parent / "api.env")
+
+# List of models and the temperature values to use for each.
+# GPT-5 models do not support a custom temperature parameter, so while a
+# default value is recorded here for completeness, it is not passed to the
+# client when initialising those models.
+models = [
+    {"name": "gpt-4.1", "temperature": 0},
+    {"name": "gpt-4.1-mini", "temperature": 0},
+    {"name": "gpt-4.1-nano", "temperature": 0},
+    {"name": "gpt-5", "temperature": 1},
+    {"name": "gpt-5-mini", "temperature": 1},
+    {"name": "gpt-5-nano", "temperature": 1},
+]
+
+user_query = """
+Where the deceased is 18 or younger *at the time of death* AND the death was due to suicide.
+Age may not be explicitly noted, but could be implied through recent use of child services (e.g. CAMHS),
+mention of being "in Year 10", etc.
+"""
+
+# ---------------------------------------------------------------------
+# 3. Run screener for each model and compute metrics
+# ---------------------------------------------------------------------
+out_path = Path(__file__).resolve().parent / "model_comparison.txt"
+with out_path.open("w", encoding="utf-8") as fh:
+    for spec in models:
+        model = spec["name"]
+        temp = spec["temperature"]
+        print(f"Testing model: {model}")
+
+        llm_kwargs = {
+            "api_key": os.getenv("OPENAI_API_KEY"),
+            "max_workers": 10,
+            "model": model,
+            "seed": 12345,
[…]
