### 2025-08-16 526db90
Add functionality to skip already tested models in performance comparison
