### 2025-08-16 4aa782b
Model comparison allow for custom kwargs
