### 2025-08-17 a3d71c7
Add gemma and llava tests
