+++ tests/model_comparison.py
@@ -67,20 +67,50 @@ load_dotenv(Path(__file__).resolve().parent.parent / "api.env")
 # client when initialising those models.
 models = [
     {"name": "gpt-4.1", "temperature": 0},
+    {"name": "gpt-4o", "temperature": 0},
     {"name": "gpt-4.1-mini", "temperature": 0},
     {"name": "gpt-4.1-nano", "temperature": 0},
     {"name": "gpt-5", "temperature": 1},
     {"name": "gpt-5-mini", "temperature": 1},
     {"name": "gpt-5-nano", "temperature": 1},
+    {"name": "o4-mini", "temperature": 1},
     {
-        "name": "olmo2:13b",
+        "name": "gemma3:12b", 
         "temperature": 0,
         "base_url": "http://localhost:11434/v1",
         "api_key": "ollama",
         "timeout": 10**9,
-    },
+     },
+    # {
+    #     "name": "gpt-oss:20b", 
+    #     "temperature": 0,
+    #     "base_url": "http://localhost:11434/v1",
+    #     "api_key": "ollama",
+    #     "timeout": 10**9,
+    #  },
     {
-        "name": "gemma3:12b", 
+        "name": "gemma2:27b", 
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+     },
+    {
+        "name": "qwen3:30b", 
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+     },
+    {
+        "name": "llava:34b", 
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+     },
+    {
+        "name": "gemma3:27b", 
         "temperature": 0,
         "base_url": "http://localhost:11434/v1",
         "api_key": "ollama",
