+++ docs/index.md
@@ -1,6 +1,6 @@
 ---
-title: PFD Toolkit - AI toolkit for analysing Prevention of Future Deaths (PFD) reports
-description: PFD Toolkit - AI toolkit for analysing Prevention of Future Deaths (PFD) reports
+title: PFD Toolkit
+description: AI toolkit for analysing Prevention of Future Deaths (PFD) reports
 image: assets/header.png
 ---
 
