### 2025-08-18 3c0da74
Add phi, llama and qwen to model comparison
