### 2025-08-19 55c7cce
Add _strip_json_markdown function and tests for markdown-wrapped JSON
