### 2025-08-19 9464484
Add OpenRouter testing
