### 2025-08-19 a2c2507
Expand OpenRouter testing suite
