### 2025-08-19 649f65f
Switch model comparison to csv & include local and params info
