+++ tests/model_comparison.py
@@ -3,7 +3,7 @@
 This script reads report sections and ground-truth labels directly from
 ``ons_replication/ONS_master_spreadsheet.xlsx`` and measures the accuracy,
 sensitivity, and specificity for several LLM models. Results are written to
-``model_comparison.txt``.
+``model_comparison.csv``.
 """
 
 from __future__ import annotations
@@ -79,9 +79,9 @@ models = [
     {"name": "gpt-5-mini", "temperature": 1},
     {"name": "gpt-5-nano", "temperature": 1},
     {"name": "o4-mini", "temperature": 1},
-    
-    # OPENROUTER API MODELS
+    {"name": "o3", "temperature": 1},
 
+    # OPENROUTER MISTRAL MODELS
     {
         "name": "mistralai/devstral-small",
         "temperature": 0,
@@ -94,7 +94,6 @@ models = [
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
-    
     {
         "name": "mistralai/mistral-medium-3.1",
         "temperature": 0,
@@ -113,6 +112,8 @@ models = [
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
+
+    # OPENROUTER OTHER MODELS
     {
         "name": "google/gemma-3-4b-it",
         "temperature": 0,
@@ -125,162 +126,156 @@ models = [
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
-    
-    # {
-    #     "name": "deepseek/deepseek-r1-0528-qwen3-8b:free",
-    #     "temperature": 0,
-    #     "base_url": OPENROUTER_URL,
-    #     "api_key": OPENROUTER_KEY,
-    # },
-    
     {
         "name": "google/gemini-2.5-flash",
         "temperature": 0,
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
-    
     {
         "name": "google/gemini-2.0-flash-001",
         "temperature": 0,
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
-    
     {
         "name": "deepseek/deepseek-chat-v3-0324",
         "temperature": 0,
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
-    
     {
         "name": "moonshotai/kimi-k2",
         "temperature": 0,
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
-    
-    # {
-    #     "name": "openai/gpt-oss-120b",
-    #     "temperature": 1,
-    #     "base_url": OPENROUTER_URL,
-    #     "api_key": OPENROUTER_KEY,
-    # },
-    
     {
         "name": "qwen/qwen3-235b-a22b-2507",
         "temperature": 0,
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
-    
     {
         "name": "meta-llama/llama-4-maverick",
         "temperature": 0,
         "base_url": OPENROUTER_URL,
         "api_key": OPENROUTER_KEY,
     },
-    
-    # Local Ollama models (all quantised)
     {
-        "name": "gemma3:12b", 
+        "name": "x-ai/grok-3-mini",
+        "temperature": 0,
+        "base_url": OPENROUTER_URL,
+        "api_key": OPENROUTER_KEY,
+    },
+    {
+        "name": "x-ai/grok-4",
+        "temperature": 0,
+        "base_url": OPENROUTER_URL,
+        "api_key": OPENROUTER_KEY,
+    },
+    {
+        "name": "x-ai/grok-3",
[…]
