### 2025-08-20 cab3b1c
Scatter shows all local models' params and accuracy on ons research comp w/ GPT 4.1
