+++ tests/model_comparison_vis.py
@@ -1,0 +1,176 @@
+#!/usr/bin/env python3
+
+from __future__ import annotations
+
+import argparse
+from pathlib import Path
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+from adjustText import adjust_text
+import sys
+from matplotlib.ticker import PercentFormatter
+
+# ----------------- Tuning knobs -----------------
+X_PAD_RIGHT = 10.0
+X_MIN_PAD   = 2.0
+INIT_DX     = 1    # smaller horizontal offset
+INIT_DY     = 0.00   # smaller vertical jitter
+
+# Scatter jitter to spread near-duplicates
+JITTER_X    = 0.001      # billions
+JITTER_Y    = 0.0005    # accuracy units
+
+# Strength of collision handling (bigger → more space)
+EXP_TEXT    = (1.02, 1.55)  # expand text boxes in x,y during solving
+EXP_POINTS  = (1.05, 1.20)  # expand point boxes in x,y
+FORCE_TEXT  = 1
+FORCE_POINTS= 0.4
+N_STEPS     = 1000          # iterations for the solver (balance quality/speed)
+
+# Arrow aesthetics
+ARROW = dict(arrowstyle='-', color='lightgray', lw=1, alpha=0.7)
+
+# Short names for models displayed in labels
+SHORT_LABELS = {
+    "mistralai/devstral-small": "DevStral Small",
+    "mistralai/mistral-large-2411": "Mistral Large",
+    "google/gemma-3-4b-it": "Gemma 3 (4B)",
+    "deepseek/deepseek-chat-v3-0324": "DeepSeek v3",
+    "moonshotai/kimi-k2": "Kimi K2",
+    "qwen/qwen3-235b-a22b-2507": "Qwen3 A22B",
+    "meta-llama/llama-4-maverick": "Llama 4 Mav.",
+    "mistral-nemo:12b": "Mistral Nemo",
+    "mistral-small:22b": "Mistral Small (22B)",
+    "mistral-small:24b": "Mistral Small (24B)",
+    "gemma3:12b": "Gemma 3 (12B)",
+    "gemma3:27b": "Gemma 3 (27B)",
+    "gemma2:27b": "Gemma 2 (27B)",
+    "qwen3:32b": "Qwen3 (32B)",
+    "qwen3:30b": "Qwen3 (30B)",
+    "qwen2.5:72b": "Qwen 2.5 (72B)",
+    "qwen2.5:32b": "Qwen 2.5 (32B)",
+    "llava:34b": "LLaVA",
+    "phi4:14b": "Phi-4",
+    "llama3:70b": "Llama 3",
+    "google/gemma-3-12b-it": "Gemma 3 (12B)",
+    "cohere/command-a": "Command A",
+    "mistralai/codestral-2508": "Codestral"
+}
+
+# Baseline (non-local) references to show as horizontal lines if present
+FRONTIER_NAMES = {"gpt-4.1": "GPT-4.1"}
+                #   "gpt-5": "GPT-5", "x-ai/grok-4": "Grok-4"}
+
+
+def load_data(csv_path: str):
+    df = pd.read_csv(csv_path)
+    df["local"] = df["local"].astype(str).str.lower()
+
+    local_df = df[df["local"] == "yes"].copy()
+    local_df["params"] = pd.to_numeric(local_df["params"], errors="coerce")
+    local_df["accuracy"] = pd.to_numeric(local_df["accuracy"], errors="coerce")
+    local_df = local_df.dropna(subset=["params", "accuracy"]).reset_index(drop=True)
+
+    # Jitter points slightly to reveal duplicates
+    rng = np.random.default_rng(123)
[…]
