+++ pyproject.toml
@@ -25,6 +25,7 @@ classifiers = [
 
 requires-python = ">=3.10"
 dependencies = [
+    "adjusttext>=1.3.0",
     "backoff>=2.2.1",
     "beautifulsoup4>=4.13.4",
     "dotenv>=0.9.9",
@@ -61,6 +62,8 @@ dev = [
     "ruff>=0.12.0",
     "twine>=6.1.0",
     "openpyxl>=3.1.5",
+    "vl-convert-python>=1.8.0",
+    "matplotlib>=3.10.5",
 ]
 
 [project.urls]
