+++ tests/model_comparison_vis.py
@@ -8,28 +8,32 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 from adjustText import adjust_text
-import sys
 from matplotlib.ticker import PercentFormatter
 
 # ----------------- Tuning knobs -----------------
 X_PAD_RIGHT = 10.0
 X_MIN_PAD   = 2.0
-INIT_DX     = 1    # smaller horizontal offset
-INIT_DY     = 0.00   # smaller vertical jitter
+INIT_DX     = 1
+INIT_DY     = 0.00
 
 # Scatter jitter to spread near-duplicates
-JITTER_X    = 0.001      # billions
-JITTER_Y    = 0.0005    # accuracy units
+JITTER_X    = 0.001
+JITTER_Y    = 0.0005
 
-# Strength of collision handling (bigger → more space)
-EXP_TEXT    = (1.02, 1.55)  # expand text boxes in x,y during solving
-EXP_POINTS  = (1.05, 1.20)  # expand point boxes in x,y
+# Strength of collision handling
+EXP_TEXT    = (1.02, 1.55)
+EXP_POINTS  = (1.05, 1.20)
 FORCE_TEXT  = 1
 FORCE_POINTS= 0.4
-N_STEPS     = 1000          # iterations for the solver (balance quality/speed)
+N_STEPS     = 1000
 
 # Arrow aesthetics
 ARROW = dict(arrowstyle='-', color='lightgray', lw=1, alpha=0.7)
+
+# Colours
+BLUE  = "#3B82F6"   # dense
+GREEN = "#10B981"   # MoE / sparse
+EDGE  = "black"
 
 # Short names for models displayed in labels
 SHORT_LABELS = {
@@ -58,10 +62,34 @@ SHORT_LABELS = {
     "mistralai/codestral-2508": "Codestral"
 }
 
-# Baseline (non-local) references to show as horizontal lines if present
+# Baseline references (optional)
 FRONTIER_NAMES = {"gpt-4.1": "GPT-4.1"}
-                #   "gpt-5": "GPT-5", "x-ai/grok-4": "Grok-4"}
 
+# ----------------- MoE identification -----------------
+# Explicit MoE models by id in your CSV (exact match)
+MOE_EXPLICIT = {
+    "deepseek/deepseek-chat-v3-0324",
+    "qwen/qwen3-235b-a22b-2507",
+    "moonshotai/kimi-k2",
+    "meta-llama/llama-4-maverick",
+}
+
+# Heuristic keywords that imply MoE/activated-params variants
+# e.g., "...-A22B", "...-A3B", or strings that uniquely identify MoE families
+import re
+MOE_PATTERNS = [
+    re.compile(r"(?i)\bA\d+B\b"),         # ...-A22B, -A3B, etc.
+    re.compile(r"(?i)\bmaverick\b"),      # Llama 4 Maverick
+    re.compile(r"(?i)deepseek.*v3"),      # DeepSeek V3 variants
+    re.compile(r"(?i)kimi[-_]?k2"),       # Moonshot K2 variants
+]
+
+def is_moe_model(model_id: str) -> bool:
+    if model_id in MOE_EXPLICIT:
+        return True
+    return any(p.search(model_id) for p in MOE_PATTERNS)
+
+# ------------------------------------------------------
 
 def load_data(csv_path: str):
     df = pd.read_csv(csv_path)
@@ -74,9 +102,13 @@ def load_data(csv_path: str):
 
     # Jitter points slightly to reveal duplicates
     rng = np.random.default_rng(123)
-    local_df["params_jit"] = local_df["params"] + rng.uniform(-JITTER_X, JITTER_X, size=len(local_df))
[…]
