### 2025-09-02 7005218
PDF link extraction to support new and old HTML structures
