### 2025-09-02 c9bf865
Update changelog, readme and uv files
