### 2025-10-10 a78fb17
Add streamlit app
