+++ streamlit_app.py
@@ -1,0 +1,697 @@
+"""Interactive Streamlit dashboard for the PFD Toolkit API."""
+from __future__ import annotations
+
+import json
+import sys
+from datetime import date
+from pathlib import Path
+from typing import Any, Dict, Optional, Tuple
+
+import pandas as pd
+import streamlit as st
+from pydantic import Field, create_model
+
+ROOT_DIR = Path(__file__).resolve().parent
+SRC_DIR = ROOT_DIR / "src"
+if SRC_DIR.exists():
+    src_str = str(SRC_DIR)
+    if src_str not in sys.path:
+        sys.path.insert(0, src_str)
+
+from pfd_toolkit.extractor import Extractor
+from pfd_toolkit.loader import load_reports
+from pfd_toolkit.llm import LLM
+from pfd_toolkit.screener import Screener
+
+LOGO_PATH = Path("docs/assets/badge-circle.png")
+DEFAULT_OPENROUTER_URL = "https://openrouter.ai/api/v1"
+
+
+# ---------------------------------------------------------------------------
+# Helper utilities
+# ---------------------------------------------------------------------------
+
+def _init_session_state() -> None:
+    """Ensure default keys exist in ``st.session_state``."""
+    defaults = {
+        "reports_df": pd.DataFrame(),
+        "screener_result": None,
+        "extractor_result": None,
+        "summary_result": None,
+        "token_estimate": None,
+        "theme_model_schema": None,
+        "llm_client": None,
+    }
+    for key, value in defaults.items():
+        st.session_state.setdefault(key, value)
+
+
+def _styled_metric(label: str, value: Any) -> None:
+    """Render a metric with consistent styling."""
+    st.markdown(
+        f"""
+        <div class="metric-card">
+            <div class="metric-label">{label}</div>
+            <div class="metric-value">{value}</div>
+        </div>
+        """,
+        unsafe_allow_html=True,
+    )
+
+
+def _parse_optional_int(value: str) -> Optional[int]:
+    """Convert ``value`` to ``int`` when possible."""
+    value = value.strip()
+    if not value:
+        return None
+    try:
+        return int(value)
+    except ValueError:
+        st.error("Seed must be an integer if provided.")
+        return None
+
+
+def _parse_optional_float(value: str) -> Optional[float]:
+    """Convert ``value`` to ``float`` when possible."""
+    value = value.strip()
+    if not value:
+        return None
+    try:
+        return float(value)
+    except ValueError:
+        st.error("Timeout must be numeric (seconds).")
+        return None
+
+
+def _build_feature_model(schema_text: str):
+    """Return a ``pydantic`` model built from the user supplied JSON."""
+    schema_text = schema_text.strip()
+    if not schema_text:
+        raise ValueError("Please provide a JSON schema describing your features.")
+
+    try:
+        schema_dict = json.loads(schema_text)
+    except json.JSONDecodeError as exc:
+        raise ValueError(f"Schema must be valid JSON: {exc}") from exc
+
+    if not isinstance(schema_dict, dict):
+        raise ValueError("Schema must be a JSON object mapping feature names to metadata.")
+
+    type_mapping = {
+        "string": str,
+        "str": str,
+        "text": str,
+        "bool": bool,
+        "boolean": bool,
+        "int": int,
+        "integer": int,
+        "float": float,
+        "number": float,
+    }
+
+    fields: Dict[str, Tuple[type, Field]] = {}
+    for feature, config in schema_dict.items():
+        if not isinstance(config, dict):
+            raise ValueError("Each feature must map to an object containing at least a 'type'.")
+
+        feature_type = str(config.get("type", "string")).lower()
+        python_type = type_mapping.get(feature_type)
[…]
