### 2025-10-10 672502e
Simplify user interface by removing some params
