+++ streamlit_app.py
@@ -41,6 +41,12 @@ def _init_session_state() -> None:
         "token_estimate": None,
         "theme_model_schema": None,
         "llm_client": None,
+        "extractor": None,
+        "active_view": "Overview",
+        "extractor_source_signature": None,
+        "extractor_verbose": False,
+        "feature_grid": None,
+        "extractor_mode": "Discover themes in the reports",
     }
     for key, value in defaults.items():
         st.session_state.setdefault(key, value)
@@ -137,6 +143,50 @@ def _build_feature_model(schema_text: str):
     return create_model(model_name, **fields)
 
 
+def _build_feature_model_from_grid(feature_rows: pd.DataFrame):
+    """Create a ``pydantic`` model from the feature editor grid."""
+
+    if feature_rows.empty:
+        raise ValueError("Please add at least one feature to extract.")
+
+    type_mapping: Dict[str, type] = {
+        "Text": str,
+        "Boolean": bool,
+        "Integer": int,
+        "Decimal": float,
+    }
+
+    fields: Dict[str, Tuple[type, Field]] = {}
+    for row in feature_rows.to_dict(orient="records"):
+        raw_name = str(row.get("Field name", "")).strip()
+        description = str(row.get("Description", "")).strip()
+        type_label = str(row.get("Type", "")).strip()
+
+        if not raw_name:
+            continue
+
+        if type_label not in type_mapping:
+            raise ValueError(f"Select a valid type for '{raw_name}'.")
+
+        if raw_name in fields:
+            raise ValueError(f"Duplicate feature name '{raw_name}'. Each feature must be unique.")
+
+        python_type = type_mapping[type_label]
+        fields[raw_name] = (
+            python_type,
+            Field(
+                ...,
+                description=description,
+                title=raw_name.replace("_", " ").title(),
+            ),
+        )
+
+    if not fields:
+        raise ValueError("Please provide a name and type for each feature you want to extract.")
+
+    return create_model("SelectedFeatures", **fields)
+
+
 def _display_dataframe(df: pd.DataFrame, caption: str) -> None:
     """Render a ``DataFrame`` with a consistent caption."""
     if df is None or df.empty:
@@ -231,6 +281,8 @@ def _build_sidebar() -> None:
             st.session_state["summary_result"] = None
             st.session_state["token_estimate"] = None
             st.session_state["theme_model_schema"] = None
+            st.session_state["extractor"] = None
+            st.session_state["extractor_source_signature"] = None
             st.success(f"Loaded {len(df)} reports into the workspace.")
         except Exception as exc:  # pragma: no cover - UI feedback
             st.error(f"Could not load reports: {exc}")
@@ -296,15 +348,15 @@ def _render_overview() -> None:
 
 
 def _render_screener_tab() -> None:
-    st.subheader("Screener")
+    st.subheader("Screen reports")
     st.markdown(
         """
-        Use the Screener to classify reports against a topic. Provide a clear search query, choose which
-        report fields feed into the prompt, and decide whether to filter the results or append a classification column.
+        Ask the Screener to search every report for a topic or scenario. All report fields are sent to the
+        language model so you only need to describe what you are looking for.
         """
     )
     st.caption(
-        "Example query: *Identify reports relating to medication errors involving care homes.*"
+        "Example: *Find reports describing medication errors in care homes.*"
     )
 
     llm_client: Optional[LLM] = st.session_state.get("llm_client")
@@ -318,55 +370,43 @@ def _render_screener_tab() -> None:
         return
 
     with st.form("screener_form", enter_to_submit=False):
-        st.markdown("#### Configuration")
-        cols = st.columns(4)
[…]
