### 2025-10-10 142c6f5
Track progress for Screener and Extractor
