+++ streamlit_app.py
@@ -42,7 +42,6 @@ def _init_session_state() -> None:
         "theme_model_schema": None,
         "llm_client": None,
         "extractor": None,
-        "active_view": "Overview",
         "extractor_source_signature": None,
         "extractor_verbose": False,
         "feature_grid": None,
@@ -317,7 +316,7 @@ def _build_sidebar() -> None:
 # Main layout helpers
 # ---------------------------------------------------------------------------
 
-def _render_overview() -> None:
+def _render_load_reports_tab() -> None:
     st.title("PFD Toolkit AI Workbench")
     st.markdown(
         """
@@ -392,7 +391,10 @@ def _render_screener_tab() -> None:
         if not search_query.strip():
             st.error("Describe what the Screener should look for.")
             return
+        progress_placeholder = st.empty()
+        progress_bar = progress_placeholder.progress(0)
         try:
+            progress_bar.progress(10)
             screener = Screener(
                 llm=llm_client,
                 reports=reports_df,
@@ -405,17 +407,28 @@ def _render_screener_tab() -> None:
                 include_circumstances=True,
                 include_concerns=True,
             )
-            result_df = screener.screen_reports(
-                search_query=search_query or None,
-                filter_df=filter_df,
-                result_col_name=result_col_name or "matches_query",
-                produce_spans=produce_spans,
-                drop_spans=drop_spans,
-            )
+            progress_bar.progress(25)
+            with st.spinner("Running the screener..."):
+                result_df = screener.screen_reports(
+                    search_query=search_query or None,
+                    filter_df=filter_df,
+                    result_col_name=result_col_name or "matches_query",
+                    produce_spans=produce_spans,
+                    drop_spans=drop_spans,
+                )
+            progress_bar.progress(100)
             st.session_state["screener_result"] = result_df
+            st.session_state["extractor"] = None
+            st.session_state["extractor_source_signature"] = None
+            st.session_state["extractor_result"] = None
+            st.session_state["summary_result"] = None
+            st.session_state["token_estimate"] = None
+            st.session_state["theme_model_schema"] = None
             st.success("Screening complete. Review the results below.")
         except Exception as exc:  # pragma: no cover - relies on live API
             st.error(f"Screening failed: {exc}")
+        finally:
+            progress_placeholder.empty()
 
     result_df = st.session_state.get("screener_result")
     if isinstance(result_df, pd.DataFrame):
@@ -449,6 +462,33 @@ def _render_extractor_tab() -> None:
         st.info("Load reports from the sidebar before extracting insights.")
         return
 
+    screened_df = st.session_state.get("screener_result")
+    has_screened_df = isinstance(screened_df, pd.DataFrame) and not screened_df.empty
+
+    if has_screened_df:
+        dataset_label_map = {
+            "Use screened reports": "screened",
+            "Use all loaded reports": "loaded",
+        }
+        dataset_choice = st.radio(
+            "Select the reports to analyse",
+            list(dataset_label_map.keys()),
+            index=0,
+            horizontal=True,
+            key="extractor_dataset_choice",
+        )
+        active_reports_df = screened_df if dataset_label_map[dataset_choice] == "screened" else reports_df
+    else:
+        if isinstance(screened_df, pd.DataFrame) and screened_df.empty:
+            st.warning(
+                "The screener did not retain any reports. Falling back to the loaded dataset."
+            )
+        active_reports_df = reports_df
+
+    st.caption(
+        f"Working with {len(active_reports_df)} report(s) based on the current selection."
+    )
+
     with st.form("extractor_setup_form", enter_to_submit=False):
         st.markdown("#### Initial setup")
         verbose = st.checkbox(
[…]
