### 2025-10-10 eb142e6
Fix glitch where Streamlit app returned to first tab when the Screener was run
