+++ streamlit_app.py
@@ -46,6 +46,7 @@ def _init_session_state() -> None:
         "extractor_verbose": False,
         "feature_grid": None,
         "extractor_mode": "Discover themes in the reports",
+        "active_tab": "Load in reports",
     }
     for key, value in defaults.items():
         st.session_state.setdefault(key, value)
@@ -86,6 +87,23 @@ def _parse_optional_float(value: str) -> Optional[float]:
     except ValueError:
         st.error("Timeout must be numeric (seconds).")
         return None
+
+
+def _parse_optional_non_negative_int(value: str, field_name: str) -> Optional[int]:
+    """Convert ``value`` to ``int`` ensuring it is not negative."""
+
+    value = value.strip()
+    if not value:
+        return None
+    try:
+        parsed_value = int(value)
+    except ValueError as exc:
+        raise ValueError(f"{field_name} must be an integer if provided.") from exc
+
+    if parsed_value < 0:
+        raise ValueError(f"{field_name} cannot be negative.")
+
+    return parsed_value
 
 
 def _build_feature_model(schema_text: str):
@@ -586,19 +604,15 @@ def _render_extractor_tab() -> None:
                     step=1000,
                 )
                 max_col, min_col = st.columns(2)
-                max_themes = max_col.number_input(
-                    "Maximum number of themes",
-                    min_value=0,
-                    value=0,
-                    step=1,
-                    help="Leave as 0 to let the model decide.",
+                max_themes_raw = max_col.text_input(
+                    "Maximum number of themes (optional)",
+                    value="",
+                    placeholder="Leave blank to let the model decide.",
                 )
-                min_themes = min_col.number_input(
-                    "Minimum number of themes",
-                    min_value=0,
-                    value=0,
-                    step=1,
-                    help="Leave as 0 to let the model decide.",
+                min_themes_raw = min_col.text_input(
+                    "Minimum number of themes (optional)",
+                    value="",
+                    placeholder="Leave blank to let the model decide.",
                 )
 
             themes_submitted = st.form_submit_button(
@@ -606,48 +620,76 @@ def _render_extractor_tab() -> None:
             )
 
         if themes_submitted:
-            progress_placeholder = st.empty()
-            progress_bar = progress_placeholder.progress(0)
+            input_error = False
             try:
-                progress_bar.progress(10)
-                with st.spinner("Summarising the reports..."):
-                    summary_df = extractor.summarise(
-                        result_col_name=summary_column or "summary",
-                        trim_intensity=trim_labels[trim_choice],
-                    )
-                    st.session_state["summary_result"] = summary_df
-                progress_bar.progress(55)
-
-                seed_topics: Optional[Any] = None
-                if seed_topics_text.strip():
-                    try:
-                        seed_topics = json.loads(seed_topics_text)
-                    except json.JSONDecodeError:
-                        seed_topics = [
-                            line.strip()
-                            for line in seed_topics_text.splitlines()
-                            if line.strip()
-                        ]
-
-                progress_bar.progress(70)
-                with st.spinner("Identifying themes..."):
-                    ThemeModel = extractor.discover_themes(
-                        warn_exceed=int(warning_threshold),
-                        error_exceed=int(error_threshold),
-                        max_themes=int(max_themes) or None,
-                        min_themes=int(min_themes) or None,
-                        extra_instructions=extra_theme_instructions or None,
-                        seed_topics=seed_topics,
-                    )
-                st.session_state["theme_model_schema"] = ThemeModel.model_json_schema()
-                progress_bar.progress(100)
-                st.success(
-                    "Theme discovery complete. The schema below can be reused when tagging reports."
+                max_themes_value = _parse_optional_non_negative_int(
+                    max_themes_raw,
+                    "Maximum number of themes",
[…]
