### 2025-10-13 c6c6548
Add OpenRouter integration
