+++ streamlit_app.py
@@ -1,12 +1,19 @@
 """Interactive Streamlit dashboard for the PFD Toolkit API."""
 from __future__ import annotations
 
+import base64
+import hashlib
 import json
+import os
+import secrets
 import sys
+import time
 from datetime import date
 from pathlib import Path
-from typing import Any, Dict, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
+from urllib.parse import urlencode
 
+import httpx
 import pandas as pd
 import streamlit as st
 from pydantic import Field, create_model
@@ -24,7 +31,13 @@ from pfd_toolkit.llm import LLM
 from pfd_toolkit.screener import Screener
 
 LOGO_PATH = Path("docs/assets/badge-circle.png")
-DEFAULT_OPENROUTER_URL = "https://openrouter.ai/api/v1"
+OPENROUTER_AUTH_URL = "https://openrouter.ai/auth"
+OPENROUTER_EXCHANGE_URL = "https://openrouter.ai/api/v1/auth/keys"
+OPENROUTER_API_BASE = "https://openrouter.ai/api/v1"
+
+
+# Stores pending PKCE verifiers keyed by OAuth ``state`` plus creation timestamp.
+OPENROUTER_PKCE_PENDING: Dict[str, Tuple[str, float]] = {}
 
 
 # ---------------------------------------------------------------------------
@@ -47,9 +60,269 @@ def _init_session_state() -> None:
         "feature_grid": None,
         "extractor_mode": "Discover themes in the reports",
         "active_tab": "Load in reports",
+        "openrouter_api_key": None,
+        "openrouter_balance_info": None,
+        "openrouter_pkce_verifier": None,
+        "openrouter_pkce_state": None,
+        "openrouter_manual_api_key": "",
+        "openrouter_callback_override": "",
+        "openrouter_base_url": OPENROUTER_API_BASE,
+        "openai_api_key": "",
+        "openai_base_url": "",
+        "provider_override": "OpenRouter OAuth",
     }
     for key, value in defaults.items():
         st.session_state.setdefault(key, value)
+
+
+def _get_secret(name: str) -> Optional[str]:
+    """Return ``name`` from ``st.secrets`` when available."""
+
+    secrets_obj = getattr(st, "secrets", None)
+    if secrets_obj is None:
+        return None
+    try:
+        if hasattr(secrets_obj, "get"):
+            return secrets_obj.get(name)
+        return secrets_obj[name]
+    except Exception:  # pragma: no cover - depends on environment
+        return None
+
+
+def _default_callback_url() -> str:
+    """Return the callback URL used for the OpenRouter PKCE flow."""
+
+    override = st.session_state.get("openrouter_callback_override") or ""
+    if override.strip():
+        return override.strip()
+
+    for key in ("PUBLIC_URL", "public_url"):
+        secret = _get_secret(key)
+        if secret:
+            return secret
+
+    env_url = os.getenv("PUBLIC_URL")
+    if env_url:
+        return env_url
+
+    return "http://localhost:8501"
+
+
+def _new_pkce_verifier() -> str:
+    """Return a freshly generated PKCE code verifier."""
+
+    return base64.urlsafe_b64encode(secrets.token_bytes(32)).decode().rstrip("=")
+
+
+def _pkce_challenge(verifier: str) -> str:
+    """Return the S256 PKCE challenge for ``verifier``."""
+
+    digest = hashlib.sha256(verifier.encode()).digest()
+    return base64.urlsafe_b64encode(digest).decode().rstrip("=")
+
+
+def _new_pkce_state() -> str:
[…]
