### 2025-10-13 8bce06c
Update OpenRouter OAuth flow to use a top-level JavaScript redirect with a no-script fallback link
