### 2025-10-14 67517d5
Remove OpenRouter sign in link due to bugs
