### 2025-10-14 10d5670
Remove params to simplify UI
