+++ streamlit_app.py
@@ -35,15 +35,17 @@ def _init_session_state() -> None:
     """Ensure default keys exist in ``st.session_state``."""
     defaults = {
         "reports_df": pd.DataFrame(),
+        "reports_df_initial": None,
+        "reports_df_modified": False,
         "screener_result": None,
         "extractor_result": None,
         "summary_result": None,
-        "token_estimate": None,
         "theme_model_schema": None,
         "llm_client": None,
         "extractor": None,
         "extractor_source_signature": None,
         "extractor_verbose": False,
+        "seed_topics_last": None,
         "feature_grid": None,
         "extractor_mode": "Discover themes in the reports",
         "active_tab": "Load in reports",
@@ -69,30 +71,6 @@ def _styled_metric(label: str, value: Any) -> None:
         """,
         unsafe_allow_html=True,
     )
-
-
-def _parse_optional_int(value: str) -> Optional[int]:
-    """Convert ``value`` to ``int`` when possible."""
-    value = value.strip()
-    if not value:
-        return None
-    try:
-        return int(value)
-    except ValueError:
-        st.error("Seed must be an integer if provided.")
-        return None
-
-
-def _parse_optional_float(value: str) -> Optional[float]:
-    """Convert ``value`` to ``float`` when possible."""
-    value = value.strip()
-    if not value:
-        return None
-    try:
-        return float(value)
-    except ValueError:
-        st.error("Timeout must be numeric (seconds).")
-        return None
 
 
 def _parse_optional_non_negative_int(value: str, field_name: str) -> Optional[int]:
@@ -219,6 +197,50 @@ def _display_dataframe(df: pd.DataFrame, caption: str) -> None:
     st.dataframe(df, use_container_width=True, hide_index=True)
 
 
+def _render_reports_overview(
+    caption: str = "Loaded Prevention of Future Death reports", *, key_suffix: str = ""
+) -> None:
+    """Display the active reports with an optional revert button."""
+
+    reports_df: pd.DataFrame = st.session_state.get("reports_df", pd.DataFrame())
+    initial_df: Optional[pd.DataFrame] = st.session_state.get("reports_df_initial")
+
+    if initial_df is None and not reports_df.empty:
+        initial_df = reports_df.copy(deep=True)
+        st.session_state["reports_df_initial"] = initial_df
+
+    modified = False
+    if initial_df is not None:
+        if reports_df.empty and not initial_df.empty:
+            modified = True
+        elif not reports_df.empty:
+            try:
+                modified = not reports_df.equals(initial_df)
+            except Exception:
+                modified = True
+
+    st.session_state["reports_df_modified"] = modified
+
+    if modified:
+        if st.button(
+            "Revert changes",
+            key=f"revert_reports_{key_suffix}" if key_suffix else "revert_reports",
+            use_container_width=True,
+        ):
+            st.session_state["reports_df"] = initial_df.copy()
+            st.session_state["reports_df_modified"] = False
+            st.session_state["screener_result"] = None
+            st.session_state["summary_result"] = None
+            st.session_state["extractor_result"] = None
+            st.session_state["theme_model_schema"] = None
+            st.session_state["seed_topics_last"] = None
+            st.session_state["extractor"] = None
+            st.session_state["extractor_source_signature"] = None
+            reports_df = st.session_state["reports_df"]
+
+    _display_dataframe(reports_df, caption)
+
+
 # ---------------------------------------------------------------------------
 # Sidebar controls
 # ---------------------------------------------------------------------------
@@ -296,12 +318,24 @@ def _build_sidebar() -> None:
 
     st.sidebar.markdown("---")
[…]
