### 2025-10-14 37eefc0
Fix bug where active dataframe displayed twice
