### 2025-10-14 b90ee8d
Themes now display as table, not raw json
