### 2025-10-14 f36f571
Discover themes now assigns themes to reports automatically
