### 2025-10-14 98cdf8e
Remove some options for a cleaner interface
