+++ streamlit_app.py
@@ -46,7 +46,6 @@ def _init_session_state() -> None:
         "llm_client": None,
         "extractor": None,
         "extractor_source_signature": None,
-        "extractor_verbose": False,
         "seed_topics_last": None,
         "feature_grid": None,
         "extractor_mode": "Discover themes in the reports",
@@ -351,12 +350,18 @@ def _build_sidebar() -> None:
 
     st.sidebar.markdown("---")
     st.sidebar.markdown("### Model configuration")
-    model_display_options = ["GPT 4.1", "GPT 4.1 Mini"]
-    model_value_map = {"GPT 4.1": "gpt-4.1", "GPT 4.1 Mini": "gpt-4.1-mini"}
-    current_model_value = st.session_state.get("model_name", "gpt-4.1")
+    model_display_options = [
+        "GPT 4.1 Mini (+speed, -accuracy)",
+        "GPT 4.1 (+accuracy, -speed)",
+    ]
+    model_value_map = {
+        "GPT 4.1 Mini (+speed, -accuracy)": "gpt-4.1-mini",
+        "GPT 4.1 (+accuracy, -speed)": "gpt-4.1",
+    }
+    current_model_value = st.session_state.get("model_name", "gpt-4.1-mini")
     current_display = next(
         (label for label, value in model_value_map.items() if value == current_model_value),
-        "GPT 4.1",
+        model_display_options[0],
     )
     model_display = st.sidebar.selectbox(
         "Chat model",
@@ -367,23 +372,36 @@ def _build_sidebar() -> None:
     )
     model_name = model_value_map[model_display]
     st.session_state["model_name"] = model_name
-    max_workers = st.sidebar.number_input("Max parallel workers", min_value=1, max_value=32, value=8)
+    with st.sidebar.expander("Advanced options"):
+        max_workers = st.number_input(
+            "Max parallel workers",
+            min_value=1,
+            max_value=32,
+            value=int(st.session_state.get("max_parallel_workers", 8)),
+            key="max_parallel_workers_input",
+        )
+    st.session_state["max_parallel_workers"] = int(max_workers)
     validation_attempts = 2
 
     st.sidebar.markdown("---")
     st.sidebar.markdown("### Report window")
     default_start = date(2013, 1, 1)
-    date_range = st.sidebar.date_input(
-        "Report date range",
-        value=(default_start, date.today()),
+    start_date = st.sidebar.date_input(
+        "Report start date",
+        value=st.session_state.get("report_start_date", default_start),
         min_value=date(2000, 1, 1),
         max_value=date.today(),
     )
-    if isinstance(date_range, tuple):
-        start_date, end_date = date_range
-    else:
-        start_date = date_range
-        end_date = date.today()
+    st.session_state["report_start_date"] = start_date
+    end_date = st.sidebar.date_input(
+        "Report end date",
+        value=st.session_state.get("report_end_date", date.today()),
+        min_value=start_date,
+        max_value=date.today(),
+    )
+    st.session_state["report_end_date"] = end_date
+    if end_date < start_date:
+        st.sidebar.error("End date must be on or after the start date.")
 
     n_reports_raw = st.sidebar.text_input(
         "Limit number of recent reports (optional)",
@@ -409,6 +427,9 @@ def _build_sidebar() -> None:
     if load_button:
         if n_reports_raw.strip() and n_reports is None:
             st.sidebar.error("Enter a valid integer before loading reports.")
+            return
+        if end_date < start_date:
+            st.sidebar.error("Fix the report date range before loading reports.")
             return
         try:
             with st.spinner("Downloading and filtering reports..."):
@@ -523,10 +544,8 @@ def _render_screener_tab() -> None:
         )
         cols3 = st.columns(2)
[…]
