### 2025-10-15 c589967
Refactor app to single page with 'next step' option toggle
