+++ streamlit_app.py
@@ -2,6 +2,7 @@
 from __future__ import annotations
 
 import ast
+import copy
 import json
 import sys
 from datetime import date
@@ -10,6 +11,7 @@ from typing import Any, Dict, Optional, Tuple
 
 import pandas as pd
 import streamlit as st
+from streamlit.delta_generator import DeltaGenerator
 from pydantic import Field, create_model
 
 ROOT_DIR = Path(__file__).resolve().parent
@@ -37,6 +39,9 @@ def _init_session_state() -> None:
     defaults = {
         "reports_df": pd.DataFrame(),
         "reports_df_initial": None,
+        "history": [],
+        "active_action": None,
+        "preview_state": None,
         "reports_df_modified": False,
         "screener_result": None,
         "extractor_result": None,
@@ -48,8 +53,7 @@ def _init_session_state() -> None:
         "extractor_source_signature": None,
         "seed_topics_last": None,
         "feature_grid": None,
-        "extractor_mode": "Discover themes in the reports",
-        "active_tab": "Load in reports",
+        FLASH_MESSAGE_KEY: None,
         "openrouter_api_key": "",
         "openrouter_base_url": OPENROUTER_API_BASE,
         "openai_api_key": "",
@@ -72,6 +76,68 @@ def _styled_metric(label: str, value: Any) -> None:
         """,
         unsafe_allow_html=True,
     )
+
+
+FLASH_MESSAGE_KEY = "flash_message"
+
+
+SNAPSHOT_KEYS = [
+    "reports_df",
+    "screener_result",
+    "extractor_result",
+    "summary_result",
+    "theme_model_schema",
+    "theme_summary_table",
+    "seed_topics_last",
+    "extractor",
+    "extractor_source_signature",
+    "feature_grid",
+    "active_action",
+    "preview_state",
+]
+
+
+def _deepcopy_value(value: Any) -> Any:
+    """Return a deep copy of ``value`` preserving DataFrame contents."""
+
+    if isinstance(value, pd.DataFrame):
+        return value.copy(deep=True)
+    try:
+        return copy.deepcopy(value)
+    except Exception:
+        return value
+
+
+def snapshot_state() -> Dict[str, Any]:
+    """Capture a snapshot of the interactive state for undo support."""
+
+    snapshot: Dict[str, Any] = {}
+    for key in SNAPSHOT_KEYS:
+        snapshot[key] = _deepcopy_value(st.session_state.get(key))
+    return snapshot
+
+
+def restore_state(snapshot: Dict[str, Any]) -> None:
+    """Restore Streamlit session state from ``snapshot``."""
+
+    for key, value in snapshot.items():
+        st.session_state[key] = _deepcopy_value(value)
+
+
+def push_history_snapshot() -> None:
+    """Push the current state onto the undo history (max depth 10)."""
+
+    history = list(st.session_state.get("history", []))
+    history.append(snapshot_state())
+    if len(history) > 10:
+        history = history[-10:]
+    st.session_state["history"] = history
+
+
+def clear_preview_state() -> None:
+    """Remove any pending preview artefacts."""
+
+    st.session_state["preview_state"] = None
 
 
 def _parse_optional_non_negative_int(value: str, field_name: str) -> Optional[int]:
@@ -446,8 +512,13 @@ def _build_sidebar() -> None:
             st.session_state["extractor_result"] = None
             st.session_state["summary_result"] = None
             st.session_state["theme_model_schema"] = None
+            st.session_state["theme_summary_table"] = None
             st.session_state["extractor"] = None
             st.session_state["extractor_source_signature"] = None
+            st.session_state["seed_topics_last"] = None
+            st.session_state["history"] = []
+            st.session_state["active_action"] = None
[…]
