### 2025-10-15 ba3b7b3
Improve user flow by removing complex params
