+++ streamlit_app.py
@@ -7,7 +7,7 @@ import json
 import sys
 from datetime import date
 from pathlib import Path
-from typing import Any, Dict, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
 import pandas as pd
 import streamlit as st
@@ -67,11 +67,12 @@ def _init_session_state() -> None:
 
 def _styled_metric(label: str, value: Any) -> None:
     """Render a metric with consistent styling."""
+
     st.markdown(
         f"""
         <div class="metric-card">
-            <div class="metric-label">{label}</div>
-            <div class="metric-value">{value}</div>
+            <span class="metric-label">{label}</span>
+            <span class="metric-value">{value}</span>
         </div>
         """,
         unsafe_allow_html=True,
@@ -559,21 +560,9 @@ def _get_reports_df() -> pd.DataFrame:
 
 
 def _render_header(container: Optional[DeltaGenerator] = None) -> None:
-    """Render the pinned header containing KPIs and the active dataset."""
+    """Render the hero header and dataset snapshot."""
 
     ctx = container or st
-
-    ctx.title("PFD Toolkit AI Workbench")
-    ctx.markdown(
-        """
-        <p class="lead">
-            Explore Prevention of Future Death (PFD) reports, screen them against custom topics,
-            and extract structured insights using the Screener and Extractor APIs.
-            Configure your data and model in the sidebar, then work through the guided flows below.
-        </p>
-        """,
-        unsafe_allow_html=True,
-    )
 
     reports_df = _get_reports_df()
     reports_count = len(reports_df)
@@ -596,21 +585,128 @@ def _render_header(container: Optional[DeltaGenerator] = None) -> None:
         if pd.notna(latest):
             latest_display = latest.strftime("%d %b %Y")
 
-    col1, col2, col3 = ctx.columns(3)
+    hero_html = f"""
+    <section class="hero-section">
+        <div class="hero-copy">
+            <span class="hero-badge">PFD Toolkit · AI Workbench</span>
+            <h1>Command the narrative of Prevention of Future Death reports.</h1>
+            <p>
+                Load official reports, triage them against strategic focus areas, and extract
+                structured evidence in minutes. The streamlined workspace keeps your data,
+                models, and insights aligned from discovery through delivery.
+            </p>
+            <div class="hero-checklist">
+                <span class="hero-pill">① Load and curate source data</span>
+                <span class="hero-pill">② Screen topics and themes with LLMs</span>
+                <span class="hero-pill">③ Capture structured insights for reporting</span>
+            </div>
+        </div>
+        <div class="hero-visual">
+            <div class="hero-orb orb-primary"></div>
+            <div class="hero-orb orb-secondary"></div>
+            <div class="hero-kpi">
+                <span class="hero-kpi-value">{reports_count:,}</span>
+                <span class="hero-kpi-label">Reports currently in view</span>
+                <span class="hero-kpi-range">{earliest_display} – {latest_display}</span>
+            </div>
+        </div>
+    </section>
+    """
+
+    ctx.markdown(hero_html, unsafe_allow_html=True)
+
+    metric_row = ctx.container()
+    metric_row.markdown("<div class='metric-row'>", unsafe_allow_html=True)
+    col1, col2, col3 = metric_row.columns(3)
     with col1:
         _styled_metric("Reports in view", f"{reports_count:,}")
     with col2:
-        _styled_metric("Earliest date", earliest_display)
+        _styled_metric("Earliest report", earliest_display)
     with col3:
-        _styled_metric("Latest date", latest_display)
+        _styled_metric("Latest report", latest_display)
+    metric_row.markdown("</div>", unsafe_allow_html=True)
 
-    ctx.markdown("#### Current working dataset")
+    dataset_card = ctx.container()
+    dataset_card.markdown(
[…]
