### 2025-10-20 b3736c7
Improve formatting for next steps buttons
