### 2025-10-20 2b9f759
Clean up header with minimalist start over and undo buttons
