+++ streamlit_app.py
@@ -654,58 +654,94 @@ def _render_header(container: Optional[DeltaGenerator] = None) -> None:
     dataset_card.markdown("</div>", unsafe_allow_html=True)
 
 
-def _render_status_bar(container: Optional[DeltaGenerator] = None) -> None:
-    """Display a floating status ribbon with workspace health."""
+def _value_has_content(value: Any) -> bool:
+    """Return ``True`` when ``value`` contains meaningful state."""
 
-    ctx = container or st
-    reports_df = _get_reports_df()
-    dataset_modified = st.session_state.get("reports_df_modified", False)
-    llm_ready = st.session_state.get("llm_client") is not None
-    history_depth = len(st.session_state.get("history", []))
+    if value is None:
+        return False
+    if isinstance(value, (pd.DataFrame, pd.Series)):
+        return not value.empty
+    if isinstance(value, (list, tuple, set, dict)):
+        return len(value) > 0
+    try:
+        return bool(value)
+    except ValueError:
+        return True
+    except TypeError:
+        return True
 
-    if reports_df.empty:
-        dataset_status = "No dataset loaded"
-        dataset_tone = "muted"
-    else:
-        dataset_status = f"{len(reports_df):,} reports"
-        dataset_tone = "success" if not dataset_modified else "alert"
-        if dataset_modified:
-            dataset_status += " · unsaved edits"
 
-    statuses: List[Dict[str, str]] = [
-        {
-            "label": "Dataset",
-            "value": dataset_status,
-            "tone": dataset_tone,
-        },
-        {
-            "label": "Model",
-            "value": "Connected" if llm_ready else "Awaiting key",
-            "tone": "success" if llm_ready else "muted",
-        },
-        {
-            "label": "Undo",
-            "value": f"{history_depth} step(s) available" if history_depth else "Nothing to undo",
-            "tone": "info" if history_depth else "muted",
-        },
+def _workspace_has_activity() -> bool:
+    """Return ``True`` if the workspace has actions beyond the initial load."""
+
+    if st.session_state.get("reports_df_modified", False):
+        return True
+    if len(st.session_state.get("history", [])):
+        return True
+
+    activity_keys = [
+        "screener_result",
+        "extractor_result",
+        "summary_result",
+        "theme_model_schema",
+        "theme_summary_table",
+        "seed_topics_last",
+        "extractor",
+        "feature_grid",
+        "preview_state",
     ]
 
-    badges = "".join(
-        f"""
-        <div class="status-pill status-{item['tone']}">
-            <span class="status-label">{item['label']}</span>
-            <span class="status-value">{item['value']}</span>
-        </div>
-        """
-        for item in statuses
-    )
+    for key in activity_keys:
+        if _value_has_content(st.session_state.get(key)):
+            return True
 
-    ctx.markdown(
-        f"""
-        <div class="status-bar">{badges}</div>
-        """,
+    return False
+
+
+def _render_status_bar(container: Optional[DeltaGenerator] = None) -> None:
+    """Display a floating ribbon containing undo and start-over actions."""
+
+    ctx = container or st
+    history_depth = len(st.session_state.get("history", []))
+    initial_df = st.session_state.get("reports_df_initial")
+
+    undo_disabled = history_depth == 0
+    if undo_disabled:
+        undo_message = ""
+    elif history_depth == 1:
+        undo_message = "(1 step available)"
+    else:
+        undo_message = f"{history_depth} steps available"
+
+    has_initial_dataset = isinstance(initial_df, pd.DataFrame) and not initial_df.empty
+    workspace_active = _workspace_has_activity()
+    start_over_disabled = not (has_initial_dataset and workspace_active)
+
+    if not has_initial_dataset:
+        start_over_message = ""
+    elif not workspace_active:
[…]
