+++ streamlit_app.py
@@ -5,6 +5,7 @@ import ast
 import copy
 import json
 import sys
+from html import escape
 from datetime import date
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple
@@ -698,50 +699,67 @@ def _workspace_has_activity() -> bool:
     return False
 
 
-def _render_status_bar(container: Optional[DeltaGenerator] = None) -> None:
-    """Display a floating ribbon containing undo and start-over actions."""
+def _render_workspace_footer(
+    flash_payload: Optional[Tuple[str, str]] = None,
+) -> None:
+    """Render the bottom workspace footer with flash messaging and utilities."""
 
-    ctx = container or st
     history_depth = len(st.session_state.get("history", []))
     initial_df = st.session_state.get("reports_df_initial")
 
     undo_disabled = history_depth == 0
-    if undo_disabled:
-        undo_message = ""
-    elif history_depth == 1:
-        undo_message = "(1 step available)"
+    if history_depth == 1:
+        undo_hint = "(1 step available)"
+    elif history_depth > 1:
+        undo_hint = f"({history_depth} steps available)"
     else:
-        undo_message = f"{history_depth} steps available"
+        undo_hint = ""
 
     has_initial_dataset = isinstance(initial_df, pd.DataFrame) and not initial_df.empty
     workspace_active = _workspace_has_activity()
     start_over_disabled = not (has_initial_dataset and workspace_active)
 
-    if not has_initial_dataset:
-        start_over_message = ""
-    elif not workspace_active:
-        start_over_message = ""
-    else:
-        start_over_message = ""
+    footer = st.container()
+    footer.markdown("<div class='workspace-footer'>", unsafe_allow_html=True)
 
-    shell = ctx.container()
-    shell.markdown(
-        "<div class='status-shell'><div class='status-actions'>",
-        unsafe_allow_html=True,
-    )
-    if shell.button(
-        f"↶ Undo\n{undo_message}",
-        key="status_undo",
-        disabled=undo_disabled,
-    ):
-        _undo_last_change()
-    if shell.button(
-        f"↻ Start over\n{start_over_message}",
-        key="status_reset",
-        disabled=start_over_disabled,
-    ):
-        _start_again()
-    shell.markdown("</div></div>", unsafe_allow_html=True)
+    if flash_payload:
+        message, level = flash_payload
+        icon_map = {"success": "✅", "info": "ℹ️", "warning": "⚠️", "error": "❌"}
+        level_key = level if level in icon_map else "info"
+        icon = icon_map[level_key]
+        safe_message = escape(message)
+        footer.markdown(
+            f"""
+            <div class="workspace-message workspace-message--{level_key}">
+                <span class="workspace-message-icon">{icon}</span>
+                <span class="workspace-message-text">{safe_message}</span>
+            </div>
+            """,
+            unsafe_allow_html=True,
+        )
+
+    undo_label = "↶ Undo" if not undo_hint else f"↶ Undo\n{undo_hint}"
+    undo_col, start_over_col = footer.columns(2, gap="large")
+
+    with undo_col:
+        if st.button(
+            undo_label,
+            key="footer_undo",
+            use_container_width=True,
+            disabled=undo_disabled,
+        ):
+            _undo_last_change()
+
+    with start_over_col:
+        if st.button(
+            "↻ Start over",
+            key="footer_reset",
+            use_container_width=True,
+            disabled=start_over_disabled,
+        ):
+            _start_again()
+
+    footer.markdown("</div>", unsafe_allow_html=True)
 
 
 def _build_theme_summary_table(
@@ -820,20 +838,19 @@ def _trigger_rerun() -> None:
[…]
