### 2025-10-21 9e4bd42
Fix bug where feature grid was reverting every second update
