### 2025-10-21 56fbf65
Remove deprecated container sizing arg
