### 2025-10-21 1843ed6
Add beta message & icons
