+++ streamlit_app.py
@@ -368,7 +368,7 @@ def _render_reports_overview(
 def _build_sidebar() -> None:
     st.sidebar.image(str(LOGO_PATH), width=120)
     st.sidebar.markdown(
-        "<h2 style='margin-bottom:0.2rem;'>PFD Toolkit Workbench</h2>",
+        "<h2 class='sidebar-title'>PFD Toolkit Workbench <span class='sidebar-beta'>Beta</span></h2>",
         unsafe_allow_html=True,
     )
     st.sidebar.caption(
@@ -607,9 +607,25 @@ def _render_header(container: Optional[DeltaGenerator] = None) -> None:
             latest_display = latest.strftime("%d %b %Y")
 
     hero_html = f"""
+    <div class="beta-banner">
+        <div class="beta-banner__glow"></div>
+        <div class="beta-banner__content">
+            <span class="beta-banner__pill">Beta</span>
+            <div class="beta-banner__text">
+                <span class="beta-banner__headline">Welcome to the beta version of PFD Toolkit Workbench.</span>
+                <span class="beta-banner__subtext">
+                Because this app is still in development, there might be some bugs we haven't squashed yet. Please bare with us while we iron out the kinks! </span>
+            </div>
+            <div class="beta-banner__sparks">
+                <span class="beta-banner__spark"></span>
+                <span class="beta-banner__spark"></span>
+                <span class="beta-banner__spark"></span>
+            </div>
+        </div>
+    </div>
     <section class="hero-section">
         <div class="hero-copy">
-            <span class="hero-badge">PFD Toolkit · AI Workbench</span>
+            <span class="hero-badge">PFD Toolkit · AI Workbench <span class="hero-badge-beta">Beta</span></span>
             <h1>Command the narrative of Prevention of Future Death reports.</h1>
             <p>
                 Load official reports, triage them against strategic focus areas, and extract
@@ -1581,7 +1597,7 @@ def _render_extract_action() -> None:
 
 def main() -> None:
     st.set_page_config(
-        page_title="PFD Toolkit AI Workbench",
+        page_title="PFD Toolkit AI Workbench · Beta",
         page_icon="📊",
         layout="wide",
     )
@@ -1663,6 +1679,27 @@ def main() -> None:
                 background: rgba(8, 12, 36, 0.85);
                 backdrop-filter: blur(20px);
                 border-right: 1px solid rgba(148, 163, 255, 0.2);
+            }
+
+            section[data-testid="stSidebar"] h2.sidebar-title {
+                margin-bottom: 0.2rem;
+                display: inline-flex;
+                align-items: center;
+                gap: 0.5rem;
+            }
+
+            .sidebar-beta {
+                display: inline-flex;
+                align-items: center;
+                padding: 0.2rem 0.55rem;
+                border-radius: 999px;
+                background: linear-gradient(135deg, rgba(56, 189, 248, 0.65), rgba(192, 132, 252, 0.7));
+                font-size: 0.6em;
+                letter-spacing: 0.12em;
+                text-transform: uppercase;
+                color: #050b24;
+                font-weight: 700;
+                box-shadow: 0 8px 18px rgba(7, 10, 34, 0.35);
             }
 
             section[data-testid="stSidebar"] .stMarkdown, section[data-testid="stSidebar"] label, section[data-testid="stSidebar"] span {
@@ -1823,6 +1860,120 @@ def main() -> None:
                 color: #f8faff;
             }
 
+            .beta-banner {
+                margin-top: 1.4rem;
+                position: relative;
+                overflow: hidden;
+                padding: 1.4rem 1.8rem;
+                border-radius: 28px;
[…]
