### 2025-10-21 4cc49d8
Replace loading progress bar with animated dots
