+++ streamlit_app.py
@@ -82,6 +82,37 @@ def _styled_metric(label: str, value: Any) -> None:
     )
 
 
+class LoadingIndicator:
+    """Utility to display the animated loading indicator."""
+
+    def __init__(self, placeholder: DeltaGenerator, message: str = "Loading…") -> None:
+        self._placeholder = placeholder
+        self.update(message)
+
+    def update(self, message: str) -> None:
+        """Update the indicator with ``message``."""
+
+        safe_message = escape(message)
+        self._placeholder.markdown(
+            f"""
+            <div class="loading-indicator">
+                <div class="loading-indicator__dots">
+                    <span class="loading-indicator__dot"></span>
+                    <span class="loading-indicator__dot"></span>
+                    <span class="loading-indicator__dot"></span>
+                </div>
+                <span class="loading-indicator__label">{safe_message}</span>
+            </div>
+            """,
+            unsafe_allow_html=True,
+        )
+
+    def close(self) -> None:
+        """Remove the indicator from the UI."""
+
+        self._placeholder.empty()
+
+
 FLASH_MESSAGE_KEY = "flash_message"
 
 
@@ -1181,10 +1212,10 @@ def _render_filter_action() -> None:
     initial_report_count = len(reports_df)
     match_column_name = "matches_query"
 
-    progress_placeholder = st.empty()
-    progress_bar = progress_placeholder.progress(0)
+    loading_placeholder = st.empty()
+    loading_indicator = LoadingIndicator(loading_placeholder, "Preparing screener…")
     try:
-        progress_bar.progress(10)
+        loading_indicator.update("Configuring screener…")
         screener = Screener(
             llm=llm_client,
             reports=reports_df,
@@ -1197,21 +1228,20 @@ def _render_filter_action() -> None:
             include_circumstances=True,
             include_concerns=True,
         )
-        progress_bar.progress(25)
-        with st.spinner("Running the screener..."):
-            result_df = screener.screen_reports(
-                search_query=search_query or None,
-                filter_df=filter_df,
-                result_col_name=match_column_name,
-                produce_spans=produce_spans,
-                drop_spans=drop_spans,
-            )
-        progress_bar.progress(100)
+        loading_indicator.update("Running the screener…")
+        result_df = screener.screen_reports(
+            search_query=search_query or None,
+            filter_df=filter_df,
+            result_col_name=match_column_name,
+            produce_spans=produce_spans,
+            drop_spans=drop_spans,
+        )
+        loading_indicator.update("Finalising results…")
     except Exception as exc:  # pragma: no cover - relies on live API
         st.error(f"Screening failed: {exc}")
         return
     finally:
-        progress_placeholder.empty()
+        loading_indicator.close()
 
     st.session_state["screener_result"] = result_df
     st.session_state["reports_df"] = result_df.copy(deep=True)
@@ -1308,7 +1338,7 @@ def _render_discover_action() -> None:
             )
 
         preview_requested = st.form_submit_button(
-            "Preview recurring themes", width="stretch"
+            "Discover recurring themes", width="stretch"
         )
 
     if preview_requested:
@@ -1330,17 +1360,16 @@ def _render_discover_action() -> None:
         if min_themes_value is not None or max_themes_value is not None or (
             not min_themes_raw.strip() and not max_themes_raw.strip()
         ):
-            progress_placeholder = st.empty()
-            progress_bar = progress_placeholder.progress(0)
+            loading_placeholder = st.empty()
+            loading_indicator = LoadingIndicator(
+                loading_placeholder, "Summarising the reports…"
+            )
             try:
-                progress_bar.progress(10)
                 summary_col_name = extractor.summary_col or "summary"
-                with st.spinner("Summarising the reports..."):
-                    summary_df = extractor.summarise(
-                        result_col_name=summary_col_name,
[…]
