### 2025-10-21 f54259c
Create onboarding hero when no reports are loaded
