+++ streamlit_app.py
@@ -63,6 +63,7 @@ def _init_session_state() -> None:
         "openai_base_url": "",
         "provider_override": "OpenAI",
         "provider_override_select": "OpenAI",
+        REPORTS_LOADING_FLAG_KEY: False,
     }
     for key, value in defaults.items():
         st.session_state.setdefault(key, value)
@@ -114,6 +115,57 @@ class LoadingIndicator:
 
 
 FLASH_MESSAGE_KEY = "flash_message"
+
+REPORTS_LOADING_OVERLAY_KEY = "_reports_loading_overlay_placeholder"
+REPORTS_LOADING_FLAG_KEY = "reports_loading_active"
+
+
+def _get_loading_overlay_placeholder() -> DeltaGenerator:
+    """Return the placeholder used for the reports loading scrim."""
+
+    placeholder: Optional[DeltaGenerator] = st.session_state.get(REPORTS_LOADING_OVERLAY_KEY)
+    if placeholder is None:
+        placeholder = st.empty()
+        st.session_state[REPORTS_LOADING_OVERLAY_KEY] = placeholder
+    return placeholder
+
+
+def _ensure_loading_overlay_placeholder() -> DeltaGenerator:
+    """Ensure the loading overlay placeholder exists and is reset when idle."""
+
+    placeholder = _get_loading_overlay_placeholder()
+    if not st.session_state.get(REPORTS_LOADING_FLAG_KEY, False):
+        placeholder.empty()
+    return placeholder
+
+
+def _show_reports_loading_overlay(message: str = "Loading reports…") -> None:
+    """Display the blur scrim while reports are loading."""
+
+    placeholder = _get_loading_overlay_placeholder()
+    st.session_state[REPORTS_LOADING_FLAG_KEY] = True
+    safe_message = escape(message)
+    placeholder.markdown(
+        f"""
+        <div class="reports-loading-scrim reports-loading-scrim--active" role="status" aria-live="assertive">
+            <div class="reports-loading-scrim__content">
+                <span class="reports-loading-scrim__halo"></span>
+                <span class="reports-loading-scrim__ring"></span>
+                <span class="reports-loading-scrim__label">{safe_message}</span>
+            </div>
+        </div>
+        """,
+        unsafe_allow_html=True,
+    )
+
+
+def _hide_reports_loading_overlay() -> None:
+    """Remove the reports loading scrim from view."""
+
+    placeholder = st.session_state.get(REPORTS_LOADING_OVERLAY_KEY)
+    if placeholder is not None:
+        placeholder.empty()
+    st.session_state[REPORTS_LOADING_FLAG_KEY] = False
 
 
 SNAPSHOT_KEYS = [
@@ -551,7 +603,8 @@ def _build_sidebar() -> None:
             st.sidebar.error("Fix the report date range before loading reports.")
             return
         try:
-            with st.spinner("Downloading and filtering reports..."):
+            _show_reports_loading_overlay()
+            with st.spinner("Fetching reports from server..."):
                 df = load_reports(
                     start_date=start_date.isoformat(),
                     end_date=end_date.isoformat(),
@@ -575,6 +628,8 @@ def _build_sidebar() -> None:
             st.success(f"Loaded {len(df)} reports into the workspace.")
         except Exception as exc:  # pragma: no cover - UI feedback
             st.error(f"Could not load reports: {exc}")
+        finally:
+            _hide_reports_loading_overlay()
 
     if api_key:
         llm_kwargs: Dict[str, Any] = {
@@ -683,6 +738,33 @@ def _render_header(container: Optional[DeltaGenerator] = None) -> None:
 
     ctx.markdown(hero_html, unsafe_allow_html=True)
 
+    if reports_df.empty:
[…]
