### 2025-10-22 cc3c777
Update favicon
