### 2025-10-23 de6c77d
Change field types to use natural language (not Pythonian)
