### 2025-10-23 924a1b7
Add theme discover tab after 'discover themes' has been run
