+++ streamlit_app.py
@@ -792,6 +792,79 @@ def _render_header(container: Optional[DeltaGenerator] = None) -> None:
 
     dataset_card.markdown("</div>", unsafe_allow_html=True)
 
+    _render_theme_summary_panel(ctx)
+
+
+def _render_theme_summary_panel(container: Optional[DeltaGenerator] = None) -> None:
+    """Render a collapsible card summarising accepted themes."""
+
+    theme_df = _coerce_dataframe(st.session_state.get("theme_summary_table"))
+    if theme_df is None or theme_df.empty:
+        return
+
+    ctx = container or st
+
+    display_df = theme_df.copy(deep=True)
+    if {"Count", "Theme"}.issubset(display_df.columns):
+        display_df = display_df.sort_values(
+            by=["Count", "Theme"], ascending=[False, True]
+        ).reset_index(drop=True)
+
+    reports_df = _get_reports_df()
+    total_reports = len(reports_df)
+    theme_count = len(display_df)
+
+    top_theme = display_df.iloc[0] if not display_df.empty else {}
+    top_theme_name = escape(str(top_theme.get("Theme", "—"))) if isinstance(top_theme, pd.Series) else "—"
+    top_theme_count = (
+        int(top_theme.get("Count", 0)) if isinstance(top_theme, pd.Series) else 0
+    )
+    top_theme_percentage_raw = (
+        float(top_theme.get("%", 0.0)) if isinstance(top_theme, pd.Series) else 0.0
+    )
+    top_theme_percentage = f"{top_theme_percentage_raw:.1f}%"
+
+    theme_label = "theme" if theme_count == 1 else "themes"
+    reports_label = "report" if total_reports == 1 else "reports"
+    top_theme_reports_label = "report" if top_theme_count == 1 else "reports"
+
+    panel = ctx.container()
+    panel.markdown(
+        f"""
+        <div class="section-card theme-summary-card">
+            <div class="section-card-header theme-summary-card__header">
+                <span class="section-kicker">Theme discovery</span>
+                <div class="theme-summary-card__title">
+                    <h3>Thematic snapshot</h3>
+                    <p>Revisit the thematic patterns surfaced by PFD Toolkit.</p>
+                </div>
+                <div class="theme-summary-card__meta">
+                    <span class="theme-summary-chip">{theme_count} {theme_label}</span>
+                    <span class="theme-summary-chip theme-summary-chip--accent">{total_reports:,} {reports_label} analysed</span>
+                </div>
+            </div>
+            <div class="theme-summary-highlight">
+                <div class="theme-summary-highlight__icon">✨</div>
+                <div class="theme-summary-highlight__content">
+                    <span class="theme-summary-highlight__label">Most prominent theme</span>
+                    <span class="theme-summary-highlight__value">{top_theme_name}</span>
+                    <span class="theme-summary-highlight__meta">{top_theme_count} {top_theme_reports_label} · {top_theme_percentage}</span>
+                </div>
+            </div>
+            <div class="theme-summary-card__expander">
+        """,
+        unsafe_allow_html=True,
+    )
+
+    expander = panel.expander("See all identified themes", expanded=False)
+    expander.markdown(
+        "<div class='theme-summary-table-caption'>Theme assignments across the working dataset</div>",
+        unsafe_allow_html=True,
+    )
+    expander.dataframe(display_df, width="stretch", hide_index=True)
+
+    panel.markdown("""</div></div>""", unsafe_allow_html=True)
+
 
 def _value_has_content(value: Any) -> bool:
     """Return ``True`` when ``value`` contains meaningful state."""
@@ -2717,6 +2790,141 @@ def main() -> None:
 
[…]
