### 2025-10-24 #8701 @galessiorob
👋 @KennethEnevoldsen- Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).
