### 2025-10-24 07a4395
Added caching of API calls and optional checkbox for elements to export
