+++ streamlit_app.py
@@ -10,7 +10,7 @@ from html import escape
 from datetime import date
 from io import BytesIO
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, Iterable, List, Optional, Tuple
 
 import pandas as pd
 import streamlit as st
@@ -31,6 +31,92 @@ from pfd_toolkit.screener import Screener
 
 LOGO_PATH = Path("docs/assets/badge-circle.png")
 OPENROUTER_API_BASE = "https://openrouter.ai/api/v1"
+
+
+# Keys for the reproducible script caching feature
+REPRO_SCRIPT_KEY = "repro_script_lines"
+REPRO_ACTION_COUNTS_KEY = "repro_action_counts"
+LLM_SIGNATURE_KEY = "llm_config_signature"
+
+
+# ---------------------------------------------------------------------------
+# Reproducible script helpers
+# ---------------------------------------------------------------------------
+
+def _initial_repro_script_lines() -> List[str]:
+    """Return the initial lines for the reproducible workspace script."""
+
+    return [
+        "# -----------------------------------------------------------------------------",
+        "# Reproducible workspace script",
+        "# This script contains the Python code for the various actions you performed in",
+        "# PFD Toolkit Workbench.",
+        "# To replay them locally, first install the Toolkit before running the script:",
+        "#     pip install pfd_toolkit",
+        "# Workbench does not save your API key. Make sure to add this to the script."
+        ""
+        "from pfd_toolkit import load_reports",
+        "from pfd_toolkit import LLM",
+        "from pfd_toolkit import Screener",
+        "from pfd_toolkit import Extractor",
+        "",
+    ]
+
+
+def _ensure_repro_script(reset: bool = False) -> List[str]:
+    """Ensure the reproducible script list exists and optionally reset it."""
+
+    lines = st.session_state.get(REPRO_SCRIPT_KEY)
+    if reset or not isinstance(lines, list):
+        lines = _initial_repro_script_lines()
+        st.session_state[REPRO_SCRIPT_KEY] = lines
+    return lines
+
+
+def _reset_repro_tracking() -> None:
+    """Reset the cached script and action counters."""
+
+    _ensure_repro_script(reset=True)
+    st.session_state[REPRO_ACTION_COUNTS_KEY] = {}
+    st.session_state[LLM_SIGNATURE_KEY] = None
+
+
+def _format_call(prefix: str, kwargs: Dict[str, Any], raw_parameters: Iterable[str] = ()) -> str:
+    """Return a formatted multi-line call string for the reproducible script."""
+
+    lines = [f"{prefix}("]
+    for key, value in kwargs.items():
+        if key in raw_parameters:
+            rendered = value
+        else:
+            rendered = repr(value)
+        lines.append(f"    {key}={rendered},")
+    lines.append(")")
+    return "\n".join(lines)
+
+
+def _record_repro_action(action_key: str, base_comment: str, code_block: str) -> None:
+    """Append an action to the reproducible script with run counters."""
+
+    lines = _ensure_repro_script()
+    counters: Dict[str, int] = st.session_state.setdefault(
+        REPRO_ACTION_COUNTS_KEY, {}
+    )
+    count = counters.get(action_key, 0) + 1
+    counters[action_key] = count
+    comment = base_comment if count == 1 else f"{base_comment} (run {count})"
+
+    if lines and lines[-1] != "":
+        lines.append("")
+    lines.append(f"# {comment}")
+    lines.extend(code_block.splitlines())
+
+
+def _get_repro_script_text() -> str:
+    """Return the reproducible script as a single text blob."""
+
[…]
