### 2025-10-24 b45a77e
Add LLM-powered custom emoji setting for 'most prominent theme'
