### 2025-10-24 770ee9b
Add more bann ed emojis and prompt engineered to ensure better sensitivity
