### 2025-10-24 3f4aec2
Add reproducible script keys to snapshot keys
