### 2025-10-26 #8701 @PetrKorab
@galessiorob Of course, I am happy to review the paper.
