### 2025-10-27 f3eebba
Copy dependencies into app dir
