### 2025-10-27 3f89e91
Add missing readme to app dir
