### 2025-10-27 c3f8c00
Update streamlit dependencies to install pfd_toolkit from PyPI, not local src dir
