+++ app/pyproject.toml
@@ -3,77 +3,12 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "pfd_toolkit"
-version = "0.3.7"
-description = "A toolkit for using Prevention of Future Death (PFD) Reports for Research and Policy Projects."
-readme = "README.md"
-license = "AGPL-3.0-or-later"
-license-files = ["LICENCE"]
-authors = [
-  { name = "Sam Osian", email = "samoand@liverpool.ac.uk" },
-  { name = "Jonathan Pytches", email = "jonathan.pytches@gmail.com" }
-]
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
-    "Operating System :: OS Independent",
-    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)"
-]
-
+name = "pfd-toolkit-workbench"
+version = "0.1.0"
 requires-python = ">=3.10"
+description = "Streamlit workbench for PFD Toolkit."
+
 dependencies = [
-    "adjusttext>=1.3.0",
-    "backoff>=2.2.1",
-    "beautifulsoup4>=4.13.4",
-    "dotenv>=0.9.9",
-    "httpx>=0.28.1",
-    "openai>=1.88.0",
-    "pandas>=2.2.2,<2.2.3; python_version<'3.13'",
-    "pandas>=2.2.3; python_version>='3.13'",
-    "pydantic>=2.11.7",
-    "pymupdf>=1.26.1",
-    "python-dateutil>=2.9.0.post0",
-    "requests>=2.32.3",
-    "tiktoken>=0.9.0",
-    "tqdm>=4.67.1",
+  "pfd-toolkit>=0.3.7",
+  "streamlit>=1.41.1"
 ]
-
-[project.optional-dependencies]
-app = [
-    "streamlit>=1.41.1",
-]
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/pfd_toolkit"]
-include = ["src/pfd_toolkit/data/*.csv"]
-
-[dependency-groups]
-dev = [
-    "pytest>=8.4.0",
-    "pytest-cov>=6.1.1",
-    "mkdocs-git-revision-date-localized-plugin>=1.4.7",
-    "pymdown-extensions>=10.15",
-    "mkdocs>=1.6.1",
-    "mkdocs-autorefs>=1.4.2",
-    "mkdocs-material[imaging]>=9.6.14",
-    "mkdocstrings[python]>=0.29.1",
-    "ipykernel>=6.29.5",
-    "ipython-pygments-lexers>=1.1.1",
-    "jupyter-client>=8.6.3",
-    "jupyter-core>=5.8.1",
-    "ruff>=0.12.0",
-    "twine>=6.1.0",
-    "openpyxl>=3.1.5",
-    "vl-convert-python>=1.8.0",
-    "matplotlib>=3.10.5",
-]
-app = [
-    "streamlit>=1.41.1",
-]
-
-[project.urls]
-"Homepage" = "https://github.com/Sam-Osian/PFD-toolkit"
-"Issues" = "https://github.com/Sam-Osian/PFD-toolkit/issues"
