### 2025-10-27 305c441
Try to stop hatchling from building package
