+++ app/pyproject.toml
@@ -7,8 +7,14 @@ name = "pfd-toolkit-workbench"
 version = "0.1.0"
 requires-python = ">=3.10"
 description = "Streamlit workbench for PFD Toolkit."
-
 dependencies = [
   "pfd-toolkit>=0.3.7",
   "streamlit>=1.41.1"
+]
+
+# Hatch assumes Python package; below limit the wheel to only the app files
+[tool.hatch.build.targets.wheel]
+only-include = [
+  "streamlit_app.py",
+  ".streamlit/config.toml"
 ]
