### 2025-10-27 e179adf
Add assets sub-dir to app/
