### 2025-10-29 8cfeef6
Add streamlit pytest; check for runtime exceptions & UI rendering
