### 2025-10-29 6fbed3f
Bump Workbench version to 0.1.1
