+++ app/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "pfd-toolkit-workbench"
-version = "0.1.0"
+version = "0.1.1"
 requires-python = ">=3.10"
 description = "Streamlit workbench for PFD Toolkit."
 dependencies = [
