+++ mkdocs.yml
@@ -7,19 +7,9 @@ theme:
   icon:
     repo: fontawesome/brands/github
   palette:
-    # Palette toggle for light mode
-    - media: "(prefers-color-scheme: light)"
-      scheme: default
-      toggle:
-        icon: material/brightness-4
-        name: Switch to dark mode
-
-    # Palette toggle for dark mode
-    - media: "(prefers-color-scheme: dark)"
-      scheme: slate
-      toggle:
-        icon: material/brightness-7
-        name: Switch to light mode
+    - scheme: slate
+      primary: custom
+      accent: custom
 
   name: material 
   logo: assets/badge-circle.png
