### 2025-11-03 ce62b1c
Pin Python 3.12 in pfd update workflow
