### 2025-11-04 97bc768
Pin dependencies for Workbench for better production reliability
