### 2025-11-06 f041e5b
Integrate the discover_themes workflow so it refreshes summaries when instructions or trim intensity change, relays thematic guidance into the summariser prompt, and records the settings for reuse
