+++ tests/test_extractor.py
@@ -450,12 +450,28 @@ def test_token_cache_export_import(tmp_path):
     assert ext2.token_cache == ext1.token_cache
 
 
-def test_discover_themes_basic():
-    df = pd.DataFrame({"summary": ["one", "two"]})
-    llm = DummyLLM(values={"safety": "Cases about safety"})
+def test_discover_themes_basic(monkeypatch):
+    df = pd.DataFrame(
+        {
+            GeneralConfig.COL_INVESTIGATION: ["one", "two"],
+            GeneralConfig.COL_CIRCUMSTANCES: ["circ1", "circ2"],
+            GeneralConfig.COL_CONCERNS: ["conc1", "conc2"],
+        }
+    )
+    llm = DummyLLM()
     ext = Extractor(llm=llm, reports=df)
-    ext.summarised_reports = df
-    ext.summary_col = "summary"
+
+    def fake_generate(prompts, response_format=None, **kwargs):
+        llm.called += len(prompts)
+        outputs = []
+        for prompt in prompts:
+            if "Report excerpt" in prompt:
+                outputs.append("summary text")
+            else:
+                outputs.append(json.dumps({"safety": "Cases about safety"}))
+        return outputs
+
+    monkeypatch.setattr(llm, "generate", fake_generate)
 
     theme_model = ext.discover_themes()
 
@@ -464,12 +480,22 @@ def test_discover_themes_basic():
     assert ext.feature_model is theme_model
 
 
-def test_discover_themes_handles_code_fence():
-    df = pd.DataFrame({"summary": ["one"]})
-    llm = DummyLLM(values="```json\n{\n  \"fence\": \"ok\"\n}\n```")
+def test_discover_themes_handles_code_fence(monkeypatch):
+    df = pd.DataFrame({GeneralConfig.COL_CONCERNS: ["one"]})
+    llm = DummyLLM()
     ext = Extractor(llm=llm, reports=df)
-    ext.summarised_reports = df
-    ext.summary_col = "summary"
+
+    def fake_generate(prompts, response_format=None, **kwargs):
+        llm.called += len(prompts)
+        outputs = []
+        for prompt in prompts:
+            if "Report excerpt" in prompt:
+                outputs.append("summary text")
+            else:
+                outputs.append("```json\n{\n  \"fence\": \"ok\"\n}\n```")
+        return outputs
+
+    monkeypatch.setattr(llm, "generate", fake_generate)
 
     theme_model = ext.discover_themes()
 
@@ -477,25 +503,49 @@ def test_discover_themes_handles_code_fence():
     assert ext.identified_themes == "```json\n{\n  \"fence\": \"ok\"\n}\n```"
 
 
-def test_discover_themes_bool_fields():
-    df = pd.DataFrame({"summary": ["one"]})
-    llm = DummyLLM(values={"example": "desc"})
+def test_discover_themes_bool_fields(monkeypatch):
+    df = pd.DataFrame({GeneralConfig.COL_CONCERNS: ["one"]})
+    llm = DummyLLM()
     ext = Extractor(llm=llm, reports=df)
-    ext.summarised_reports = df
-    ext.summary_col = "summary"
+
+    def fake_generate(prompts, response_format=None, **kwargs):
+        llm.called += len(prompts)
+        outputs = []
+        for prompt in prompts:
+            if "Report excerpt" in prompt:
+                outputs.append("summary text")
+            else:
+                outputs.append(json.dumps({"example": "desc"}))
+        return outputs
+
+    monkeypatch.setattr(llm, "generate", fake_generate)
 
     theme_model = ext.discover_themes()
 
     assert theme_model.model_fields["example"].annotation is bool
 
 
[…]
