+++ ons_replication/scripts/kappa_results.R
@@ -1,0 +1,172 @@
+# install.packages(c("readxl","dplyr","tidyr","purrr","irr"))
+library(readxl)
+library(dplyr)
+library(tidyr)
+library(purrr)
+library(irr)
+
+# irr doesn’t return a confidence interval for k.
+# It gives k and its z-statistic (z = k / SE), but not SE or variance.
+# We therefore recover SE as 'k / z' and build a 95% Wald CI: k +/- 1.96*SE.
+
+
+# Read data
+# ...data is an Excel spreadsheet with multiple tabs, but 'Randomised w 
+#    Annotations' is the only tab we need.
+
+file_path  <- "Downloads/PFD Toolkit Consensus Annotations.xlsx"
+sheet_name <- "Consensus annotations"
+
+raw <- read_excel(file_path, sheet = sheet_name)
+
+# Keep only the annotation cols + Toolkit output and put them in a tidier form 
+
+annotation_cols <- c("Dan: child & suicide", 
+                     "Arpan: child & suicide", 
+                     "Sahil: child & suicide")
+
+df <- raw %>% 
+  select(`PFD Toolkit: child suicide?`,
+         `Post-consensus verdict: Is this a child suicide case? (Yes or No)`,
+         all_of(annotation_cols)) %>% 
+  # trim leading/trailing spaces in all selected columns (handles stray spaces)
+  mutate(across(everything(), ~ trimws(as.character(.x)))) %>% 
+  mutate(across(everything(),
+                ~factor(.x, levels = c("No","Yes")))) %>% 
+  # short, convenient aliases for analysis while preserving original names above
+  rename(Toolkit  = `PFD Toolkit: child suicide?`,
+         Consensus = `Post-consensus verdict: Is this a child suicide case? (Yes or No)`)
+
+
+# Calculations!
+
+## For each clin, raw agreement & Cohen's k vs. Toolkit 
+
+### Documentation for kappa2(): https://www.rdocumentation.org/packages/irr/versions/0.84.1/topics/kappa2
+
+agreement_vs_toolkit <- map_dfr(
+  annotation_cols,
+  function(cln){
+    rater_col <- cln
+    k_out     <- kappa2(df[, c(rater_col, "Toolkit")], weight = "unweighted")
+    se        <- abs(k_out$value / k_out$statistic)
+    ci        <- k_out$value + c(-1.96, 1.96) * se
+    tibble(
+      Clinician       = cln,
+      Raw_Agreement   = mean(df[[rater_col]] == df$Toolkit, na.rm = TRUE),
+      Cohen_Kappa     = k_out$value,
+      CI_lower        = ci[1],
+      CI_upper        = ci[2]
+    )
+  }
+)
+
+## Consensus vs. Toolkit: raw agreement & Cohen's kappa
+
+### Documentation for kappa2(): https://www.rdocumentation.org/packages/irr/versions/0.84.1/topics/kappa2
+
+consensus_out <- kappa2(df[, c("Consensus", "Toolkit")], weight = "unweighted")
+consensus_se  <- abs(consensus_out$value / consensus_out$statistic)
+consensus_ci  <- consensus_out$value + c(-1.96, 1.96) * consensus_se
+
+consensus_vs_toolkit <- tibble(
+  Raw_Agreement = mean(df$Consensus == df$Toolkit, na.rm = TRUE),
+  Cohen_Kappa   = consensus_out$value,
+  CI_lower      = consensus_ci[1],
+  CI_upper      = consensus_ci[2]
+)
+
+## Overall three-rater Fleiss' kappa (inter-rater agreement)
+
[…]
