### 2025-11-10 #8701 @isdanni
> 👋 Hi [@isdanni](https://github.com/isdanni) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).

Hi! Appreciate the invite. I’m currently tied up with two other JOSS reviews, so I’ll have to pass this time.
