### 2025-11-12 728a7c8
Attempt to make the LLM less over-zealous when assigning a report with a list of candidate themes
