### 2025-11-24 1555966
Add truncation system as alternative to summarisation for thematic discovery pipeline
