+++ src/pfd_toolkit/extractor.py
@@ -8,6 +8,7 @@ import re
 import warnings
 from typing import Dict, List, Optional, Type, Union, Literal
 
+import matplotlib.pyplot as plt
 import pandas as pd
 from pydantic import BaseModel, Field, create_model, ConfigDict
 
@@ -119,6 +120,38 @@ class Extractor:
         self.token_cache: Dict[str, List[int]] = {}
         # Store raw LLM output from discover_themes for debugging
         self.identified_themes = None
+        self._last_summary_trim_intensity: Optional[str] = None
+        self._last_summary_discover_extra: Optional[str] = None
+        self._last_summary_truncate: Optional[int] = None
+
+    # ------------------------------------------------------------------
+    def _get_field_parts(
+        self, row: pd.Series, *, include_labels: bool = True
+    ) -> List[str]:
+        fields = [
+            (self.include_coroner, self.COL_CORONER_NAME, "Coroner name"),
+            (self.include_area, self.COL_AREA, "Area"),
+            (self.include_receiver, self.COL_RECEIVER, "Receiver"),
+            (
+                self.include_investigation,
+                self.COL_INVESTIGATION,
+                "Investigation and Inquest",
+            ),
+            (
+                self.include_circumstances,
+                self.COL_CIRCUMSTANCES,
+                "Circumstances of Death",
+            ),
+            (self.include_concerns, self.COL_CONCERNS, "Matters of Concern"),
+        ]
+
+        parts: List[str] = []
+        for flag, col, label in fields:
+            if flag and col in row.index:
+                val = row.get(col)
+                if pd.notna(val):
+                    parts.append(f"{label}: {str(val)}" if include_labels else str(val))
+        return parts
 
     # ------------------------------------------------------------------
     def _build_prompt_template(self) -> str:
@@ -499,6 +532,7 @@ Here is the report excerpt:
         trim_intensity: str = "medium",
         extra_instructions: Optional[str] = None,
         discover_themes_extra_instructions: Optional[str] = None,
+        truncate: Optional[int] = None,
     ) -> pd.DataFrame:
         """Summarise selected report fields into one column using the LLM.
 
@@ -506,8 +540,10 @@ Here is the report excerpt:
         ----------
         result_col_name : str, optional
             Name of the summary column. Defaults to ``"summary"``.
-        trim_intensity : {"low", "medium", "high", "very high"}, optional
+        trim_intensity : {"none", "low", "medium", "high", "very high"}, optional
             Controls how concise the summary should be. Defaults to ``"medium"``.
+            When set to ``"none"``, no LLM summarisation occurs and the selected
+            report fields are concatenated instead.
         extra_instructions : str, optional
             Additional instructions to append to the prompt before the report
             excerpt.
@@ -517,15 +553,24 @@ Here is the report excerpt:
             prompt reminds the model that downstream theme discovery will follow
             specific instructions. The value should typically mirror the
             ``extra_instructions`` argument passed to ``discover_themes``.
+        truncate : int, optional
+            Word limit applied when ``trim_intensity`` is ``"none"``. Text will
+            be truncated to the first ``truncate`` words after concatenation.
+            Providing this argument with any other ``trim_intensity`` raises a
+            ``ValueError``.
 
         Returns
-        ------- 
+        -------
         pandas.DataFrame
             A new DataFrame identical to the one provided at initialisation with
             an extra summary column.
         """
-        if trim_intensity not in {"low", "medium", "high", "very high"}:
-            raise ValueError("trim_intensity must be 'low', 'medium', 'high', or 'very high'")
+        if trim_intensity not in {"none", "low", "medium", "high", "very high"}:
+            raise ValueError(
+                "trim_intensity must be 'none', 'low', 'medium', 'high', or 'very high'"
+            )
+        if truncate is not None and trim_intensity != "none":
[…]
