+++ src/pfd_toolkit/extractor.py
@@ -11,6 +11,7 @@ from typing import Dict, List, Optional, Type, Union, Literal
 import matplotlib.pyplot as plt
 import pandas as pd
 from pydantic import BaseModel, Field, create_model, ConfigDict
+import tiktoken
 
 from .llm import LLM
 from .config import GeneralConfig
@@ -120,9 +121,11 @@ class Extractor:
         self.token_cache: Dict[str, List[int]] = {}
         # Store raw LLM output from discover_themes for debugging
         self.identified_themes = None
-        self._last_summary_trim_intensity: Optional[str] = None
+        self._last_summary_summarise_intensity: Optional[str] = None
         self._last_summary_discover_extra: Optional[str] = None
-        self._last_summary_truncate: Optional[int] = None
+        self._last_summary_trim_approach: Optional[str] = None
+        self._last_summary_max_tokens: Optional[int] = None
+        self._last_summary_max_words: Optional[int] = None
 
     # ------------------------------------------------------------------
     def _get_field_parts(
@@ -529,10 +532,12 @@ Here is the report excerpt:
     def summarise(
         self,
         result_col_name: str = "summary",
-        trim_intensity: str = "medium",
+        trim_approach: Literal["summarise", "truncate"] = "truncate",
+        summarise_intensity: Optional[str] = None,
         extra_instructions: Optional[str] = None,
         discover_themes_extra_instructions: Optional[str] = None,
-        truncate: Optional[int] = None,
+        max_tokens: Optional[int] = None,
+        max_words: Optional[int] = None,
     ) -> pd.DataFrame:
         """Summarise selected report fields into one column using the LLM.
 
@@ -540,10 +545,15 @@ Here is the report excerpt:
         ----------
         result_col_name : str, optional
             Name of the summary column. Defaults to ``"summary"``.
-        trim_intensity : {"none", "low", "medium", "high", "very high"}, optional
-            Controls how concise the summary should be. Defaults to ``"medium"``.
-            When set to ``"none"``, no LLM summarisation occurs and the selected
-            report fields are concatenated instead.
+        trim_approach : {"summarise", "truncate"}, optional
+            Choose whether to use LLM-driven summarisation (``"summarise"``) or
+            concatenation with optional trimming (``"truncate"``). Defaults to
+            ``"truncate"`` which performs no trimming unless ``max_tokens`` or
+            ``max_words`` is provided.
+        summarise_intensity : {"low", "medium", "high", "very high"} or None, optional
+            Controls how concise the summary should be when
+            ``trim_approach="summarise"``. Defaults to ``"medium"`` when
+            omitted. Cannot be used alongside ``trim_approach="truncate"``.
         extra_instructions : str, optional
             Additional instructions to append to the prompt before the report
             excerpt.
@@ -553,11 +563,14 @@ Here is the report excerpt:
             prompt reminds the model that downstream theme discovery will follow
             specific instructions. The value should typically mirror the
             ``extra_instructions`` argument passed to ``discover_themes``.
-        truncate : int, optional
-            Word limit applied when ``trim_intensity`` is ``"none"``. Text will
-            be truncated to the first ``truncate`` words after concatenation.
-            Providing this argument with any other ``trim_intensity`` raises a
-            ``ValueError``.
+        max_tokens : int, optional
+            Token limit applied when ``trim_approach="truncate"``. The
+            concatenated text will be trimmed to approximately this number of
+            tokens. Cannot be combined with ``max_words``.
+        max_words : int, optional
+            Word limit applied when ``trim_approach="truncate"``. The
+            concatenated text will be trimmed to the first ``max_words`` words.
+            Cannot be combined with ``max_tokens``.
 
         Returns
         -------
@@ -565,12 +578,23 @@ Here is the report excerpt:
[…]
