### 2025-11-25 47941f3
Improve count() display for stats
