### 2025-11-25 cb87af5
Upgrade Workbench to use Toolkit v0.4.0
