+++ app/streamlit_app.py
@@ -532,7 +532,7 @@ def _display_dataframe(df: pd.DataFrame, caption: str) -> None:
         st.info("No rows to display yet. Load or generate data to see results here.")
         return
     st.markdown(f"<div class='section-caption'>{caption}</div>", unsafe_allow_html=True)
-    st.dataframe(df, width="stretch", hide_index=True)
+    st.dataframe(df, use_container_width=True, hide_index=True)
 
 
 def _render_reports_overview(
@@ -563,7 +563,7 @@ def _render_reports_overview(
         if st.button(
             "Revert changes",
             key=f"revert_reports_{key_suffix}" if key_suffix else "revert_reports",
-            width="stretch",
+            use_container_width=True,
         ):
             st.session_state["reports_df"] = initial_df.copy()
             st.session_state["reports_df_modified"] = False
@@ -737,7 +737,7 @@ def _build_sidebar() -> None:
         "Force refresh from remote dataset", value=True, help="Disable to reuse the cached CSV if available."
     )
 
-    load_button = st.sidebar.button("Load in reports", width="stretch")
+    load_button = st.sidebar.button("Load in reports", use_container_width=True)
 
     if load_button:
         if n_reports_raw.strip() and n_reports is None:
@@ -959,7 +959,7 @@ def _render_header(container: Optional[DeltaGenerator] = None) -> None:
         unsafe_allow_html=True,
     )
 
-    dataset_card.dataframe(reports_df, width="stretch", hide_index=True)
+    dataset_card.dataframe(reports_df, use_container_width=True, hide_index=True)
 
     dataset_card.markdown("</div>", unsafe_allow_html=True)
 
@@ -1091,7 +1091,7 @@ def _render_theme_summary_panel(container: Optional[DeltaGenerator] = None) -> N
         "<div class='theme-summary-table-caption'>Theme assignments across the working dataset</div>",
         unsafe_allow_html=True,
     )
-    expander.dataframe(display_df, width="stretch", hide_index=True)
+    expander.dataframe(display_df, use_container_width=True, hide_index=True)
 
     panel.markdown("""</div></div>""", unsafe_allow_html=True)
 
@@ -1206,7 +1206,7 @@ def _render_workspace_footer(
         if st.button(
             undo_label,
             key="footer_undo",
-            width="stretch",
+            use_container_width=True,
             disabled=undo_disabled,
         ):
             _undo_last_change()
@@ -1215,7 +1215,7 @@ def _render_workspace_footer(
         if st.button(
             redo_label,
             key="footer_redo",
-            width="stretch",
+            use_container_width=True,
             disabled=redo_disabled,
         ):
             _redo_last_change()
@@ -1224,7 +1224,7 @@ def _render_workspace_footer(
         if st.button(
             "↻ Start over",
             key="footer_reset",
-            width="stretch",
+            use_container_width=True,
             disabled=start_over_disabled,
         ):
             _start_again()
@@ -1405,7 +1405,7 @@ def _render_action_tiles() -> None:
                 if card.button(
                     f"{action['icon']} {action['label']}",
                     key=action["key"],
-                    width="stretch",
+                    use_container_width=True,
                     disabled=action["disabled"],
                 ):
                     st.session_state["active_action"] = action["target"]
@@ -1592,7 +1592,7 @@ def _render_save_action() -> None:
         data=zip_bytes,
         file_name="pfd_research_bundle.zip",
         mime="application/zip",
-        width="stretch",
+        use_container_width=True,
         disabled=not has_selection,
     )
 
@@ -1697,7 +1697,7 @@ def _render_filter_action() -> None:
         with st.container():
             submitted = st.form_submit_button(
                 "Run Screener",
-                width="stretch",
[…]
