### 2025-11-25 5ad5bc2
In Workbench, remove reference to local package
