### 2025-11-26 c967ccd
Convert model_comparison.py to notebook
