+++ tests/open_weight_experiment/model_comparison.py
@@ -1,0 +1,231 @@
+"""Compare screener performance across local and proprietary LLMs."""
+
+from pathlib import Path
+import os
+
+import pandas as pd
+from dotenv import load_dotenv
+
+from pfd_toolkit import LLM, Screener
+from pfd_toolkit.config import GeneralConfig
+
+DATA_PATH = Path("../../ons_replication/PFD Toolkit--Consensus Comparison.xlsx")
+RESULTS_PATH = Path("model_comparison.csv")
+SHEET_NAME = "Consensus annotations"
+
+MODEL_SPECS = [
+    # OpenAI API models
+    {"name": "gpt-4.1", "temperature": 0},
+    {"name": "gpt-4o", "temperature": 0},
+    {"name": "gpt-4.1-mini", "temperature": 0},
+    {"name": "gpt-4.1-nano", "temperature": 0},
+    {"name": "gpt-5", "temperature": 1},
+    {"name": "gpt-5-mini", "temperature": 1},
+    {"name": "gpt-5-nano", "temperature": 1},
+    {"name": "o4-mini", "temperature": 1},
+    {"name": "o3", "temperature": 1},
+
+    # Ollama-hosted models
+    {
+        "name": "mistral-nemo:12b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "mistral-small:22b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "mistral-small:24b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "gemma3:12b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "gemma3:27b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "gemma2:27b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "qwen3:32b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "qwen3:30b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "qwen2.5:72b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "qwen2.5:32b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "llava:34b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "phi4:14b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
+        "api_key": "ollama",
+        "timeout": 10**9,
+    },
+    {
+        "name": "llama3:70b",
+        "temperature": 0,
+        "base_url": "http://localhost:11434/v1",
[…]
