+++ tests/model_comparison.py
@@ -1,378 +1,0 @@
-"""Compare Screener performance across multiple LLM models against ONS consensus labels.
-
-This script reads report sections and ground-truth labels directly from
-``ons_replication/ONS_master_spreadsheet.xlsx`` and measures the accuracy,
-sensitivity, and specificity for several LLM models. Results are written to
-``model_comparison.csv``.
-"""
-
-from __future__ import annotations
-
-import os
-from pathlib import Path
-
-import pandas as pd
-from dotenv import load_dotenv
-
-from pfd_toolkit import LLM, Screener
-from pfd_toolkit.config import GeneralConfig
-
-# ---------------------------------------------------------------------
-# 1. Load ONS ground-truth data
-# ---------------------------------------------------------------------
-ons_path = (
-    Path(__file__).resolve().parent.parent
-    / "ons_replication"
-    / "ONS_master_spreadsheet.xlsx"
-)
-ons_df = pd.read_excel(ons_path, sheet_name=0)
-
-ons_df = ons_df.rename(
-    columns={
-        "Investigation section": GeneralConfig.COL_INVESTIGATION,
-        "Circumstances of death section": GeneralConfig.COL_CIRCUMSTANCES,
-        "Matters of concern section": GeneralConfig.COL_CONCERNS,
-        "Consensus": "consensus",
-        "Ref": GeneralConfig.COL_ID,
-    }
-)
-
-keep_cols = [
-    GeneralConfig.COL_ID,
-    GeneralConfig.COL_INVESTIGATION,
-    GeneralConfig.COL_CIRCUMSTANCES,
-    GeneralConfig.COL_CONCERNS,
-    "consensus",
-]
-reports = ons_df[keep_cols].copy()
-
-reports["consensus"] = (
-    reports["consensus"]
-    .astype(str)
-    .str.strip()
-    .str.lower()
-    .map({"yes": True, "no": False, "1": True, "0": False})
-)
-
-print(f"Loaded {len(reports)} reports from ONS spreadsheet.")
-
-# ---------------------------------------------------------------------
-# 2. Initialise models and query
-# ---------------------------------------------------------------------
-load_dotenv(Path(__file__).resolve().parent.parent / "api.env")
-
-# Settings for OpenRouter models
-OPENROUTER_KEY = os.getenv("OPENROUTER_API_KEY")
-OPENROUTER_URL = "https://openrouter.ai/api/v1"
-
-# List of models and the temperature values to use for each.
-# GPT-5 models do not support a custom temperature parameter, so while a
-# default value is recorded here for completeness, it is not passed to the
-# client when initialising those models.
-models = [
-    # OPENAI API MODELS
-    {"name": "gpt-4.1", "temperature": 0},
-    {"name": "gpt-4o", "temperature": 0},
-    {"name": "gpt-4.1-mini", "temperature": 0},
-    {"name": "gpt-4.1-nano", "temperature": 0},
-    {"name": "gpt-5", "temperature": 1},
-    {"name": "gpt-5-mini", "temperature": 1},
-    {"name": "gpt-5-nano", "temperature": 1},
-    {"name": "o4-mini", "temperature": 1},
-    {"name": "o3", "temperature": 1},
-
-    # OPENROUTER MISTRAL MODELS
-    {
-        "name": "mistralai/devstral-small",
-        "temperature": 0,
-        "base_url": OPENROUTER_URL,
-        "api_key": OPENROUTER_KEY,
-    },
-    {
-        "name": "mistralai/devstral-medium",
-        "temperature": 0,
-        "base_url": OPENROUTER_URL,
-        "api_key": OPENROUTER_KEY,
-    },
-    {
-        "name": "mistralai/mistral-medium-3.1",
-        "temperature": 0,
[…]
