### 2025-12-03 ffe0b8a
Add functionality for scraping responses
