+++ src/pfd_toolkit/scraper/scraper.py
@@ -10,6 +10,8 @@ from tqdm import tqdm
 import requests
 from itertools import count
 from datetime import datetime
+from urllib.parse import urljoin
+from pydantic import BaseModel, ConfigDict
 
 from .html_extractor import HtmlExtractor
 from .pdf_extractor import PdfExtractor
@@ -131,6 +133,8 @@ class Scraper:
     COL_CIRCUMSTANCES = GeneralConfig.COL_CIRCUMSTANCES
     COL_CONCERNS = GeneralConfig.COL_CONCERNS
     COL_DATE_SCRAPED = GeneralConfig.COL_DATE_SCRAPED
+
+    RESPONSE_ATTRIBUTION_ERROR = GeneralConfig.RESPONSE_ATTRIBUTION_ERROR
 
     # Keys used for LLM interaction when requesting missing fields
     LLM_KEY_DATE = ScraperConfig.LLM_KEY_DATE
@@ -346,6 +350,9 @@ class Scraper:
             "concerns": self.include_concerns,
         }
 
+        # Cache for response document bytes
+        self._response_cache: dict[str, bytes] = {}
+
 
 
     def _parse_scraping_strategy(self, strategy: list[int] | tuple[int, int, int]) -> None:
@@ -512,6 +519,232 @@ class Scraper:
         
         self.reports = updated_df.copy()
         return updated_df
+
+    def scrape_responses(self, reports: pd.DataFrame | None = None) -> pd.DataFrame:
+        """Scrape responses to previously scraped PFD reports using the LLM only.
+
+        Parameters
+        ----------
+        reports : pandas.DataFrame | None
+            DataFrame of scraped PFD reports. Defaults to ``self.reports``.
+
+        Returns
+        -------
+        pandas.DataFrame
+            A DataFrame containing ``url``, ``respondent`` and ``response``
+            columns for validated responses.
+        """
+
+        if reports is None:
+            if self.reports is None:
+                raise ValueError(
+                    "No reports provided. Please supply a reports DataFrame or run scrape_reports() first."
+                )
+            reports = self.reports
+
+        if self.llm is None:
+            raise ValueError("LLM client (self.llm) must be provided to scrape responses.")
+
+        if reports.empty:
+            logger.info("No reports available to scrape responses from.")
+            return pd.DataFrame(columns=[self.COL_URL, "respondent", "response"])
+
+        responses: list[dict[str, str]] = []
+        candidates: list[dict[str, Any]] = []
+
+        class ResponseCheckModel(BaseModel):
+            is_response: bool
+
+            model_config = ConfigDict(extra="forbid")
+
+        class ResponseExtractionModel(BaseModel):
+            respondent: str | None
+            response: str | None
+
+            model_config = ConfigDict(extra="forbid")
+
+        for _, row in tqdm(
+            reports.iterrows(),
+            total=len(reports),
+            desc="Collecting response documents",
+            leave=True,
+        ):
+            report_url = row.get(self.COL_URL)
+            receivers = row.get(self.COL_RECEIVER)
+
+            if pd.isna(report_url) or pd.isna(receivers):
+                continue
+
+            receiver_options = [r.strip() for r in str(receivers).split(";") if r.strip()]
+            if not receiver_options:
+                continue
+
+            pdf_links = self._collect_pdf_links(str(report_url))
+            if len(pdf_links) <= 1:
+                continue
+
+            for response_link in pdf_links[1:]:
+                pdf_bytes = self._fetch_response_bytes(response_link)
+                if not pdf_bytes:
+                    continue
+
+                try:
+                    base64_images = self.llm._pdf_bytes_to_base64_images(pdf_bytes, dpi=200)
+                except Exception as exc:  # pragma: no cover - defensive
[…]
