+++ src/pfd_toolkit/scraper/scraper.py
@@ -520,20 +520,34 @@ class Scraper:
         self.reports = updated_df.copy()
         return updated_df
 
-    def scrape_responses(self, reports: pd.DataFrame | None = None) -> pd.DataFrame:
+    def scrape_responses(
+        self, reports: pd.DataFrame | None = None, *, verify_responses: bool = True
+    ) -> pd.DataFrame:
         """Scrape responses to previously scraped PFD reports using the LLM only.
 
         Parameters
         ----------
         reports : pandas.DataFrame | None
             DataFrame of scraped PFD reports. Defaults to ``self.reports``.
+        verify_responses : bool, optional
+            When ``True``, confirm each document is a response before scraping.
+            When ``False``, assume all discovered PDFs are responses and skip the
+            validation stage.
 
         Returns
         -------
         pandas.DataFrame
-            A DataFrame containing ``url``, ``respondent`` and ``response``
-            columns for validated responses.
+            A DataFrame containing ``id``, ``parent_url``, ``response_url``,
+            ``respondent`` and ``response`` columns for validated responses.
         """
+
+        responses_columns = [
+            self.COL_ID,
+            "parent_url",
+            "response_url",
+            "respondent",
+            "response",
+        ]
 
         if reports is None:
             if self.reports is None:
@@ -547,9 +561,9 @@ class Scraper:
 
         if reports.empty:
             logger.info("No reports available to scrape responses from.")
-            return pd.DataFrame(columns=[self.COL_URL, "respondent", "response"])
+            return pd.DataFrame(columns=responses_columns)
 
-        responses: list[dict[str, str]] = []
+        responses: list[dict[str, Any]] = []
         candidates: list[dict[str, Any]] = []
 
         class ResponseCheckModel(BaseModel):
@@ -570,6 +584,7 @@ class Scraper:
             leave=True,
         ):
             report_url = row.get(self.COL_URL)
+            report_id = row.get(self.COL_ID)
             receivers = row.get(self.COL_RECEIVER)
 
             if pd.isna(report_url) or pd.isna(receivers):
@@ -584,19 +599,28 @@ class Scraper:
                 continue
 
             for response_link in pdf_links[1:]:
-                pdf_bytes = self._fetch_response_bytes(response_link)
+                pdf_bytes = self._fetch_response_bytes(
+                    response_link, parent_url=report_url, report_id=report_id
+                )
                 if not pdf_bytes:
                     continue
 
                 try:
                     base64_images = self.llm._pdf_bytes_to_base64_images(pdf_bytes, dpi=200)
                 except Exception as exc:  # pragma: no cover - defensive
-                    logger.error("Failed to convert PDF at %s to images: %s", response_link, exc)
+                    logger.error(
+                        "Failed to convert response PDF %s from %s (%s) to images: %s",
+                        response_link,
+                        report_url,
+                        report_id,
+                        exc,
+                    )
                     continue
 
                 candidates.append(
                     {
-                        "report_url": str(report_url),
+                        "report_id": report_id,
+                        "parent_url": str(report_url),
                         "response_url": response_link,
                         "receiver_options": receiver_options,
                         "base64_images": base64_images,
@@ -604,29 +628,36 @@ class Scraper:
                 )
 
         if not candidates:
-            return pd.DataFrame(columns=[self.COL_URL, "respondent", "response"])
+            return pd.DataFrame(columns=responses_columns)
 
-        check_prompts = [self._build_response_check_prompt() for _ in candidates]
-        check_images = [candidate["base64_images"] for candidate in candidates]
+        if verify_responses:
+            check_prompts = [self._build_response_check_prompt() for _ in candidates]
[…]
