+++ src/pfd_toolkit/__init__.py
@@ -21,6 +21,13 @@ from pfd_toolkit.llm import LLM
 from pfd_toolkit.loader import load_reports
 from pfd_toolkit.screener import Screener
 from pfd_toolkit.extractor import Extractor
+from pfd_toolkit.paired_statements import (
+    ConcernItem,
+    ConcernPairGenerator,
+    ConcernResults,
+    ConcernSet,
+    ReportConcerns,
+)
 from pfd_toolkit.config import (
     GeneralConfig,
     ScraperConfig,
@@ -33,6 +40,11 @@ __all__ = [
     "load_reports",
     "Screener",
     "Extractor",
+    "ConcernItem",
+    "ConcernPairGenerator",
+    "ConcernResults",
+    "ConcernSet",
+    "ReportConcerns",
     "GeneralConfig",
     "ScraperConfig",
     "HtmlFieldConfig",
