+++ src/pfd_toolkit/paired_statements.py
@@ -1,0 +1,279 @@
+"""LLM-powered concern pairing for Prevention of Future Deaths reports.
+
+This module focuses on converting the free-text "Matters of Concern" section
+into a structured list of concise, discrete concerns suitable for downstream
+pairing and analysis.
+"""
+from __future__ import annotations
+
+from typing import Any, Iterable, List, Mapping, Sequence
+
+import pandas as pd
+from pydantic import BaseModel, Field, field_validator
+
+from .config import GeneralConfig
+from .llm import LLM
+
+
+class ConcernItem(BaseModel):
+    """A single distilled concern from a coroner's report."""
+
+    concern: str = Field(
+        ..., description="One short concern capturing a single specific risk or failing."
+    )
+
+    @field_validator("concern")
+    @classmethod
+    def _normalise_text(cls, value: str) -> str:
+        cleaned = " ".join((value or "").split()).strip()
+        if not cleaned:
+            raise ValueError("concern text cannot be empty")
+        return cleaned
+
+
+class ConcernSet(BaseModel):
+    """Ordered collection of concerns extracted from a PFD report."""
+
+    concerns: List[ConcernItem] = Field(
+        default_factory=list,
+        description=(
+            "A list of discrete concerns, kept in the same order as the "
+            "original Matters of Concern section."
+        ),
+    )
+
+
+class ReportConcerns(BaseModel):
+    """Concerns for a single report plus propagated metadata."""
+
+    url: str = Field(..., description="Canonical URL for the PFD report.")
+    report_id: str | None = Field(
+        default=None, description="Unique identifier scraped from the report URL."
+    )
+    recipients: List[str] = Field(
+        default_factory=list, description="Recipient list split from the reports CSV."
+    )
+    concern_set: ConcernSet = Field(
+        ..., description="Structured concerns parsed from the Matters of Concern section."
+    )
+    responses: List[str] = Field(
+        default_factory=list,
+        description="All response texts linked to this report's URL (one-to-many).",
+    )
+
+    @property
+    def concerns(self) -> List[ConcernItem]:
+        """Expose the inner concerns list for ergonomic access."""
+
+        return self.concern_set.concerns
+
+
+class ConcernResults(BaseModel):
+    """Container for multiple reports worth of concerns."""
+
+    reports: List[ReportConcerns] = Field(default_factory=list)
+
+    def as_list(self) -> List[dict[str, Any]]:
+        """Return a JSON-ready list of report-level concerns."""
+
+        serialised: List[dict[str, Any]] = []
+        for report in self.reports:
+            serialised.append(
+                {
+                    GeneralConfig.COL_URL: report.url,
+                    GeneralConfig.COL_ID: report.report_id,
+                    GeneralConfig.COL_RECEIVER: report.recipients,
+                    "responses": report.responses,
+                    "concerns": [item.concern for item in report.concerns],
+                }
+            )
+        return serialised
+
+    def as_df(self) -> pd.DataFrame:
+        """Return a flat DataFrame, one row per concern."""
+
+        rows: List[dict[str, Any]] = []
+        for report in self.reports:
+            for item in report.concerns:
+                rows.append(
+                    {
+                        GeneralConfig.COL_URL: report.url,
+                        GeneralConfig.COL_ID: report.report_id,
+                        GeneralConfig.COL_RECEIVER: report.recipients,
+                        "concern": item.concern,
+                    }
+                )
+        return pd.DataFrame(rows)
+
+
+class ConcernPairGenerator:
+    """Generate discrete concerns from narrative text using an LLM.
+
+    Parameters
+    ----------
+    llm : LLM
[…]
