+++ tests/test_paired_statements.py
@@ -1,0 +1,111 @@
+import pandas as pd
+
+from pfd_toolkit.config import GeneralConfig
+from pfd_toolkit.paired_statements import ConcernItem, ConcernPairGenerator, ConcernSet
+
+
+class DummyLLM:
+    """Minimal stub matching the LLM.generate interface used in tests."""
+
+    def __init__(self, responses: list[ConcernSet] | None = None):
+        self.responses = responses or []
+
+    def generate(self, prompts, response_format):
+        if self.responses:
+            return self.responses
+        return [response_format() for _ in prompts]
+
+
+def test_concern_item_normalises_whitespace():
+    item = ConcernItem(concern="  multiple   spaces\n and lines ")
+    assert item.concern == "multiple spaces and lines"
+
+
+def test_prompt_includes_context_and_source_text():
+    generator = ConcernPairGenerator(llm=DummyLLM())
+    sample_text = "I am concerned about how emergency calls are triaged."
+    prompt = generator.build_prompt(sample_text)
+
+    assert "Prevention of Future Deaths" in prompt
+    assert "Matters of Concern" in prompt
+    assert sample_text in prompt
+    assert "Guidance for identifying concerns" in prompt
+
+
+def test_extract_for_reports_matches_url_and_parent_url():
+    llm = DummyLLM(
+        responses=[
+            ConcernSet(concerns=[ConcernItem(concern="one")]),
+            ConcernSet(concerns=[ConcernItem(concern="two")]),
+        ]
+    )
+    generator = ConcernPairGenerator(llm=llm)
+
+    reports = [
+        {
+            GeneralConfig.COL_URL: "https://example.test/report-1",
+            GeneralConfig.COL_CONCERNS: "concern text",
+            GeneralConfig.COL_RECEIVER: "A;B",
+            GeneralConfig.COL_ID: "2024-0001",
+        },
+        {
+            GeneralConfig.COL_URL: "https://example.test/report-2",
+            GeneralConfig.COL_CONCERNS: "another",
+            GeneralConfig.COL_RECEIVER: "",
+            GeneralConfig.COL_ID: "2024-0002",
+        },
+    ]
+    responses = [
+        {"parent_url": "https://example.test/report-1", "response": "resp A"},
+        {"parent_url": "https://example.test/report-1", "response": "resp B"},
+        {"parent_url": "https://example.test/report-3", "response": "irrelevant"},
+    ]
+
+    paired = generator.build_concerns(reports=reports, responses=responses)
+
+    assert paired.reports[0].url == reports[0][GeneralConfig.COL_URL]
+    assert [item.concern for item in paired.reports[0].concerns] == ["one"]
+    assert paired.reports[0].responses == ["resp A", "resp B"]
+    assert paired.reports[0].recipients == ["A", "B"]
+    assert paired.reports[0].report_id == "2024-0001"
+
+    assert paired.reports[1].url == reports[1][GeneralConfig.COL_URL]
+    assert paired.reports[1].responses == []
+    assert paired.reports[1].recipients == []
+
+
+def test_as_df_flattens_concerns():
+    llm = DummyLLM(responses=[ConcernSet(concerns=[ConcernItem(concern="one")])])
+    generator = ConcernPairGenerator(llm=llm)
+
+    reports_df = pd.DataFrame(
+        [
+            {
+                GeneralConfig.COL_URL: "https://example.test/report-1",
+                GeneralConfig.COL_CONCERNS: "concern text",
+                GeneralConfig.COL_RECEIVER: "Receiver A; Receiver B",
+                GeneralConfig.COL_ID: "2024-0001",
[…]
