### 2025-12-04 abb13d1
Refactor ConcernsParser
