+++ src/pfd_toolkit/paired_statements.py
@@ -106,8 +106,8 @@ class ConcernResults(BaseModel):
         return pd.DataFrame(rows)
 
 
-class ConcernPairGenerator:
-    """Generate discrete concerns from narrative text using an LLM.
+class ConcernParser:
+    """LLM-based manipulation of coroners' concerns.
 
     Parameters
     ----------
@@ -115,6 +115,11 @@ class ConcernPairGenerator:
         LLM client used to call the OpenAI API. Parallelisation and retry
         settings are controlled directly by this object and should be provided
         by the caller.
+    reports : Iterable[Mapping[str, Any]] | Any
+        Table-like collection of PFD reports containing at least the URL and
+        Matters of Concern text.
+    responses : Iterable[Mapping[str, Any]] | Any | None
+        Optional table-like collection of responses keyed by parent URL.
     """
 
     REPORT_URL_COLUMN = GeneralConfig.COL_URL
@@ -124,8 +129,16 @@ class ConcernPairGenerator:
     RESPONSES_PARENT_URL_COLUMN = "parent_url"
     RESPONSES_TEXT_COLUMN = "response"
 
-    def __init__(self, llm: LLM):
+    def __init__(
+        self,
+        llm: LLM,
+        reports: Iterable[Mapping[str, Any]] | Any,
+        responses: Iterable[Mapping[str, Any]] | Any | None = None,
+    ):
         self.llm = llm
+        self.reports = reports
+        self.responses = responses
+        self._results: ConcernResults | None = None
 
     def build_prompt(self, concerns_section: str) -> str:
         """Build a detailed prompt to split a concerns section into items."""
@@ -152,6 +165,7 @@ class ConcernPairGenerator:
             "Only populate the `concerns` field of the response schema with the "
             "final list. Do not invent information beyond the supplied text.\n\n"
             "Concerns section to analyse:\n"
+            "The Coroner's Concerns text is below:\n\n"
             + concerns_section.strip()
         )
 
@@ -180,16 +194,21 @@ class ConcernPairGenerator:
                 raise ValueError(f"Unexpected LLM response at index {idx}: {result}")
         return parsed
 
-    def build_concerns(
-        self,
-        reports: Iterable[Mapping[str, Any]] | Any,
-        responses: Iterable[Mapping[str, Any]] | Any | None = None,
-    ) -> ConcernResults:
-        """Extract concerns for reports and attach recipients and matching responses."""
+    def parse_concerns(self, output: str = "json") -> ConcernResults | list[dict[str, Any]] | pd.DataFrame:
+        """Parse discrete concerns from Matters of Concern text in PFD reports.
 
-        report_records = self._records_from_table(reports)
+        Parameters
+        ----------
+        output : {"json", "dataframe", "object"}
+            Controls the return type. ``"json"`` (default) returns a list of
+            JSON-serialisable dictionaries. ``"dataframe"`` returns a pandas
+            DataFrame with one row per concern. ``"object"`` returns the
+            :class:`ConcernResults` instance for further manipulation.
+        """
+
+        report_records = self._records_from_table(self.reports)
         self._validate_report_columns(report_records)
-        response_lookup = self._response_lookup(responses)
+        response_lookup = self._response_lookup(self.responses)
 
         urls: List[str] = []
         ids: List[str | None] = []
@@ -220,7 +239,16 @@ class ConcernPairGenerator:
                 )
             )
 
-        return ConcernResults(reports=paired)
+        self._results = ConcernResults(reports=paired)
+
+        if output == "json":
+            return self._results.as_list()
+        if output == "dataframe":
+            return self._results.as_df()
+        if output == "object":
+            return self._results
+
[…]
