+++ src/pfd_toolkit/__init__.py
@@ -23,7 +23,7 @@ from pfd_toolkit.screener import Screener
 from pfd_toolkit.extractor import Extractor
 from pfd_toolkit.paired_statements import (
     ConcernItem,
-    ConcernPairGenerator,
+    ConcernParser,
     ConcernResults,
     ConcernSet,
     ReportConcerns,
@@ -41,7 +41,7 @@ __all__ = [
     "Screener",
     "Extractor",
     "ConcernItem",
-    "ConcernPairGenerator",
+    "ConcernParser",
     "ConcernResults",
     "ConcernSet",
     "ReportConcerns",
