+++ src/pfd_toolkit/paired_statements.py
@@ -2,10 +2,12 @@
 
 This module focuses on converting the free-text "Matters of Concern" section
 into a structured list of concise, discrete concerns suitable for downstream
-pairing and analysis.
+pairing and analysis, and on linking those concerns to actions described in
+response documents.
 """
 from __future__ import annotations
 
+import json
 from typing import Any, Iterable, List, Mapping, Sequence
 
 import pandas as pd
@@ -31,6 +33,58 @@ class ConcernItem(BaseModel):
         return cleaned
 
 
+class ActionPhrase(BaseModel):
+    """Short description of an action or commitment from a response."""
+
+    action_phrase: str = Field(
+        ..., description="One concise action or commitment drawn from a response."
+    )
+
+    @field_validator("action_phrase")
+    @classmethod
+    def _normalise_text(cls, value: str) -> str:
+        cleaned = " ".join((value or "").split()).strip()
+        if not cleaned:
+            raise ValueError("action_phrase text cannot be empty")
+        return cleaned
+
+
+class ResponseAuthor(BaseModel):
+    """Author of a response and their associated actions."""
+
+    author: str = Field(..., description="Name of the individual or organisation.")
+    action_phrases: List[ActionPhrase] = Field(
+        default_factory=list,
+        description="List of discrete actions or commitments attributed to this author.",
+    )
+
+    @field_validator("author")
+    @classmethod
+    def _normalise_text(cls, value: str) -> str:
+        cleaned = " ".join((value or "").split()).strip()
+        if not cleaned:
+            raise ValueError("author name cannot be empty")
+        return cleaned
+
+
+class ConcernResponseItem(BaseModel):
+    """A concern paired with response authors and their action phrases."""
+
+    concern: str = Field(..., description="Concern text copied from the report.")
+    responses: List[ResponseAuthor] = Field(
+        default_factory=list,
+        description="Authors and their actions linked to this concern.",
+    )
+
+    @field_validator("concern")
+    @classmethod
+    def _normalise_text(cls, value: str) -> str:
+        cleaned = " ".join((value or "").split()).strip()
+        if not cleaned:
+            raise ValueError("concern text cannot be empty")
+        return cleaned
+
+
 class ConcernSet(BaseModel):
     """Ordered collection of concerns extracted from a PFD report."""
 
@@ -40,6 +94,15 @@ class ConcernSet(BaseModel):
             "A list of discrete concerns, kept in the same order as the "
             "original Matters of Concern section."
         ),
+    )
+
+
+class ConcernResponseSet(BaseModel):
+    """Responses mapped to concerns for a single PFD report."""
+
+    concerns: List[ConcernResponseItem] = Field(
+        default_factory=list,
+        description="List of concerns with any linked responses and action phrases.",
     )
 
 
@@ -59,6 +122,10 @@ class ReportConcerns(BaseModel):
     responses: List[str] = Field(
         default_factory=list,
         description="All response texts linked to this report's URL (one-to-many).",
+    )
+    concern_responses: ConcernResponseSet | None = Field(
+        default=None,
+        description="Structured responses paired with each concern when available.",
     )
 
     @property
@@ -89,6 +156,11 @@ class ConcernResults(BaseModel):
             )
         return serialised
 
+    def as_json(self) -> str:
+        """Return a JSON string serialising the concerns list."""
+
+        return json.dumps(self.as_list(), ensure_ascii=False, indent=2)
+
     def as_df(self) -> pd.DataFrame:
         """Return a flat DataFrame, one row per concern."""
 
@@ -103,6 +175,112 @@ class ConcernResults(BaseModel):
                         "concern": item.concern,
                     }
                 )
+        return pd.DataFrame(rows)
+
+
[…]
