+++ tests/test_paired_statements.py
@@ -1,16 +1,33 @@
 import pandas as pd
 
 from pfd_toolkit.config import GeneralConfig
-from pfd_toolkit.paired_statements import ConcernItem, ConcernParser, ConcernSet
+from pfd_toolkit.paired_statements import (
+    ActionPhrase,
+    ConcernItem,
+    ConcernParser,
+    ConcernResponseItem,
+    ConcernResponseSet,
+    ConcernSet,
+    ResponseAuthor,
+)
 
 
 class DummyLLM:
     """Minimal stub matching the LLM.generate interface used in tests."""
 
-    def __init__(self, responses: list[ConcernSet] | None = None):
+    def __init__(
+        self,
+        responses: list[ConcernSet] | None = None,
+        response_actions: list[ConcernResponseSet] | None = None,
+    ):
         self.responses = responses or []
+        self.response_actions = response_actions or []
 
-    def generate(self, prompts, response_format):
+    def generate(self, prompts, response_format, **_kwargs):
+        if response_format is ConcernResponseSet:
+            if self.response_actions:
+                return self.response_actions
+            return [response_format() for _ in prompts]
         if self.responses:
             return self.responses
         return [response_format() for _ in prompts]
@@ -110,6 +127,9 @@ def test_as_df_flattens_concerns():
     ]
     assert json_ready["concerns"] == ["one"]
 
+    json_str = results.as_json()
+    assert "\n" in json_str
+
 
 def test_count_methods_require_parsed_concerns_for_concern_counts():
     llm = DummyLLM(
@@ -139,3 +159,50 @@ def test_count_methods_require_parsed_concerns_for_concern_counts():
     combined = parser.count("all")
     assert combined["reports"] == 1
     assert combined["concerns"] == 1
+
+
+def test_parse_responses_aligns_to_concerns_and_counts_actions():
+    concerns_output = ConcernSet(
+        concerns=[ConcernItem(concern="concern one"), ConcernItem(concern="concern two")]
+    )
+    response_output = ConcernResponseSet(
+        concerns=[
+            ConcernResponseItem(
+                concern="concern one",
+                responses=[
+                    ResponseAuthor(
+                        author="Org A",
+                        action_phrases=[ActionPhrase(action_phrase="promise a change")],
+                    )
+                ],
+            )
+        ]
+    )
+
+    llm = DummyLLM(responses=[concerns_output], response_actions=[response_output])
+    reports = [
+        {
+            GeneralConfig.COL_URL: "https://example.test/report-1",
+            GeneralConfig.COL_CONCERNS: "concern text",
+            GeneralConfig.COL_RECEIVER: "Recipient A; Recipient B",
+            GeneralConfig.COL_ID: "2024-0001",
+        }
+    ]
+    responses = [
+        {"parent_url": "https://example.test/report-1", "response": "Letter from Org A"}
+    ]
+
+    parser = ConcernParser(llm=llm, reports=reports, responses=responses)
+    parser.parse_concerns()
+    paired = parser.parse_responses(output="object")
+
+    assert len(paired.reports) == 1
+    concern_items = paired.reports[0].concern_responses.concerns
+    assert concern_items[0].responses[0].author == "Org A"
+    assert concern_items[0].responses[0].action_phrases[0].action_phrase == "promise a change"
+    assert concern_items[1].responses[0].author == "[no response]"
+    assert concern_items[1].responses[0].action_phrases[0].action_phrase == "[no response]"
+
+    df = paired.as_df()
+    assert {"response_author", "action_phrase"}.issubset(df.columns)
+    assert parser.count("action_phrases") == {"action_phrases": 2}
