### 2025-12-04 351b09a
Modify LLM prompts for parse methods to better adhere to original wording and specificity
