+++ src/pfd_toolkit/paired_statements.py
@@ -50,26 +50,30 @@ class ActionPhrase(BaseModel):
 
 
 class ResponseAuthor(BaseModel):
-    """Author of a response and their associated actions."""
+    """Respondent of a response and their associated actions."""
 
-    author: str = Field(..., description="Name of the individual or organisation.")
+    respondent: str = Field(
+        ..., description="Name of the individual or organisation responding."
+    )
     action_phrases: List[ActionPhrase] = Field(
         default_factory=list,
-        description="List of discrete actions or commitments attributed to this author.",
+        description=(
+            "List of discrete actions or commitments attributed to this respondent."
+        ),
     )
 
-    @field_validator("author")
+    @field_validator("respondent")
     @classmethod
     def _normalise_text(cls, value: str) -> str:
         cleaned = " ".join((value or "").split()).strip()
         if not cleaned:
-            raise ValueError("author name cannot be empty")
+            raise ValueError("respondent name cannot be empty")
         return cleaned
 
     @field_validator("action_phrases")
     @classmethod
     def _limit_actions(cls, value: List[ActionPhrase]) -> List[ActionPhrase]:
-        """Ensure at most one action phrase is kept per author."""
+        """Ensure at most one action phrase is kept per respondent."""
 
         if len(value) > 1:
             return value[:1]
@@ -77,12 +81,12 @@ class ResponseAuthor(BaseModel):
 
 
 class ConcernResponseItem(BaseModel):
-    """A concern paired with response authors and their action phrases."""
+    """A concern paired with respondents and their action phrases."""
 
     concern: str = Field(..., description="Concern text copied from the report.")
     responses: List[ResponseAuthor] = Field(
         default_factory=list,
-        description="Authors and their actions linked to this concern.",
+        description="Respondents and their actions linked to this concern.",
     )
 
     @field_validator("concern")
@@ -203,7 +207,7 @@ class ReportConcernResponses(BaseModel):
     )
     concern_responses: ConcernResponseSet = Field(
         default_factory=ConcernResponseSet,
-        description="Structured concerns with linked response authors and actions.",
+        description="Structured concerns with linked respondents and actions.",
     )
 
 
@@ -217,30 +221,32 @@ class ResponseResults(BaseModel):
 
         serialised: List[dict[str, Any]] = []
         for report in self.reports:
-            serialised.append(
-                {
-                    GeneralConfig.COL_URL: report.url,
-                    GeneralConfig.COL_ID: report.report_id,
-                    GeneralConfig.COL_RECEIVER: report.recipients,
-                    "responses": report.responses,
-                    "concern_responses": [
-                        {
-                            "concern": item.concern,
-                            "responses": [
-                                {
-                                    "author": response.author,
-                                    "action_phrases": [
-                                        action.action_phrase
-                                        for action in response.action_phrases
-                                    ],
-                                }
-                                for response in item.responses
-                            ],
-                        }
-                        for item in report.concern_responses.concerns
-                    ],
-                }
-            )
+            record = {
+                GeneralConfig.COL_URL: report.url,
+                GeneralConfig.COL_ID: report.report_id,
+                GeneralConfig.COL_RECEIVER: report.recipients,
+                "concern_responses": [
+                    {
+                        "concern": item.concern,
+                        "responses": [
+                            {
+                                "respondent": response.respondent,
+                                "action_phrases": [
+                                    action.action_phrase
+                                    for action in response.action_phrases
+                                ],
+                            }
+                            for response in item.responses
+                        ],
+                    }
+                    for item in report.concern_responses.concerns
+                ],
+            }
+
[…]
