### 2025-12-11 720a3bb
Add assess actionability method which counts components of actionablility
