+++ src/pfd_toolkit/paired_statements.py
@@ -49,6 +49,94 @@ class ActionPhrase(BaseModel):
         return cleaned
 
 
+class ActionabilityActionPhraseElements(BaseModel):
+    """Per-element actionability assessment for an action phrase."""
+
+    action_phrase: str = Field(
+        ..., description="One concise action or commitment drawn from a response."
+    )
+    agent: bool = Field(..., description="Identifies who will act.")
+    action: bool = Field(..., description="Contains a verb of commitment or obligation.")
+    object_: bool = Field(
+        ...,
+        alias="object",
+        description="Specifies what will be acted upon.",
+    )
+    constraints: bool = Field(
+        ..., description="Includes time, place, quantity, or resource constraints."
+    )
+    accountability_cues: bool = Field(
+        ..., description="Mentions recipients, evidence, follow-up, or escalation."
+    )
+
+    @property
+    def actionability_score(self) -> int:
+        """Compute a 0–5 score from boolean element flags."""
+
+        return int(
+            self.agent
+            + self.action
+            + self.object_
+            + self.constraints
+            + self.accountability_cues
+        )
+
+    @field_validator("action_phrase")
+    @classmethod
+    def _normalise_text(cls, value: str) -> str:
+        cleaned = " ".join((value or "").split()).strip()
+        if not cleaned:
+            raise ValueError("action_phrase text cannot be empty")
+        return cleaned
+
+    @field_validator("agent", "action", "object_", "constraints", "accountability_cues", mode="before")
+    @classmethod
+    def _coerce_bool(cls, value: Any) -> bool:
+        if isinstance(value, str):
+            lowered = value.strip().lower()
+            if lowered in {"true", "yes", "y", "1"}:
+                return True
+            if lowered in {"false", "no", "n", "0"}:
+                return False
+        return bool(value)
+
+
+class ActionabilityActionPhrase(BaseModel):
+    """An action phrase paired with an actionability score."""
+
+    action_phrase: str = Field(
+        ..., description="One concise action or commitment drawn from a response."
+    )
+    actionability_score: int = Field(
+        ...,
+        ge=0,
+        le=5,
+        description=(
+            "Actionability rating from 0 (no actionable elements) to 5 (all "
+            "elements present)."
+        ),
+    )
+
+    @field_validator("action_phrase")
+    @classmethod
+    def _normalise_text(cls, value: str) -> str:
+        cleaned = " ".join((value or "").split()).strip()
+        if not cleaned:
+            raise ValueError("action_phrase text cannot be empty")
+        return cleaned
+
+    @field_validator("actionability_score")
+    @classmethod
+    def _clamp_score(cls, value: int) -> int:
+        if not isinstance(value, int):
+            raise TypeError("actionability_score must be an integer")
+        if value < 0:
+            return 0
+        if value > 5:
+            return 5
+        return value
+
+
 class ResponseAuthor(BaseModel):
     """Respondent of a response and their associated actions."""
 
@@ -98,6 +186,42 @@ class ConcernResponseItem(BaseModel):
         return cleaned
 
 
+class ActionabilityConcern(BaseModel):
+    """Actionability assessments for all action phrases linked to a concern."""
+
+    concern: str = Field(..., description="Concern text copied from the report.")
+    action_phrases: List[ActionabilityActionPhrase] = Field(
+        default_factory=list,
+        description="Action phrases with associated actionability scores.",
+    )
+
+    @field_validator("concern")
+    @classmethod
+    def _normalise_text(cls, value: str) -> str:
+        cleaned = " ".join((value or "").split()).strip()
+        if not cleaned:
+            raise ValueError("concern text cannot be empty")
+        return cleaned
+
+
+class ActionabilityConcernElements(BaseModel):
[…]
