### 2025-12-12 c613d73
Improve LLM prompt & change score to 0-4 range
