### 2025-12-12 47d6ed4
Make scoring more conservative
