### 2025-12-12 9713d13
Add bar chart functionality for action scores
