+++ pyproject.toml
@@ -28,6 +28,7 @@ dependencies = [
     "adjusttext>=1.3.0",
     "backoff>=2.2.1",
     "beautifulsoup4>=4.13.4",
+    "django>=5.2.11",
     "dotenv>=0.9.9",
     "httpx>=0.28.1",
     "openai>=1.88.0",
