### 2026-02-09 47f76f9
Attempt to copy Streamlit app into Django framework
