+++ django_workbench/workbench/views.py
@@ -1,0 +1,794 @@
+"""Views for the standalone Django Workbench."""
+from __future__ import annotations
+
+import json
+import zipfile
+from datetime import date
+from io import BytesIO, StringIO
+from typing import Any, Optional
+
+import pandas as pd
+from django.contrib import messages
+from django.http import HttpRequest, HttpResponse
+from django.shortcuts import redirect, render
+from django.views.decorators.http import require_http_methods
+
+from .services import (
+    build_extractor,
+    build_feature_model_from_rows,
+    build_llm,
+    build_screener,
+    load_reports_dataframe,
+    parse_seed_topics,
+    resolve_theme_emoji,
+)
+from .state import (
+    build_theme_summary_table,
+    clear_outputs_for_modified_dataset,
+    clear_outputs_for_new_dataset,
+    clear_preview_state,
+    dataframe_to_payload,
+    format_call,
+    get_dataframe,
+    get_repro_script_text,
+    has_api_key,
+    init_state,
+    parse_optional_non_negative_int,
+    push_history_snapshot,
+    record_repro_action,
+    redo_last_change,
+    reset_repro_tracking,
+    set_dataframe,
+    undo_last_change,
+    workspace_has_activity,
+)
+
+
+def _bool_from_post(request: HttpRequest, key: str, default: bool = False) -> bool:
+    if key not in request.POST:
+        return default
+    value = request.POST.get(key)
+    if value is None:
+        return True
+    return str(value).strip().lower() in {"1", "true", "on", "yes"}
+
+
+def _parse_date(value: str, default: date) -> date:
+    try:
+        return date.fromisoformat((value or "").strip())
+    except ValueError:
+        return default
+
+
+def _df_to_html(df: pd.DataFrame, table_class: str = "data-table") -> str:
+    if df.empty:
+        return ""
+    display_df = df.copy()
+    return display_df.to_html(
+        index=False,
+        classes=table_class,
+        border=0,
+        na_rep="",
+        escape=False,
+    )
+
+
+def _serialize_preview_state(preview_state: dict[str, Any]) -> dict[str, Any]:
+    payload = dict(preview_state)
+    for key in ("summary_df", "preview_df", "theme_summary"):
+        if key in payload and isinstance(payload[key], pd.DataFrame):
+            payload[key] = dataframe_to_payload(payload[key])
+    return payload
+
+
+def _deserialize_preview_state(preview_state: Any) -> Optional[dict[str, Any]]:
+    if not isinstance(preview_state, dict):
+        return None
+    payload = dict(preview_state)
+    for key in ("summary_df", "preview_df", "theme_summary"):
+        payload[key] = pd.DataFrame()
+        if isinstance(preview_state.get(key), str):
+            try:
+                payload[key] = pd.read_json(StringIO(preview_state[key]), orient="split")
+            except ValueError:
+                payload[key] = pd.DataFrame()
+    return payload
+
+
+def _default_feature_rows() -> list[dict[str, str]]:
+    return [
+        {
+            "Field name": "age",
+            "Description": "Age of the deceased (in years), if provided.",
+            "Type": "Whole number",
+        },
+        {
+            "Field name": "is_healthcare",
+            "Description": "True if the report involves a healthcare setting.",
+            "Type": "Conditional (True/False)",
+        },
+    ]
+
+
+def _get_feature_grid_df(request: HttpRequest, *, use_default: bool = True) -> pd.DataFrame:
+    payload = request.session.get("feature_grid")
+    if isinstance(payload, str):
+        try:
+            df = pd.read_json(payload, orient="split")
+        except ValueError:
+            df = pd.DataFrame()
+    else:
+        df = pd.DataFrame()
+
+    required = ["Field name", "Description", "Type"]
[…]
