+++ django_workbench/workbench/state.py
@@ -1,0 +1,366 @@
+"""Session-backed workspace state helpers for the Django workbench."""
+from __future__ import annotations
+
+import copy
+import json
+from datetime import date
+from io import StringIO
+from typing import Any, Dict, Iterable, Optional
+
+import pandas as pd
+
+REPRO_SCRIPT_KEY = "repro_script_lines"
+REPRO_ACTION_COUNTS_KEY = "repro_action_counts"
+LLM_SIGNATURE_KEY = "llm_config_signature"
+
+SNAPSHOT_KEYS = [
+    "reports_df",
+    "reports_df_initial",
+    "reports_df_modified",
+    "screener_result",
+    "extractor_result",
+    "summary_result",
+    "theme_model_schema",
+    "theme_summary_table",
+    "seed_topics_last",
+    "feature_grid",
+    "active_action",
+    "preview_state",
+    REPRO_SCRIPT_KEY,
+    REPRO_ACTION_COUNTS_KEY,
+    LLM_SIGNATURE_KEY,
+]
+
+
+def _initial_repro_script_lines() -> list[str]:
+    return [
+        "# -----------------------------------------------------------------------------",
+        "# Reproducible workspace script",
+        "# This script contains the Python code for the actions you performed in",
+        "# PFD Toolkit Workbench.",
+        "# To replay locally, first install the Toolkit:",
+        "#     pip install pfd_toolkit",
+        "# Workbench does not save your API key. Add it manually before running.",
+        "",
+        "from pfd_toolkit import load_reports",
+        "from pfd_toolkit import LLM",
+        "from pfd_toolkit import Screener",
+        "from pfd_toolkit import Extractor",
+        "",
+    ]
+
+
+def init_state(session: dict[str, Any]) -> None:
+    """Ensure the workspace keys exist in session."""
+
+    defaults: Dict[str, Any] = {
+        "reports_df": None,
+        "reports_df_initial": None,
+        "history": [],
+        "redo_history": [],
+        "active_action": None,
+        "preview_state": None,
+        "reports_df_modified": False,
+        "screener_result": None,
+        "extractor_result": None,
+        "summary_result": None,
+        "theme_model_schema": None,
+        "theme_summary_table": None,
+        "seed_topics_last": None,
+        "feature_grid": None,
+        "openrouter_api_key": "",
+        "openrouter_base_url": "https://openrouter.ai/api/v1",
+        "openai_api_key": "",
+        "openai_base_url": "",
+        "provider_override": "OpenAI",
+        "model_name": "gpt-4.1-mini",
+        "max_parallel_workers": 8,
+        "report_start_date": date(2013, 1, 1).isoformat(),
+        "report_end_date": date.today().isoformat(),
+        REPRO_SCRIPT_KEY: _initial_repro_script_lines(),
+        REPRO_ACTION_COUNTS_KEY: {},
+        LLM_SIGNATURE_KEY: None,
+        "theme_emoji_cache": {},
+    }
+    for key, value in defaults.items():
+        session.setdefault(key, value)
+
+
+def dataframe_to_payload(df: Optional[pd.DataFrame]) -> Optional[str]:
+    if df is None:
+        return None
+    if not isinstance(df, pd.DataFrame):
+        try:
+            df = pd.DataFrame(df)
+        except ValueError:
+            return None
+    return df.to_json(orient="split", date_format="iso")
+
+
+def dataframe_from_payload(payload: Any) -> pd.DataFrame:
+    if not payload or not isinstance(payload, str):
+        return pd.DataFrame()
+    try:
+        return pd.read_json(StringIO(payload), orient="split")
+    except ValueError:
+        return pd.DataFrame()
+
[…]
