### 2026-02-10 db22d45
Modify cell expansion pop-up sizing rules
