+++ django_workbench/workbench/static/workbench/workbench.js
@@ -90,11 +90,259 @@
         bindRemoveButtons();
     }
 
+    function setupDatasetCellPreview() {
+        const datasetTable = document.querySelector(".data-card .data-table");
+        if (!datasetTable) {
+            return;
+        }
+
+        const headerCells = Array.from(datasetTable.querySelectorAll("thead th"));
+        const bodyRows = Array.from(datasetTable.querySelectorAll("tbody tr"));
+        if (!headerCells.length) {
+            return;
+        }
+
+        function getColumnClass(headerText) {
+            const label = (headerText || "").trim().toLowerCase();
+            if (label === "id") {
+                return "dataset-col-id";
+            }
+            if (label.includes("date")) {
+                return "dataset-col-date";
+            }
+            if (label.includes("url")) {
+                return "dataset-col-url";
+            }
+            if (
+                label.includes("investigation") ||
+                label.includes("circumstances") ||
+                label.includes("concern")
+            ) {
+                return "dataset-col-long";
+            }
+            if (
+                label.includes("coroner") ||
+                label.includes("area") ||
+                label.includes("receiver")
+            ) {
+                return "dataset-col-meta";
+            }
+            return "dataset-col-standard";
+        }
+
+        const columnClasses = headerCells.map((th) => {
+            const columnClass = getColumnClass(th.innerText);
+            th.classList.add(columnClass);
+            return columnClass;
+        });
+
+        bodyRows.forEach((row) => {
+            Array.from(row.children).forEach((cell, index) => {
+                const columnClass = columnClasses[index];
+                if (columnClass) {
+                    cell.classList.add(columnClass);
+                }
+            });
+        });
+
+        const cells = datasetTable.querySelectorAll("tbody td");
+        if (!cells.length) {
+            return;
+        }
+
+        const popup = document.createElement("div");
+        popup.className = "dataset-cell-popup";
+        popup.setAttribute("role", "dialog");
+        popup.setAttribute("aria-modal", "false");
+        popup.setAttribute("aria-label", "Full cell text");
+
+        const popupContent = document.createElement("div");
+        popupContent.className = "dataset-cell-popup__content";
+        popup.appendChild(popupContent);
+        document.body.appendChild(popup);
+        const textMeasureCanvas = document.createElement("canvas");
+        const textMeasureContext = textMeasureCanvas.getContext("2d");
+
+        let activeCell = null;
+
+        function decodeEscapedNewlines(value) {
+            return String(value || "")
+                .replace(/\\r\\n/g, "\n")
+                .replace(/\\n/g, "\n")
+                .replace(/\\r/g, "\n");
+        }
+
+        function clampNumber(value, min, max) {
+            return Math.min(max, Math.max(min, value));
+        }
+
+        function measureLongestLinePx(text) {
+            if (!textMeasureContext) {
+                return 0;
+            }
+            const font = window.getComputedStyle(popupContent).font;
+            if (font) {
+                textMeasureContext.font = font;
+            }
+            const lines = String(text || "").split("\n");
+            let widest = 0;
+            lines.forEach((line) => {
+                const width = textMeasureContext.measureText(line).width;
+                widest = Math.max(widest, width);
+            });
+            return widest;
+        }
+
+        function computePopupWidthPx(text, cell) {
+            const totalLength = String(text || "").trim().length;
+            const isVeryShortField =
+                cell.classList.contains("dataset-col-id") ||
+                cell.classList.contains("dataset-col-date") ||
+                cell.classList.contains("dataset-col-meta");
+            const textPx = measureLongestLinePx(text);
+            const contentPaddingPx = 34;
+
+            // Keep short metadata popups compact but still proportional to content length.
[…]
