### 2026-02-10 97c9a4c
Alter page layout - just saving progress
