+++ django_workbench/workbench/urls.py
@@ -7,4 +7,6 @@ app_name = "workbench"
 
 urlpatterns = [
     path("", views.index, name="index"),
+    path("home/", views.home, name="home"),
+    path("explore-pfds/", views.explore, name="explore"),
 ]
