### 2026-02-10 94e3350
Change layout congif and add icons to nav bar
