+++ django_workbench/workbench/urls.py
@@ -9,4 +9,5 @@ urlpatterns = [
     path("", views.index, name="index"),
     path("home/", views.home, name="home"),
     path("explore-pfds/", views.explore, name="explore"),
+    path("settings/", views.settings_page, name="settings"),
 ]
