### 2026-02-10 1c12bca
Stopped refresh on tab switch
