+++ django_workbench/workbench/static/workbench/workbench.js
@@ -1,8 +1,40 @@
 (function () {
     document.documentElement.classList.add("js-enabled");
+    const PAGE_CLASS_PREFIX = "page-";
+    const KNOWN_PAGES = ["home", "explore", "settings"];
 
     function byId(id) {
         return document.getElementById(id);
+    }
+
+    function getPageFromPath(pathname) {
+        if (pathname.startsWith("/explore-pfds")) {
+            return "explore";
+        }
+        if (pathname.startsWith("/settings")) {
+            return "settings";
+        }
+        return "home";
+    }
+
+    function applyBodyPageClass(page) {
+        KNOWN_PAGES.forEach((name) => {
+            document.body.classList.remove(`${PAGE_CLASS_PREFIX}${name}`);
+        });
+        document.body.classList.add(`${PAGE_CLASS_PREFIX}${page}`);
+    }
+
+    function syncSidebarActive(page) {
+        const links = document.querySelectorAll(".sidebar-nav [data-page-link]");
+        links.forEach((link) => {
+            const isActive = link.getAttribute("data-page-link") === page;
+            link.classList.toggle("is-active", isActive);
+            if (isActive) {
+                link.setAttribute("aria-current", "page");
+            } else {
+                link.removeAttribute("aria-current");
+            }
+        });
     }
 
     function toggleProviderFields() {
@@ -150,6 +182,10 @@
         if (!datasetTable) {
             return;
         }
+        if (datasetTable.dataset.previewBound === "1") {
+            return;
+        }
+        datasetTable.dataset.previewBound = "1";
 
         const headerCells = Array.from(datasetTable.querySelectorAll("thead th"));
         const bodyRows = Array.from(datasetTable.querySelectorAll("tbody tr"));
@@ -400,6 +436,10 @@
         if (!toggleButton) {
             return;
         }
+        if (toggleButton.dataset.bound === "1") {
+            return;
+        }
+        toggleButton.dataset.bound = "1";
 
         const storageKey = "workbench.sidebarCollapsed";
 
@@ -469,7 +509,99 @@
         });
     }
 
-    document.addEventListener("DOMContentLoaded", function () {
+    function setupStartOverConfirm() {
+        const openButton = document.querySelector("[data-start-over-open]");
+        const cancelButton = document.querySelector("[data-start-over-cancel]");
+        const modal = byId("start-over-confirm");
+        if (!openButton || !cancelButton || !modal) {
+            return;
+        }
+        if (openButton.dataset.bound === "1") {
+            return;
+        }
+        openButton.dataset.bound = "1";
+
+        function openModal() {
+            modal.classList.remove("hidden");
+            document.body.classList.add("modal-open");
+            cancelButton.focus();
+        }
+
+        function closeModal() {
+            modal.classList.add("hidden");
+            document.body.classList.remove("modal-open");
+            openButton.focus();
+        }
+
+        openButton.addEventListener("click", openModal);
+        cancelButton.addEventListener("click", closeModal);
+
+        modal.addEventListener("click", function (event) {
+            if (event.target === modal) {
+                closeModal();
+            }
+        });
+
+        document.addEventListener("keydown", function (event) {
+            if (event.key === "Escape" && !modal.classList.contains("hidden")) {
+                closeModal();
+            }
+        });
+    }
+
+    function setupDownloadBundleModal() {
+        const openButton = document.querySelector("[data-download-open]");
+        const cancelButton = document.querySelector("[data-download-cancel]");
+        const popover = byId("download-popover");
+        if (!openButton || !cancelButton || !popover) {
+            return;
+        }
+        if (openButton.dataset.bound === "1") {
+            return;
+        }
+        openButton.dataset.bound = "1";
+
[…]
