### 2026-02-11 25432e9
Split AI features into their own section/tab and add pagnation for reports
