### 2026-02-11 2363482
Create loaders for LLM operations
