### 2026-02-11 82e097a
Add dashboarding elements to Explore tab
