+++ django_workbench/workbench/static/workbench/workbench.js
@@ -619,6 +619,13 @@
         }
         document.body.dataset.paginationBound = "1";
 
+        function buildUrlWithSharedQuery(baseUrl, sharedQuery) {
+            if (!sharedQuery) {
+                return baseUrl;
+            }
+            return `${baseUrl}${baseUrl.includes("?") ? "&" : "?"}${sharedQuery}`;
+        }
+
         function replaceDatasetFromHtml(html, targetUrl) {
             const wrapper = document.createElement("div");
             wrapper.innerHTML = html.trim();
@@ -663,6 +670,10 @@
                     window.location.href = targetUrl;
                 });
         }
+
+        window.WorkbenchDatasetPanel = {
+            fetchAndSwapDataset: fetchAndSwapDataset,
+        };
 
         document.addEventListener("click", function (event) {
             const link = event.target.closest(".dataset-pagination__actions a[data-dataset-panel-url]");
@@ -709,10 +720,560 @@
 
             const panelBase = form.getAttribute("data-dataset-panel-base") || "/dataset-panel/";
             const browserBase = form.getAttribute("data-dataset-browser-base") || "?page=";
-            const panelUrl = `${panelBase}?page=${page}`;
-            const targetUrl = `${browserBase}${page}`;
+            const sharedQuery = (form.getAttribute("data-dataset-shared-query") || "").trim();
+            const panelUrl = buildUrlWithSharedQuery(`${panelBase}?page=${page}`, sharedQuery);
+            const targetUrl = buildUrlWithSharedQuery(`${browserBase}${page}`, sharedQuery);
             fetchAndSwapDataset(panelUrl, targetUrl);
         });
+    }
+
+    function setupExploreDashboard() {
+        const root = byId("explore-dashboard");
+        if (!root) {
+            return;
+        }
+
+        const dataNode = byId("explore-dashboard-data");
+        if (!dataNode) {
+            return;
+        }
+
+        const resetButton = byId("dashboard-filter-reset");
+        const statusNode = byId("dashboard-status");
+        const statReports = byId("dashboard-stat-reports");
+        const statCoroners = byId("dashboard-stat-coroners");
+        const statReceiverLinks = byId("dashboard-stat-receiver-links");
+        const searchCoroner = byId("dashboard-filter-coroner-search");
+        const searchArea = byId("dashboard-filter-area-search");
+        const searchReceiver = byId("dashboard-filter-receiver-search");
+        const optionsCoroner = byId("dashboard-options-coroner");
+        const optionsArea = byId("dashboard-options-area");
+        const optionsReceiver = byId("dashboard-options-receiver");
+        const badgesCoroner = byId("dashboard-badges-coroner");
+        const badgesArea = byId("dashboard-badges-area");
+        const badgesReceiver = byId("dashboard-badges-receiver");
+        const filterCoronerField = root.querySelector("[data-dashboard-filter='coroner']");
+        const filterAreaField = root.querySelector("[data-dashboard-filter='area']");
+        const filterReceiverField = root.querySelector("[data-dashboard-filter='receiver']");
+
+        const chartRoots = {
+            monthly: byId("dashboard-chart-monthly"),
+            coroner: byId("dashboard-chart-coroner"),
+            area: byId("dashboard-chart-area"),
+            receiver: byId("dashboard-chart-receiver"),
+        };
+
+        if (
+            !resetButton ||
+            !statReports ||
+            !statCoroners ||
+            !statReceiverLinks ||
+            !searchCoroner ||
+            !searchArea ||
+            !searchReceiver ||
+            !optionsCoroner ||
+            !optionsArea ||
+            !optionsReceiver ||
+            !badgesCoroner ||
+            !badgesArea ||
[…]
