+++ django_workbench/workbench/views.py
@@ -5,6 +5,7 @@ import json
 import zipfile
 from datetime import date
 from io import BytesIO
+from urllib.parse import urlencode
 from typing import Any, Optional
 
 import pandas as pd
@@ -72,6 +73,184 @@ def _df_to_html(df: pd.DataFrame, table_class: str = "data-table") -> str:
         na_rep="",
         escape=False,
     )
+
+
+def _normalise_dashboard_value(value: Any) -> str:
+    if value is None:
+        return ""
+    if isinstance(value, str):
+        cleaned = value.strip()
+    else:
+        try:
+            if pd.isna(value):
+                return ""
+        except TypeError:
+            pass
+        cleaned = str(value).strip()
+
+    lowered = cleaned.lower()
+    if lowered in {"nan", "nat", "none"}:
+        return ""
+    return cleaned
+
+
+def _split_dashboard_receivers(value: Any) -> list[str]:
+    raw = _normalise_dashboard_value(value)
+    if not raw:
+        return []
+
+    seen: set[str] = set()
+    receivers: list[str] = []
+    for chunk in raw.split(";"):
+        cleaned = _normalise_dashboard_value(chunk)
+        if not cleaned:
+            continue
+        lowered = cleaned.casefold()
+        if lowered in seen:
+            continue
+        seen.add(lowered)
+        receivers.append(cleaned)
+    return receivers
+
+
+def _build_explore_dashboard_payload(reports_df: pd.DataFrame) -> dict[str, Any]:
+    if reports_df.empty:
+        return {
+            "rows": [],
+            "options": {
+                "coroners": [],
+                "areas": [],
+                "receivers": [],
+            },
+        }
+
+    row_count = len(reports_df)
+    default_series = pd.Series([""] * row_count, index=reports_df.index, dtype=object)
+
+    if "date" in reports_df.columns:
+        date_series = pd.to_datetime(reports_df["date"], errors="coerce")
+    else:
+        date_series = pd.Series([pd.NaT] * row_count, index=reports_df.index, dtype="datetime64[ns]")
+
+    coroner_series = reports_df["coroner"] if "coroner" in reports_df.columns else default_series
+    area_series = reports_df["area"] if "area" in reports_df.columns else default_series
+    receiver_series = reports_df["receiver"] if "receiver" in reports_df.columns else default_series
+
+    rows: list[dict[str, Any]] = []
+    coroners: set[str] = set()
+    areas: set[str] = set()
+    receivers: set[str] = set()
+
+    for date_value, coroner_value, area_value, receiver_value in zip(
+        date_series,
+        coroner_series,
+        area_series,
+        receiver_series,
+    ):
+        date_iso = ""
+        if pd.notna(date_value):
+            date_iso = date_value.strftime("%Y-%m-%d")
+
+        coroner = _normalise_dashboard_value(coroner_value)
+        area = _normalise_dashboard_value(area_value)
+        receiver_parts = _split_dashboard_receivers(receiver_value)
+
+        if coroner:
+            coroners.add(coroner)
+        if area:
+            areas.add(area)
+        for receiver_name in receiver_parts:
+            receivers.add(receiver_name)
+
+        rows.append(
+            {
+                "date": date_iso,
+                "year_month": date_iso[:7] if date_iso else "",
+                "coroner": coroner,
+                "area": area,
+                "receivers": receiver_parts,
+            }
+        )
+
+    sort_key = lambda value: value.casefold()
+    return {
+        "rows": rows,
+        "options": {
+            "coroners": sorted(coroners, key=sort_key),
+            "areas": sorted(areas, key=sort_key),
+            "receivers": sorted(receivers, key=sort_key),
+        },
[…]
