### 2026-02-11 24bbf3d
Update upper bar with interactive elements & move report filtering back into Explore view
