+++ django_workbench/workbench/static/workbench/workbench.js
@@ -1,7 +1,7 @@
 (function () {
     document.documentElement.classList.add("js-enabled");
     const PAGE_CLASS_PREFIX = "page-";
-    const KNOWN_PAGES = ["home", "explore", "filter", "themes", "extract", "settings"];
+    const KNOWN_PAGES = ["home", "explore", "themes", "extract", "settings"];
 
     function byId(id) {
         return document.getElementById(id);
@@ -12,7 +12,7 @@
             return "explore";
         }
         if (pathname.startsWith("/filter")) {
-            return "filter";
+            return "explore";
         }
         if (pathname.startsWith("/analyse-themes")) {
             return "themes";
@@ -613,6 +613,308 @@
         });
     }
 
+    function setupTopbarQuickSettings() {
+        const openButtons = document.querySelectorAll("[data-topbar-open]");
+        const closeButtons = document.querySelectorAll("[data-topbar-close]");
+        const popovers = Array.from(document.querySelectorAll(".topbar-popover-backdrop"));
+        if (!openButtons.length || !popovers.length) {
+            return;
+        }
+        if (document.body.dataset.topbarQuickSettingsBound === "1") {
+            return;
+        }
+        document.body.dataset.topbarQuickSettingsBound = "1";
+
+        const popoverByKey = {
+            "report-limit": byId("report-limit-popover"),
+            "date-range": byId("date-range-popover"),
+            "llm-settings": byId("llm-settings-popover"),
+        };
+
+        function closeAllPopovers(restoreFocusTarget) {
+            popovers.forEach((popover) => {
+                if (popover) {
+                    popover.classList.add("hidden");
+                }
+            });
+            if (restoreFocusTarget) {
+                restoreFocusTarget.focus();
+            }
+        }
+
+        openButtons.forEach((button) => {
+            button.addEventListener("click", function () {
+                const key = button.getAttribute("data-topbar-open");
+                const targetPopover = popoverByKey[key];
+                if (!targetPopover) {
+                    return;
+                }
+                closeAllPopovers();
+                targetPopover.classList.remove("hidden");
+            });
+        });
+
+        closeButtons.forEach((button) => {
+            button.addEventListener("click", function () {
+                closeAllPopovers();
+            });
+        });
+
+        popovers.forEach((popover) => {
+            popover.addEventListener("click", function (event) {
+                if (event.target === popover) {
+                    closeAllPopovers();
+                }
+            });
+        });
+
+        document.addEventListener("keydown", function (event) {
+            if (event.key === "Escape") {
+                closeAllPopovers();
+            }
+        });
+    }
+
+    function setupTopbarLimitControls() {
+        const slider = document.querySelector("[data-topbar-limit-slider]");
+        const numberInput = document.querySelector("[data-topbar-limit-number]");
+        if (!slider || !numberInput) {
+            return;
+        }
+        if (slider.dataset.bound === "1") {
+            return;
+        }
+        slider.dataset.bound = "1";
+
+        function clampToSlider(value) {
+            const parsed = Number(value);
+            if (!Number.isFinite(parsed)) {
+                return Number(slider.min || 1);
+            }
+            const min = Number(slider.min || 1);
+            const max = Number(slider.max || 5000);
+            return Math.max(min, Math.min(max, Math.round(parsed)));
+        }
+
+        slider.addEventListener("input", function () {
+            numberInput.value = String(clampToSlider(slider.value));
+        });
+
+        numberInput.addEventListener("input", function () {
+            slider.value = String(clampToSlider(numberInput.value));
+        });
+    }
+
+    function setupTopbarLLMProviderToggle() {
+        const provider = byId("topbar_provider_override");
[…]
