### 2026-02-12 eee8f5b
Move Discover themes and Extract data back into Explore view and add check for running AI on manually filtered data
