+++ django_workbench/workbench/static/workbench/workbench.js
@@ -630,13 +630,43 @@
             "date-range": byId("date-range-popover"),
             "llm-settings": byId("llm-settings-popover"),
         };
+        let active = null;
+
+        function positionPopover(popover, button) {
+            const panel = popover.querySelector(".topbar-popover");
+            if (!panel) {
+                return;
+            }
+
+            const margin = 10;
+            const buttonRect = button.getBoundingClientRect();
+            const panelRect = panel.getBoundingClientRect();
+
+            let left = buttonRect.left;
+            if (left + panelRect.width > window.innerWidth - margin) {
+                left = window.innerWidth - panelRect.width - margin;
+            }
+            left = Math.max(margin, left);
+
+            let top = buttonRect.bottom + 10;
+            if (top + panelRect.height > window.innerHeight - margin) {
+                top = buttonRect.top - panelRect.height - 10;
+            }
+            top = Math.max(margin, top);
+
+            popover.style.left = `${left}px`;
+            popover.style.top = `${top}px`;
+        }
 
         function closeAllPopovers(restoreFocusTarget) {
             popovers.forEach((popover) => {
                 if (popover) {
                     popover.classList.add("hidden");
+                    popover.style.left = "";
+                    popover.style.top = "";
                 }
             });
+            active = null;
             if (restoreFocusTarget) {
                 restoreFocusTarget.focus();
             }
@@ -649,8 +679,14 @@
                 if (!targetPopover) {
                     return;
                 }
+                if (active && active.popover === targetPopover && !targetPopover.classList.contains("hidden")) {
+                    closeAllPopovers(button);
+                    return;
+                }
                 closeAllPopovers();
                 targetPopover.classList.remove("hidden");
+                positionPopover(targetPopover, button);
+                active = { popover: targetPopover, button: button };
             });
         });
 
@@ -660,19 +696,34 @@
             });
         });
 
-        popovers.forEach((popover) => {
-            popover.addEventListener("click", function (event) {
-                if (event.target === popover) {
-                    closeAllPopovers();
-                }
-            });
-        });
+        document.addEventListener("pointerdown", function (event) {
+            if (!active || !active.popover || active.popover.classList.contains("hidden")) {
+                return;
+            }
+            if (active.popover.contains(event.target) || active.button.contains(event.target)) {
+                return;
+            }
+            closeAllPopovers();
+        }, true);
 
         document.addEventListener("keydown", function (event) {
             if (event.key === "Escape") {
-                closeAllPopovers();
+                const focusTarget = active ? active.button : null;
+                closeAllPopovers(focusTarget);
             }
         });
+
+        window.addEventListener("resize", function () {
+            if (active && active.popover && active.button && !active.popover.classList.contains("hidden")) {
+                positionPopover(active.popover, active.button);
+            }
+        });
+
+        window.addEventListener("scroll", function () {
+            if (active && active.popover && active.button && !active.popover.classList.contains("hidden")) {
+                positionPopover(active.popover, active.button);
+            }
+        }, { passive: true, capture: true });
     }
 
     function setupTopbarLimitControls() {
@@ -692,7 +743,7 @@
                 return Number(slider.min || 1);
             }
             const min = Number(slider.min || 1);
-            const max = Number(slider.max || 5000);
+            const max = Number(slider.max || 7000);
             return Math.max(min, Math.min(max, Math.round(parsed)));
         }
 
@@ -863,55 +914,248 @@
         });
     }
 
-    function setupAdvancedAIFilterPopover() {
[…]
