### 2026-02-12 af9d59b
Changed layout of settings & add user warning when reports are re-loaded
