+++ django_workbench/workbench/static/workbench/workbench.js
@@ -914,6 +914,117 @@
         });
     }
 
+    function setupSettingsLoadReportsConfirm() {
+        const form = document.querySelector(".settings-workbench .config-form");
+        const openButton = document.querySelector("[data-settings-load-open]");
+        const confirmModal = byId("settings-load-reports-confirm");
+        if (!form || !openButton) {
+            return;
+        }
+        if (form.dataset.loadConfirmBound === "1") {
+            return;
+        }
+        form.dataset.loadConfirmBound = "1";
+
+        if (!confirmModal) {
+            openButton.addEventListener("click", function (event) {
+                if (form.dataset.loadReportsConfirmed === "1") {
+                    return;
+                }
+                const allowed = window.confirm(
+                    "Loading reports will refresh your existing workspace. Any unsaved workflow changes and generated outputs will be lost. Continue?"
+                );
+                if (!allowed) {
+                    event.preventDefault();
+                    return;
+                }
+                form.dataset.loadReportsConfirmed = "1";
+            });
+            return;
+        }
+
+        const cancelButton = confirmModal.querySelector("[data-settings-load-cancel]");
+        const confirmButton = confirmModal.querySelector("[data-settings-load-confirm]");
+        let pendingSubmitter = null;
+
+        function closeConfirm(restoreFocus) {
+            confirmModal.classList.add("hidden");
+            document.body.classList.remove("modal-open");
+            if (restoreFocus && pendingSubmitter && typeof pendingSubmitter.focus === "function") {
+                pendingSubmitter.focus();
+            }
+            pendingSubmitter = null;
+        }
+
+        function openConfirm(submitter) {
+            pendingSubmitter = submitter;
+            confirmModal.classList.remove("hidden");
+            document.body.classList.add("modal-open");
+            if (cancelButton) {
+                cancelButton.focus();
+            }
+        }
+
+        if (openButton) {
+            openButton.addEventListener("click", function (event) {
+                if (form.dataset.loadReportsConfirmed === "1") {
+                    return;
+                }
+                event.preventDefault();
+                openConfirm(openButton);
+            });
+        }
+
+        form.addEventListener("submit", function (event) {
+            const submitter = event.submitter;
+            const actionValue = submitter ? submitter.value : "";
+            if (actionValue !== "load_reports") {
+                return;
+            }
+            if (form.dataset.loadReportsConfirmed === "1") {
+                delete form.dataset.loadReportsConfirmed;
+                return;
+            }
+            event.preventDefault();
+            openConfirm(submitter);
+        });
+
+        if (cancelButton) {
+            cancelButton.addEventListener("click", function () {
+                closeConfirm(true);
+            });
+        }
+
+        if (confirmButton) {
+            confirmButton.addEventListener("click", function () {
+                if (!pendingSubmitter) {
+                    closeConfirm(false);
+                    return;
+                }
+                form.dataset.loadReportsConfirmed = "1";
+                const submitter = pendingSubmitter;
+                closeConfirm(false);
+                if (typeof form.requestSubmit === "function") {
+                    form.requestSubmit(submitter);
+                    return;
+                }
+                submitter.click();
+            });
+        }
+
+        confirmModal.addEventListener("click", function (event) {
+            if (event.target === confirmModal) {
+                closeConfirm(true);
+            }
+        });
+
+        document.addEventListener("keydown", function (event) {
+            if (event.key === "Escape" && !confirmModal.classList.contains("hidden")) {
+                closeConfirm(true);
+            }
+        });
+    }
+
     function setupAdvancedAIPopovers() {
[…]
