### 2026-02-12 7d0a73f
Create sharable, named, read-only worksheet functionality
