+++ django_workbench/workbench/views.py
@@ -2,18 +2,23 @@
 from __future__ import annotations
 
 import json
+import random
+import re
 import zipfile
 from datetime import date, datetime
-from io import BytesIO
+from io import BytesIO, StringIO
 from urllib.parse import urlencode
 from typing import Any, Optional
+from uuid import UUID
 
 import pandas as pd
 from django.contrib import messages
-from django.http import HttpRequest, HttpResponse
-from django.shortcuts import redirect, render
+from django.http import HttpRequest, HttpResponse, JsonResponse
+from django.shortcuts import get_object_or_404, redirect, render
+from django.utils.text import slugify
 from django.views.decorators.http import require_http_methods
 
+from .models import Workbook
 from .services import (
     build_extractor,
     build_feature_model_from_rows,
@@ -299,6 +304,118 @@ def _resolve_reports_for_ai_action(request: HttpRequest) -> tuple[pd.DataFrame, 
     return _apply_explore_dashboard_filters(reports_df, post_filters), post_filters
 
 
+WORKBOOK_TITLE_MAX_LENGTH = 120
+WORKBOOK_SNAPSHOT_MAX_BYTES = 2_500_000
+WORKBOOK_TITLE_ALLOWED_PATTERN = re.compile(r"^[A-Za-z0-9 -]+$")
+
+
+def _normalise_workbook_id(raw_value: str) -> Optional[str]:
+    cleaned = (raw_value or "").strip()
+    if not cleaned:
+        return None
+    try:
+        parsed = UUID(cleaned)
+    except ValueError:
+        return None
+    return str(parsed)
+
+
+def _normalise_workbook_token(raw_value: str) -> Optional[str]:
+    cleaned = (raw_value or "").strip()
+    if not cleaned:
+        return None
+    try:
+        parsed = UUID(cleaned)
+    except ValueError:
+        return None
+    return str(parsed)
+
+
+def _normalise_workbook_title(raw_value: Any) -> str:
+    title = str(raw_value or "").strip()
+    if len(title) > WORKBOOK_TITLE_MAX_LENGTH:
+        title = title[:WORKBOOK_TITLE_MAX_LENGTH].rstrip()
+    return title
+
+
+def _workbook_title_error(title: str) -> Optional[str]:
+    if not title:
+        return "Worksheet name is required."
+    if not WORKBOOK_TITLE_ALLOWED_PATTERN.fullmatch(title):
+        return "Worksheet name can use letters, numbers, spaces, and hyphens only."
+    return None
+
+
+def _workbook_title_slug(title: str) -> str:
+    slug = slugify(title)
+    return slug or "worksheet"
+
+
+def _allocate_workbook_share_number() -> int:
+    for _ in range(30):
+        candidate = random.randint(100000, 999999)
+        if not Workbook.objects.filter(share_number=candidate).exists():
+            return candidate
+    raise RuntimeError("Could not allocate a unique worksheet share number.")
+
+
+def _normalise_workbook_filters(raw_filters: Any) -> dict[str, list[str]]:
+    if not isinstance(raw_filters, dict):
+        return {"coroner": [], "area": [], "receiver": []}
+    return {
+        "coroner": _parse_explore_filter_values(raw_filters.get("coroner", [])),
+        "area": _parse_explore_filter_values(raw_filters.get("area", [])),
+        "receiver": _parse_explore_filter_values(raw_filters.get("receiver", [])),
+    }
+
+
+def _snapshot_dataframe_to_payload(df: pd.DataFrame) -> str:
+    if df.empty:
+        return ""
+    return df.to_json(orient="split", date_format="iso")
+
+
+def _snapshot_dataframe_from_payload(payload: Any) -> pd.DataFrame:
+    if not payload or not isinstance(payload, str):
+        return pd.DataFrame()
+    try:
[…]
