+++ django_workbench/workbench/migrations/0002_workbook_snapshot_delete_workbookrevision.py
@@ -1,0 +1,41 @@
+# Generated by Django 5.2.11 on 2026-02-12 16:07
+
+from django.db import migrations, models
+
+
+def copy_latest_revision_to_workbook_snapshot(apps, schema_editor):
+    Workbook = apps.get_model("workbench", "Workbook")
+    WorkbookRevision = apps.get_model("workbench", "WorkbookRevision")
+    for workbook in Workbook.objects.all():
+        latest = (
+            WorkbookRevision.objects.filter(workbook_id=workbook.id)
+            .order_by("-revision_number")
+            .first()
+        )
+        if latest is None:
+            continue
+        workbook.snapshot = latest.snapshot
+        workbook.save(update_fields=["snapshot"])
+
+
+def noop_reverse(apps, schema_editor):
+    return None
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("workbench", "0001_initial"),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name="workbook",
+            name="snapshot",
+            field=models.JSONField(default=dict),
+        ),
+        migrations.RunPython(copy_latest_revision_to_workbook_snapshot, noop_reverse),
+        migrations.DeleteModel(
+            name="WorkbookRevision",
+        ),
+    ]
