+++ pyproject.toml
@@ -28,18 +28,22 @@ dependencies = [
     "adjusttext>=1.3.0",
     "backoff>=2.2.1",
     "beautifulsoup4>=4.13.4",
+    "dj-database-url>=3.1.0",
     "django>=5.2.11",
     "dotenv>=0.9.9",
+    "gunicorn>=25.0.3",
     "httpx>=0.28.1",
     "openai>=1.88.0",
     "pandas>=2.2.2,<2.2.3; python_version<'3.13'",
     "pandas>=2.2.3; python_version>='3.13'",
+    "psycopg[binary]>=3.3.2",
     "pydantic>=2.11.7",
     "pymupdf>=1.26.1",
     "python-dateutil>=2.9.0.post0",
     "requests>=2.32.3",
     "tiktoken>=0.9.0",
     "tqdm>=4.67.1",
+    "whitenoise>=6.11.0",
 ]
 
 [project.optional-dependencies]

[…]
