### 2026-02-12 0e2a55a
Add functionality to allow the user to make a local editable version of a shared worksheet
