### 2026-02-12 9c7445c
Re-running thematic discovery now prompts the user to discard existing themes before re-running
