+++ django_workbench/workbench/static/workbench/workbench.js
@@ -1035,7 +1035,11 @@
         const confirmCancel = document.querySelector("[data-advanced-ai-manual-filter-cancel]");
         const confirmKeep = document.querySelector("[data-advanced-ai-manual-filter-keep]");
         const confirmDiscard = document.querySelector("[data-advanced-ai-manual-filter-discard]");
+        const themeRerunBackdrop = byId("advanced-ai-theme-rerun-confirm-popover");
+        const themeRerunCancel = document.querySelector("[data-advanced-ai-theme-rerun-cancel]");
+        const themeRerunConfirm = document.querySelector("[data-advanced-ai-theme-rerun-confirm]");
         let pendingManualFilterConfirm = null;
+        let pendingThemeRerunConfirm = null;
 
         function getDashboardManualFilters() {
             if (
@@ -1078,6 +1082,14 @@
             });
         }
 
+        function hasExistingThemes(form) {
+            return form && form.dataset && form.dataset.hasExistingThemes === "1";
+        }
+
+        function getThemeRerunInput(form) {
+            return form ? form.querySelector("input[data-confirm-rerun-themes]") : null;
+        }
+
         function openManualFilterConfirm(config) {
             if (!confirmBackdrop || !confirmList || !confirmMessage) {
                 return false;
@@ -1111,6 +1123,21 @@
             return true;
         }
 
+        function openThemeRerunConfirm(config) {
+            if (!themeRerunBackdrop) {
+                return false;
+            }
+            if (window.WorkbenchLoading && typeof window.WorkbenchLoading.hide === "function") {
+                window.WorkbenchLoading.hide();
+            }
+            pendingThemeRerunConfirm = config;
+            if (config.sourceBackdrop) {
+                config.sourceBackdrop.classList.add("hidden");
+            }
+            themeRerunBackdrop.classList.remove("hidden");
+            return true;
+        }
+
         function closeManualFilterConfirm(restoreSource) {
             if (confirmBackdrop) {
                 confirmBackdrop.classList.add("hidden");
@@ -1123,6 +1150,20 @@
                 pendingManualFilterConfirm.sourceBackdrop.classList.remove("hidden");
             }
             pendingManualFilterConfirm = null;
+        }
+
+        function closeThemeRerunConfirm(restoreSource) {
+            if (themeRerunBackdrop) {
+                themeRerunBackdrop.classList.add("hidden");
+            }
+            if (
+                restoreSource &&
+                pendingThemeRerunConfirm &&
+                pendingThemeRerunConfirm.sourceBackdrop
+            ) {
+                pendingThemeRerunConfirm.sourceBackdrop.classList.remove("hidden");
+            }
+            pendingThemeRerunConfirm = null;
         }
 
         if (confirmCancel && !confirmCancel.dataset.bound) {
@@ -1163,6 +1204,36 @@
                 pending.form.requestSubmit(pending.submitter || undefined);
             });
         }
+        if (themeRerunCancel && !themeRerunCancel.dataset.bound) {
+            themeRerunCancel.dataset.bound = "1";
+            themeRerunCancel.addEventListener("click", function () {
+                closeThemeRerunConfirm(true);
+            });
+        }
+        if (themeRerunBackdrop && !themeRerunBackdrop.dataset.bound) {
+            themeRerunBackdrop.dataset.bound = "1";
+            themeRerunBackdrop.addEventListener("click", function (event) {
+                if (event.target === themeRerunBackdrop) {
+                    closeThemeRerunConfirm(true);
+                }
+            });
+        }
+        if (themeRerunConfirm && !themeRerunConfirm.dataset.bound) {
+            themeRerunConfirm.dataset.bound = "1";
+            themeRerunConfirm.addEventListener("click", function () {
[…]
