+++ django_workbench/workbench/views.py
@@ -10,7 +10,7 @@ from datetime import date, datetime
 from io import BytesIO, StringIO
 from urllib.parse import urlencode
 from typing import Any, Optional
-from uuid import UUID
+from uuid import UUID, uuid4
 
 import pandas as pd
 from django.contrib import messages
@@ -100,7 +100,10 @@ def _format_date_ddmmyyyy(value: str, default: date) -> str:
 def _df_to_html(df: pd.DataFrame, table_class: str = "data-table") -> str:
     if df.empty:
         return ""
-    display_df = df.copy()
+    display_df = df.copy().drop(
+        columns=[WORKBENCH_ROW_ID_COL, EXCLUSION_REASON_COL],
+        errors="ignore",
+    )
     return display_df.to_html(
         index=False,
         classes=table_class,
@@ -108,6 +111,159 @@ def _df_to_html(df: pd.DataFrame, table_class: str = "data-table") -> str:
         na_rep="",
         escape=False,
     )
+
+
+def _normalise_row_id(raw_value: Any) -> str:
+    return str(raw_value or "").strip()
+
+
+def _ensure_workbench_row_ids(df: pd.DataFrame) -> tuple[pd.DataFrame, bool]:
+    if df.empty:
+        return df.copy(), False
+
+    result_df = df.copy()
+    changed = False
+
+    if WORKBENCH_ROW_ID_COL not in result_df.columns:
+        result_df[WORKBENCH_ROW_ID_COL] = [uuid4().hex for _ in range(len(result_df))]
+        return result_df, True
+
+    seen: set[str] = set()
+    normalised_ids: list[str] = []
+    for value in result_df[WORKBENCH_ROW_ID_COL].tolist():
+        row_id = _normalise_row_id(value)
+        if not row_id or row_id in seen:
+            row_id = uuid4().hex
+            changed = True
+        seen.add(row_id)
+        normalised_ids.append(row_id)
+
+    if list(result_df[WORKBENCH_ROW_ID_COL].astype(str)) != normalised_ids:
+        changed = True
+    result_df[WORKBENCH_ROW_ID_COL] = normalised_ids
+    return result_df, changed
+
+
+def _normalise_excluded_reports_df(df: pd.DataFrame) -> tuple[pd.DataFrame, bool]:
+    if df.empty:
+        return pd.DataFrame(columns=[WORKBENCH_ROW_ID_COL, EXCLUSION_REASON_COL]), False
+
+    result_df, changed = _ensure_workbench_row_ids(df)
+    if EXCLUSION_REASON_COL not in result_df.columns:
+        result_df[EXCLUSION_REASON_COL] = ""
+        changed = True
+    else:
+        reasons = result_df[EXCLUSION_REASON_COL].map(lambda value: str(value or "").strip())
+        if not reasons.equals(result_df[EXCLUSION_REASON_COL]):
+            changed = True
+        result_df[EXCLUSION_REASON_COL] = reasons
+    return result_df, changed
+
+
+def _get_reports_df_with_row_ids(session: dict[str, Any]) -> pd.DataFrame:
+    reports_df = get_dataframe(session, "reports_df")
+    reports_df, changed = _ensure_workbench_row_ids(reports_df)
+    if changed:
+        set_dataframe(session, "reports_df", reports_df)
+    return reports_df
+
+
+def _set_reports_df_with_row_ids(session: dict[str, Any], key: str, df: pd.DataFrame) -> pd.DataFrame:
+    normalised_df, _ = _ensure_workbench_row_ids(df)
+    set_dataframe(session, key, normalised_df)
+    return normalised_df
+
+
+def _get_excluded_reports_df(session: dict[str, Any]) -> pd.DataFrame:
[…]
