+++ django_workbench/workbench/static/workbench/workbench.js
@@ -190,26 +190,16 @@
     }
 
     function setupDatasetCellPreview() {
-        const datasetTable = document.querySelector(".data-card .data-table");
-        if (!datasetTable) {
-            return;
-        }
-        if (datasetTable.dataset.previewBound === "1") {
-            return;
-        }
-        datasetTable.dataset.previewBound = "1";
-
-        const headerCells = Array.from(datasetTable.querySelectorAll("thead th"));
-        const bodyRows = Array.from(datasetTable.querySelectorAll("tbody tr"));
-        if (bodyRows.length > 120) {
-            return;
-        }
-        if (!headerCells.length) {
+        const datasetTables = Array.from(document.querySelectorAll(".data-card .data-table"));
+        if (!datasetTables.length) {
             return;
         }
 
         function getColumnClass(headerText) {
             const label = (headerText || "").trim().toLowerCase();
+            if (label === "actions" || label === "restore") {
+                return "dataset-col-action";
+            }
             if (label === "id") {
                 return "dataset-col-id";
             }
@@ -236,23 +226,41 @@
             return "dataset-col-standard";
         }
 
-        const columnClasses = headerCells.map((th) => {
-            const columnClass = getColumnClass(th.innerText);
-            th.classList.add(columnClass);
-            return columnClass;
-        });
+        const previewCells = [];
+        datasetTables.forEach((datasetTable) => {
+            if (datasetTable.dataset.previewBound === "1") {
+                return;
+            }
+            datasetTable.dataset.previewBound = "1";
 
-        bodyRows.forEach((row) => {
-            Array.from(row.children).forEach((cell, index) => {
-                const columnClass = columnClasses[index];
-                if (columnClass) {
-                    cell.classList.add(columnClass);
-                }
+            const headerCells = Array.from(datasetTable.querySelectorAll("thead th"));
+            const bodyRows = Array.from(datasetTable.querySelectorAll("tbody tr"));
+            if (bodyRows.length > 120 || !headerCells.length) {
+                return;
+            }
+
+            const columnClasses = headerCells.map((th) => {
+                const columnClass = getColumnClass(th.innerText);
+                th.classList.add(columnClass);
+                return columnClass;
             });
+
+            bodyRows.forEach((row) => {
+                Array.from(row.children).forEach((cell, index) => {
+                    const columnClass = columnClasses[index];
+                    if (columnClass) {
+                        cell.classList.add(columnClass);
+                    }
+                });
+            });
+
+            const cells = datasetTable.querySelectorAll(
+                "tbody td:not(.dataset-row-actions-cell):not(.dataset-row-restore-cell)"
+            );
+            cells.forEach((cell) => previewCells.push(cell));
         });
 
-        const cells = datasetTable.querySelectorAll("tbody td");
-        if (!cells.length) {
+        if (!previewCells.length) {
             return;
         }
 
@@ -377,7 +385,7 @@
             positionPopup(cell);
         }
 
-        cells.forEach((cell) => {
+        previewCells.forEach((cell) => {
             const fullText = (cell.innerText || "").trim();
             if (!fullText) {
                 return;
@@ -421,7 +429,9 @@
             if (popup.contains(event.target)) {
                 return;
             }
-            const clickedCell = event.target.closest(".data-card .data-table tbody td");
+            const clickedCell = event.target.closest(
+                ".data-card .data-table tbody td:not(.dataset-row-actions-cell):not(.dataset-row-restore-cell)"
+            );
             if (clickedCell) {
                 return;
             }
@@ -434,16 +444,204 @@
             }
         });
 
-        const datasetWrap = document.querySelector(".data-card .table-wrap");
[…]

[…]
