### 2026-02-13 b2ca6d8
Add mkdocs site on to new Django app via markdown rendering
