+++ django_workbench/workbench/views.py
@@ -2,19 +2,27 @@
 from __future__ import annotations
 
 import json
+import mimetypes
 import random
 import re
 import zipfile
 import copy
+import importlib
+import importlib.util
+import inspect
 from datetime import date, datetime
 from io import BytesIO, StringIO
-from urllib.parse import urlencode
+from pathlib import Path
+from urllib.parse import parse_qs, urlencode, urljoin, urlparse
 from typing import Any, Optional
 from uuid import UUID, uuid4
 
+import markdown
 import pandas as pd
+import yaml
+from bs4 import BeautifulSoup
 from django.contrib import messages
-from django.http import HttpRequest, HttpResponse, JsonResponse
+from django.http import FileResponse, Http404, HttpRequest, HttpResponse, JsonResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 from django.utils.text import slugify
@@ -487,6 +495,426 @@ WORKBOOK_TITLE_ALLOWED_PATTERN = re.compile(r"^[A-Za-z0-9 -]+$")
 WORKBENCH_ROW_ID_COL = "_workbench_row_id"
 EXCLUSION_REASON_COL = "_exclusion_reason"
 EXCLUSION_REASON_LABEL = "Reason for exclusion"
+DOCS_SOURCE_ROOT = Path(__file__).resolve().parents[2] / "docs"
+DOCS_CONFIG_FILE = Path(__file__).resolve().parents[2] / "mkdocs.yml"
+DOCS_LINK_HOSTS = {"127.0.0.1", "localhost", "pfdtoolkit.org", "www.pfdtoolkit.org"}
+DOCS_SITE_FILE_PREFIXES = ("assets/", "stylesheets/", "search/")
+DOCS_REMOVED_PATHS = {"contact"}
+DOCS_MARKDOWN_EXTENSIONS = [
+    "admonition",
+    "attr_list",
+    "codehilite",
+    "fenced_code",
+    "tables",
+    "toc",
+    "abbr",
+    "md_in_html",
+    "pymdownx.highlight",
+    "pymdownx.inlinehilite",
+    "pymdownx.superfences",
+    "pymdownx.tabbed",
+    "pymdownx.details",
+]
+
+
+def _path_within_root(path: Path, root: Path) -> bool:
+    try:
+        path.relative_to(root)
+    except ValueError:
+        return False
+    return True
+
+
+def _normalise_docs_path(raw_path: Optional[str]) -> str:
+    cleaned = str(raw_path or "").strip().strip("/")
+    if cleaned in {"", "index", "index.html"}:
+        return ""
+    if cleaned.endswith("/index"):
+        cleaned = cleaned[: -len("/index")]
+    if cleaned.endswith(".html"):
+        cleaned = cleaned[: -len(".html")]
+    return cleaned.strip("/")
+
+
+def _docs_markdown_file_path(doc_path: str) -> Path:
+    if not doc_path:
+        return DOCS_SOURCE_ROOT / "index.md"
+    return DOCS_SOURCE_ROOT / f"{doc_path}.md"
+
+
+def _mkdocs_path_to_doc_path(path_value: str) -> Optional[str]:
+    raw_path = str(path_value or "").strip()
+    if not raw_path:
+        return ""
+
+    if raw_path.startswith("/pfd-toolkit/"):
+        raw_path = raw_path[len("/pfd-toolkit/") :]
+
+    cleaned = raw_path.lstrip("/")
+    if not cleaned:
+        return ""
+    if cleaned.startswith(DOCS_SITE_FILE_PREFIXES):
+        return None
+    if cleaned in {
+        "CNAME",
+        "favicon.ico",
+        "googlef12640be708032bb.html",
+        "objects.inv",
[…]
