+++ pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
     "dotenv>=0.9.9",
     "gunicorn>=25.0.3",
     "httpx>=0.28.1",
+    "markdown>=3.8.2",
     "openai>=1.88.0",
     "pandas>=2.2.2,<2.2.3; python_version<'3.13'",
     "pandas>=2.2.3; python_version>='3.13'",
@@ -40,6 +41,7 @@ dependencies = [
     "pydantic>=2.11.7",
     "pymupdf>=1.26.1",
     "python-dateutil>=2.9.0.post0",
+    "pyyaml>=6.0.2",
     "requests>=2.32.3",
     "tiktoken>=0.9.0",
     "tqdm>=4.67.1",
