+++ django_workbench/workbench/views.py
@@ -10,6 +10,7 @@ import copy
 import importlib
 import importlib.util
 import inspect
+from html import escape
 from datetime import date, datetime
 from io import BytesIO, StringIO
 from pathlib import Path
@@ -59,6 +60,53 @@ from .state import (
     undo_last_change,
     workspace_has_activity,
 )
+
+SEO_PAGE_METADATA: dict[str, dict[str, str]] = {
+    "home": {
+        "title": "PFD Toolkit - Analyse Prevention of Future Deaths Reports with AI",
+        "description": (
+            "Analyse Prevention of Future Deaths reports with AI. "
+            "Filter reports, discover themes, and extract structured evidence in one workflow."
+        ),
+        "robots": "index,follow",
+    },
+    "explore": {
+        "title": "Explore PFD Reports | PFD Toolkit",
+        "description": (
+            "Explore and filter Prevention of Future Deaths reports with interactive dashboards and AI-assisted screening."
+        ),
+        "robots": "index,follow",
+    },
+    "themes": {
+        "title": "Analyse Themes in PFD Reports | PFD Toolkit",
+        "description": (
+            "Discover recurring themes across Prevention of Future Deaths reports and preview assignments before applying."
+        ),
+        "robots": "index,follow",
+    },
+    "extract": {
+        "title": "Extract Structured Data from PFD Reports | PFD Toolkit",
+        "description": (
+            "Define custom fields and extract structured information from Prevention of Future Deaths reports using AI."
+        ),
+        "robots": "index,follow",
+    },
+    "settings": {
+        "title": "Workbench Settings | PFD Toolkit",
+        "description": "Configure model providers, API keys, and workspace defaults for PFD Toolkit.",
+        "robots": "noindex,nofollow",
+    },
+    "for_coders": {
+        "title": "For Coders | PFD Toolkit",
+        "description": "Technical documentation and implementation details for PFD Toolkit.",
+        "robots": "index,follow",
+    },
+    "workbook_public": {
+        "title": "Shared PFD Workbook | PFD Toolkit",
+        "description": "View a shared PFD Toolkit workbook with report-level filtering and summary outputs.",
+        "robots": "index,follow",
+    },
+}
 
 
 def _bool_from_post(request: HttpRequest, key: str, default: bool = False) -> bool:
@@ -1332,6 +1380,92 @@ def _build_context(request: HttpRequest) -> dict[str, Any]:
     }
 
 
+def _build_seo_context(
+    request: HttpRequest,
+    page_key: str,
+    *,
+    canonical_path: str,
+    title_override: Optional[str] = None,
+    description_override: Optional[str] = None,
+    robots_override: Optional[str] = None,
+    og_type: str = "website",
+) -> dict[str, str]:
+    defaults = SEO_PAGE_METADATA.get(page_key, {})
+    title = title_override or defaults.get("title") or "PFD Toolkit AI Workbench · Beta"
+    description = (
+        description_override
+        or defaults.get("description")
+        or "Analyse Prevention of Future Deaths reports with AI workflows."
+    )
+    robots_value = robots_override or defaults.get("robots") or "index,follow"
+    canonical_url = request.build_absolute_uri(canonical_path)
+    return {
+        "seo_title": title,
+        "seo_description": description,
+        "seo_robots": robots_value,
+        "seo_canonical_url": canonical_url,
+        "seo_og_type": og_type,
+    }
+
+
+def _render_sitemap_xml(request: HttpRequest) -> HttpResponse:
+    entries: list[tuple[str, str, str, Optional[str]]] = []
+    entries.extend(
+        [
+            (reverse("workbench:index"), "daily", "1.0", None),
[…]
