### 2026-02-13 727ed5a
Add theme setting
