### 2026-02-14 125e6d9
Changing theme no longer requires settings to be saved
