### 2026-02-14 8837762
Memory efficiency tweaks in the Explore tab
