+++ django_workbench/workbench/views.py
@@ -126,6 +126,8 @@ UI_THEME_CHOICES: tuple[dict[str, str], ...] = (
     },
 )
 UI_THEME_IDS = {choice["id"] for choice in UI_THEME_CHOICES}
+DASHBOARD_TOP_N = 12
+DASHBOARD_UNKNOWN_LABEL = "Not specified"
 
 
 def _normalise_ui_theme(raw_value: Any) -> str:
@@ -388,44 +390,31 @@ def _split_dashboard_receivers(value: Any) -> list[str]:
     return receivers
 
 
-def _build_explore_dashboard_payload(reports_df: pd.DataFrame) -> dict[str, Any]:
+def _sorted_dashboard_count_rows(counter: dict[str, int], *, limit: Optional[int] = None) -> list[dict[str, Any]]:
+    rows = sorted(counter.items(), key=lambda item: (-item[1], item[0].casefold()))
+    if limit is not None:
+        rows = rows[:limit]
+    return [{"name": name, "value": int(value)} for name, value in rows]
+
+
+def _build_explore_dashboard_options(reports_df: pd.DataFrame) -> dict[str, list[str]]:
     if reports_df.empty:
-        return {
-            "rows": [],
-            "options": {
-                "coroners": [],
-                "areas": [],
-                "receivers": [],
-            },
-        }
+        return {"coroners": [], "areas": [], "receivers": []}
 
     row_count = len(reports_df)
     default_series = pd.Series([""] * row_count, index=reports_df.index, dtype=object)
-
-    if "date" in reports_df.columns:
-        date_series = pd.to_datetime(reports_df["date"], errors="coerce")
-    else:
-        date_series = pd.Series([pd.NaT] * row_count, index=reports_df.index, dtype="datetime64[ns]")
-
     coroner_series = reports_df["coroner"] if "coroner" in reports_df.columns else default_series
     area_series = reports_df["area"] if "area" in reports_df.columns else default_series
     receiver_series = reports_df["receiver"] if "receiver" in reports_df.columns else default_series
-
-    rows: list[dict[str, Any]] = []
     coroners: set[str] = set()
     areas: set[str] = set()
     receivers: set[str] = set()
 
-    for date_value, coroner_value, area_value, receiver_value in zip(
-        date_series,
+    for coroner_value, area_value, receiver_value in zip(
         coroner_series,
         area_series,
         receiver_series,
     ):
-        date_iso = ""
-        if pd.notna(date_value):
-            date_iso = date_value.strftime("%Y-%m-%d")
-
         coroner = _normalise_dashboard_value(coroner_value)
         area = _normalise_dashboard_value(area_value)
         receiver_parts = _split_dashboard_receivers(receiver_value)
@@ -437,25 +426,104 @@ def _build_explore_dashboard_payload(reports_df: pd.DataFrame) -> dict[str, Any]
         for receiver_name in receiver_parts:
             receivers.add(receiver_name)
 
-        rows.append(
-            {
-                "date": date_iso,
-                "year_month": date_iso[:7] if date_iso else "",
-                "coroner": coroner,
-                "area": area,
-                "receivers": receiver_parts,
-            }
-        )
-
     sort_key = lambda value: value.casefold()
     return {
-        "rows": rows,
-        "options": {
-            "coroners": sorted(coroners, key=sort_key),
-            "areas": sorted(areas, key=sort_key),
-            "receivers": sorted(receivers, key=sort_key),
-        },
+        "coroners": sorted(coroners, key=sort_key),
+        "areas": sorted(areas, key=sort_key),
+        "receivers": sorted(receivers, key=sort_key),
     }
+
+
[…]
