+++ django_workbench/workbench/static/workbench/workbench.js
@@ -2244,15 +2244,12 @@
             return;
         }
 
-        const rows = Array.isArray(payload.rows) ? payload.rows : [];
         const options = payload && typeof payload.options === "object" ? payload.options : {};
         const selected = payload && typeof payload.selected === "object" ? payload.selected : {};
+        let summary = payload && typeof payload.summary === "object" ? payload.summary : {};
         const coronerOptions = Array.isArray(options.coroners) ? options.coroners : [];
         const areaOptions = Array.isArray(options.areas) ? options.areas : [];
         const receiverOptions = Array.isArray(options.receivers) ? options.receivers : [];
-
-        const TOP_N = 12;
-        const UNKNOWN_LABEL = "Not specified";
 
         function chartPalette() {
             return {
@@ -2275,15 +2272,6 @@
                 return String(value).trim();
             }
             return value.trim();
-        }
-
-        function normaliseReceivers(rawReceivers) {
-            if (!Array.isArray(rawReceivers) || !rawReceivers.length) {
-                return [];
-            }
-            return rawReceivers
-                .map((value) => normaliseValue(value))
-                .filter((value) => Boolean(value));
         }
 
         const charts = {};
@@ -2352,81 +2340,6 @@
             const panelUrl = `${panelBase}?page=1${query ? `&${query}` : ""}`;
             const targetUrl = `${browserBase}1${query ? `&${query}` : ""}`;
             window.WorkbenchDatasetPanel.fetchAndSwapDataset(panelUrl, targetUrl);
-        }
-
-        function toSortedCountRows(counter) {
-            return Array.from(counter.entries())
-                .sort((left, right) => {
-                    if (right[1] !== left[1]) {
-                        return right[1] - left[1];
-                    }
-                    return left[0].localeCompare(right[0], undefined, { sensitivity: "base" });
-                })
-                .map(([name, value]) => ({ name, value }));
-        }
-
-        function filterRows(sourceRows) {
-            return sourceRows.filter((row) => {
-                const rowCoroner = normaliseValue(row.coroner);
-                const rowArea = normaliseValue(row.area);
-                const rowReceivers = normaliseReceivers(row.receivers);
-
-                if (selectedFilters.coroner.size && !selectedFilters.coroner.has(rowCoroner)) {
-                    return false;
-                }
-                if (selectedFilters.area.size && !selectedFilters.area.has(rowArea)) {
-                    return false;
-                }
-                if (selectedFilters.receiver.size) {
-                    const hasMatch = rowReceivers.some((value) => selectedFilters.receiver.has(value));
-                    if (!hasMatch) {
-                        return false;
-                    }
-                }
-                if (selectedFilters.receiver.size && !rowReceivers.length) {
-                    return false;
-                }
-                return true;
-            });
-        }
-
-        function computeMonthlySeries(sourceRows) {
-            const monthlyCounts = new Map();
-            sourceRows.forEach((row) => {
-                const month = normaliseValue(row.year_month);
-                if (!month) {
-                    return;
-                }
-                monthlyCounts.set(month, (monthlyCounts.get(month) || 0) + 1);
-            });
-
-            return Array.from(monthlyCounts.entries())
-                .sort(([left], [right]) => left.localeCompare(right))
-                .map(([name, value]) => ({ name, value }));
-        }
-
-        function computeTopValues(sourceRows, fieldName) {
-            const counts = new Map();
-            sourceRows.forEach((row) => {
-                const value = normaliseValue(row[fieldName]) || UNKNOWN_LABEL;
-                counts.set(value, (counts.get(value) || 0) + 1);
-            });
[…]
