### 2026-02-14 babc975
Disable AI workflow buttons when LLM is not initialised
