### 2026-02-14 dd83dcf
Add Explore loading icon on to side bar element
