+++ django_workbench/RAILWAY_DEPLOYMENT_GUIDE.md
@@ -1,0 +1,220 @@
+# Django Deployment Guide (Railway) for pfdtoolkit.org
+
+This guide migrates you from:
+- MkDocs at `https://pfdtoolkit.org/`
+- Streamlit at `https://workbench.pfdtoolkit.org/`
+
+to a **single Django deployment** on Railway at `https://pfdtoolkit.org/`.
+
+## 1. Target Architecture
+
+- One Railway service running Django + Gunicorn.
+- One Railway Postgres database.
+- Django serves app pages, docs pages (`/for-coders/`), `robots.txt`, and `sitemap.xml`.
+- `workbench.pfdtoolkit.org` removed from DNS and Railway.
+
+## 2. Important Pre-Cutover Risks
+
+### Risk A: SQLite on Railway is not suitable
+Your current Django settings default to SQLite (`django_workbench/pfd_workbench/settings.py`). Railway containers are ephemeral, so SQLite data can be lost.
+
+Use Railway Postgres and `DATABASE_URL`.
+
+### Risk B: Old MkDocs URLs will 404 after cutover
+Current Django routes do not serve old root docs paths like `/llm_setup/` or `/reference/loader/`.
+
+Before cutover, add redirects from old MkDocs paths to your Django docs routes (`/for-coders/?doc=...`), or keep equivalent pages at old paths.
+
+## 3. Update Django Settings for Production
+
+Edit `django_workbench/pfd_workbench/settings.py` to support env-based production config.
+
+## 3.1 Add imports
+```python
+import os
+import dj_database_url
+```
+
+## 3.2 Replace secret/debug/hosts
+```python
+SECRET_KEY = os.environ["DJANGO_SECRET_KEY"]
+DEBUG = os.getenv("DJANGO_DEBUG", "False").lower() == "true"
+
+ALLOWED_HOSTS = [
+    host.strip()
+    for host in os.getenv(
+        "DJANGO_ALLOWED_HOSTS",
+        "pfdtoolkit.org,www.pfdtoolkit.org"
+    ).split(",")
+    if host.strip()
+]
+
+CSRF_TRUSTED_ORIGINS = [
+    origin.strip()
+    for origin in os.getenv(
+        "DJANGO_CSRF_TRUSTED_ORIGINS",
+        "https://pfdtoolkit.org,https://www.pfdtoolkit.org"
+    ).split(",")
+    if origin.strip()
+]
+```
+
+## 3.3 Replace DATABASES
+```python
+DATABASES = {
+    "default": dj_database_url.config(
+        default=f"sqlite:///{BASE_DIR / 'db.sqlite3'}",
+        conn_max_age=600,
+        ssl_require=not DEBUG,
+    )
+}
+```
+
+## 3.4 Static files with WhiteNoise
+
+Add WhiteNoise middleware directly after `SecurityMiddleware`:
+```python
+"django.middleware.security.SecurityMiddleware",
+"whitenoise.middleware.WhiteNoiseMiddleware",
+```
+
+Add static config:
+```python
+STATIC_URL = "/static/"
+STATIC_ROOT = BASE_DIR / "staticfiles"
+STATICFILES_DIRS = [BASE_DIR / "static"]
+STATICFILES_STORAGE = "whitenoise.storage.CompressedManifestStaticFilesStorage"
+```
+
+## 3.5 HTTPS/proxy settings
+```python
+SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")
+SECURE_SSL_REDIRECT = os.getenv("DJANGO_SECURE_SSL_REDIRECT", "True").lower() == "true"
+SESSION_COOKIE_SECURE = True
+CSRF_COOKIE_SECURE = True
+```
+
[…]
