+++ django_workbench/workbench/urls.py
@@ -8,6 +8,7 @@ app_name = "workbench"
 urlpatterns = [
     path("", views.index, name="index"),
     path("home/", views.home, name="home"),
+    path("privacy/", views.privacy_policy, name="privacy_policy"),
     path("robots.txt", views.robots_txt, name="robots_txt"),
     path("sitemap.xml", views.sitemap_xml, name="sitemap_xml"),
     path("explore-pfds/", views.explore, name="explore"),
