### 2026-02-24 #10074 @PetrKorab
@Sam-Osian Thanks for the useful package.
Implementing open LLMs alongside OpenAI API would help wider use among researchers. Would it be possible to add Mistral models to the llm_client?
