### 2026-03-01 ec021bf
Update settings for Railway deployment
