### 2026-03-02 276bd2f
Fix favicon issue
