### 2026-03-02 52aa0db
Rewrite stale test to no longer reference rows element
