### 2026-03-02 6fd03f6
Add shared base dataset caching, lightweight request logging, and cut repeated dashboard recomputation
