+++ django_workbench/workbench/views.py
@@ -1,6 +1,8 @@
 """Views for the standalone Django Workbench."""
 from __future__ import annotations
 
+from collections import OrderedDict
+import logging
 import json
 import mimetypes
 import random
@@ -14,6 +16,7 @@ from html import escape
 from datetime import date, datetime
 from io import BytesIO, StringIO
 from pathlib import Path
+from time import perf_counter
 from urllib.parse import parse_qs, urlencode, urljoin, urlparse
 from typing import Any, Optional
 from uuid import UUID, uuid4
@@ -37,6 +40,7 @@ from .services import (
     build_llm,
     build_screener,
     load_reports_dataframe,
+    normalise_parallel_workers,
     parse_seed_topics,
     resolve_theme_emoji,
 )
@@ -61,6 +65,8 @@ from .state import (
     undo_last_change,
     workspace_has_activity,
 )
+
+logger = logging.getLogger(__name__)
 
 SEO_PAGE_METADATA: dict[str, dict[str, str]] = {
     "home": {
@@ -134,6 +140,8 @@ UI_THEME_CHOICES: tuple[dict[str, str], ...] = (
         "description": "Infrared Ops-inspired command palette.",
     },
 )
+DASHBOARD_PAYLOAD_CACHE: "OrderedDict[tuple[str, tuple[str, ...], tuple[str, ...], tuple[str, ...], bool], dict[str, Any]]" = OrderedDict()
+DASHBOARD_PAYLOAD_CACHE_MAX = 64
 
 
 def favicon(request: HttpRequest) -> HttpResponse:
@@ -143,6 +151,76 @@ def favicon(request: HttpRequest) -> HttpResponse:
 UI_THEME_IDS = {choice["id"] for choice in UI_THEME_CHOICES}
 DASHBOARD_TOP_N = 12
 DASHBOARD_UNKNOWN_LABEL = "Not specified"
+
+
+def _normalise_report_limit(raw_value: Any) -> Optional[int]:
+    if raw_value in {None, ""}:
+        return None
+    try:
+        parsed = int(raw_value)
+    except (TypeError, ValueError):
+        return None
+    if parsed <= 0:
+        return None
+    return parsed
+
+
+def _dashboard_payload_cache_key(
+    dataset_token: str,
+    selected_filters: dict[str, list[str]],
+    include_options: bool,
+) -> tuple[str, tuple[str, ...], tuple[str, ...], tuple[str, ...], bool]:
+    return (
+        dataset_token,
+        tuple(selected_filters.get("coroner", [])),
+        tuple(selected_filters.get("area", [])),
+        tuple(selected_filters.get("receiver", [])),
+        include_options,
+    )
+
+
+def _get_cached_dashboard_payload(
+    dataset_token: str,
+    reports_df: pd.DataFrame,
+    *,
+    selected_filters: dict[str, list[str]],
+    include_options: bool,
+) -> dict[str, Any]:
+    if dataset_token:
+        cache_key = _dashboard_payload_cache_key(dataset_token, selected_filters, include_options)
+        cached_payload = DASHBOARD_PAYLOAD_CACHE.get(cache_key)
+        if isinstance(cached_payload, dict):
+            DASHBOARD_PAYLOAD_CACHE.move_to_end(cache_key)
+            return copy.deepcopy(cached_payload)
+
+    payload = _build_explore_dashboard_payload(
+        reports_df,
+        selected_filters=selected_filters,
+        include_options=include_options,
+    )
+    if dataset_token:
+        DASHBOARD_PAYLOAD_CACHE[cache_key] = copy.deepcopy(payload)
[…]
