### 2026-03-02 9457382
Update allowed areas list for cleaning, introduce fuzzy matching for non-matches, update changelog
