+++ src/pfd_toolkit/config.py
@@ -56,6 +56,7 @@ class GeneralConfig:
         "Bedfordshire and Luton",
         "Berkshire",
         "Birmingham and Solihull",
+        "Black Country",
         "Blackpool and Fylde",
         "Buckinghamshire",
         "Cambridgeshire and Peterborough",
@@ -69,26 +70,20 @@ class GeneralConfig:
         "Cumbria",
         "Derby and Derbyshire",
         "Dorset",
-        "East Riding of Yorkshire and Kingston-upon-Hull",
+        "East Riding and Hull",
         "East Sussex",
-        "Essex",
+        "Essex and Thurrock",
         "Gateshead and South Tyneside",
         "Gloucestershire",
         "Greater Lincolnshire",
-        "Greater Manchester North",
-        "Greater Manchester South",
-        "Greater Manchester West",
         "Gwent",
         "Hampshire, Portsmouth and Southampton",
         "Herefordshire",
         "Hertfordshire",
         "Isle of Wight",
-        "Kent Central and South East",
-        "Kent Mid and Medway",
-        "Kent North East",
-        "Kent North West",
-        "Lancashire and Blackburn with Darwen",
-        "Leicester City and South Leicestershire",
+        "Kent and Medway",
+        "Lancashire with Blackburn and Darwen",
+        "Leicester and South Leicestershire",
         "Liverpool and the Wirral",
         "London City",
         "London East",
@@ -99,17 +94,19 @@ class GeneralConfig:
         "London South",
         "London West",
         "Manchester City",
+        "Manchester North",
+        "Manchester South",
+        "Manchester West",
         "Milton Keynes",
-        "Newcastle and North Tyneside",
+        "Newcastle Upon Tyne and North Tyneside",
         "Norfolk",
         "North Wales (East and Central)",
         "North West Wales",
         "North Yorkshire and York",
         "Northamptonshire",
         "Northumberland",
-        "Nottinghamshire",
+        "Nottinghamshire and Nottingham",
         "Oxfordshire",
-        "Powys, Bridgend and Glamorgan",
         "Rutland and North Leicestershire",
         "Sefton, Knowsley and St Helens",
         "Shropshire, Telford and Wrekin",
@@ -121,7 +118,6 @@ class GeneralConfig:
         "Surrey",
         "Swansea and Neath Port Talbot",
         "Teesside and Hartlepool",
-        "The Black Country Jurisdiction",
         "Warwickshire",
         "West Sussex, Brighton and Hove",
         "Wiltshire and Swindon",
@@ -144,6 +140,25 @@ class GeneralConfig:
         "Inner South London": "London Inner South",
         "Cardiff and the Vale of Glamorgan": "South Wales Central",
         "Cardiff and Vale of Glamorgan": "South Wales Central",
+    }
+
+    # Legacy coroner areas that should be recoded to current canonical areas
+    LEGACY_AREA_SYNONYMS: Dict[str, str] = {
+        "The Black Country Jurisdiction": "Black Country",
+        "East Riding of Yorkshire and Kingston-upon-Hull": "East Riding and Hull",
+        "Essex": "Essex and Thurrock",
+        "Greater Manchester North": "Manchester North",
+        "Greater Manchester South": "Manchester South",
+        "Greater Manchester West": "Manchester West",
+        "Kent Central and South East": "Kent and Medway",
+        "Kent Mid and Medway": "Kent and Medway",
+        "Kent North East": "Kent and Medway",
+        "Kent North West": "Kent and Medway",
+        "Lancashire and Blackburn with Darwen": "Lancashire with Blackburn and Darwen",
+        "Leicester City and South Leicestershire": "Leicester and South Leicestershire",
[…]
