+++ src/pfd_toolkit/cleaner.py
@@ -1,4 +1,8 @@
 import logging
+from dataclasses import dataclass
+from difflib import SequenceMatcher
+from functools import lru_cache
+import re
 import warnings
 
 import pandas as pd
@@ -23,32 +27,206 @@ logging.getLogger("openai").setLevel(logging.WARNING)
 warnings.filterwarnings("ignore", category=TqdmWarning)
 
 
+def _normalise_area_for_matching(area: str) -> str:
+    """Return a simplified representation used only for area matching."""
+    if not isinstance(area, str):
+        return ""
+
+    normalised = area.strip().casefold()
+    normalised = normalised.replace("&", " and ")
+    normalised = re.sub(r"[^a-z0-9]+", " ", normalised)
+    return " ".join(normalised.split())
+
+
+DEFAULT_AREA_MATCH_STRATEGY = "exact_then_fuzzy"
+AREA_MATCH_STRATEGIES = ("exact_only", "exact_then_fuzzy", "fuzzy_only")
+AREA_FUZZY_MIN_SCORE = 0.88
+AREA_FUZZY_MIN_MARGIN = 0.03
+
+
+@dataclass(frozen=True)
+class AreaMatchResult:
+    canonical_area: str | None
+    match_method: str
+    score: float | None = None
+    matched_value: str | None = None
+
+
+@lru_cache(maxsize=1)
+def _build_area_match_lookups() -> tuple[dict[str, str], dict[str, str], dict[str, str]]:
+    """Build cached normalized lookups for canonical areas and synonyms."""
+    allowed_lookup = {
+        _normalise_area_for_matching(candidate): candidate
+        for candidate in GeneralConfig.ALLOWED_AREAS
+    }
+    synonym_lookup = {
+        _normalise_area_for_matching(alias): canonical
+        for alias, canonical in GeneralConfig.AREA_SYNONYMS.items()
+        if canonical in GeneralConfig.ALLOWED_AREAS
+    }
+    legacy_lookup = {
+        _normalise_area_for_matching(alias): canonical
+        for alias, canonical in GeneralConfig.LEGACY_AREA_SYNONYMS.items()
+        if canonical in GeneralConfig.ALLOWED_AREAS
+    }
+    return allowed_lookup, synonym_lookup, legacy_lookup
+
+
+def _exact_area_match(area: str) -> AreaMatchResult:
+    """Resolve exact and normalized matches against canonical areas."""
+    if not isinstance(area, str):
+        return AreaMatchResult(None, "invalid_input")
+
+    stripped = area.strip()
+    if not stripped:
+        return AreaMatchResult(None, "empty")
+
+    direct_synonym = GeneralConfig.AREA_SYNONYMS.get(stripped)
+    if direct_synonym in GeneralConfig.ALLOWED_AREAS:
+        return AreaMatchResult(direct_synonym, "exact_synonym", 1.0, stripped)
+
+    direct_legacy = GeneralConfig.LEGACY_AREA_SYNONYMS.get(stripped)
+    if direct_legacy in GeneralConfig.ALLOWED_AREAS:
+        return AreaMatchResult(direct_legacy, "legacy_synonym", 1.0, stripped)
+
+    if stripped in GeneralConfig.ALLOWED_AREAS:
+        return AreaMatchResult(stripped, "exact_allowed", 1.0, stripped)
+
+    allowed_lookup, synonym_lookup, legacy_lookup = _build_area_match_lookups()
+    match_key = _normalise_area_for_matching(stripped)
+    if not match_key:
+        return AreaMatchResult(None, "empty")
+
+    if match_key in synonym_lookup:
+        return AreaMatchResult(
+            synonym_lookup[match_key],
+            "normalised_synonym",
+            1.0,
[…]

[…]
