### 2026-03-02 f223ac0
Change receiver parsing cleaning prompt & add text normalisation
