### 2026-03-02 7b6441d
Recipient changes: normalise formatting, remove boilerplate roles, apply canonical mappings
