+++ src/pfd_toolkit/cleaner.py
@@ -89,6 +89,25 @@ RECEIVER_ORG_CUES = (
     "cardinal healthcare",
 )
 
+RECEIVER_CANONICAL_MAP = {
+    "department of health": "Department of Health and Social Care",
+    "department of health and social care": "Department of Health and Social Care",
+    "nhs england and nhs improvement": "NHS England",
+    "highways agency": "National Highways",
+    "highways england": "National Highways",
+}
+
+RECEIVER_ROLE_PREFIX_TEXTS = (
+    "the chief executive officer of ",
+    "chief executive officer of ",
+    "the chief executive of ",
+    "chief executive of ",
+    "the chief executive officer ",
+    "chief executive officer ",
+    "the chief executive ",
+    "chief executive ",
+)
+
 
 def _normalise_receiver_match_key(text: str) -> str:
     """Return a permissive key used only for receiver cleanup rules."""
@@ -106,11 +125,44 @@ def _looks_like_receiver_organisation(text: str) -> bool:
     return any(cue in key for cue in RECEIVER_ORG_CUES) or text.isupper()
 
 
+def _normalise_receiver_formatting(text: str) -> str:
+    """Apply low-risk receiver formatting cleanup while preserving casing."""
+    cleaned = " ".join((text or "").split()).strip(" ,;:.")
+    if not cleaned:
+        return ""
+
+    cleaned = cleaned.replace("’", "'").replace("‘", "'")
+    cleaned = re.sub(r"^[\-\–\)\(:\s]+", "", cleaned)
+    cleaned = re.sub(r"^\s*The\s+", "", cleaned, flags=re.IGNORECASE)
+    cleaned = re.sub(r"\bDept\.?\b", "Department", cleaned, flags=re.IGNORECASE)
+    cleaned = re.sub(r"\s*&\s*", " and ", cleaned)
+    cleaned = re.sub(
+        r"^Secretary of State for (?:the )?Department(?: of)? ",
+        "Secretary of State for ",
+        cleaned,
+        flags=re.IGNORECASE,
+    )
+    cleaned = re.sub(r"\s{2,}", " ", cleaned)
+    cleaned = re.sub(r"\s*\(([A-Za-z0-9&./' -]{2,15})\)\s*$", "", cleaned)
+    cleaned = re.sub(r'\s*\((?:"?the Trust"?|Highways Department|Local Government)\)\s*$', "", cleaned, flags=re.IGNORECASE)
+    return cleaned.strip(" ,;:.")
+
+
+def _canonicalise_receiver_segment(text: str) -> str:
+    """Apply explicit receiver canonical mappings while preserving formatting."""
+    key = _normalise_receiver_match_key(text)
+    return RECEIVER_CANONICAL_MAP.get(key, text)
+
+
 def _clean_receiver_segment(segment: str) -> str:
     """Normalise a single receiver segment while preserving useful casing."""
     text = " ".join((segment or "").replace("\n", " ").split()).strip(" ,;:.")
     if not text:
         return ""
+
+    text = re.sub(r"^\[REDACTED\][,:\- ]*", "", text, flags=re.IGNORECASE)
+    text = re.sub(r'^[A-Z][^,;]{0,80}\bMP\b[,:\- ]*', "", text)
+    text = _normalise_receiver_formatting(text)
 
     if _normalise_receiver_match_key(text) in RECEIVER_DROP_VALUES:
         return ""
@@ -129,9 +181,28 @@ def _clean_receiver_segment(segment: str) -> str:
             text = parts[-1]
             parts = [text]
 
+    normalised_role_form = _normalise_receiver_formatting(text)
+    role_form_key = _normalise_receiver_match_key(normalised_role_form)
+    for prefix in RECEIVER_ROLE_PREFIX_TEXTS:
[…]
