### 2026-03-02 84762b6
Add expand acronym function to cleaner
