+++ tests/test_cleaner_screener.py
@@ -37,6 +37,22 @@ class DummyLLM:
         return outputs
 
 
+class EchoAreaLLM(DummyLLM):
+    def generate(self, prompts, response_format=None, **kwargs):
+        outputs = []
+        for p in prompts:
+            text = p.split("\n")[-1]
+            if response_format is None:
+                outputs.append(text.upper())
+            else:
+                field_name = next(iter(response_format.model_fields))
+                if field_name == "area":
+                    outputs.append(response_format(area=text))
+                else:
+                    outputs.append(response_format(**{field_name: text.upper()}))
+        return outputs
+
+
 def test_cleaner_basic():
     df = pd.DataFrame(
         {
@@ -218,11 +234,11 @@ def test_cleaner_unrecognised_area_defaults_to_other():
     assert cleaned[GeneralConfig.COL_AREA].iloc[0] == "Other"
 
 
-def test_cleaner_area_synonyms():
+def test_cleaner_area_prompt_includes_allowed_area_guidance():
     df = pd.DataFrame(
         {
             GeneralConfig.COL_CORONER_NAME: ["john"],
-            GeneralConfig.COL_AREA: ["area"],
+            GeneralConfig.COL_AREA: ["County Durham and Darlington"],
             GeneralConfig.COL_RECEIVER: ["x"],
             GeneralConfig.COL_INVESTIGATION: ["inv"],
             GeneralConfig.COL_CIRCUMSTANCES: ["circ"],
@@ -230,22 +246,58 @@ def test_cleaner_area_synonyms():
         }
     )
 
-    class SynonymLLM(DummyLLM):
-        def generate(self, prompts, response_format=None, **kwargs):
-            outputs = []
-            for p in prompts:
-                text = p.split("\n")[-1]
-                if response_format is None:
-                    outputs.append(text.upper())
-                else:
-                    field_name = next(iter(response_format.model_fields))
-                    if field_name == "area":
-                        outputs.append(response_format(area="West London"))
-                    else:
-                        outputs.append(response_format(**{field_name: text.upper()}))
-            return outputs
+    cleaner = Cleaner(df, DummyLLM())
+    prompt = cleaner.area_prompt_template
 
-    cleaner = Cleaner(df, SynonymLLM())
+    assert "Allowed area labels:" in prompt
+    assert "County Durham and Darlington" in prompt
+    assert "London West" in prompt
+    assert "Return exactly one label from this list." in prompt
+    assert "If no logical match exists, return exactly: <NA>." in prompt
+
+
+@pytest.mark.parametrize(
+    ("raw_area", "expected"),
+    [
+        ("West Yorkshire (E)", "Yorkshire West Eastern"),
+        ("South Yorkshire (West)", "Yorkshire South West"),
+        (
+            "City of Kingston Upon Hull and the County of the East Riding of Yorkshire",
+            "East Riding and Hull",
+        ),
+    ],
+)
+def test_cleaner_area_uses_deterministic_matching_for_near_matches(raw_area, expected):
+    df = pd.DataFrame(
+        {
+            GeneralConfig.COL_CORONER_NAME: ["john"],
+            GeneralConfig.COL_AREA: [raw_area],
+            GeneralConfig.COL_RECEIVER: ["x"],
+            GeneralConfig.COL_INVESTIGATION: ["inv"],
+            GeneralConfig.COL_CIRCUMSTANCES: ["circ"],
+            GeneralConfig.COL_CONCERNS: ["conc"],
+        }
+    )
+
[…]

[…]
