### 2026-03-04 b80a2af
Django dataset to refresh on initial dataset load for a new session, then freeze until start over
