### 2026-03-04 7c65d4b
Create 5 receiver 'collections' and add functionality to load_reports()
