### 2026-03-04 25f9dc5
GitHub workflow now updates collection assignment of new reports
