+++ .github/workflows/pfd_update.yml
@@ -48,6 +48,7 @@ jobs:
         run: |
           set +x
           uv run python scripts/update_reports.py
+          uv run python scripts/update_collections.py
           set -x
 
       - name: Upload dataset to release

[…]
