### 2026-03-04 725f6bc
Add explore view for individual collections
