+++ django_workbench/workbench/views.py
@@ -44,6 +44,8 @@ from .services import (
     parse_seed_topics,
     resolve_theme_emoji,
 )
+from pfd_toolkit.loader import load_reports
+from pfd_toolkit.collections import COLLECTION_COLUMNS, apply_collection_columns
 from .state import (
     build_theme_summary_table,
     clear_outputs_for_modified_dataset,
@@ -187,6 +189,41 @@ BROWSE_COLLECTIONS: tuple[dict[str, str], ...] = (
         "accent": "mint",
     },
 )
+
+
+def _browse_collection_meta(collection_slug: str) -> dict[str, str]:
+    for collection in BROWSE_COLLECTIONS:
+        if collection["slug"] == collection_slug:
+            return collection
+    raise Http404("Collection not found.")
+
+
+def _ensure_collection_columns_in_reports(reports_df: pd.DataFrame) -> pd.DataFrame:
+    missing_columns = [column for column in COLLECTION_COLUMNS.values() if column not in reports_df.columns]
+    if not missing_columns or reports_df.empty:
+        return reports_df
+    reports_with_collections = reports_df.copy()
+    apply_collection_columns(reports_with_collections)
+    return reports_with_collections
+
+
+def _load_browse_reports_df() -> pd.DataFrame:
+    reports_df = load_reports(refresh=False)
+    reports_df, _ = _ensure_workbench_row_ids(reports_df)
+    return _ensure_collection_columns_in_reports(reports_df)
+
+
+def _reports_for_browse_collection(reports_df: pd.DataFrame, collection_slug: str) -> pd.DataFrame:
+    reports_with_collections = _ensure_collection_columns_in_reports(reports_df)
+    collection_column = COLLECTION_COLUMNS.get(collection_slug)
+    if not collection_column or collection_column not in reports_with_collections.columns:
+        return reports_with_collections.iloc[0:0].copy()
+    collection_mask = reports_with_collections[collection_column].fillna(False).astype(bool)
+    filtered_reports = reports_with_collections.loc[collection_mask].reset_index(drop=True)
+    visible_columns = [
+        column for column in filtered_reports.columns if not str(column).startswith("theme_")
+    ]
+    return filtered_reports.loc[:, visible_columns].copy()
 
 UI_THEME_CHOICES: tuple[dict[str, str], ...] = (
     {
@@ -2953,10 +2990,23 @@ def privacy_policy(request: HttpRequest) -> HttpResponse:
 @require_http_methods(["GET"])
 def browse_page(request: HttpRequest) -> HttpResponse:
     init_state(request.session)
-    _ensure_workspace_reports_loaded(request)
     context = _build_context(request)
+    reports_df = _load_browse_reports_df()
+    browse_collections: list[dict[str, str | int]] = []
+    for collection in BROWSE_COLLECTIONS:
+        collection_column = COLLECTION_COLUMNS.get(collection["slug"], "")
+        collection_count = 0
+        if collection_column and collection_column in reports_df.columns:
+            collection_count = int(reports_df[collection_column].fillna(False).astype(bool).sum())
+        browse_collections.append(
+            {
+                **collection,
+                "count": collection_count,
+                "url": reverse("workbench:browse_collection", kwargs={"collection_slug": collection["slug"]}),
+            }
+        )
     context["current_page"] = "browse"
-    context["browse_collections"] = BROWSE_COLLECTIONS
+    context["browse_collections"] = browse_collections
     context.update(
         _build_seo_context(
             request,
@@ -2965,6 +3015,159 @@ def browse_page(request: HttpRequest) -> HttpResponse:
         )
     )
     return render(request, "workbench/browse.html", context)
+
+
+@require_http_methods(["GET"])
[…]
