### 2026-03-04 3f12ede
Add custom collection view in browse
