### 2026-03-04 d659075
Remove legacy Streamlit files
