+++ app/streamlit_app.py
@@ -1,3785 +1,0 @@
-"""Interactive Streamlit dashboard for the PFD Toolkit API."""
-from __future__ import annotations
-
-import ast
-import copy
-import json
-import sys
-import zipfile
-from html import escape
-from datetime import date
-from io import BytesIO
-from pathlib import Path
-from typing import Any, Dict, Iterable, List, Optional, Tuple
-
-import pandas as pd
-import streamlit as st
-from streamlit.delta_generator import DeltaGenerator
-from pydantic import Field, create_model
-
-ROOT_DIR = Path(__file__).resolve().parent
-SRC_DIR = ROOT_DIR / "src"
-if SRC_DIR.exists():
-    src_str = str(SRC_DIR)
-    if src_str not in sys.path:
-        sys.path.insert(0, src_str)
-
-from pfd_toolkit.extractor import Extractor
-from pfd_toolkit.loader import load_reports
-from pfd_toolkit.llm import LLM
-from pfd_toolkit.screener import Screener
-
-LOGO_PATH = ROOT_DIR / "assets" / "badge-circle.png"
-OPENROUTER_API_BASE = "https://openrouter.ai/api/v1"
-
-DEFAULT_THEME_EMOJI = "💡"
-THEME_EMOJI_BANNED_TOKENS = {
-    "🩸",
-    "🩹",
-    "🔪",
-    "🗡",
-    "🗡️",
-    "🪓",
-    "⚔",
-    "⚔️",
-    "💣",
-    "🧨",
-    "🔫",
-    "☠",
-    "☠️",
-    "💀",
-    "🧟",
-    "⚰️",
-    "🪦",
-    "🧛", "🧛‍♂️", "🧛‍♀️",
-    "🥀",
-    "⚱️",
-    "🕷️", "🕸️",
-    "🕊️",
-    "🕯️",
-}
-
-ThemeEmojiModel = create_model(
-    "ThemeEmojiModel",
-    emoji=(
-        str,
-        Field(
-            min_length=1,
-            max_length=8,
-            pattern=r"^\S+$",
-            description="A single emoji with no surrounding text or whitespace.",
-            json_schema_extra={"examples": ["🗣️"]},
-        ),
-    ),
-)
-
-
-# Keys for the reproducible script caching feature
-REPRO_SCRIPT_KEY = "repro_script_lines"
-REPRO_ACTION_COUNTS_KEY = "repro_action_counts"
-LLM_SIGNATURE_KEY = "llm_config_signature"
-
-
-# ---------------------------------------------------------------------------
-# Reproducible script helpers
-# ---------------------------------------------------------------------------
-
-def _initial_repro_script_lines() -> List[str]:
-    """Return the initial lines for the reproducible workspace script."""
-
-    return [
-        "# -----------------------------------------------------------------------------",
-        "# Reproducible workspace script",
-        "# This script contains the Python code for the various actions you performed in",
-        "# PFD Toolkit Workbench.",
-        "# To replay them locally, first install the Toolkit before running the script:",
-        "#     pip install pfd_toolkit",
-        "# Workbench does not save your API key. Make sure to add this to the script."
-        ""
-        "from pfd_toolkit import load_reports",
-        "from pfd_toolkit import LLM",
-        "from pfd_toolkit import Screener",
-        "from pfd_toolkit import Extractor",
-        "",
-    ]
-
-
-def _ensure_repro_script(reset: bool = False) -> List[str]:
-    """Ensure the reproducible script list exists and optionally reset it."""
-
-    lines = st.session_state.get(REPRO_SCRIPT_KEY)
-    if reset or not isinstance(lines, list):
[…]
