+++ app/README.md
@@ -1,3 +1,0 @@
-# PFD Toolkit - Streamlit App
-
-This directory contains the source code for PFD Toolkit Workbench
