+++ app/pyproject.toml
@@ -1,25 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "pfd-toolkit-workbench"
-version = "0.1.2"
-requires-python = ">=3.10"
-description = "Streamlit workbench for PFD Toolkit."
-dependencies = [
-  "pfd_toolkit==0.4.1",
-  "streamlit==1.41.1"
-]
-
-[tool.hatch.metadata]
-allow-direct-references = true
-
-
-# Hatch assumes Python package; below limit the wheel to only the app files
-[tool.hatch.build.targets.wheel]
-only-include = [
-  "streamlit_app.py",
-  ".streamlit/config.toml",
-  "assets"
-]
