### 2026-03-05 1594c45
Adjust Django version constraints
