+++ pyproject.toml
@@ -29,7 +29,7 @@ dependencies = [
     "backoff>=2.2.1",
     "beautifulsoup4>=4.13.4",
     "dj-database-url>=3.1.0",
-    "django>=5.2.11",
+    "django>=5.2.12,<6.0",
     "dotenv>=0.9.9",
     "gunicorn>=25.0.3",
     "httpx>=0.28.1",
