### 2026-03-05 0406351
Stop reports from being loaded on initial state
