### 2026-03-05 860fba8
Fix bug where loading screen for discover themes was terminating too early
