### 2026-03-05 c31cb6f
Run discover themes pipeline on report corpus and tie changes to load_reports()
