+++ scripts/update_theme_collections.py
@@ -1,0 +1,205 @@
+#!/usr/bin/env python3
+"""Apply approved theme collections to the report dataset."""
+
+from __future__ import annotations
+
+import argparse
+import json
+import os
+from pathlib import Path
+from typing import Any
+
+import pandas as pd
+from pydantic import Field, create_model
+
+from pfd_toolkit import Extractor, LLM
+
+
+def _load_openai_key() -> str:
+    env_value = (os.getenv("OPENAI_API_KEY") or os.getenv("OPEN_API_KEY") or "").strip()
+    if env_value:
+        return env_value
+
+    project_root = Path(__file__).resolve().parents[1]
+    candidate_paths = [
+        project_root / "api.env",
+        project_root / "src" / "pfd_toolkit" / "api.env",
+    ]
+    for path in candidate_paths:
+        if not path.exists():
+            continue
+        raw = path.read_text(encoding="utf-8")
+        for line in raw.splitlines():
+            token = line.strip()
+            if not token or token.startswith("#"):
+                continue
+            if "=" not in token:
+                continue
+            left, right = token.split("=", 1)
+            key_name = left.strip()
+            if key_name not in {"OPENAI_API_KEY", "OPEN_API_KEY"}:
+                continue
+            value = right.strip().strip('"').strip("'")
+            if value:
+                return value
+
+    raise ValueError("OPENAI_API_KEY is required.")
+
+
+def _parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description="Apply curated themes to all reports using the LLM extractor."
+    )
+    parser.add_argument(
+        "--input",
+        type=Path,
+        default=Path("scripts/data/all_reports_with_collections.csv"),
+    )
+    parser.add_argument("--schema", type=Path, default=Path("scripts/theme_collections/approved_themes.json"))
+    parser.add_argument("--model", default="gpt-4.1")
+    parser.add_argument("--max-workers", type=int, default=18)
+    parser.add_argument(
+        "--recompute-all",
+        action="store_true",
+        help="Recompute theme columns for every row, not only rows with missing values.",
+    )
+    parser.add_argument(
+        "--summary-file",
+        type=Path,
+        default=Path(".github/workflows/update_summary.txt"),
+    )
+    return parser.parse_args()
+
+
+def _load_theme_schema(path: Path) -> dict[str, str]:
+    if not path.exists():
+        raise FileNotFoundError(f"Theme schema file not found: {path}")
+    payload = json.loads(path.read_text(encoding="utf-8"))
+    themes = payload.get("themes")
+    if not isinstance(themes, dict) or not themes:
+        raise ValueError("Theme schema JSON must include a non-empty 'themes' object.")
+    cleaned: dict[str, str] = {}
+    for key, value in themes.items():
+        name = str(key or "").strip()
+        if not name:
+            continue
+        cleaned[name] = str(value or "").strip()
+    if not cleaned:
+        raise ValueError("No valid themes found in schema file.")
+    return cleaned
+
+
+def _build_theme_model(themes: dict[str, str]):
+    fields = {
+        name: (bool, Field(description=description))
+        for name, description in themes.items()
+    }
+    return create_model("ApprovedThemeCollections", **fields)
+
+
+def _coerce_boolean_series(series: pd.Series) -> pd.Series:
+    def _coerce(value: Any) -> bool:
+        if isinstance(value, bool):
+            return value
+        if value is None:
+            return False
+        if isinstance(value, (int, float)):
+            return bool(value)
+        token = str(value).strip().lower()
+        if token in {"true", "1", "yes", "y", "t"}:
+            return True
[…]
