+++ .github/workflows/update_summary.txt
@@ -1,0 +1,2 @@
+
+Theme collections updated for 6211 row(s) using 24 approved themes.

[…]
