### 2026-03-05 95e0691
Add new themes to collections browse in app, and update Git workflow to tag newly scraped reports with themes
