+++ tests/test_loader.py
@@ -23,8 +23,9 @@ def use_local_dataset(tmp_path, monkeypatch):
                 "2024-12-30",
                 "2025-01-01",
             ],
-            "theme_patient_safety": [True, False, True, False],
-            "theme_mental_health": [False, True, True, False],
+            "theme_medication_safety": [True, False, True, False],
+            "theme_suicide_risk": [False, True, True, False],
+            "theme_care_home_safety": [False, True, False, True],
             "theme_sent_to_nhs_bodies": [True, False, True, False],
             "theme_sent_to_government_departments": [False, True, False, False],
             "theme_sent_to_prisons": [False, False, True, False],
@@ -92,27 +93,32 @@ def test_load_reports_reuses_process_cache(monkeypatch):
     assert call_count["value"] == 1
 
 
-def test_load_reports_filters_single_theme():
-    df = loader.load_reports(theme="patient_safety", refresh=False)
+def test_load_reports_filters_single_thematic_collection():
+    df = loader.load_reports(collection="medication_safety", refresh=False)
     assert set(df["url"]) == {"u1", "u3"}
 
 
-def test_load_reports_filters_multiple_themes_with_or_semantics():
+def test_load_reports_filters_multiple_thematic_collections_with_or_semantics():
     df = loader.load_reports(
-        theme=["patient_safety", "mental_health"],
+        collection=["medication_safety", "suicide"],
         refresh=False,
     )
     assert set(df["url"]) == {"u1", "u2", "u3"}
 
 
-def test_load_reports_filters_theme_with_prefix():
-    df = loader.load_reports(theme="theme_patient_safety", refresh=False)
+def test_load_reports_filters_collection_with_theme_prefix():
+    df = loader.load_reports(collection="theme_medication_safety", refresh=False)
     assert set(df["url"]) == {"u1", "u3"}
 
 
-def test_load_reports_invalid_theme_raises():
-    with pytest.raises(ValueError, match="Unknown theme"):
-        loader.load_reports(theme="not_a_real_theme", refresh=False)
+def test_load_reports_supports_collection_aliases():
+    df = loader.load_reports(collection="care_home", refresh=False)
+    assert set(df["url"]) == {"u2", "u4"}
+
+
+def test_load_reports_invalid_collection_still_raises():
+    with pytest.raises(ValueError, match="Unknown collection"):
+        loader.load_reports(collection="not_a_real_collection", refresh=False)
 
 
 def test_load_reports_filters_single_collection():
@@ -123,14 +129,14 @@ def test_load_reports_filters_single_collection():
 
 def test_load_reports_filters_multiple_collections_with_or_semantics():
     df = loader.load_reports(
-        collection=["gov_department", "local_gov"],
+        collection=["gov_department", "local_gov", "health_regulators"],
         refresh=False,
     )
-    assert set(df["url"]) == {"u2", "u4"}
+    assert set(df["url"]) == {"u2", "u3", "u4"}
     assert "theme_sent_to_government_departments" in df.columns
     assert "theme_sent_to_local_government" in df.columns
 
 
-def test_load_reports_invalid_collection_raises():
-    with pytest.raises(ValueError, match="Unknown collection"):
-        loader.load_reports(collection="not_a_real_collection", refresh=False)
+def test_load_reports_receiver_collection_alias_health_reg():
+    df = loader.load_reports(collection="health_reg", refresh=False)
+    assert set(df["url"]) == {"u3"}
