+++ django_workbench/workbench/views.py
@@ -178,7 +178,7 @@ BROWSE_COLLECTIONS: tuple[dict[str, str], ...] = (
         "accent": "rose",
     },
     {
-        "slug": "health_reg",
+        "slug": "health_regulators",
         "title": "Health Regulators",
         "description": (
             "Reports sent to national health regulators and oversight bodies including "
@@ -201,9 +201,114 @@ BROWSE_COLLECTIONS: tuple[dict[str, str], ...] = (
     },
 )
 
+THEME_COLLECTION_SLUG_PREFIX = "theme"
+THEME_COLLECTION_NAME_OVERRIDES: dict[str, str] = {
+    "suicide_risk": "suicide",
+    "care_home_safety": "care_home",
+}
+THEME_COLLECTION_TITLE_OVERRIDES: dict[str, str] = {
+    "suicide_risk": "Suicide",
+    "care_home_safety": "Care home deaths",
+}
+THEME_COLLECTION_DESCRIPTIONS: dict[str, str] = {
+    "access_to_care": "Delays, referral barriers, and service gaps that block timely care.",
+    "ambulance_response": "Dispatch, triage, and handover failures in ambulance pathways.",
+    "care_home_safety": "Preventable harm risks in care homes, including supervision and safeguarding gaps.",
+    "discharge_planning": "Unsafe discharges, weak follow-up, and poor handover into community care.",
+    "environmental_safety": "Physical environment hazards in care and public settings.",
+    "equipment_safety": "Safety failures involving medical devices, alarms, beds, or monitoring equipment.",
+    "falls_prevention": "Missed falls risk assessment, prevention measures, or care plan review.",
+    "family_involvement": "Insufficient communication with or involvement of families and carers.",
+    "hospital_care": "Inpatient care failures in monitoring, escalation, treatment, or handover.",
+    "infection_control": "Breakdowns in infection prevention, recognition, and outbreak control.",
+    "interagency_communication": "Cross-service information-sharing and coordination failures.",
+    "medication_safety": "Prescribing, dispensing, monitoring, or medicines communication errors.",
+    "mental_health_care": "Gaps in access, continuity, or coordination of mental health support.",
+    "observation_failures": "Inadequate checks, supervision, or patient monitoring processes.",
+    "online_hazards": "Harms linked to unregulated online access to dangerous products or information.",
+    "physical_health_in_mental_health": "Physical health needs missed or poorly managed in mental health settings.",
+    "record_keeping": "Incomplete, inaccurate, or inaccessible documentation and records.",
+    "road_safety": "Road design, signage, lighting, or maintenance risks contributing to fatalities.",
+    "safeguarding": "Missed safeguarding action for vulnerable adults and children.",
+    "staff_shortages": "Safety risks from understaffing, workload pressure, or workforce instability.",
+    "staff_training": "Insufficient staff training, induction, or competency refreshers.",
+    "substance_misuse": "Risk management failures around substance misuse and related prescribing.",
+    "suicide_risk": "Missed identification or management of suicide and self-harm risk.",
+    "vulnerable_groups": "Inadequate support for people with elevated vulnerability or complex needs.",
+}
 
-def _browse_collection_meta(collection_slug: str) -> dict[str, str]:
-    for collection in BROWSE_COLLECTIONS:
+
+def _theme_collection_slug(collection_name: str) -> str:
+    return f"{THEME_COLLECTION_SLUG_PREFIX}-{slugify(str(collection_name or '')).replace('-', '_')}"
+
+
+def _theme_collection_description(theme_name: str, pretty_title: str) -> str:
+    key = str(theme_name or "").strip().lower()
[…]
