### 2026-03-05 8159855
Update SVG generation for collections
