+++ django_workbench/workbench/views.py
@@ -206,6 +206,33 @@ THEME_COLLECTION_NAME_OVERRIDES: dict[str, str] = {
     "suicide_risk": "suicide",
     "care_home_safety": "care_home",
 }
+THEME_COLLECTION_SVG_MAP: dict[str, str] = {
+    "access_to_care": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><circle cx="18" cy="36" r="6"></circle><circle cx="36" cy="20" r="6"></circle><circle cx="54" cy="36" r="6"></circle><path d="M24 33L30 24"></path><path d="M42 24L48 33"></path><path d="M24 39L48 39"></path></svg>',
+    "ambulance_response": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><rect x="12" y="24" width="48" height="24" rx="6"></rect><path d="M20 36H30"></path><path d="M25 31V41"></path><path d="M42 36H52"></path><circle cx="24" cy="52" r="3"></circle><circle cx="48" cy="52" r="3"></circle></svg>',
+    "care_home": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M14 33L36 16L58 33"></path><path d="M20 31V57H52V31"></path><rect x="30" y="41" width="12" height="16" rx="2"></rect><path d="M25 45H20"></path><path d="M52 45H47"></path></svg>',
+    "discharge_planning": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><rect x="16" y="14" width="30" height="44" rx="4"></rect><path d="M24 26H38"></path><path d="M24 34H38"></path><path d="M24 42H34"></path><path d="M46 36H56"></path><path d="M52 30L58 36L52 42"></path></svg>',
+    "environmental_safety": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M36 14L58 54H14L36 14Z"></path><path d="M36 30V41"></path><circle cx="36" cy="48" r="1.8"></circle></svg>',
+    "equipment_safety": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><rect x="14" y="20" width="26" height="32" rx="4"></rect><rect x="44" y="26" width="14" height="20" rx="3"></rect><path d="M23 36L29 42L36 31"></path></svg>',
+    "falls_prevention": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M20 54H54"></path><path d="M24 48L34 30L44 40L54 24"></path><circle cx="24" cy="48" r="3"></circle><circle cx="34" cy="30" r="3"></circle><circle cx="44" cy="40" r="3"></circle><circle cx="54" cy="24" r="3"></circle></svg>',
+    "family_involvement": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><circle cx="24" cy="26" r="6"></circle><circle cx="48" cy="26" r="6"></circle><path d="M16 46C16 40 20 36 24 36C28 36 32 40 32 46"></path><path d="M40 46C40 40 44 36 48 36C52 36 56 40 56 46"></path><path d="M32 46H40"></path></svg>',
[…]
