### 2026-03-05 2b1f9e5
Add network tab to app
