+++ django_workbench/workbench/views.py
@@ -1,7 +1,7 @@
 """Views for the standalone Django Workbench."""
 from __future__ import annotations
 
-from collections import OrderedDict
+from collections import Counter, OrderedDict
 import logging
 import json
 import mimetypes
@@ -18,7 +18,7 @@ from io import BytesIO, StringIO
 from pathlib import Path
 from time import perf_counter
 from urllib.parse import parse_qs, urlencode, urljoin, urlparse
-from typing import Any, Optional
+from typing import Any, Callable, Optional
 from uuid import UUID, uuid4
 
 import markdown
@@ -90,6 +90,13 @@ SEO_PAGE_METADATA: dict[str, dict[str, str]] = {
         "title": "Explore PFD Reports | PFD Toolkit",
         "description": (
             "Explore and filter Prevention of Future Deaths reports with interactive dashboards and AI-assisted screening."
+        ),
+        "robots": "index,follow",
+    },
+    "network": {
+        "title": "Network Analysis for PFD Reports | PFD Toolkit",
+        "description": (
+            "Map relationships between report themes, receivers, coroners, and areas with an interactive network graph."
         ),
         "robots": "index,follow",
     },
@@ -318,6 +325,18 @@ def _theme_collection_map_from_reports(reports_df: pd.DataFrame) -> dict[str, di
     return collections
 
 
+def _display_dataset_df(df: pd.DataFrame) -> pd.DataFrame:
+    """Return dataset columns intended for table display (hide internal theme flags)."""
+    if df.empty:
+        return df.copy()
+    visible_columns = [
+        column
+        for column in df.columns
+        if not str(column).startswith("theme_") and not str(column).startswith("collection_")
+    ]
+    return df.loc[:, visible_columns].copy()
+
+
 def _all_browse_collections(reports_df: pd.DataFrame) -> list[dict[str, str]]:
     combined = [dict(collection) for collection in BROWSE_COLLECTIONS]
     combined.extend(_theme_collection_map_from_reports(reports_df).values())
@@ -359,10 +378,7 @@ def _load_browse_reports_df() -> pd.DataFrame:
 def _reports_for_browse_collection(reports_df: pd.DataFrame, collection_slug: str) -> pd.DataFrame:
     reports_with_collections = _ensure_collection_columns_in_reports(reports_df)
     if collection_slug == "custom":
-        visible_columns = [
-            column for column in reports_with_collections.columns if not str(column).startswith("theme_")
-        ]
-        return reports_with_collections.loc[:, visible_columns].reset_index(drop=True).copy()
+        return reports_with_collections.reset_index(drop=True).copy()
 
     if collection_slug in COLLECTION_COLUMNS:
         filtered_reports = load_reports(refresh=False, collection=collection_slug)
@@ -377,10 +393,7 @@ def _reports_for_browse_collection(reports_df: pd.DataFrame, collection_slug: st
 
     filtered_reports, _ = _ensure_workbench_row_ids(filtered_reports)
     filtered_reports = _ensure_collection_columns_in_reports(filtered_reports)
-    visible_columns = [
-        column for column in filtered_reports.columns if not str(column).startswith("theme_")
-    ]
-    return filtered_reports.loc[:, visible_columns].copy()
+    return filtered_reports.copy()
 
 UI_THEME_CHOICES: tuple[dict[str, str], ...] = (
     {
@@ -410,6 +423,25 @@ def favicon(request: HttpRequest) -> HttpResponse:
[…]

[…]
