### 2026-03-05 d30dadf
Add full screen and tweak filters
