+++ django_workbench/workbench/static/workbench/workbench.js
@@ -2818,6 +2818,7 @@
         root.dataset.bound = "1";
 
         const dataNode = byId("network-graph-data");
+        const graphCard = byId("network-graph-card");
         const graphRoot = byId("network-graph-canvas");
         const statusNode = byId("network-status");
         const searchQueryInput = byId("network-search-query");
@@ -2830,6 +2831,9 @@
         const minEdgeInput = byId("network-min-edge");
         const minEdgeValue = byId("network-min-edge-value");
         const typeInputs = Array.from(root.querySelectorAll("[data-network-type]"));
+        const fullscreenEnterButton = byId("network-graph-fullscreen-enter");
+        const fullscreenExitButton = byId("network-graph-fullscreen-exit");
+        const applySelectionButton = byId("network-apply-selection");
         const refreshButton = byId("network-refresh-data");
         const resetButton = byId("network-reset-controls");
         const fitButton = byId("network-fit-graph");
@@ -2852,6 +2856,7 @@
 
         if (
             !dataNode ||
+            !graphCard ||
             !graphRoot ||
             !statusNode ||
             !searchQueryInput ||
@@ -2864,6 +2869,9 @@
             !minEdgeInput ||
             !minEdgeValue ||
             !typeInputs.length ||
+            !fullscreenEnterButton ||
+            !fullscreenExitButton ||
+            !applySelectionButton ||
             !refreshButton ||
             !resetButton ||
             !fitButton ||
@@ -2889,6 +2897,40 @@
 
         function setStatus(text) {
             statusNode.textContent = text;
+        }
+
+        function fullscreenElement() {
+            return document.fullscreenElement || document.webkitFullscreenElement || null;
+        }
+
+        function isGraphFullscreen() {
+            return fullscreenElement() === graphCard;
+        }
+
+        function syncFullscreenButtons() {
+            const active = isGraphFullscreen();
+            fullscreenEnterButton.classList.toggle("hidden", active);
+            fullscreenExitButton.classList.toggle("hidden", !active);
+        }
+
+        async function enterGraphFullscreen() {
+            if (graphCard.requestFullscreen) {
+                await graphCard.requestFullscreen();
+                return;
+            }
+            if (graphCard.webkitRequestFullscreen) {
+                graphCard.webkitRequestFullscreen();
+            }
+        }
+
+        async function exitGraphFullscreen() {
+            if (document.exitFullscreen) {
+                await document.exitFullscreen();
+                return;
+            }
+            if (document.webkitExitFullscreen) {
+                document.webkitExitFullscreen();
+            }
         }
 
         function escapeHtml(value) {
@@ -2922,6 +2964,8 @@
         let sourceNodes = [];
         let sourceEdges = [];
         let sourceNodesById = new Map();
+        let sourceSearchNodes = [];
+        let sourceSearchNodesById = new Map();
         let currentNodesById = new Map();
         let currentEdges = [];
         let searchDropdownIntentOpen = false;
@@ -2939,6 +2983,7 @@
         };
         const initialParams = new URLSearchParams(window.location.search);
         let selectedSearchNodeId = (initialParams.get("network_focus") || "").trim();
+        let selectedGraphNodeId = (initialParams.get("network_selected_node") || "").trim();
 
         function typePalette() {
             return {
@@ -3008,12 +3053,22 @@
                 .map(normalisePayloadEdge)
                 .filter(Boolean);
             sourceNodesById = new Map(sourceNodes.map((node) => [node.id, node]));
+            const searchIndex = payload.search_index && typeof payload.search_index === "object"
+                ? payload.search_index
[…]
