+++ django_workbench/workbench/views.py
@@ -431,6 +431,7 @@ NETWORK_EDGE_MAX = 2600
 NETWORK_SCOPE_QUERY_KEYS: tuple[str, ...] = (
     "network_search",
     "network_focus",
+    "network_selected_node",
     "network_min_edge",
     "network_node_limit",
     "network_type",
@@ -940,6 +941,18 @@ def _network_node_id(node_type: str, value: str) -> str:
     return f"{node_type}::{value}"
 
 
+def _network_parse_node_id(node_id: str) -> Optional[tuple[str, str]]:
+    cleaned = str(node_id or "").strip()
+    if "::" not in cleaned:
+        return None
+    node_type, raw_value = cleaned.split("::", 1)
+    parsed_type = str(node_type).strip().lower()
+    parsed_value = str(raw_value).strip()
+    if parsed_type not in NETWORK_NODE_TYPES or not parsed_value:
+        return None
+    return parsed_type, parsed_value
+
+
 def _network_theme_columns(reports_df: pd.DataFrame) -> list[str]:
     receiver_collection_columns = set(COLLECTION_COLUMNS.values())
     thematic_columns: list[str] = []
@@ -966,10 +979,31 @@ def _network_scale_node_size(
     return round(size_floor + ((size_ceiling - size_floor) * eased_ratio), 2)
 
 
+def _network_search_nodes_from_counter(
+    node_type: str,
+    counter: Counter[str],
+    *,
+    label_fn: Optional[Callable[[str], str]] = None,
+) -> list[dict[str, Any]]:
+    labels = label_fn or (lambda value: value)
+    rows = sorted(counter.items(), key=lambda item: (-int(item[1]), str(item[0]).casefold()))
+    return [
+        {
+            "id": _network_node_id(node_type, str(value)),
+            "name": str(labels(str(value))),
+            "type": node_type,
+            "raw_key": str(value),
+            "value": int(count),
+        }
+        for value, count in rows
+    ]
+
+
 def _build_network_graph_payload(
     reports_df: pd.DataFrame,
     *,
     selected_filters: Optional[dict[str, list[str]]] = None,
+    forced_node_ids: Optional[list[str]] = None,
 ) -> dict[str, Any]:
     filters = selected_filters or {"coroner": [], "area": [], "receiver": []}
     normalised_filters = {
@@ -980,6 +1014,18 @@ def _build_network_graph_payload(
     filtered_df = _apply_explore_dashboard_filters(reports_df, normalised_filters)
     reports_total = int(len(reports_df))
     reports_shown = int(len(filtered_df))
+    forced_nodes: list[tuple[str, str]] = []
+    seen_forced: set[str] = set()
+    for raw_id in forced_node_ids or []:
+        parsed = _network_parse_node_id(raw_id)
+        if not parsed:
+            continue
+        node_type, node_value = parsed
+        key = f"{node_type}::{node_value.casefold()}"
+        if key in seen_forced:
+            continue
+        seen_forced.add(key)
+        forced_nodes.append((node_type, node_value))
 
     theme_columns = _network_theme_columns(filtered_df)
     if filtered_df.empty:
@@ -995,6 +1041,7 @@ def _build_network_graph_payload(
                 "top_theme": {"name": "-", "count": 0},
             },
             "graph": {"nodes": [], "edges": []},
+            "search_index": {"nodes": []},
             "options": {
                 "max_edge_weight": 1,
                 "default_min_edge_weight": 1,
@@ -1054,6 +1101,24 @@ def _build_network_graph_payload(
         selected_receivers = set(top_receivers)
         selected_coroners = set(top_coroners)
         selected_areas = set(top_areas)
+        forced_ids = {
+            _network_node_id(node_type, node_value)
[…]
