### 2026-03-05 77935e9
Remove max nodes slider
