### 2026-03-06 f5adb8f
Cleaner layout for network diagram
