+++ django_workbench/workbench/static/workbench/workbench.js
@@ -2829,6 +2829,7 @@
         const minEdgeInput = byId("network-min-edge");
         const minEdgeValue = byId("network-min-edge-value");
         const typeInputs = Array.from(root.querySelectorAll("[data-network-type]"));
+        const normalizeThemeCooccurrenceInput = byId("network-normalise-theme-cooccurrence");
         const fullscreenEnterButton = byId("network-graph-fullscreen-enter");
         const fullscreenExitButton = byId("network-graph-fullscreen-exit");
         const applySelectionButton = byId("network-apply-selection");
@@ -2867,6 +2868,7 @@
             !minEdgeInput ||
             !minEdgeValue ||
             !typeInputs.length ||
+            !normalizeThemeCooccurrenceInput ||
             !fullscreenEnterButton ||
             !fullscreenExitButton ||
             !applySelectionButton ||
@@ -2985,6 +2987,7 @@
         const initialParams = new URLSearchParams(window.location.search);
         let selectedSearchNodeId = (initialParams.get("network_focus") || "").trim();
         let selectedGraphNodeId = (initialParams.get("network_selected_node") || "").trim();
+        const initialNormaliseThemeCooccurrence = (initialParams.get("network_normalise_theme_cooccurrence") || "").trim().toLowerCase();
 
         function typePalette() {
             return {
@@ -3000,6 +3003,7 @@
                 theme_receiver: getCssVar("--network-edge-theme-receiver", "rgba(121, 210, 176, 0.58)"),
                 theme_coroner: getCssVar("--network-edge-theme-coroner", "rgba(255, 178, 126, 0.56)"),
                 theme_area: getCssVar("--network-edge-theme-area", "rgba(184, 162, 255, 0.58)"),
+                theme_theme: getCssVar("--network-edge-theme-theme", "rgba(137, 171, 255, 0.66)"),
                 fallback: getCssVar("--network-edge-fallback", "rgba(164, 185, 255, 0.48)"),
             };
         }
@@ -3035,10 +3039,17 @@
             const target = typeof edge.target === "string" ? edge.target : "";
             const value = Number(edge.value || 0);
             const kind = typeof edge.kind === "string" ? edge.kind : "";
+            const normalizedValue = Number(edge.normalized_value || 0);
             if (!source || !target || !Number.isFinite(value) || value <= 0) {
                 return null;
             }
-            return { source: source, target: target, value: value, kind: kind };
+            return {
+                source: source,
+                target: target,
+                value: value,
+                kind: kind,
+                normalizedValue: Number.isFinite(normalizedValue) && normalizedValue > 0 ? normalizedValue : 0,
+            };
         }
 
         function applyPayload(newPayload, preserveValues) {
@@ -3075,7 +3086,7 @@
             const maxEdgeWeight = Math.max(1, Number(options.max_edge_weight || 1));
             const defaultMinEdgeWeight = Math.max(
                 1,
-                Math.min(maxEdgeWeight, Number(options.default_min_edge_weight || 1))
+                Math.min(maxEdgeWeight, 20)
             );
             const maxNodeLimit = Math.max(24, Number(options.max_node_limit || sourceNodes.length || 120));
             const defaultNodeLimit = Math.max(
@@ -3090,6 +3101,29 @@
             if (!preserveValues) {
                 minEdgeInput.value = String(defaultMinEdgeWeight);
                 currentNodeLimit = defaultNodeLimit;
+
+                const defaultTypes = Array.isArray(options.default_types) ? options.default_types : [];
+                const defaultTypeSet = new Set(
+                    defaultTypes
+                        .map((value) => String(value || "").trim())
+                        .filter((value) => categoryOrder.includes(value))
+                );
[…]
