+++ scripts/update_theme_collections.py
@@ -6,8 +6,10 @@ from __future__ import annotations
 import argparse
 import json
 import os
+import hashlib
 from pathlib import Path
 from typing import Any
+import math
 
 import pandas as pd
 from pydantic import Field, create_model
@@ -59,6 +61,12 @@ def _parse_args() -> argparse.Namespace:
     parser.add_argument("--model", default="gpt-4.1")
     parser.add_argument("--max-workers", type=int, default=18)
     parser.add_argument(
+        "--theme-batches",
+        type=int,
+        default=3,
+        help="Number of theme batches to process sequentially. Defaults to 3.",
+    )
+    parser.add_argument(
         "--recompute-all",
         action="store_true",
         help="Recompute theme columns for every row, not only rows with missing values.",
@@ -67,6 +75,23 @@ def _parse_args() -> argparse.Namespace:
         "--summary-file",
         type=Path,
         default=Path(".github/workflows/update_summary.txt"),
+    )
+    parser.add_argument(
+        "--checkpoint-file",
+        type=Path,
+        default=Path("scripts/theme_collections/update_theme_collections_checkpoint.csv"),
+        help="CSV file used for incremental checkpointing and resume.",
+    )
+    parser.add_argument(
+        "--checkpoint-metadata-file",
+        type=Path,
+        default=Path("scripts/theme_collections/update_theme_collections_checkpoint.meta.json"),
+        help="Metadata file used to validate checkpoint compatibility on resume.",
+    )
+    parser.add_argument(
+        "--no-resume",
+        action="store_true",
+        help="Ignore any existing checkpoint and recompute from scratch.",
     )
     return parser.parse_args()
 
@@ -95,6 +120,20 @@ def _build_theme_model(themes: dict[str, str]):
         for name, description in themes.items()
     }
     return create_model("ApprovedThemeCollections", **fields)
+
+
+def _chunk_theme_map(themes: dict[str, str], batches: int) -> list[dict[str, str]]:
+    if batches <= 0:
+        raise ValueError("--theme-batches must be at least 1.")
+    items = list(themes.items())
+    if not items:
+        return []
+    batch_size = max(1, math.ceil(len(items) / batches))
+    chunks: list[dict[str, str]] = []
+    for start in range(0, len(items), batch_size):
+        chunk_items = items[start : start + batch_size]
+        chunks.append(dict(chunk_items))
+    return chunks
 
 
 def _coerce_boolean_series(series: pd.Series) -> pd.Series:
@@ -136,6 +175,74 @@ def _append_summary(summary_file: Path, message: str) -> None:
             handle.write("\n")
 
 
+def _metadata_signature(
+    input_path: Path,
+    schema_path: Path,
+    theme_columns: list[str],
+    source_indices: pd.Index,
+) -> dict[str, Any]:
+    digest = hashlib.sha256()
+    digest.update(str(input_path.resolve()).encode("utf-8"))
+    digest.update(str(schema_path.resolve()).encode("utf-8"))
+    digest.update("|".join(theme_columns).encode("utf-8"))
+    digest.update(",".join(str(int(idx)) for idx in source_indices.tolist()).encode("utf-8"))
+    return {
+        "input_path": str(input_path.resolve()),
+        "schema_path": str(schema_path.resolve()),
+        "theme_columns": theme_columns,
+        "source_index_hash": digest.hexdigest(),
+        "row_count": int(len(source_indices)),
+    }
+
+
+def _save_checkpoint(
+    checkpoint_file: Path,
+    metadata_file: Path,
+    source_indices: pd.Index,
+    assigned_subset: pd.DataFrame,
+    theme_columns: list[str],
+    metadata: dict[str, Any],
+) -> None:
+    checkpoint_file.parent.mkdir(parents=True, exist_ok=True)
[…]
