+++ .github/workflows/update_summary.txt
@@ -1,2 +1,4 @@
 
 Theme collections updated for 6211 row(s) using 24 approved themes.
+
+Theme collections updated for 6211 row(s) using 90 approved themes across 3 batch(es).
