+++ scripts/update_collections.py
@@ -1,46 +1,488 @@
 #!/usr/bin/env python3
-"""Update rule-based collection columns in the current all_reports.csv dataset."""
+"""Update rule-based and LLM-based collection columns in the report dataset."""
 
 from __future__ import annotations
 
+import argparse
+import hashlib
+import json
+import math
+import os
 from pathlib import Path
+from typing import Any
 
 import pandas as pd
+from pydantic import Field, create_model
 
+from pfd_toolkit import Extractor, LLM
 from pfd_toolkit.collections import COLLECTION_COLUMNS, apply_collection_columns
 
 
-DATA_PATH = Path("all_reports.csv")
+def _load_openai_key() -> str:
+    env_value = (os.getenv("OPENAI_API_KEY") or os.getenv("OPEN_API_KEY") or "").strip()
+    if env_value:
+        return env_value
+
+    project_root = Path(__file__).resolve().parents[1]
+    candidate_paths = [
+        project_root / "api.env",
+        project_root / "src" / "pfd_toolkit" / "api.env",
+    ]
+    for path in candidate_paths:
+        if not path.exists():
+            continue
+        raw = path.read_text(encoding="utf-8")
+        for line in raw.splitlines():
+            token = line.strip()
+            if not token or token.startswith("#"):
+                continue
+            if "=" not in token:
+                continue
+            left, right = token.split("=", 1)
+            key_name = left.strip()
+            if key_name not in {"OPENAI_API_KEY", "OPEN_API_KEY"}:
+                continue
+            value = right.strip().strip('"').strip("'")
+            if value:
+                return value
+
+    raise ValueError("OPENAI_API_KEY is required.")
+
+
+def _parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description="Update regex/rule collections and approved LLM theme collections.",
+    )
+    parser.add_argument("--input", type=Path, default=Path("all_reports.csv"))
+    parser.add_argument(
+        "--schema",
+        type=Path,
+        default=Path("scripts/theme_collections/approved_themes.json"),
+    )
+    parser.add_argument("--model", default="gpt-4.1")
+    parser.add_argument("--max-workers", type=int, default=18)
+    parser.add_argument(
+        "--theme-batches",
+        type=int,
+        default=3,
+        help="Number of LLM theme batches to process sequentially.",
+    )
+    parser.add_argument(
+        "--recompute-all",
+        action="store_true",
+        help="Recompute both regex and LLM collection columns for all rows.",
+    )
+    parser.add_argument(
+        "--recompute-all-rules",
+        action="store_true",
+        help="Recompute regex/rule-based collection columns for all rows.",
+    )
+    parser.add_argument(
+        "--recompute-all-llm",
+        action="store_true",
+        help="Recompute approved LLM theme columns for all rows.",
+    )
+    parser.add_argument(
+        "--skip-rule-based",
+        action="store_true",
+        help="Skip regex/rule-based collection updates.",
+    )
+    parser.add_argument(
+        "--skip-llm",
+        action="store_true",
+        help="Skip LLM-based collection updates.",
+    )
+    parser.add_argument(
+        "--summary-file",
+        type=Path,
+        default=Path(".github/workflows/update_summary.txt"),
+    )
+    parser.add_argument(
+        "--checkpoint-file",
+        type=Path,
+        default=Path("scripts/theme_collections/update_theme_collections_checkpoint.csv"),
+        help="CSV file used for incremental checkpointing and resume.",
+    )
+    parser.add_argument(
+        "--checkpoint-metadata-file",
+        type=Path,
[…]
