+++ .github/workflows/update_summary.txt
@@ -2,3 +2,5 @@
 Theme collections updated for 6211 row(s) using 24 approved themes.
 
 Theme collections updated for 6211 row(s) using 90 approved themes across 3 batch(es).
+
+Rule-based collection columns updated for 6211 row(s); LLM theme collections updated for 0 row(s).
