+++ .github/workflows/pfd_update.yml
@@ -49,7 +49,6 @@ jobs:
           set +x
           uv run python scripts/update_reports.py
           uv run python scripts/update_collections.py
-          uv run python scripts/update_theme_collections.py
           set -x
 
       - name: Upload dataset to release
@@ -92,6 +91,7 @@ jobs:
         uses: EndBug/add-and-commit@v9
         with:
           add: |
+            all_reports.csv
             docs/index.md
           default_author: github_actions
           message: "🤖 Auto PFD dataset top-up"

[…]
