### 2026-03-10 1494f93
Update loader with new collections
