+++ src/pfd_toolkit/loader.py
@@ -31,21 +31,15 @@ _THEME_COLLECTION_NAME_OVERRIDES: Final[dict[str, str]] = {
     "care_home_safety": "care_home",
 }
 _COLLECTION_ALIASES: Final[dict[str, str]] = {
+    "welsh": "wales",
     "health_reg": "health_regulators",
     "suicide_risk": "suicide",
     "care_home_safety": "care_home",
 }
 
 
-def _normalise_theme_name(raw_theme: str) -> str:
-    value = str(raw_theme or "").strip()
-    if value.startswith(_THEME_PREFIX):
-        return value[len(_THEME_PREFIX):]
-    return value
-
-
 def _normalise_collection_name(raw_value: str) -> str:
-    value = _normalise_theme_name(raw_value).strip()
+    value = str(raw_value or "").strip()
     return _COLLECTION_ALIASES.get(value, value)
 
 def _ensure_cached_dataset(
@@ -137,29 +131,144 @@ def load_reports(
     ----------
     start_date : str, optional
         Inclusive lower bound for the report date in ``YYYY-MM-DD`` format.
-        Defaults to ``"2000-01-01"``.
+        Default: ``"2000-01-01"``.
     end_date : str, optional
         Inclusive upper bound for the report date in ``YYYY-MM-DD`` format.
-        Defaults to ``"2050-01-01"``.
+        Default: ``"2050-01-01"``.
     n_reports : int or None, optional
         Keep only the most recent ``n_reports`` rows after filtering by date.
-        ``None`` (the default) returns all rows.
+        Default: ``None`` (all rows).
     refresh : bool, optional
-        If ``True`` (the default), force a fresh download of the dataset. Set to
-        ``False`` to reuse the previously cached copy.
+        If ``True``, force a fresh download of the dataset. If ``False``, reuse
+        the cached copy.
+        Default: ``True``.
     collection : str | list[str] | None, optional
-        Filter rows using packaged collections. This includes receiver-based
-        collections (e.g. ``"nhs"``, ``"gov_department"``,
-        ``"health_regulators"``) and thematic collections inferred from
-        packaged ``theme_*`` columns (e.g. ``"medication_safety"``,
-        ``"suicide"``, ``"care_home"``). Values may be supplied with or
-        without the ``theme_`` prefix for thematic collections. Pass a single
-        collection name or a list of collection names. When multiple
-        collections are supplied, rows matching **any** requested collection are
-        returned. When a single collection is supplied, its boolean helper
-        column is dropped from the returned DataFrame. When two or more
-        collections are supplied, the requested boolean columns are retained for
-        comparison.
+        One collection name or a list of collection names.
+        Collection matching uses OR semantics across provided names.
+        When one collection is provided, the helper boolean column is dropped from
+        output. When multiple collections are provided, helper columns are kept.
+        Supported collection names:
+
+        - wales
+        - nhs
+        - gov_department
+        - prisons
+        - health_regulators
+        - health_reg
+        - local_gov
+        - access_to_care
+        - ambulance_response
+        - care_home
+        - discharge_planning
+        - environmental_safety
+        - equipment_safety
+        - falls_prevention
+        - family_involvement
+        - hospital_care
+        - infection_control
+        - interagency_communication
+        - medication_safety
+        - mental_health_care
+        - observation_failures
+        - online_hazards
+        - physical_health_in_mental_health
+        - record_keeping
+        - road_safety
+        - safeguarding
+        - staff_shortages
+        - staff_training
+        - substance_misuse
+        - suicide
+        - vulnerable_groups
+        - emergency_departments
+        - ambulance_services
+        - primary_care
+        - out_of_hours_care
+        - acute_hospital_wards
+        - intensive_care
+        - surgical_care
+        - maternity_neonatal_perinatal_care
+        - mental_health_services
[…]
