### 2026-03-10 4280c98
Update network tab to use new collections
