+++ django_workbench/workbench/views.py
@@ -12,6 +12,7 @@ import copy
 import importlib
 import importlib.util
 import inspect
+from threading import Lock
 from html import escape
 from datetime import date, datetime
 from io import BytesIO, StringIO
@@ -142,73 +143,82 @@ SEO_PAGE_METADATA: dict[str, dict[str, str]] = {
     },
 }
 
-BROWSE_COLLECTIONS: tuple[dict[str, str], ...] = (
-    {
-        "slug": "custom",
-        "title": "Custom Collection",
+BROWSE_CUSTOM_COLLECTION: dict[str, str] = {
+    "slug": "custom",
+    "title": "Custom Collection",
+    "description": (
+        "Open the full report archive as a starting point for building your own "
+        "editable collection and dashboard workflow."
+    ),
+    "badge": "Browse starter",
+    "icon": "hgi-search-01",
+    "accent": "slate",
+}
+
+RULE_COLLECTION_CARD_OVERRIDES: dict[str, dict[str, str]] = {
+    "wales": {
+        "title": "Wales",
         "description": (
-            "Open the full report archive as a starting point for building your own "
-            "editable collection and dashboard workflow."
+            "Reports where the coroner area is a Welsh jurisdiction, including "
+            "legacy area names recoded to current Welsh areas."
         ),
-        "badge": "Browse starter",
-        "icon": "hgi-search-01",
-        "accent": "slate",
+        "badge": "Area collection",
+        "accent": "mint",
+        "icon_svg": '<svg viewBox="0 0 72 72" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M14 52L24 30L36 44L46 22L58 52H14Z"></path><path d="M20 24L28 18L34 26"></path><path d="M42 16L50 22L44 30"></path></svg>',
     },
-    {
-        "slug": "nhs",
+    "nhs": {
         "title": "NHS Bodies",
         "description": (
             "Reports sent to NHS organisations including trusts, foundation trusts, "
             "integrated care boards, and health boards."
         ),
         "badge": "Receiver collection",
-        "icon": "hgi-dashboard-square-01",
         "accent": "sky",
     },
-    {
-        "slug": "gov_department",
+    "gov_department": {
         "title": "Government Departments",
         "description": (
             "Reports addressed to UK government departments and central offices such as "
             "the Home Office, Cabinet Office, and ministerial departments."
         ),
         "badge": "Receiver collection",
-        "icon": "hgi-chart-bar-line",
         "accent": "amber",
     },
-    {
-        "slug": "prisons",
+    "prisons": {
         "title": "Prisons",
         "description": (
             "Reports sent directly to prisons, prison institutions, young offender "
             "settings, or HM Prison and Probation Service."
         ),
         "badge": "Receiver collection",
-        "icon": "hgi-filter-horizontal",
         "accent": "rose",
     },
-    {
-        "slug": "health_regulators",
+    "health_regulators": {
         "title": "Health Regulators",
         "description": (
             "Reports sent to national health regulators and oversight bodies including "
[…]
