+++ pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+addopts = --ignore=tests/test_streamlit_app.py
