### 2026-03-10 6329e8b
Introduce theme cap parameter in network view
