+++ django_workbench/workbench/static/workbench/workbench.js
@@ -2827,7 +2827,9 @@
         const searchClearButton = byId("network-search-clear");
         const layoutSelect = byId("network-layout");
         const minEdgeInput = byId("network-min-edge");
-        const minEdgeValue = byId("network-min-edge-value");
+        const minEdgeApplyButton = byId("network-min-edge-apply");
+        const themeCapInput = byId("network-theme-cap");
+        const themeCapApplyButton = byId("network-theme-cap-apply");
         const typeInputs = Array.from(root.querySelectorAll("[data-network-type]"));
         const normalizeThemeCooccurrenceInput = byId("network-normalise-theme-cooccurrence");
         const fullscreenEnterButton = byId("network-graph-fullscreen-enter");
@@ -2866,7 +2868,9 @@
             !searchClearButton ||
             !layoutSelect ||
             !minEdgeInput ||
-            !minEdgeValue ||
+            !minEdgeApplyButton ||
+            !themeCapInput ||
+            !themeCapApplyButton ||
             !typeInputs.length ||
             !normalizeThemeCooccurrenceInput ||
             !fullscreenEnterButton ||
@@ -2970,8 +2974,12 @@
         let sourceSearchNodesById = new Map();
         let currentNodesById = new Map();
         let currentEdges = [];
+        const DEFAULT_NETWORK_EDGE_THRESHOLD = 50;
+        const MIN_NETWORK_NODE_LIMIT = 50;
         let searchDropdownIntentOpen = false;
         let currentNodeLimit = 120;
+        let appliedMinEdgeWeight = DEFAULT_NETWORK_EDGE_THRESHOLD;
+        let appliedThemeCap = null;
         let currentCentrality = {
             degree: new Map(),
             betweenness: new Map(),
@@ -2985,9 +2993,71 @@
             area: "Area",
         };
         const initialParams = new URLSearchParams(window.location.search);
+        const initialMinEdgeParam = (initialParams.get("network_min_edge") || "").trim();
+        const initialThemeCapParam = (initialParams.get("network_theme_cap") || "").trim();
+        let initialMinEdgeApplied = false;
+        let initialThemeCapApplied = false;
         let selectedSearchNodeId = (initialParams.get("network_focus") || "").trim();
         let selectedGraphNodeId = (initialParams.get("network_selected_node") || "").trim();
         const initialNormaliseThemeCooccurrence = (initialParams.get("network_normalise_theme_cooccurrence") || "").trim().toLowerCase();
+
+        function normaliseMinEdgeThreshold(rawValue, fallback) {
+            const cleaned = String(rawValue || "").trim();
+            if (!cleaned) {
+                return fallback;
+            }
+            const parsed = Number(cleaned);
+            if (!Number.isFinite(parsed)) {
+                return fallback;
+            }
+            const rounded = Math.round(parsed);
+            if (rounded <= 0) {
+                return fallback;
+            }
+            return rounded;
+        }
+
+        function normaliseThemeCap(rawValue, fallback) {
+            const cleaned = String(rawValue || "").trim();
+            if (!cleaned) {
+                return null;
+            }
+            const parsed = Number(cleaned);
+            if (!Number.isFinite(parsed)) {
+                return fallback;
+            }
+            const rounded = Math.round(parsed);
+            if (rounded <= 0) {
+                return fallback;
+            }
+            return rounded;
+        }
+
+        function applyMinEdgeInputValue(options) {
+            const settings = options && typeof options === "object" ? options : {};
+            const fallback = Number.isFinite(settings.fallback)
+                ? Number(settings.fallback)
+                : appliedMinEdgeWeight;
[…]
