+++ django_workbench/workbench/views.py
@@ -637,10 +637,11 @@ def favicon(request: HttpRequest) -> HttpResponse:
 UI_THEME_IDS = {choice["id"] for choice in UI_THEME_CHOICES}
 DASHBOARD_TOP_N = 12
 DASHBOARD_UNKNOWN_LABEL = "Not specified"
-NETWORK_THEME_TOP_N = 36
 NETWORK_RECEIVER_TOP_N = 72
 NETWORK_CORONER_TOP_N = 44
 NETWORK_AREA_TOP_N = 36
+NETWORK_MIN_NODE_LIMIT = 50
+NETWORK_DEFAULT_EDGE_THRESHOLD = 50
 NETWORK_EDGE_MAX = 2600
 NETWORK_THEME_THEME_EDGE_MAX = 420
 NETWORK_THEME_THEME_MAX_PER_THEME = 8
@@ -649,6 +650,7 @@ NETWORK_SCOPE_QUERY_KEYS: tuple[str, ...] = (
     "network_focus",
     "network_selected_node",
     "network_min_edge",
+    "network_theme_cap",
     "network_node_limit",
     "network_type",
 )
@@ -1294,7 +1296,7 @@ def _build_network_graph_payload(
             "search_index": {"nodes": []},
             "options": {
                 "max_edge_weight": 1,
-                "default_min_edge_weight": 1,
+                "default_min_edge_weight": int(NETWORK_DEFAULT_EDGE_THRESHOLD),
                 "max_node_limit": 120,
                 "default_node_limit": 120,
                 "available_types": ["theme", "receiver", "coroner", "area"],
@@ -1505,7 +1507,7 @@ def _build_network_graph_payload(
             if str(edge["source"]) in node_ids and str(edge["target"]) in node_ids
         ]
         max_edge_weight = max((int(edge["value"]) for edge in edges_filtered), default=1)
-        default_node_limit = min(180, max(24, len(nodes)))
+        default_node_limit = min(180, max(NETWORK_MIN_NODE_LIMIT, len(nodes)))
         max_node_limit = max(default_node_limit, min(420, max(80, len(nodes))))
         search_nodes = []
         search_nodes.extend(_network_search_nodes_from_counter("receiver", receiver_counter))
@@ -1530,7 +1532,7 @@ def _build_network_graph_payload(
             "search_index": {"nodes": search_nodes},
             "options": {
                 "max_edge_weight": int(max_edge_weight),
-                "default_min_edge_weight": int(min_count_for_edge),
+                "default_min_edge_weight": int(NETWORK_DEFAULT_EDGE_THRESHOLD),
                 "max_node_limit": int(max_node_limit),
                 "default_node_limit": int(default_node_limit),
                 "available_types": ["receiver", "coroner", "area"],
@@ -1600,7 +1602,7 @@ def _build_network_graph_payload(
 
     top_themes = _network_top_items(
         theme_counter,
-        limit=NETWORK_THEME_TOP_N,
+        limit=max(1, len(theme_counter)),
         min_value=min_count_for_node,
     )
     top_receivers = _network_top_items(
@@ -1838,7 +1840,7 @@ def _build_network_graph_payload(
     ]
 
     max_edge_weight = max((int(edge["value"]) for edge in edges), default=1)
-    default_node_limit = min(180, max(24, len(nodes)))
+    default_node_limit = min(180, max(NETWORK_MIN_NODE_LIMIT, len(nodes)))
     max_node_limit = max(default_node_limit, min(420, max(80, len(nodes))))
     top_theme_key = top_themes[0] if top_themes else ""
     search_nodes = []
@@ -1870,7 +1872,7 @@ def _build_network_graph_payload(
         "search_index": {"nodes": search_nodes},
[…]
