### 2026-03-10 a3d70a6
Change loader to always use the current date as the max
