### 2026-03-10 dbd00ea
Improve collections naming & display
