+++ django_workbench/workbench/views.py
@@ -308,6 +308,37 @@ THEME_SEMANTIC_ICON_RULES: tuple[tuple[tuple[str, ...], str], ...] = (
 THEME_COLLECTION_TITLE_OVERRIDES: dict[str, str] = {
     "suicide_risk": "Suicide",
     "care_home_safety": "Care home deaths",
+    "acute_hospital_wards": "Acute Hospital Care",
+    "alarm_alert_failures": "Alarms and Alerts",
+    "capacity_best_interests_failures": "Capacity and Best Interests",
+    "consent_decision_making_failures": "Consent and Decision-making",
+    "failure_recognise_escalate_deterioration": "Recognising and Escalating Deterioration",
+    "family_carer_concerns_not_acted_on": "Family and Carer Concerns",
+    "failure_learn_previous_deaths_incidents": "Failure to Learn",
+    "follow_up_failures": "Follow-up",
+    "housing_homelessness": "Housing and Homelessness",
+    "inter_agency_working": "Inter-agency Working",
+    "investigation_incident_review_failures": "Investigations and Incident Reviews",
+    "it_digital_system_failures": "IT and Digital Systems",
+    "language_interpreter_barriers": "Language and Interpreter Barriers",
+    "maternity_neonatal_perinatal_care": "Maternity, Neonatal and Perinatal Care",
+    "missed_appointments_non_attendance": "Missed Appointments and Non-attendance",
+    "observation_monitoring_failures": "Observation and Monitoring",
+    "policy_procedure_failures": "Policy and Procedure",
+    "record_sharing_failures": "Record Sharing",
+    "referral_failures": "Referrals",
+    "risk_assessment_failures": "Risk Assessment",
+    "roads_highways": "Roads and Highways",
+    "sepsis_infection": "Sepsis and Infection",
+    "staffing_shortages_workload_pressure": "Staffing Shortages and Workload Pressure",
+    "suicide_self_harm": "Suicide and Self-harm",
+    "test_result_management_failures": "Test Result Management",
+    "training_competence_gaps": "Training and Competence",
+    "transitions_discharge_failures": "Transitions and Discharge",
+    "violence_homicide_related_systems_failures": "Violence and Homicide-related Systems",
+    "falls_frailty": "Falls and Frailty",
+    "equipment_failures": "Equipment",
+    "environmental_design_failures": "Environmental Design",
 }
 APPROVED_THEME_SCHEMA_PATH = (
     Path(__file__).resolve().parents[2] / "scripts" / "theme_collections" / "approved_themes.json"
@@ -535,6 +566,13 @@ def _collection_column_for_slug(collection_slug: str, reports_df: pd.DataFrame) 
     if theme_meta:
         return str(theme_meta.get("collection_column", "")).strip()
     return ""
+
+
+def _collection_excluded_theme_key(collection_slug: str, reports_df: pd.DataFrame) -> Optional[str]:
+    collection_column = _collection_column_for_slug(collection_slug, reports_df)
+    if not collection_column:
+        return None
+    return _network_theme_key_from_column(collection_column)
 
 
 def _browse_collection_meta(collection_slug: str, reports_df: pd.DataFrame) -> dict[str, str]:
@@ -1037,7 +1075,41 @@ def _build_explore_dashboard_options(reports_df: pd.DataFrame) -> dict[str, list
     }
 
 
-def _build_explore_dashboard_summary(filtered_df: pd.DataFrame, *, total_reports: int) -> dict[str, Any]:
+def _build_explore_dashboard_theme_counts(
+    filtered_df: pd.DataFrame,
+    *,
+    excluded_theme_key: Optional[str] = None,
+) -> dict[str, int]:
+    if filtered_df.empty:
+        return {}
+
+    excluded_key = str(excluded_theme_key or "").strip().lower()
[…]
