### 2026-03-13 #10074 @PetrKorab
@arfon Hello editor, this paper has been submitted for some time. I asked the author for clarification on some points, but did not receive a response. What is the status, and how should I proceed now?
