### 2026-03-16 26a772a
Implement search feature
