+++ django_workbench/workbench/tests.py
@@ -5,10 +5,16 @@ from unittest.mock import patch
 
 from .models import Workbook
 from .state import dataframe_from_payload, dataframe_to_payload
+from . import views
 
 
 class WorkbenchViewTests(TestCase):
     """Smoke tests for the Workbench page."""
+
+    def setUp(self) -> None:
+        super().setUp()
+        views.BROWSE_REPORTS_CACHE_DF = None
+        views.BROWSE_REPORTS_CACHE_UPDATED_AT = None
 
     def _unlock_network_tab(self) -> None:
         response = self.client.post(
@@ -104,6 +110,161 @@ class WorkbenchViewTests(TestCase):
         response = self.client.get(reverse("workbench:browse_collection", kwargs={"collection_slug": "wales"}))
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "Wales")
+
+    @patch("workbench.views.load_reports")
+    def test_browse_page_renames_custom_collection_and_adds_custom_search_panel(self, mock_load_reports) -> None:
+        mock_load_reports.return_value = pd.DataFrame(
+            [
+                {
+                    "date": "2024-01-10",
+                    "coroner": "A. Example",
+                    "area": "London Inner South",
+                    "receiver": "NHS England",
+                    "investigation": "Inquest conclusion",
+                    "circumstances": "Circumstances text",
+                    "concerns": "Concerns text",
+                }
+            ]
+        )
+
+        response = self.client.get(reverse("workbench:browse"))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "All reports")
+        self.assertContains(response, "Custom collection")
+        self.assertContains(response, "Create a custom collection")
+        self.assertContains(response, 'name="q"')
+
+    @patch("workbench.views.load_reports")
+    def test_custom_collection_search_returns_ranked_results(self, mock_load_reports) -> None:
+        mock_load_reports.return_value = pd.DataFrame(
+            [
+                {
+                    "date": "2024-01-10",
+                    "coroner": "A. Example",
+                    "area": "London Inner South",
+                    "receiver": "NHS England",
+                    "investigation": "Conclusion narrative",
+                    "circumstances": "A patient developed methotrexate toxicity after delayed monitoring.",
+                    "concerns": "Methotrexate guidance was outdated and monitoring failed.",
+                },
+                {
+                    "date": "2024-02-20",
+                    "coroner": "B. Example",
+                    "area": "Merseyside",
+                    "receiver": "Department of Health and Social Care",
+                    "investigation": "Road traffic death",
+                    "circumstances": "A collision happened on the motorway.",
+                    "concerns": "Road signage was unclear.",
+                },
+            ]
+        )
+
+        response = self.client.get(
+            reverse("workbench:browse_collection", kwargs={"collection_slug": "custom-search"}),
+            data={"q": "methotrexate monitoring"},
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, 'Search query</span>')
+        self.assertContains(response, '"methotrexate monitoring"')
+        self.assertContains(response, "Search results for")
+        self.assertContains(response, "Make editable copy")
+        self.assertEqual(response.context["reports_count"], 1)
+        self.assertContains(response, "NHS England")
+        self.assertNotContains(response, "Road signage was unclear")
+
+    @patch("workbench.views.load_reports")
+    def test_custom_collection_dashboard_data_respects_query(self, mock_load_reports) -> None:
+        mock_load_reports.return_value = pd.DataFrame(
+            [
+                {
+                    "date": "2024-01-10",
+                    "coroner": "A. Example",
+                    "area": "London Inner South",
+                    "receiver": "NHS England",
+                    "investigation": "Medication review",
[…]

[…]
