+++ .github/workflows/pfd_update.yml
@@ -86,14 +86,16 @@ jobs:
           fi
 
 
-      - name: Commit & Push Changes
-        if: github.ref == 'refs/heads/main'
-        uses: EndBug/add-and-commit@v9
-        with:
-          add: |
-            all_reports.csv
-            docs/index.md
-          default_author: github_actions
-          message: "🤖 Auto PFD dataset top-up"
-          pull: '--rebase --autostash'
-          push: true
+      # Disabled: this workflow should only update the release asset and
+      # should not commit generated data back to the repository.
+      # - name: Commit & Push Changes
+      #   if: github.ref == 'refs/heads/main'
+      #   uses: EndBug/add-and-commit@v9
+      #   with:
+      #     add: |
+      #       all_reports.csv
+      #       docs/index.md
+      #     default_author: github_actions
+      #     message: "🤖 Auto PFD dataset top-up"
+      #     pull: '--rebase --autostash'
+      #     push: true
