### 2026-03-18 6108c8f
Change update frequency to twice-weekly
