### 2026-03-24 804cba7
Stream AI workflow progress via server-sent events
