+++ django_workbench/workbench/views.py
@@ -5392,3 +5392,356 @@ def workbook_dashboard_data(request: HttpRequest, share_number: int, title_slug:
         include_options=False,
     )
     return JsonResponse(payload)
+
+
+# ──────────────────────────────────────────────────────────────────────
+# SSE endpoints for long-running AI workflows
+# ──────────────────────────────────────────────────────────────────────
+
+from .sse import SSEProgressBridge, run_in_background, sse_response  # noqa: E402
+
+
+def _patch_llm_progress(llm_client, bridge: SSEProgressBridge):
+    """Monkey-patch *llm_client.generate* to inject the SSE progress callback."""
+    original = llm_client.generate
+
+    def _generate_with_progress(*args, **kwargs):
+        kwargs.setdefault("progress_callback", bridge.progress_callback)
+        return original(*args, **kwargs)
+
+    llm_client.generate = _generate_with_progress
+
+
+@require_http_methods(["POST"])
+def sse_filter_reports(request: HttpRequest) -> HttpResponse:
+    """SSE endpoint for the screening workflow."""
+    init_state(request.session)
+    _ensure_workspace_reports_loaded(request)
+    _update_sidebar_state(request)
+    session = request.session
+
+    reports_df, dashboard_filters = _resolve_reports_for_ai_action(request)
+    if reports_df.empty:
+        return JsonResponse({"error": "No reports loaded."}, status=400)
+
+    search_query = (request.POST.get("search_query") or "").strip()
+    if not search_query:
+        return JsonResponse({"error": "Describe what the Screener should look for."}, status=400)
+
+    filter_df = _bool_from_post(request, "filter_df", default=True)
+    produce_spans = _bool_from_post(request, "produce_spans", default=False)
+    drop_spans = _bool_from_post(request, "drop_spans", default=False)
+
+    try:
+        llm_client = build_llm(session)
+    except Exception as exc:
+        return JsonResponse({"error": f"LLM setup failed: {exc}"}, status=400)
+
+    bridge = SSEProgressBridge()
+    _patch_llm_progress(llm_client, bridge)
+    initial_report_count = len(reports_df)
+
+    def _work():
+        started_at = perf_counter()
+        bridge.stage("screen", "Screening reports...")
+        push_history_snapshot(session)
+        screener = build_screener(llm_client, reports_df, session)
+        result_df = screener.screen_reports(
+            search_query=search_query,
+            filter_df=filter_df,
+            result_col_name="matches_query",
+            produce_spans=produce_spans,
+            drop_spans=drop_spans,
+        )
+        record_repro_action(
+            session, "run_screener", "Screen the reports",
+            format_call("result_df = screener.screen_reports", {
+                "search_query": search_query, "filter_df": filter_df,
+                "result_col_name": "matches_query",
+                "produce_spans": produce_spans, "drop_spans": drop_spans,
+            }),
+        )
+        set_dataframe(session, "screener_result", result_df)
+        _set_reports_df_with_row_ids(session, "reports_df", result_df)
+        clear_outputs_for_modified_dataset(session)
+        session["active_action"] = None
+
+        if filter_df:
+            matched = len(result_df)
+        elif isinstance(result_df, pd.DataFrame) and "matches_query" in result_df.columns:
+            matched = int(result_df["matches_query"].fillna(False).astype(bool).sum())
+        else:
+            matched = len(result_df)
+
+        messages.success(
+            request,
+            f"Screening successful! From the initial {initial_report_count:,} reports, "
+            f"{matched:,} matched your search query.",
+        )
+        session.modified = True
+        session.save()
+        _log_timed_event(
+            "screen_reports", started_at, force_info=True,
[…]

[…]
