### 2026-03-24 c22498d
Move pymdown to main dependency list
