+++ pyproject.toml
@@ -34,6 +34,7 @@ dependencies = [
     "gunicorn>=25.0.3",
     "httpx>=0.28.1",
     "markdown>=3.8.2",
+    "pymdown-extensions>=10.15",
     "openai>=1.88.0",
     "pandas>=2.2.2,<2.2.3; python_version<'3.13'",
     "pandas>=2.2.3; python_version>='3.13'",
@@ -62,7 +63,6 @@ dev = [
     "pytest>=8.4.0",
     "pytest-cov>=6.1.1",
     "mkdocs-git-revision-date-localized-plugin>=1.4.7",
-    "pymdown-extensions>=10.15",
     "mkdocs>=1.6.1",
     "mkdocs-autorefs>=1.4.2",
     "mkdocs-material[imaging]>=9.6.14",
