### 2026-03-25 e8901a5
Fix issue where pull out structured data workflow redirected to wrong view
