### 2026-03-25 1b6715f
For coders section no longer tries to import mkdocs-material
