+++ django_workbench/workbench/views.py
@@ -2966,7 +2966,7 @@ def _build_docs_nav_items(entries: list[Any], current_doc_path: str) -> list[dic
 def _load_docs_nav_from_config(current_doc_path: str) -> list[dict[str, Any]]:
     if not DOCS_CONFIG_FILE.is_file():
         return []
-    config_data = yaml.load(DOCS_CONFIG_FILE.read_text(encoding="utf-8"), Loader=yaml.UnsafeLoader) or {}
+    config_data = yaml.safe_load(DOCS_CONFIG_FILE.read_text(encoding="utf-8")) or {}
     nav_entries = config_data.get("nav") or []
     if not isinstance(nav_entries, list):
         return []
