### 2026-03-27 c04f0d9
Make snapshot bytes configurable
