### 2026-03-27 #10074 @RISbO001
@arfon
PFD-toolkit meets all JOSS requirements for code availability, licensing, authorship, reproducibility, functionality, documentation, and paper quality.
Minor improvements could include more extensive API documentation and additional automated tests, but these are optional.
