### 2026-04-06 c8cc9d9
Fix themes workflow not returning to Explore tab
