### 2026-04-06 30c2417
Remove Nutrition from collection
