### 2026-04-06 16dd75b
Prevent thematic snapshot from showing JSON-like strings
