### 2026-04-18 a0c9fa9
Set-up data model and apps
