+++ django_workbench_v02/wb_runs/migrations/0001_initial.py
@@ -1,0 +1,335 @@
+# Generated by Django 5.2.12 on 2026-04-18 21:30
+
+import django.db.models.deletion
+import django.utils.timezone
+import uuid
+from django.conf import settings
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    initial = True
+
+    dependencies = [
+        ("wb_investigations", "0001_initial"),
+        ("wb_workspaces", "0001_initial"),
+        migrations.swappable_dependency(settings.AUTH_USER_MODEL),
+    ]
+
+    operations = [
+        migrations.CreateModel(
+            name="InvestigationRun",
+            fields=[
+                (
+                    "id",
+                    models.UUIDField(
+                        default=uuid.uuid4,
+                        editable=False,
+                        primary_key=True,
+                        serialize=False,
+                    ),
+                ),
+                (
+                    "run_type",
+                    models.CharField(
+                        choices=[
+                            ("filter", "Filter"),
+                            ("themes", "Themes"),
+                            ("extract", "Extract"),
+                            ("export", "Export"),
+                        ],
+                        max_length=16,
+                    ),
+                ),
+                (
+                    "status",
+                    models.CharField(
+                        choices=[
+                            ("queued", "Queued"),
+                            ("starting", "Starting"),
+                            ("running", "Running"),
+                            ("cancelling", "Cancelling"),
+                            ("cancelled", "Cancelled"),
+                            ("succeeded", "Succeeded"),
+                            ("failed", "Failed"),
+                            ("timed_out", "Timed out"),
+                        ],
+                        default="queued",
+                        max_length=16,
+                    ),
+                ),
+                (
+                    "progress_percent",
+                    models.PositiveSmallIntegerField(blank=True, null=True),
+                ),
+                ("queued_at", models.DateTimeField(default=django.utils.timezone.now)),
+                ("started_at", models.DateTimeField(blank=True, null=True)),
+                ("finished_at", models.DateTimeField(blank=True, null=True)),
+                ("cancel_requested_at", models.DateTimeField(blank=True, null=True)),
+                ("cancel_reason", models.TextField(blank=True)),
+                ("worker_id", models.CharField(blank=True, max_length=255)),
+                ("error_code", models.CharField(blank=True, max_length=64)),
+                ("error_message", models.TextField(blank=True)),
+                ("input_config_json", models.JSONField(default=dict)),
+                ("query_start_date", models.DateField(blank=True, null=True)),
+                ("query_end_date", models.DateField(blank=True, null=True)),
+                (
+                    "created_at",
+                    models.DateTimeField(
+                        default=django.utils.timezone.now, editable=False
+                    ),
+                ),
+                ("updated_at", models.DateTimeField(auto_now=True)),
+                (
+                    "cancel_requested_by",
+                    models.ForeignKey(
+                        blank=True,
+                        null=True,
+                        on_delete=django.db.models.deletion.SET_NULL,
+                        related_name="cancel_requested_runs",
+                        to=settings.AUTH_USER_MODEL,
+                    ),
+                ),
+                (
+                    "investigation",
+                    models.ForeignKey(
+                        on_delete=django.db.models.deletion.CASCADE,
+                        related_name="runs",
+                        to="wb_investigations.investigation",
+                    ),
+                ),
+                (
+                    "requested_by",
+                    models.ForeignKey(
+                        on_delete=django.db.models.deletion.PROTECT,
+                        related_name="requested_runs",
+                        to=settings.AUTH_USER_MODEL,
+                    ),
+                ),
+                (
+                    "workspace",
+                    models.ForeignKey(
+                        on_delete=django.db.models.deletion.CASCADE,
+                        related_name="runs",
+                        to="wb_workspaces.workspace",
[…]
