+++ django_workbench_v02/wb_notifications/admin.py
@@ -1,0 +1,20 @@
+from django.contrib import admin
+
+from .models import NotificationRequest
+
+
+@admin.register(NotificationRequest)
+class NotificationRequestAdmin(admin.ModelAdmin):
+    list_display = [
+        "run",
+        "user",
+        "channel",
+        "notify_on",
+        "status",
+        "sent_at",
+        "created_at",
+    ]
+    list_filter = ["channel", "notify_on", "status", "created_at"]
+    search_fields = ["run__id", "user__username", "user__email", "error_message"]
+    autocomplete_fields = ["run", "user"]
+    readonly_fields = ["created_at", "updated_at", "sent_at"]
