+++ django_workbench_v02/wb_runs/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class WbRunsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "wb_runs"
