+++ django_workbench_v02/wb_investigations/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class WbInvestigationsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "wb_investigations"
