### 2026-04-18 b92835e
Set up Auth0
