+++ django_workbench_v02/accounts/models.py
@@ -1,0 +1,18 @@
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+from .managers import UserManager
+
+
+class User(AbstractUser):
+    username = None
+    email = models.EmailField(unique=True)
+    auth0_sub = models.CharField(max_length=128, unique=True, null=True, blank=True)
+
+    USERNAME_FIELD = "email"
+    REQUIRED_FIELDS = ["first_name", "last_name"]
+
+    objects = UserManager()
+
+    def __str__(self) -> str:
+        return self.email
