+++ django_workbench_v02/accounts/admin.py
@@ -1,0 +1,52 @@
+from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
+
+from .forms import UserChangeForm, UserCreationForm
+from .models import User
+
+
+@admin.register(User)
+class UserAdmin(BaseUserAdmin):
+    add_form = UserCreationForm
+    form = UserChangeForm
+    model = User
+
+    list_display = [
+        "email",
+        "first_name",
+        "last_name",
+        "is_active",
+        "is_staff",
+        "is_superuser",
+        "last_login",
+    ]
+    list_filter = ["is_active", "is_staff", "is_superuser", "groups"]
+    ordering = ["email"]
+    search_fields = ["email", "first_name", "last_name", "auth0_sub"]
+
+    fieldsets = (
+        (None, {"fields": ("email", "password", "auth0_sub")}),
+        ("Personal info", {"fields": ("first_name", "last_name")}),
+        (
+            "Permissions",
+            {
+                "fields": (
+                    "is_active",
+                    "is_staff",
+                    "is_superuser",
+                    "groups",
+                    "user_permissions",
+                )
+            },
+        ),
+        ("Important dates", {"fields": ("last_login", "date_joined")}),
+    )
+    add_fieldsets = (
+        (
+            None,
+            {
+                "classes": ("wide",),
+                "fields": ("email", "first_name", "last_name", "password1", "password2"),
+            },
+        ),
+    )
