### 2026-04-18 a9330d9
service-layer invariants and endpoints for membership management
