+++ django_workbench_v02/wb_workspaces/tests.py
@@ -1,9 +1,18 @@
+from django.core.exceptions import PermissionDenied, ValidationError
 from django.contrib.auth import get_user_model
 from django.test import TestCase
+from django.urls import reverse
+
+from wb_auditlog.models import AuditEvent
 
 from .models import MembershipAccessMode, MembershipRole, Workspace, WorkspaceMembership, WorkspaceVisibility
 from .permissions import can_edit_workspace, can_manage_members, can_run_workflows, can_view_workspace
-from .services import create_workspace_for_user
+from .services import (
+    add_workspace_member,
+    create_workspace_for_user,
+    remove_workspace_member,
+    update_workspace_member,
+)
 
 
 User = get_user_model()
@@ -79,17 +88,184 @@ class WorkspacePermissionTests(TestCase):
 
 
 class WorkspaceServiceTests(TestCase):
-    def test_create_workspace_for_user_creates_owner_membership(self):
-        user = User.objects.create_user(email="new-owner@example.com", password="x")
-        workspace = create_workspace_for_user(
-            user=user,
-            title="New Workspace",
-            slug="new-workspace",
+    def setUp(self):
+        self.owner = User.objects.create_user(email="owner2@example.com", password="x")
+        self.editor = User.objects.create_user(email="editor2@example.com", password="x")
+        self.viewer = User.objects.create_user(email="viewer2@example.com", password="x")
+        self.workspace = create_workspace_for_user(
+            user=self.owner,
+            title="Service Workspace",
+            slug="service-workspace",
             description="Desc",
         )
 
-        membership = WorkspaceMembership.objects.get(workspace=workspace, user=user)
+    def test_create_workspace_for_user_creates_owner_membership(self):
+        membership = WorkspaceMembership.objects.get(workspace=self.workspace, user=self.owner)
         self.assertEqual(membership.role, MembershipRole.OWNER)
         self.assertEqual(membership.access_mode, MembershipAccessMode.EDIT)
         self.assertTrue(membership.can_manage_members)
         self.assertTrue(membership.can_manage_shares)
+        self.assertTrue(
+            AuditEvent.objects.filter(
+                workspace=self.workspace, action_type="workspace.created"
+            ).exists()
+        )
+
+    def test_add_workspace_member_creates_membership_and_audit(self):
+        membership = add_workspace_member(
+            actor=self.owner,
+            workspace=self.workspace,
+            target_user=self.editor,
+            role=MembershipRole.EDITOR,
+            access_mode=MembershipAccessMode.EDIT,
+            can_run_workflows=True,
+            can_manage_members_flag=False,
+            can_manage_shares_flag=False,
+        )
+        self.assertEqual(membership.user, self.editor)
+        self.assertTrue(
+            AuditEvent.objects.filter(
+                workspace=self.workspace, action_type="workspace.member_added"
+            ).exists()
+        )
+
+    def test_non_owner_cannot_add_workspace_member(self):
+        add_workspace_member(
+            actor=self.owner,
+            workspace=self.workspace,
+            target_user=self.editor,
+            role=MembershipRole.EDITOR,
+            access_mode=MembershipAccessMode.EDIT,
+            can_run_workflows=True,
+            can_manage_members_flag=False,
+            can_manage_shares_flag=False,
+        )
+
+        with self.assertRaises(PermissionDenied):
+            add_workspace_member(
+                actor=self.editor,
+                workspace=self.workspace,
+                target_user=self.viewer,
+                role=MembershipRole.VIEWER,
+                access_mode=MembershipAccessMode.READ_ONLY,
+                can_run_workflows=False,
+                can_manage_members_flag=False,
+                can_manage_shares_flag=False,
+            )
+
+    def test_update_member_enforces_owner_invariants(self):
+        owner_membership = WorkspaceMembership.objects.get(
[…]
