### 2026-04-18 dfb8c51
Build investigation/run endpoints
