+++ django_workbench_v02/wb_investigations/tests.py
@@ -1,3 +1,150 @@
+from django.contrib.auth import get_user_model
+from django.core.exceptions import PermissionDenied
 from django.test import TestCase
+from django.urls import reverse
 
-# Create your tests here.
+from wb_auditlog.models import AuditEvent
+from wb_workspaces.models import MembershipAccessMode, MembershipRole, WorkspaceMembership, WorkspaceVisibility
+from wb_workspaces.services import create_workspace_for_user
+
+from .models import Investigation, InvestigationStatus
+from .services import create_investigation, update_investigation
+
+
+User = get_user_model()
+
+
+class InvestigationServiceTests(TestCase):
+    def setUp(self):
+        self.owner = User.objects.create_user(email="inv-owner@example.com", password="x")
+        self.viewer = User.objects.create_user(email="inv-viewer@example.com", password="x")
+        self.workspace = create_workspace_for_user(
+            user=self.owner,
+            title="Inv Workspace",
+            slug="inv-workspace",
+            description="desc",
+        )
+        WorkspaceMembership.objects.create(
+            workspace=self.workspace,
+            user=self.viewer,
+            role=MembershipRole.VIEWER,
+            access_mode=MembershipAccessMode.READ_ONLY,
+            can_manage_members=False,
+            can_manage_shares=False,
+            can_run_workflows=False,
+        )
+
+    def test_owner_can_create_investigation(self):
+        inv = create_investigation(
+            actor=self.owner,
+            workspace=self.workspace,
+            title="Q1",
+            question_text="What is happening?",
+            scope_json={"dataset": "all"},
+            method_json={"model": "x"},
+            status=InvestigationStatus.DRAFT,
+        )
+        self.assertEqual(inv.created_by, self.owner)
+        self.assertTrue(
+            AuditEvent.objects.filter(
+                action_type="investigation.created",
+                target_id=str(inv.id),
+            ).exists()
+        )
+
+    def test_viewer_cannot_create_investigation(self):
+        with self.assertRaises(PermissionDenied):
+            create_investigation(
+                actor=self.viewer,
+                workspace=self.workspace,
+                title="Q2",
+                question_text="No",
+                scope_json={},
+                method_json={},
+                status=InvestigationStatus.DRAFT,
+            )
+
+    def test_owner_can_update_investigation(self):
+        inv = create_investigation(
+            actor=self.owner,
+            workspace=self.workspace,
+            title="Base",
+            question_text="Q",
+            scope_json={},
+            method_json={},
+            status=InvestigationStatus.DRAFT,
+        )
+        update_investigation(
+            actor=self.owner,
+            investigation=inv,
+            title="Updated",
+            question_text="Q2",
+            scope_json={"a": 1},
+            method_json={"b": 2},
+            status=InvestigationStatus.ACTIVE,
+        )
+        inv.refresh_from_db()
+        self.assertEqual(inv.title, "Updated")
+        self.assertEqual(inv.status, InvestigationStatus.ACTIVE)
+        self.assertTrue(
+            AuditEvent.objects.filter(
+                action_type="investigation.updated",
+                target_id=str(inv.id),
+            ).exists()
+        )
+
+
+class InvestigationViewTests(TestCase):
+    def setUp(self):
+        self.owner = User.objects.create_user(email="inv-owner2@example.com", password="x")
+        self.stranger = User.objects.create_user(email="inv-stranger2@example.com", password="x")
+        self.workspace = create_workspace_for_user(
+            user=self.owner,
+            title="Inv Workspace2",
+            slug="inv-workspace2",
+            description="desc",
+        )
+        self.investigation = create_investigation(
+            actor=self.owner,
+            workspace=self.workspace,
+            title="Q1",
[…]
