### 2026-04-18 4db24bf
Hook worker-side run transitions (starting, running, succeeded/failed/timed_out/cancelled) into wb_runs.services.set_run_status
