+++ django_workbench_v02/wb_runs/worker.py
@@ -1,0 +1,241 @@
+from __future__ import annotations
+
+import logging
+import time
+import uuid
+from datetime import timedelta
+
+from django.db import transaction
+from django.utils import timezone
+
+from .models import (
+    ArtifactStatus,
+    ArtifactStorageBackend,
+    ArtifactType,
+    InvestigationRun,
+    RunStatus,
+    RunType,
+    RunArtifact,
+)
+from .services import is_terminal_status, set_run_status
+
+
+logger = logging.getLogger(__name__)
+
+
+RUN_STAGE_MAP = {
+    RunType.FILTER: [
+        ("Preparing filter inputs", 10),
+        ("Applying report filters", 45),
+        ("Compiling filtered dataset", 80),
+    ],
+    RunType.THEMES: [
+        ("Preparing theme discovery", 10),
+        ("Discovering themes", 50),
+        ("Summarizing theme outputs", 85),
+    ],
+    RunType.EXTRACT: [
+        ("Preparing extraction prompts", 10),
+        ("Running extraction", 55),
+        ("Compiling extraction table", 85),
+    ],
+    RunType.EXPORT: [
+        ("Preparing export bundle", 20),
+        ("Generating package", 60),
+        ("Finalizing export output", 90),
+    ],
+}
+
+ARTIFACT_TYPE_BY_RUN_TYPE = {
+    RunType.FILTER: ArtifactType.FILTERED_DATASET,
+    RunType.THEMES: ArtifactType.THEME_SUMMARY,
+    RunType.EXTRACT: ArtifactType.EXTRACTION_TABLE,
+    RunType.EXPORT: ArtifactType.BUNDLE_EXPORT,
+}
+
+
+def _reload_run(run_id):
+    return InvestigationRun.objects.select_related("workspace", "investigation").get(id=run_id)
+
+
+def _cancel_requested(run) -> bool:
+    current = _reload_run(run.id)
+    return current.status == RunStatus.CANCELLING or current.cancel_requested_at is not None
+
+
+@transaction.atomic
+def claim_next_runnable_run(worker_id: str) -> InvestigationRun | None:
+    run = (
+        InvestigationRun.objects.select_for_update(skip_locked=True)
+        .filter(status__in=[RunStatus.QUEUED, RunStatus.CANCELLING])
+        .order_by("queued_at", "created_at")
+        .first()
+    )
+    if run is None:
+        return None
+
+    run.worker_id = worker_id
+    run.save(update_fields=["worker_id", "updated_at"])
+    return run
+
+
+def _create_success_artifact(run: InvestigationRun) -> None:
+    artifact_type = ARTIFACT_TYPE_BY_RUN_TYPE.get(run.run_type, ArtifactType.PREVIEW)
+    RunArtifact.objects.create(
+        run=run,
+        workspace=run.workspace,
+        artifact_type=artifact_type,
+        status=ArtifactStatus.READY,
+        storage_backend=ArtifactStorageBackend.DB,
+        storage_uri=f"db://run-artifacts/{run.id}",
+        metadata_json={
+            "generated_by_worker": run.worker_id or "",
+            "run_type": run.run_type,
+            "investigation_id": str(run.investigation_id),
+            "completed_at": timezone.now().isoformat(),
+        },
+        expires_at=timezone.now() + timedelta(days=365),
+    )
+
+
+def _execute_run(run: InvestigationRun, sleep_between_stages_seconds: float = 0.0) -> InvestigationRun:
+    run = _reload_run(run.id)
+
+    if run.status == RunStatus.CANCELLING:
+        return set_run_status(
+            run=run,
+            status=RunStatus.CANCELLED,
+            message="Run was cancelled before processing started.",
+        )
+
[…]

[…]
