### 2026-04-18 853a6e0
Create adapters for filter themes and extract pipelines
