+++ django_workbench_v02/wb_runs/tests.py
@@ -1,3 +1,6 @@
+from pathlib import Path
+from unittest.mock import patch
+
 from django.contrib.auth import get_user_model
 from django.core.exceptions import PermissionDenied, ValidationError
 from django.test import TestCase
@@ -9,7 +12,7 @@ from wb_investigations.services import create_investigation
 from wb_workspaces.models import MembershipAccessMode, MembershipRole, WorkspaceMembership
 from wb_workspaces.services import create_workspace_for_user
 
-from .models import RunStatus, RunType
+from .models import ArtifactStorageBackend, ArtifactType, RunStatus, RunType
 from .services import queue_run, request_run_cancellation, set_run_status
 from .worker import process_single_available_run
 
@@ -242,7 +245,7 @@ class RunWorkerTests(TestCase):
             actor=self.owner,
             investigation=self.investigation,
             run_type=RunType.FILTER,
-            input_config_json={},
+            input_config_json={"execution_mode": "simulate"},
         )
         processed = process_single_available_run(worker_id="test-worker")
         self.assertIsNotNone(processed)
@@ -259,7 +262,7 @@ class RunWorkerTests(TestCase):
             actor=self.owner,
             investigation=self.investigation,
             run_type=RunType.FILTER,
-            input_config_json={},
+            input_config_json={"execution_mode": "simulate"},
         )
         request_run_cancellation(actor=self.owner, run=run, reason="cancel early")
         process_single_available_run(worker_id="test-worker")
@@ -271,7 +274,11 @@ class RunWorkerTests(TestCase):
             actor=self.owner,
             investigation=self.investigation,
             run_type=RunType.FILTER,
-            input_config_json={"simulate_failure": True, "simulate_failure_stage": 1},
+            input_config_json={
+                "execution_mode": "simulate",
+                "simulate_failure": True,
+                "simulate_failure_stage": 1,
+            },
         )
         process_single_available_run(worker_id="test-worker")
         run.refresh_from_db()
@@ -284,9 +291,146 @@ class RunWorkerTests(TestCase):
             actor=self.owner,
             investigation=self.investigation,
             run_type=RunType.FILTER,
-            input_config_json={"simulate_timeout": True, "simulate_timeout_stage": 1},
+            input_config_json={
+                "execution_mode": "simulate",
+                "simulate_timeout": True,
+                "simulate_timeout_stage": 1,
+            },
         )
         process_single_available_run(worker_id="test-worker")
         run.refresh_from_db()
         self.assertEqual(run.status, RunStatus.TIMED_OUT)
         self.assertEqual(run.error_code, "SIMULATED_TIMEOUT")
+
+    def test_worker_uses_real_filter_adapter_when_enabled(self):
+        output_path = Path("/tmp/test-run-filter-output.csv")
+        output_path.write_text("id,matches_query\n1,True\n", encoding="utf-8")
+
+        run = queue_run(
+            actor=self.owner,
+            investigation=self.investigation,
+            run_type=RunType.FILTER,
+            input_config_json={"execution_mode": "real", "search_query": "medication safety"},
+        )
+
+        with patch(
+            "wb_runs.worker.execute_filter_workflow",
+            return_value={
+                "output_path": str(output_path),
+                "total_reports": 10,
+                "matched_reports": 3,
+                "output_reports": 3,
+                "search_query": "medication safety",
+                "filter_df": True,
+                "produce_spans": False,
+                "drop_spans": False,
+                "start_date": "2024-01-01",
+                "end_date": "2024-12-31",
+                "report_limit": None,
+            },
+        ) as mocked:
[…]
