+++ django_workbench_v02/wb_runs/worker.py
@@ -4,6 +4,7 @@ import logging
 import time
 import uuid
 from datetime import timedelta
+from pathlib import Path
 
 from django.db import transaction
 from django.utils import timezone
@@ -18,6 +19,13 @@ from .models import (
     RunArtifact,
 )
 from .services import is_terminal_status, set_run_status
+from .pfd_toolkit_adapter import (
+    AdapterCancelledError,
+    AdapterConfigurationError,
+    execute_extract_workflow,
+    execute_filter_workflow,
+    execute_themes_workflow,
+)
 
 
 logger = logging.getLogger(__name__)
@@ -53,6 +61,20 @@ ARTIFACT_TYPE_BY_RUN_TYPE = {
     RunType.EXPORT: ArtifactType.BUNDLE_EXPORT,
 }
 
+REAL_ADAPTER_RUN_TYPES = {RunType.FILTER, RunType.THEMES, RunType.EXTRACT}
+
+RUN_LABEL_BY_TYPE = {
+    RunType.FILTER: "filter",
+    RunType.THEMES: "themes",
+    RunType.EXTRACT: "extract",
+}
+
+EXECUTION_ERROR_CODE_BY_RUN_TYPE = {
+    RunType.FILTER: "FILTER_EXECUTION_ERROR",
+    RunType.THEMES: "THEMES_EXECUTION_ERROR",
+    RunType.EXTRACT: "EXTRACT_EXECUTION_ERROR",
+}
+
 
 def _reload_run(run_id):
     return InvestigationRun.objects.select_related("workspace", "investigation").get(id=run_id)
@@ -79,26 +101,195 @@ def claim_next_runnable_run(worker_id: str) -> InvestigationRun | None:
     return run
 
 
-def _create_success_artifact(run: InvestigationRun) -> None:
-    artifact_type = ARTIFACT_TYPE_BY_RUN_TYPE.get(run.run_type, ArtifactType.PREVIEW)
+def _create_success_artifact(
+    run: InvestigationRun,
+    *,
+    artifact_type: str | None = None,
+    storage_backend: str = ArtifactStorageBackend.DB,
+    storage_uri: str | None = None,
+    metadata: dict | None = None,
+) -> None:
+    selected_artifact_type = artifact_type or ARTIFACT_TYPE_BY_RUN_TYPE.get(
+        run.run_type, ArtifactType.PREVIEW
+    )
+    size_bytes = None
+    if storage_backend == ArtifactStorageBackend.FILE and storage_uri:
+        try:
+            size_bytes = Path(storage_uri).stat().st_size
+        except OSError:
+            size_bytes = None
+
     RunArtifact.objects.create(
         run=run,
         workspace=run.workspace,
-        artifact_type=artifact_type,
+        artifact_type=selected_artifact_type,
         status=ArtifactStatus.READY,
-        storage_backend=ArtifactStorageBackend.DB,
-        storage_uri=f"db://run-artifacts/{run.id}",
+        storage_backend=storage_backend,
+        storage_uri=storage_uri or f"db://run-artifacts/{run.id}",
+        size_bytes=size_bytes,
         metadata_json={
             "generated_by_worker": run.worker_id or "",
             "run_type": run.run_type,
             "investigation_id": str(run.investigation_id),
             "completed_at": timezone.now().isoformat(),
+            **(metadata or {}),
         },
         expires_at=timezone.now() + timedelta(days=365),
     )
 
 
-def _execute_run(run: InvestigationRun, sleep_between_stages_seconds: float = 0.0) -> InvestigationRun:
+def _is_real_adapter_mode(run: InvestigationRun) -> bool:
+    config = run.input_config_json or {}
[…]

[…]
