### 2026-04-18 c1766fa
add real export bundling and audited artifact downloads with human-view keepalive baseline
