+++ pyproject.toml
@@ -47,6 +47,7 @@ dependencies = [
     "tiktoken>=0.9.0",
     "tqdm>=4.67.1",
     "whitenoise>=6.11.0",
+    "boto3>=1.42.91",
 ]
 
 [project.optional-dependencies]

[…]
