### 2026-04-19 f6556cc
add human-only keepalive and scheduler-based inactivity expiry/archive jobs
