### 2026-04-19 78a4cb0
add ADR set and async completion email dispatcher with queue-time subscriptions
