+++ django_workbench_v02/wb_notifications/services.py
@@ -1,0 +1,254 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from time import sleep
+
+from django.conf import settings
+from django.core.exceptions import ValidationError
+from django.core.mail import send_mail
+from django.utils import timezone
+
+from wb_auditlog.services import log_audit_event
+from wb_runs.models import RunStatus
+
+from .models import NotificationChannel, NotificationRequest, NotificationStatus, NotificationTrigger
+
+
+TERMINAL_STATUSES = {
+    RunStatus.CANCELLED,
+    RunStatus.SUCCEEDED,
+    RunStatus.FAILED,
+    RunStatus.TIMED_OUT,
+}
+FAILURE_STATUSES = {
+    RunStatus.CANCELLED,
+    RunStatus.FAILED,
+    RunStatus.TIMED_OUT,
+}
+
+
+class NotificationRequestError(ValidationError):
+    pass
+
+
+@dataclass
+class NotificationDispatchResult:
+    scanned: int = 0
+    sent: int = 0
+    failed: int = 0
+    cancelled: int = 0
+
+
+def _should_send_for_trigger(*, notify_on: str, run_status: str) -> bool:
+    if notify_on == NotificationTrigger.ANY:
+        return run_status in TERMINAL_STATUSES
+    if notify_on == NotificationTrigger.SUCCESS:
+        return run_status == RunStatus.SUCCEEDED
+    if notify_on == NotificationTrigger.FAILURE:
+        return run_status in FAILURE_STATUSES
+    return False
+
+
+def _run_detail_url(run) -> str:
+    return (
+        f"{settings.WORKBENCH_BASE_URL}/workspaces/{run.workspace_id}/runs/{run.id}/"
+    )
+
+
+def _build_subject(notification: NotificationRequest) -> str:
+    run = notification.run
+    return f"[PFD Toolkit] Run {run.status}: {run.investigation.title}"
+
+
+def _build_body(notification: NotificationRequest) -> str:
+    run = notification.run
+    return (
+        "Your run has completed.\n\n"
+        f"Investigation: {run.investigation.title}\n"
+        f"Run ID: {run.id}\n"
+        f"Run type: {run.run_type}\n"
+        f"Status: {run.status}\n"
+        f"Queued at: {run.queued_at}\n"
+        f"Started at: {run.started_at}\n"
+        f"Finished at: {run.finished_at}\n\n"
+        f"View run details: {_run_detail_url(run)}\n"
+    )
+
+
+def create_notification_request(
+    *,
+    run,
+    user,
+    notify_on: str = NotificationTrigger.ANY,
+    channel: str = NotificationChannel.EMAIL,
+    request=None,
+) -> NotificationRequest:
+    if channel != NotificationChannel.EMAIL:
+        raise NotificationRequestError("Only email notifications are currently supported.")
+    if notify_on not in NotificationTrigger.values:
+        raise NotificationRequestError("Invalid notification trigger.")
+    if not user.email:
+        raise NotificationRequestError("User account has no email address for notifications.")
+
+    existing = NotificationRequest.objects.filter(
+        run=run,
+        user=user,
+        channel=channel,
+        notify_on=notify_on,
+        status__in=[NotificationStatus.PENDING, NotificationStatus.SENT],
+    ).first()
+    if existing:
+        return existing
+
+    notification = NotificationRequest.objects.create(
+        run=run,
+        user=user,
+        channel=channel,
+        notify_on=notify_on,
+        status=NotificationStatus.PENDING,
+    )
+    log_audit_event(
+        action_type="notification.requested",
+        target_type="notification_request",
+        target_id=str(notification.id),
+        workspace=run.workspace,
+        user=user,
+        payload={
+            "run_id": str(run.id),
[…]
