+++ django_workbench_v02/wb_runs/forms.py
@@ -1,5 +1,7 @@
 from django import forms
 
+from wb_notifications.models import NotificationTrigger
+
 from .models import RunType
 
 
@@ -8,6 +10,17 @@ class RunQueueForm(forms.Form):
     input_config_json = forms.JSONField(required=False, initial=dict)
     query_start_date = forms.DateField(required=False, widget=forms.DateInput(attrs={"type": "date"}))
     query_end_date = forms.DateField(required=False, widget=forms.DateInput(attrs={"type": "date"}))
+    request_completion_email = forms.BooleanField(
+        required=False,
+        initial=False,
+        help_text="Email me when this run reaches a terminal status.",
+    )
+    notify_on = forms.ChoiceField(
+        required=False,
+        choices=NotificationTrigger.choices,
+        initial=NotificationTrigger.ANY,
+        help_text="When to send email if completion notification is enabled.",
+    )
 
 
 class RunCancelForm(forms.Form):
