### 2026-04-19 c9c8417
Support encrypted, workspace-scoped API credentials
