+++ django_workbench_v02/wb_workspaces/migrations/0002_workspacecredential.py
@@ -1,0 +1,80 @@
+# Generated by Django 5.2.12 on 2026-04-19 13:25
+
+import django.db.models.deletion
+import django.utils.timezone
+import uuid
+from django.conf import settings
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("wb_workspaces", "0001_initial"),
+        migrations.swappable_dependency(settings.AUTH_USER_MODEL),
+    ]
+
+    operations = [
+        migrations.CreateModel(
+            name="WorkspaceCredential",
+            fields=[
+                (
+                    "id",
+                    models.UUIDField(
+                        default=uuid.uuid4,
+                        editable=False,
+                        primary_key=True,
+                        serialize=False,
+                    ),
+                ),
+                (
+                    "provider",
+                    models.CharField(
+                        choices=[("openai", "OpenAI"), ("openrouter", "OpenRouter")],
+                        max_length=16,
+                    ),
+                ),
+                ("encrypted_api_key", models.TextField()),
+                ("base_url", models.URLField(blank=True)),
+                ("key_last4", models.CharField(max_length=4)),
+                ("last_used_at", models.DateTimeField(blank=True, null=True)),
+                (
+                    "created_at",
+                    models.DateTimeField(
+                        default=django.utils.timezone.now, editable=False
+                    ),
+                ),
+                ("updated_at", models.DateTimeField(auto_now=True)),
+                (
+                    "user",
+                    models.ForeignKey(
+                        on_delete=django.db.models.deletion.CASCADE,
+                        related_name="workspace_credentials",
+                        to=settings.AUTH_USER_MODEL,
+                    ),
+                ),
+                (
+                    "workspace",
+                    models.ForeignKey(
+                        on_delete=django.db.models.deletion.CASCADE,
+                        related_name="credentials",
+                        to="wb_workspaces.workspace",
+                    ),
+                ),
+            ],
+            options={
+                "indexes": [
+                    models.Index(
+                        fields=["workspace", "user"], name="idx_ws_cred_scope"
+                    ),
+                    models.Index(fields=["provider"], name="idx_ws_cred_provider"),
+                ],
+                "constraints": [
+                    models.UniqueConstraint(
+                        fields=("workspace", "user", "provider"),
+                        name="uniq_workspace_credential_scope",
+                    )
+                ],
+            },
+        ),
+    ]
