### 2026-04-19 6a9a189
add per-workspace credential settings UI with save/delete endpoints and permission checks
