### 2026-04-19 b2d712b
make workbook the canonical app surface and enforce single-investigation-per-workbook
