### 2026-04-19 8ecab90
add read-only share copy flow with login gate and full workbook history clone
