+++ django_workbench_v02/wb_sharing/services.py
@@ -1,11 +1,18 @@
 from __future__ import annotations
 
+from copy import deepcopy
+
 from django.core.exceptions import PermissionDenied, ValidationError
 from django.db import transaction
 from django.utils import timezone
+from django.utils.text import slugify
 
 from wb_auditlog.services import log_audit_event
+from wb_investigations.models import Investigation
+from wb_runs.models import InvestigationRun, RunArtifact, RunEvent
 from wb_workspaces.activity import is_human_view_request, should_update_last_viewed
+from wb_workspaces.models import WorkspaceVisibility, WorkspaceRevision
+from wb_workspaces.services import create_workspace_for_user
 from wb_workspaces.permissions import can_manage_shares
 
 from .models import ShareMode, WorkspaceShareLink
@@ -13,6 +20,145 @@ from .models import ShareMode, WorkspaceShareLink
 
 class WorkspaceShareLinkError(ValidationError):
     pass
+
+
+def _next_copy_slug(*, actor, source_workspace) -> str:
+    base = slugify(source_workspace.slug or source_workspace.title or "workbook-copy")[:80]
+    if not base:
+        base = "workbook-copy"
+
+    candidate = base
+    counter = 2
+    while source_workspace.__class__.objects.filter(created_by=actor, slug=candidate).exists():
+        suffix = f"-copy-{counter}"
+        candidate = f"{base[: max(1, 100 - len(suffix))]}{suffix}"
+        counter += 1
+    return candidate
+
+
+@transaction.atomic
+def copy_share_link_to_workbook(*, actor, share_link: WorkspaceShareLink, request=None):
+    source_workspace = share_link.workspace
+    target_slug = _next_copy_slug(actor=actor, source_workspace=source_workspace)
+    target_title = f"{source_workspace.title} (Copy)"
+
+    copied_workspace = create_workspace_for_user(
+        user=actor,
+        title=target_title,
+        slug=target_slug,
+        description=source_workspace.description,
+        request=request,
+    )
+    copied_workspace.visibility = WorkspaceVisibility.PRIVATE
+    copied_workspace.is_listed = False
+    copied_workspace.last_viewed_at = None
+    copied_workspace.archived_at = None
+    copied_workspace.save(
+        update_fields=["visibility", "is_listed", "last_viewed_at", "archived_at", "updated_at"]
+    )
+
+    revision_map: dict[str, WorkspaceRevision] = {}
+    for revision in source_workspace.revisions.order_by("revision_number", "created_at"):
+        copied_revision = WorkspaceRevision.objects.create(
+            workspace=copied_workspace,
+            revision_number=revision.revision_number,
+            state_json=deepcopy(revision.state_json),
+            created_by=revision.created_by,
+            change_type=revision.change_type,
+            parent_revision=revision_map.get(str(revision.parent_revision_id))
+            if revision.parent_revision_id
+            else None,
+            created_at=revision.created_at,
+        )
+        revision_map[str(revision.id)] = copied_revision
+
+    source_investigation = (
+        Investigation.objects.filter(workspace=source_workspace).order_by("-created_at").first()
+    )
+    copied_investigation = None
+    run_map: dict[str, InvestigationRun] = {}
+    if source_investigation is not None:
+        copied_investigation = Investigation.objects.create(
+            workspace=copied_workspace,
+            created_by=source_investigation.created_by,
+            title=source_investigation.title,
+            question_text=source_investigation.question_text,
+            scope_json=deepcopy(source_investigation.scope_json),
+            method_json=deepcopy(source_investigation.method_json),
+            status=source_investigation.status,
+            last_viewed_at=None,
+            created_at=source_investigation.created_at,
+            updated_at=source_investigation.updated_at,
+        )
[…]

[…]
