### 2026-04-20 e2ec878
Add exclused reports feature and collections parity
