+++ django_workbench_v02/wb_collections/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path
+
+from . import views
+
+
+urlpatterns = [
+    path("collections/", views.collection_list, name="collection-list"),
+    path("collections/<slug:collection_slug>/", views.collection_detail, name="collection-detail"),
+    path("collections/<slug:collection_slug>/copy/", views.collection_copy, name="collection-copy"),
+]
