### 2026-04-20 40be499
Add collections to v0.2
