+++ django_workbench_v02/wb_collections/services.py
@@ -1,12 +1,15 @@
 from __future__ import annotations
 
 from datetime import datetime
+import json
+from pathlib import Path
 from threading import Lock
 from typing import Any
+import logging
 
 import pandas as pd
 from django.db import transaction
-from django.utils import timezone
+from django.utils.text import slugify
 from pfd_toolkit import load_reports
 from pfd_toolkit.collections import COLLECTION_COLUMNS, apply_collection_columns
 
@@ -20,12 +23,88 @@ _COLLECTIONS_CACHE_LOCK = Lock()
 _COLLECTIONS_CACHE_DF: pd.DataFrame | None = None
 _COLLECTIONS_CACHE_UPDATED_AT: datetime | None = None
 _COLLECTIONS_CACHE_TTL_SECONDS = 6 * 60 * 60
+logger = logging.getLogger(__name__)
+
+THEME_COLLECTION_SLUG_PREFIX = "theme"
+THEME_COLLECTION_TITLE_OVERRIDES: dict[str, str] = {
+    "suicide_risk": "Suicide",
+    "care_home_safety": "Care home deaths",
+    "acute_hospital_wards": "Acute Hospital Care",
+    "alarm_alert_failures": "Alarms and Alerts",
+    "capacity_best_interests_failures": "Capacity and Best Interests",
+    "consent_decision_making_failures": "Consent and Decision-making",
+    "failure_recognise_escalate_deterioration": "Recognising and Escalating Deterioration",
+    "family_carer_concerns_not_acted_on": "Family and Carer Concerns",
+    "failure_learn_previous_deaths_incidents": "Failure to Learn",
+    "follow_up_failures": "Follow-up",
+    "housing_homelessness": "Housing and Homelessness",
+    "inter_agency_working": "Inter-agency Working",
+    "investigation_incident_review_failures": "Investigations and Incident Reviews",
+    "it_digital_system_failures": "IT and Digital Systems",
+    "language_interpreter_barriers": "Language and Interpreter Barriers",
+    "maternity_neonatal_perinatal_care": "Maternity, Neonatal and Perinatal Care",
+    "children_young_people": "Children and Young People",
+    "missed_appointments_non_attendance": "Missed Appointments and Non-attendance",
+    "observation_monitoring_failures": "Observation and Monitoring",
+    "policy_procedure_failures": "Policy and Procedure",
+    "record_sharing_failures": "Record Sharing",
+    "referral_failures": "Referrals",
+    "risk_assessment_failures": "Risk Assessment",
+    "roads_highways": "Roads and Highways",
+    "sepsis_infection": "Sepsis and Infection",
+    "staffing_shortages_workload_pressure": "Staffing Shortages and Workload Pressure",
+    "suicide_self_harm": "Suicide and Self-harm",
+    "test_result_management_failures": "Test Result Management",
+    "training_competence_gaps": "Training and Competence",
+    "transitions_discharge_failures": "Transitions and Discharge",
+    "violence_homicide_related_systems_failures": "Violence and Homicide-related Systems",
+    "falls_frailty": "Falls and Frailty",
+    "equipment_failures": "Equipment",
+    "environmental_design_failures": "Environmental Design",
+}
+EXCLUDED_THEME_KEYS: set[str] = {
+    "policy_procedure_failures",
+    "failure_recognise_escalate_deterioration",
+    "nutrition",
+}
+APPROVED_THEME_SCHEMA_PATH = (
+    Path(__file__).resolve().parents[2] / "scripts" / "theme_collections" / "approved_themes.json"
+)
+_APPROVED_THEME_COLUMNS_CACHE: set[str] | None = None
+_APPROVED_THEME_COLUMNS_MTIME_NS: int | None = None
 
 
 def _cache_stale(now_utc: datetime) -> bool:
[…]
