### 2026-04-20 c16febf
Add workbook lifecycle controls
