+++ django_workbench_v02/wb_workspaces/lifecycle.py
@@ -14,11 +14,14 @@ from .models import Workspace
 @dataclass
 class LifecycleMaintenanceResult:
     inactivity_days: int
+    archive_retention_days: int
     artifacts_scanned: int = 0
     artifacts_expired: int = 0
     artifacts_expiry_refreshed: int = 0
     workspaces_scanned: int = 0
     workspaces_archived: int = 0
+    archived_workspaces_scanned: int = 0
+    workspaces_purged: int = 0
 
 
 def _last_activity_at(*, last_viewed_at, created_at):
@@ -29,12 +32,18 @@ def run_lifecycle_maintenance(
     *,
     now=None,
     inactivity_days: int = 365,
+    archive_retention_days: int = 60,
     dry_run: bool = False,
     archive_workspaces: bool = True,
+    purge_archived_workspaces: bool = True,
 ) -> LifecycleMaintenanceResult:
     current_time = now or timezone.now()
     inactivity_window = timedelta(days=inactivity_days)
-    result = LifecycleMaintenanceResult(inactivity_days=inactivity_days)
+    archive_retention_window = timedelta(days=archive_retention_days)
+    result = LifecycleMaintenanceResult(
+        inactivity_days=inactivity_days,
+        archive_retention_days=archive_retention_days,
+    )
 
     ready_artifacts = RunArtifact.objects.select_related("workspace").filter(status=ArtifactStatus.READY)
     for artifact in ready_artifacts.iterator():
@@ -77,36 +86,70 @@ def run_lifecycle_maintenance(
             artifact.expires_at = effective_expires_at
             artifact.save(update_fields=["expires_at", "updated_at"])
 
-    if not archive_workspaces:
+    if not archive_workspaces and not purge_archived_workspaces:
         return result
 
-    active_workspaces = Workspace.objects.filter(archived_at__isnull=True)
-    for workspace in active_workspaces.iterator():
-        result.workspaces_scanned += 1
-        last_activity_at = _last_activity_at(
-            last_viewed_at=workspace.last_viewed_at,
-            created_at=workspace.created_at,
-        )
-        if (last_activity_at + inactivity_window) > current_time:
-            continue
+    if archive_workspaces:
+        active_workspaces = Workspace.objects.filter(archived_at__isnull=True)
+        for workspace in active_workspaces.iterator():
+            result.workspaces_scanned += 1
+            last_activity_at = _last_activity_at(
+                last_viewed_at=workspace.last_viewed_at,
+                created_at=workspace.created_at,
+            )
+            if (last_activity_at + inactivity_window) > current_time:
+                continue
 
-        result.workspaces_archived += 1
-        if dry_run:
-            continue
+            result.workspaces_archived += 1
+            if dry_run:
+                continue
 
-        workspace.archived_at = current_time
-        workspace.save(update_fields=["archived_at", "updated_at"])
-        log_audit_event(
-            action_type="workspace.auto_archived_inactive",
-            target_type="workspace",
-            target_id=str(workspace.id),
-            workspace=workspace,
-            user=None,
-            payload={
-                "last_activity_at": last_activity_at.isoformat(),
-                "archived_at": current_time.isoformat(),
-                "inactivity_days": inactivity_days,
-            },
-        )
+            workspace.archived_at = current_time
+            workspace.save(update_fields=["archived_at", "updated_at"])
+            log_audit_event(
+                action_type="workspace.auto_archived_inactive",
+                target_type="workspace",
+                target_id=str(workspace.id),
+                workspace=workspace,
+                user=None,
+                payload={
+                    "last_activity_at": last_activity_at.isoformat(),
+                    "archived_at": current_time.isoformat(),
+                    "inactivity_days": inactivity_days,
+                },
+            )
+
+    if purge_archived_workspaces:
+        from .models import WorkspaceUserState
+
[…]
