### 2026-04-20 e9f55ec
Add performance tweaks, plan, and monitoring
