+++ django_workbench_v02/wb_workspaces/migrations/0005_workspace_idx_ws_arch_upd_desc.py
@@ -1,0 +1,21 @@
+# Generated by Django 5.2.12 on 2026-04-20 13:13
+
+from django.conf import settings
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("wb_workspaces", "0004_workspaceuserstate"),
+        migrations.swappable_dependency(settings.AUTH_USER_MODEL),
+    ]
+
+    operations = [
+        migrations.AddIndex(
+            model_name="workspace",
+            index=models.Index(
+                fields=["archived_at", "-updated_at"], name="idx_ws_arch_upd_desc"
+            ),
+        ),
+    ]
