+++ django_workbench_v02/wb_runs/migrations/0002_investigationrun_idx_run_stat_queue_cr_and_more.py
@@ -1,0 +1,30 @@
+# Generated by Django 5.2.12 on 2026-04-20 13:13
+
+from django.conf import settings
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("wb_investigations", "0002_investigation_uniq_investigation_per_workspace"),
+        ("wb_runs", "0001_initial"),
+        ("wb_workspaces", "0005_workspace_idx_ws_arch_upd_desc"),
+        migrations.swappable_dependency(settings.AUTH_USER_MODEL),
+    ]
+
+    operations = [
+        migrations.AddIndex(
+            model_name="investigationrun",
+            index=models.Index(
+                fields=["status", "queued_at", "created_at"],
+                name="idx_run_stat_queue_cr",
+            ),
+        ),
+        migrations.AddIndex(
+            model_name="runartifact",
+            index=models.Index(
+                fields=["status", "expires_at"], name="idx_art_stat_exp"
+            ),
+        ),
+    ]
