### 2026-04-20 0709f93
Implement workbook revision cursor and state actions (undo/redo/start-over/revert-reports) with share snapshot compatibility
