+++ django_workbench_v02/wb_workspaces/migrations/0006_workspace_current_revision.py
@@ -1,0 +1,25 @@
+# Generated by Django 5.2.12 on 2026-04-20 13:29
+
+import django.db.models.deletion
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("wb_workspaces", "0005_workspace_idx_ws_arch_upd_desc"),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name="workspace",
+            name="current_revision",
+            field=models.ForeignKey(
+                blank=True,
+                null=True,
+                on_delete=django.db.models.deletion.SET_NULL,
+                related_name="cursor_workspaces",
+                to="wb_workspaces.workspacerevision",
+            ),
+        ),
+    ]

[…]
