### 2026-04-20 c6474fa
Implement ActionCacheEvent transparency layer with major workflow hooks and workbook inspection view
