+++ django_workbench_v02/wb_auditlog/services.py
@@ -4,7 +4,7 @@ import hashlib
 
 from django.http import HttpRequest
 
-from .models import AuditEvent
+from .models import ActionCacheEvent, AuditEvent
 
 
 def _extract_ip(request: HttpRequest) -> str:
@@ -45,4 +45,31 @@ def log_audit_event(
         ip_hash=ip_hash,
         user_agent=user_agent,
         payload_json=payload or {},
+    )
+
+
+def log_action_cache_event(
+    *,
+    workspace,
+    action_key: str,
+    entity_type: str,
+    entity_id: str,
+    user=None,
+    query: dict | None = None,
+    options: dict | None = None,
+    state_before: dict | None = None,
+    state_after: dict | None = None,
+    context: dict | None = None,
+) -> ActionCacheEvent:
+    return ActionCacheEvent.objects.create(
+        workspace=workspace,
+        user=user if user and getattr(user, "is_authenticated", False) else None,
+        action_key=action_key,
+        entity_type=entity_type,
+        entity_id=str(entity_id),
+        query_json=query or {},
+        options_json=options or {},
+        state_before_json=state_before or {},
+        state_after_json=state_after or {},
+        context_json=context or {},
     )
