+++ django_workbench_v02/wb_auditlog/migrations/0002_actioncacheevent.py
@@ -1,0 +1,81 @@
+# Generated by Django 5.2.12 on 2026-04-20 13:39
+
+import django.db.models.deletion
+import django.utils.timezone
+import uuid
+from django.conf import settings
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("wb_auditlog", "0001_initial"),
+        ("wb_workspaces", "0006_workspace_current_revision"),
+        migrations.swappable_dependency(settings.AUTH_USER_MODEL),
+    ]
+
+    operations = [
+        migrations.CreateModel(
+            name="ActionCacheEvent",
+            fields=[
+                (
+                    "id",
+                    models.UUIDField(
+                        default=uuid.uuid4,
+                        editable=False,
+                        primary_key=True,
+                        serialize=False,
+                    ),
+                ),
+                ("action_key", models.CharField(max_length=100)),
+                ("entity_type", models.CharField(max_length=100)),
+                ("entity_id", models.CharField(max_length=100)),
+                ("query_json", models.JSONField(default=dict)),
+                ("options_json", models.JSONField(default=dict)),
+                ("state_before_json", models.JSONField(default=dict)),
+                ("state_after_json", models.JSONField(default=dict)),
+                ("context_json", models.JSONField(default=dict)),
+                (
+                    "created_at",
+                    models.DateTimeField(
+                        default=django.utils.timezone.now, editable=False
+                    ),
+                ),
+                (
+                    "user",
+                    models.ForeignKey(
+                        blank=True,
+                        null=True,
+                        on_delete=django.db.models.deletion.SET_NULL,
+                        related_name="action_cache_events",
+                        to=settings.AUTH_USER_MODEL,
+                    ),
+                ),
+                (
+                    "workspace",
+                    models.ForeignKey(
+                        on_delete=django.db.models.deletion.CASCADE,
+                        related_name="action_cache_events",
+                        to="wb_workspaces.workspace",
+                    ),
+                ),
+            ],
+            options={
+                "ordering": ["-created_at"],
+                "indexes": [
+                    models.Index(
+                        fields=["workspace", "-created_at"],
+                        name="idx_ac_workspace_created",
+                    ),
+                    models.Index(
+                        fields=["action_key", "-created_at"],
+                        name="idx_ac_action_created",
+                    ),
+                    models.Index(
+                        fields=["entity_type", "entity_id"], name="idx_ac_entity"
+                    ),
+                ],
+            },
+        ),
+    ]
