### 2026-04-20 805362f
Add server-side wizard flow, chained pipeline orchestration, and timeline/status UX
