+++ django_workbench_v02/wb_investigations/forms.py
@@ -1,5 +1,14 @@
-from django import forms
+from __future__ import annotations
 
+from dataclasses import dataclass, field
+from datetime import date, timedelta
+
+from django import forms
+from django.db import models
+
+from wb_notifications.models import NotificationTrigger
+from wb_runs.models import RunType
+from wb_workspaces.models import WorkspaceLLMProvider
 from .models import InvestigationStatus
 
 
@@ -20,3 +29,239 @@ class InvestigationUpdateForm(forms.Form):
     scope_json = forms.JSONField(required=False, initial=dict)
     method_json = forms.JSONField(required=False, initial=dict)
     status = forms.ChoiceField(choices=InvestigationStatus.choices)
+
+
+class TemporalScopeOption(models.TextChoices):
+    ALL_REPORTS = "all_reports", "All reports"
+    LAST_3_YEARS = "last_3_years", "Last 3 years"
+    LAST_YEAR = "last_year", "Last year"
+    LAST_6_MONTHS = "last_6_months", "Last 6 months"
+    MOST_RECENT_100 = "most_recent_100", "100 most recent reports"
+
+
+WIZARD_STAGES = (
+    "question",
+    "scope",
+    "method",
+    "themes",
+    "extract",
+    "review",
+)
+
+
+def _subtract_years(value: date, years: int) -> date:
+    try:
+        return value.replace(year=value.year - years)
+    except ValueError:
+        # Handles leap-day rollover.
+        return value.replace(month=2, day=28, year=value.year - years)
+
+
+def temporal_scope_parameters(*, scope_option: str, today: date | None = None) -> dict:
+    current_date = today or date.today()
+    option = str(scope_option or "").strip()
+
+    if option == TemporalScopeOption.LAST_3_YEARS:
+        return {
+            "query_start_date": _subtract_years(current_date, 3),
+            "query_end_date": current_date,
+            "report_limit": None,
+            "scope_option": option,
+        }
+    if option == TemporalScopeOption.LAST_YEAR:
+        return {
+            "query_start_date": _subtract_years(current_date, 1),
+            "query_end_date": current_date,
+            "report_limit": None,
+            "scope_option": option,
+        }
+    if option == TemporalScopeOption.LAST_6_MONTHS:
+        return {
+            "query_start_date": current_date - timedelta(days=183),
+            "query_end_date": current_date,
+            "report_limit": None,
+            "scope_option": option,
+        }
+    if option == TemporalScopeOption.MOST_RECENT_100:
+        return {
+            "query_start_date": None,
+            "query_end_date": None,
+            "report_limit": 100,
+            "scope_option": option,
+        }
+    return {
+        "query_start_date": None,
+        "query_end_date": None,
+        "report_limit": None,
+        "scope_option": TemporalScopeOption.ALL_REPORTS,
+    }
+
+
+def build_pipeline_plan(*, run_themes: bool, run_extract: bool) -> list[str]:
+    plan = [RunType.FILTER]
+    if run_themes:
+        plan.append(RunType.THEMES)
+    if run_extract:
+        plan.append(RunType.EXTRACT)
+    return plan
+
+
+class InvestigationWizardQuestionForm(forms.Form):
+    title = forms.CharField(max_length=255)
+    question_text = forms.CharField(widget=forms.Textarea)
+
+    def clean_question_text(self):
+        value = str(self.cleaned_data.get("question_text") or "").strip()
+        if not value:
+            raise forms.ValidationError("Research question is required.")
+        return value
+
+
+class InvestigationWizardScopeForm(forms.Form):
+    scope_option = forms.ChoiceField(
+        choices=TemporalScopeOption.choices,
+        initial=TemporalScopeOption.ALL_REPORTS,
+    )
+
+    def resolved_scope(self, *, today: date | None = None) -> dict:
[…]
