+++ django_workbench_v02/wb_investigations/views.py
@@ -4,14 +4,419 @@ from django.core.exceptions import PermissionDenied, ValidationError
 from django.shortcuts import get_object_or_404, redirect, render
 from django.views.decorators.http import require_GET, require_http_methods, require_POST
 
-from wb_runs.forms import RunQueueForm
-from wb_runs.models import InvestigationRun
+from wb_runs.models import InvestigationRun, RunStatus, RunType
+from wb_notifications.services import NotificationRequestError, create_notification_request
 from wb_workspaces.models import Workspace
 from wb_workspaces.permissions import can_edit_workspace, can_view_workspace
 
-from .forms import InvestigationCreateForm, InvestigationUpdateForm
+from .forms import (
+    InvestigationCreateForm,
+    InvestigationUpdateForm,
+    InvestigationWizardExtractConfigForm,
+    InvestigationWizardMethodForm,
+    InvestigationWizardQuestionForm,
+    InvestigationWizardReviewForm,
+    InvestigationWizardScopeForm,
+    InvestigationWizardState,
+    InvestigationWizardThemesConfigForm,
+)
 from .models import Investigation
-from .services import create_investigation, record_investigation_view, update_investigation
+from .services import (
+    InvestigationServiceError,
+    create_investigation,
+    launch_investigation_wizard_pipeline,
+    record_investigation_view,
+    update_investigation,
+)
+
+
+PIPELINE_STAGE_LABELS = {
+    RunType.FILTER: "Screen and filter",
+    RunType.THEMES: "Discover themes",
+    RunType.EXTRACT: "Extract structured data",
+}
+
+TERMINAL_STATUSES = {
+    RunStatus.SUCCEEDED,
+    RunStatus.FAILED,
+    RunStatus.TIMED_OUT,
+    RunStatus.CANCELLED,
+}
+
+FAILED_STATUSES = {RunStatus.FAILED, RunStatus.TIMED_OUT}
+
+
+def _normalise_pipeline_plan(raw_plan) -> list[str]:
+    if not isinstance(raw_plan, list):
+        return []
+    allowed = {RunType.FILTER, RunType.THEMES, RunType.EXTRACT}
+    result: list[str] = []
+    for raw in raw_plan:
+        value = str(raw or "").strip().lower()
+        if value in allowed:
+            result.append(value)
+    return result
+
+
+def _pipeline_index_for_run(run: InvestigationRun, *, plan: list[str]) -> int:
+    config = run.input_config_json if isinstance(run.input_config_json, dict) else {}
+    raw_index = config.get("pipeline_index")
+    try:
+        index = int(raw_index)
+    except (TypeError, ValueError):
+        index = -1
+    if 0 <= index < len(plan):
+        return index
+    try:
+        return plan.index(str(run.run_type or "").strip().lower())
+    except ValueError:
+        return -1
+
+
+def _previous_pipeline_run_id(run: InvestigationRun) -> str:
+    for event in run.events.all():
+        payload = event.payload_json if isinstance(event.payload_json, dict) else {}
+        previous_id = str(payload.get("pipeline_previous_run_id") or "").strip()
+        if previous_id:
+            return previous_id
+    return ""
+
+
+def _build_pipeline_timeline(*, investigation: Investigation) -> dict | None:
+    runs = list(
+        InvestigationRun.objects.filter(investigation=investigation)
+        .prefetch_related("events")
+        .order_by("created_at")
+    )
+    if not runs:
+        return None
+
+    root_candidates: list[InvestigationRun] = []
+    for run in runs:
+        config = run.input_config_json if isinstance(run.input_config_json, dict) else {}
+        plan = _normalise_pipeline_plan(config.get("pipeline_plan"))
+        if not plan:
+            continue
[…]
