### 2026-04-20 2401a11
Add Investigations to UI
