+++ django_workbench_v02/wb_investigations/tests.py
@@ -128,12 +128,22 @@ class InvestigationServiceTests(TestCase):
 class InvestigationViewTests(TestCase):
     def setUp(self):
         self.owner = User.objects.create_user(email="inv-owner2@example.com", password="x")
+        self.viewer = User.objects.create_user(email="inv-viewer2@example.com", password="x")
         self.stranger = User.objects.create_user(email="inv-stranger2@example.com", password="x")
         self.workspace = create_workspace_for_user(
             user=self.owner,
             title="Inv Workspace2",
             slug="inv-workspace2",
             description="desc",
+        )
+        WorkspaceMembership.objects.create(
+            workspace=self.workspace,
+            user=self.viewer,
+            role=MembershipRole.VIEWER,
+            access_mode=MembershipAccessMode.READ_ONLY,
+            can_manage_members=False,
+            can_manage_shares=False,
+            can_run_workflows=False,
         )
         self.investigation = create_investigation(
             actor=self.owner,
@@ -178,6 +188,46 @@ class InvestigationViewTests(TestCase):
             reverse("workbook-investigation-list", kwargs={"workbook_id": self.workspace.id})
         )
         self.assertEqual(response.status_code, 200)
+
+    def test_investigation_entry_redirects_editor_to_wizard(self):
+        self.client.force_login(self.owner)
+        response = self.client.get(
+            reverse("workbook-investigation-entry", kwargs={"workbook_id": self.workspace.id})
+        )
+        self.assertEqual(response.status_code, 302)
+        self.assertIn("/wizard/", response.url)
+
+    def test_investigation_entry_redirects_read_only_user_to_detail(self):
+        self.client.force_login(self.viewer)
+        response = self.client.get(
+            reverse("workbook-investigation-entry", kwargs={"workbook_id": self.workspace.id})
+        )
+        self.assertEqual(response.status_code, 302)
+        self.assertIn("/investigations/", response.url)
+        self.assertNotIn("/wizard/", response.url)
+
+    def test_investigation_entry_creates_investigation_for_editor_when_missing(self):
+        self.client.force_login(self.owner)
+        workspace_without_investigation = create_workspace_for_user(
+            user=self.owner,
+            title="No Investigation Yet",
+            slug="no-investigation-yet",
+            description="desc",
+        )
+        self.assertFalse(
+            Investigation.objects.filter(workspace=workspace_without_investigation).exists()
+        )
+        response = self.client.get(
+            reverse(
+                "workbook-investigation-entry",
+                kwargs={"workbook_id": workspace_without_investigation.id},
+            )
+        )
+        self.assertEqual(response.status_code, 302)
+        self.assertIn("/wizard/", response.url)
+        self.assertTrue(
+            Investigation.objects.filter(workspace=workspace_without_investigation).exists()
+        )
 
     def test_bot_investigation_detail_view_does_not_update_last_viewed(self):
         self.workspace.visibility = WorkspaceVisibility.PUBLIC
@@ -361,10 +411,14 @@ class InvestigationWizardFormTests(TestCase):
 
     def test_method_form_pipeline_plan_fixed_order(self):
         form = InvestigationWizardMethodForm(
-            data={"run_themes": "on", "run_extract": "on"}
+            data={"run_filter": "on", "run_themes": "on", "run_extract": "on"}
         )
         self.assertTrue(form.is_valid())
         self.assertEqual(form.pipeline_plan(), ["filter", "themes", "extract"])
+
+    def test_method_form_requires_at_least_one_stage(self):
+        form = InvestigationWizardMethodForm(data={})
+        self.assertFalse(form.is_valid())
 
     def test_themes_config_rejects_invalid_min_max(self):
         form = InvestigationWizardThemesConfigForm(
[…]
