### 2026-04-20 7117e89
Improve display for investigation runs
