+++ django_workbench_v02/wb_runs/views.py
@@ -1,3 +1,5 @@
+import json
+
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import PermissionDenied, ValidationError
@@ -17,13 +19,255 @@ from wb_workspaces.services import (
 
 from .artifact_storage import ArtifactStorageError, open_artifact_for_download
 from .forms import RunCancelForm, RunQueueForm
-from .models import ArtifactStatus, InvestigationRun, RunArtifact
+from .models import ArtifactStatus, InvestigationRun, RunArtifact, RunStatus
 from .services import (
     queue_run,
     request_run_cancellation,
     record_artifact_download,
     record_run_view,
 )
+
+TERMINAL_STATUSES = {
+    RunStatus.CANCELLED,
+    RunStatus.SUCCEEDED,
+    RunStatus.FAILED,
+    RunStatus.TIMED_OUT,
+}
+
+RUN_STATUS_LABELS = dict(RunStatus.choices)
+ARTIFACT_TYPE_LABELS = dict(RunArtifact._meta.get_field("artifact_type").choices)
+RUN_TYPE_LABELS = dict(InvestigationRun._meta.get_field("run_type").choices)
+
+ARTIFACT_INTENT_BY_TYPE = {
+    "filtered_dataset": "Matched report subset after filtering.",
+    "theme_summary": "Topline theme outputs from the theme workflow.",
+    "theme_assignments": "Per-report theme assignment table.",
+    "extraction_table": "Structured extract table generated from prompts.",
+    "bundle_export": "Packaged workbook export bundle.",
+    "preview": "Preview output produced during processing.",
+}
+
+CONFIG_SUMMARY_KEYS = [
+    ("execution_mode", "Execution mode"),
+    ("provider", "Provider"),
+    ("model_name", "Model"),
+    ("collection_slug", "Collection"),
+    ("collection_query", "Collection query"),
+    ("query_text", "Query text"),
+    ("search_query", "Search query"),
+    ("pipeline_plan", "Pipeline plan"),
+    ("pipeline_continue_on_fail", "Continue on fail"),
+    ("pipeline_index", "Pipeline index"),
+    ("input_artifact_id", "Upstream artifact"),
+]
+
+CONFIG_SKIP_KEYS = {
+    "report_identity_allowlist",
+    "excluded_report_identities",
+    "selected_filters",
+}
+
+
+def _format_size(size_bytes):
+    if size_bytes is None:
+        return "-"
+    units = ["B", "KB", "MB", "GB"]
+    value = float(size_bytes)
+    unit_index = 0
+    while value >= 1024 and unit_index < len(units) - 1:
+        value /= 1024
+        unit_index += 1
+    if unit_index == 0:
+        return f"{int(value)} {units[unit_index]}"
+    return f"{value:.1f} {units[unit_index]}"
+
+
+def _format_config_value(value):
+    if isinstance(value, bool):
+        return "yes" if value else "no"
+    if value is None:
+        return "-"
+    if isinstance(value, list):
+        if not value:
+            return "none"
+        if len(value) <= 4:
+            return ", ".join(str(item) for item in value)
+        return f"{len(value)} values"
+    if isinstance(value, dict):
+        if not value:
+            return "none"
+        keys = list(value.keys())
+        preview = ", ".join(str(item) for item in keys[:3])
+        if len(keys) > 3:
+            preview = f"{preview}, +{len(keys) - 3} more"
+        return f"{len(keys)} keys ({preview})"
+    return str(value)
+
+
+def _humanize_config_key(key):
+    return key.replace("_", " ").capitalize()
+
+
+def _build_run_journey(run):
+    steps = [
+        {
+            "label": "Queued",
+            "state": "done",
[…]
