### 2026-04-20 2bd99dd
Unify run scope resolution, lock thematic collection schema behavior, and preserve collection-copy scope context
