+++ django_workbench_v02/wb_collections/tests.py
@@ -5,6 +5,7 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.urls import reverse
 
+from wb_investigations.models import Investigation
 from wb_workspaces.models import Workspace
 
 User = get_user_model()
@@ -27,6 +28,7 @@ class CollectionViewTests(TestCase):
                     "collection_wales": True,
                     "collection_nhs": True,
                     "theme_medication_safety": True,
+                    "theme_unapproved_x": True,
                 },
                 {
                     "date": "2024-01-02",
@@ -40,6 +42,7 @@ class CollectionViewTests(TestCase):
                     "collection_wales": False,
                     "collection_nhs": False,
                     "theme_medication_safety": False,
+                    "theme_unapproved_x": False,
                 },
             ]
         )
@@ -65,16 +68,35 @@ class CollectionViewTests(TestCase):
         self.assertContains(response, "Circumstances")
 
     @patch("wb_collections.services.load_reports")
-    def test_collection_copy_creates_workspace_only(self, mock_load_reports):
+    def test_collection_copy_creates_workspace_and_investigation_scope(self, mock_load_reports):
         mock_load_reports.return_value = self.dataset.copy()
         self.client.force_login(self.user)
         response = self.client.post(
             reverse("collection-copy", kwargs={"collection_slug": "custom-search"}),
-            data={"q": "medication", "workbook_title": "Medication Copy"},
+            data={
+                "q": "medication",
+                "coroner": ["A"],
+                "area": ["Wales"],
+                "receiver": ["NHS Trust"],
+                "workbook_title": "Medication Copy",
+            },
         )
         self.assertEqual(response.status_code, 302)
         workspace = Workspace.objects.filter(created_by=self.user).latest("created_at")
         self.assertEqual(workspace.title, "Medication Copy")
+        investigation = Investigation.objects.get(workspace=workspace)
+        self.assertEqual(investigation.scope_json.get("collection_slug"), "custom-search")
+        self.assertEqual(investigation.scope_json.get("collection_query"), "medication")
+        self.assertEqual(
+            investigation.scope_json.get("selected_filters"),
+            {
+                "coroner": ["A"],
+                "area": ["Wales"],
+                "receiver": ["NHS Trust"],
+            },
+        )
+        allowlist = investigation.scope_json.get("report_identity_allowlist") or []
+        self.assertEqual(len(allowlist), 1)
 
     @patch("wb_collections.services.load_reports")
     def test_collection_list_includes_theme_collection_cards(self, mock_load_reports):
@@ -82,6 +104,30 @@ class CollectionViewTests(TestCase):
         response = self.client.get(reverse("collection-list"))
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "Medication Safety")
+
+    @patch("wb_collections.services._approved_theme_columns_with_status")
+    @patch("wb_collections.services.load_reports")
+    def test_collection_list_locks_to_approved_theme_schema(self, mock_load_reports, mock_approved_columns):
+        mock_load_reports.return_value = self.dataset.copy()
+        mock_approved_columns.return_value = (["theme_medication_safety"], True)
+        response = self.client.get(reverse("collection-list"))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "Medication Safety")
+        self.assertNotContains(response, "Unapproved X")
+
+    @patch("wb_collections.services._approved_theme_columns_with_status")
+    @patch("wb_collections.services.load_reports")
+    def test_collection_list_falls_back_to_discovered_when_schema_invalid(
+        self,
+        mock_load_reports,
+        mock_approved_columns,
+    ):
[…]

[…]
