### 2026-04-20 c26c9a1
Add “collection cards snapshot” (DB/JSON artifact) for Collections
