### 2026-04-20 d13a6ee
Update email notification templates
