### 2026-04-20 6ab8a37
Add base template shell and global stylesheet & tokens
