+++ django_workbench_v02/pfd_workbench_v02/settings.py
@@ -211,6 +211,7 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/5.2/howto/static-files/
 
 STATIC_URL = "static/"
+STATICFILES_DIRS = [BASE_DIR / "static"]
 
 CSRF_TRUSTED_ORIGINS = [
     origin.strip()
