+++ django_workbench_v02/docs/internal/BACKEND_CONTRACTS_UI_PORT.md
@@ -1,0 +1,322 @@
+# Backend Contracts for UI Port (v0.2)
+
+Status: Locked for Claude-design UI port  
+Last updated: 2026-04-20
+
+## 1. Purpose
+
+This document freezes the backend behavior contracts that the UI will consume during the design port.
+
+Rule:
+
+1. UI refactors should consume these contracts.
+2. Backend contract changes require explicit doc + test updates in the same PR.
+
+## 2. Contract Scope
+
+Locked in this file:
+
+1. Investigation wizard flow and launch payload behavior.
+2. Run detail page data contract.
+3. Workbook dashboard/list and investigation entry behavior.
+4. Completion notification request/dispatch contract.
+
+Out of scope (can change freely in UI work):
+
+1. Layout, typography, color, animation, component styling.
+2. Copy polish, as long as meaning and validation behavior stay intact.
+
+## 3. Investigation Wizard Contract
+
+Primary route:
+
+1. `GET/POST /workbooks/<workbook_id>/investigations/<investigation_id>/wizard/`
+
+Permission gate:
+
+1. Requires `can_edit_workspace`.
+
+Source files:
+
+1. `wb_investigations/views.py`
+2. `wb_investigations/forms.py`
+3. `templates/wb_investigations/investigation_wizard.html`
+
+### 3.1 Stage Model
+
+Canonical stage ids:
+
+1. `question`
+2. `scope`
+3. `method`
+4. `themes` (conditional)
+5. `extract` (conditional)
+6. `review`
+
+Sequence rule:
+
+1. Fixed order.
+2. `themes` appears only when `run_themes=true`.
+3. `extract` appears only when `run_extract=true`.
+
+### 3.2 Stage Inputs and Validation
+
+`question`:
+
+1. `title` (required)
+2. `question_text` (required, non-empty)
+
+`scope`:
+
+1. `scope_option` in:
+   - `all_reports`
+   - `last_3_years`
+   - `last_year`
+   - `last_6_months`
+   - `most_recent_100`
+
+`method`:
+
+1. `run_filter` (bool, default true)
+2. `run_themes` (bool, default false)
+3. `run_extract` (bool, default false)
+4. At least one of filter/themes/extract must be true.
+
+`themes`:
+
+1. `seed_topics`
+2. `min_themes`
+3. `max_themes`
+4. `extra_theme_instructions`
+5. Validation: if both min/max exist, `min_themes <= max_themes`.
+
+`extract`:
+
+1. `feature_fields` required when extract enabled.
+2. Each field row must contain:
+   - `name`
+   - `description`
+   - `type`
+3. Supports line format `name | description | type` and legacy compatibility parsing.
+4. Additional flags:
+   - `allow_multiple`
+   - `force_assign`
+   - `skip_if_present`
+
+`review`:
+
+1. `execution_mode` (`real` or `simulate`)
+2. `provider` (`openai` or `openrouter`)
+3. `model_name`
+4. `request_completion_email` (bool)
+5. `notify_on` (`success`, `failure`, `any`)
+
+### 3.3 Launch Semantics
+
+Launch action:
+
+1. Triggered from `review` with `wizard_action=launch`.
+2. Updates investigation `title`, `question_text`, `scope_json`, `method_json`.
+3. Queues first run from pipeline plan index `0`.
+4. Pipeline is chained via run config:
+   - `pipeline_plan`
+   - `pipeline_index`
+   - `pipeline_continue_on_fail=true`
+
+Temporal scope mapping:
+
+1. `all_reports` -> no date bounds, no report limit.
[…]
