### 2026-04-20 5f33da9
Static asset plumbing
