### 2026-04-20 8db1098
Add Wizard template
