+++ django_workbench_v02/docs/internal/UI_GLOBAL_SHELL_PLAN.md
@@ -7,8 +7,9 @@ This stage is visual-only. It must not change business logic, permissions, run o
 
 ## Status
 - Stage 1 (global shell): complete.
-- Stage 2 (page-by-page port): in progress (Dashboard started).
+- Stage 2 (page-by-page port): in progress (Dashboard + Investigation Wizard ported).
 - Stage 3 (static asset plumbing): complete for web runtime.
+- Stage 4 (shared includes for nav/messages): complete.
 
 ## Constraints
 - Keep backend contracts locked per `docs/internal/BACKEND_CONTRACTS_UI_PORT.md`.
