### 2026-04-20 cca1139
Add html for collections, auth and shared pages
