+++ django_workbench_v02/accounts/urls.py
@@ -5,6 +5,10 @@ from . import views
 
 urlpatterns = [
     path("", views.landing, name="landing"),
+    path("explore/", views.explore, name="explore"),
+    path("about/", views.about, name="about"),
+    path("research/", views.research, name="research"),
+    path("settings/llm/", views.llm_config, name="llm-config"),
     path("auth/login/", views.auth_login, name="accounts-login"),
     path("auth/callback/", views.auth_callback, name="accounts-callback"),
     path("auth/logout/", views.auth_logout, name="accounts-logout"),
