### 2026-04-21 134f573
Configure Explore tab
