+++ django_workbench_v02/accounts/views.py
@@ -1,19 +1,396 @@
 from __future__ import annotations
 
 import secrets
+import logging
+from math import ceil
 from urllib.parse import urlencode
+from io import StringIO
 
+import pandas as pd
 import requests
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth import login, logout
-from django.http import HttpRequest, HttpResponse, HttpResponseBadRequest
+from django.http import HttpRequest, HttpResponse, HttpResponseBadRequest, QueryDict
 from django.shortcuts import redirect, render
 from django.urls import reverse
+from django.utils import timezone
 from django.utils.http import url_has_allowed_host_and_scheme
 from django.views.decorators.http import require_GET
 
 from .services import normalize_auth0_profile, sync_user_from_auth0
+from wb_collections.services import (
+    collection_cards,
+    build_explore_metrics,
+    load_collections_dataset,
+    reports_for_collection,
+    _split_receivers,
+    _normalise_dashboard_value,
+    _theme_collection_map_from_reports,
+    _network_truthy,
+)
+
+logger = logging.getLogger(__name__)
+
+EXPLORE_AI_FILTER_EXCLUDED_SLUGS = {
+    "wales",
+    "nhs",
+    "gov_department",
+    "prisons",
+    "health_regulators",
+}
+EXPLORE_REPORTS_PAGE_SIZE = 15
+EXPLORE_FILTERS_SESSION_KEY = "explore_filters_v1"
+EXPLORE_SHOW_REPORTS_SESSION_KEY = "explore_show_reports_v1"
+
+
+def _dedupe_values(raw_values: list[str]) -> list[str]:
+    values: list[str] = []
+    seen: set[str] = set()
+    for raw in raw_values:
+        cleaned = str(raw or "").strip()
+        if not cleaned:
+            continue
+        lowered = cleaned.casefold()
+        if lowered in seen:
+            continue
+        seen.add(lowered)
+        values.append(cleaned)
+    return values
+
+
+def _parse_explore_request_params(request: HttpRequest) -> dict[str, object]:
+    filter_keys = {"q", "ai_filter", "receiver", "area", "date_start", "date_end"}
+    has_explicit_filters = any(key in request.GET for key in filter_keys)
+
+    if has_explicit_filters:
+        return {
+            "query": str(request.GET.get("q") or "").strip(),
+            "ai_filter": str(request.GET.get("ai_filter") or "custom").strip() or "custom",
+            "selected_receivers": _dedupe_values(request.GET.getlist("receiver")),
+            "selected_areas": _dedupe_values(request.GET.getlist("area")),
+            "date_start_raw": str(request.GET.get("date_start") or "").strip(),
+            "date_end_raw": str(request.GET.get("date_end") or "").strip(),
+        }
+
+    session_payload = request.session.get(EXPLORE_FILTERS_SESSION_KEY, {})
+    if not isinstance(session_payload, dict):
+        session_payload = {}
+
+    return {
+        "query": str(session_payload.get("query") or "").strip(),
+        "ai_filter": str(session_payload.get("ai_filter") or "custom").strip() or "custom",
+        "selected_receivers": _dedupe_values(
+            [str(value or "").strip() for value in session_payload.get("selected_receivers", [])]
+            if isinstance(session_payload.get("selected_receivers", []), list)
+            else []
+        ),
+        "selected_areas": _dedupe_values(
+            [str(value or "").strip() for value in session_payload.get("selected_areas", [])]
+            if isinstance(session_payload.get("selected_areas", []), list)
+            else []
+        ),
+        "date_start_raw": str(session_payload.get("date_start_raw") or "").strip(),
[…]
