+++ django_workbench/workbench/views.py
@@ -308,7 +308,7 @@ THEME_SEMANTIC_ICON_RULES: tuple[tuple[tuple[str, ...], str], ...] = (
     (("rail",), "rail"),
     (("communication", "handover", "record_sharing", "language_interpreter", "inter_agency"), "communication"),
     (("record", "referral", "consent", "capacity_best_interests", "policy_procedure"), "documents"),
-    (("follow_up", "waiting_times", "missed_appointments", "repeated_missed_opportunities", "continuity"), "clock"),
+    (("follow_up", "waiting_times", "missed_appointments", "continuity"), "clock"),
     (("staffing", "training", "competence", "workload", "workplaces", "universities"), "staffing"),
     (("it_digital", "remote_digital", "alarm_alert", "test_result"), "digital"),
     (("safeguarding", "risk_assessment", "failure_recognise_escalate_deterioration"), "safeguarding"),
@@ -342,7 +342,6 @@ THEME_COLLECTION_TITLE_OVERRIDES: dict[str, str] = {
     "policy_procedure_failures": "Policy and Procedure",
     "record_sharing_failures": "Record Sharing",
     "referral_failures": "Referrals",
-    "risk_assessment_failures": "Risk Assessment",
     "roads_highways": "Roads and Highways",
     "sepsis_infection": "Sepsis and Infection",
     "staffing_shortages_workload_pressure": "Staffing Shortages and Workload Pressure",
