### 2026-04-21 e4ecbe2
Configure Collections
