+++ django_workbench_v02/wb_collections/tests.py
@@ -55,8 +55,10 @@ class CollectionViewTests(TestCase):
         refresh_collection_cards_snapshot(force_refresh_dataset=False)
         response = self.client.get(reverse("collection-list"))
         self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "All reports")
-        self.assertContains(response, "Custom collection")
+        self.assertContains(response, "Wales")
+        self.assertContains(response, "NHS Bodies")
+        self.assertNotContains(response, "All reports")
+        self.assertNotContains(response, "Custom collection")
 
     @patch("wb_collections.services.load_reports")
     def test_collection_detail_custom_search_filters_rows(self, mock_load_reports):
@@ -66,9 +68,47 @@ class CollectionViewTests(TestCase):
             {"q": "medication"},
         )
         self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "Matching reports: 1")
-        self.assertContains(response, "Investigation")
-        self.assertContains(response, "Circumstances")
+        self.assertContains(response, "Collection ·")
+        self.assertContains(response, "Filter")
+        self.assertContains(response, "Interactive report list")
+        self.assertNotContains(response, "AI filter:")
+
+    @patch("wb_collections.services.load_reports")
+    def test_collection_reports_panel_uses_collection_scope(self, mock_load_reports):
+        mock_load_reports.return_value = self.dataset.copy()
+        response = self.client.get(
+            reverse("collection-reports-panel", kwargs={"collection_slug": "custom-search"}),
+            {"q": "medication", "show_reports": "1"},
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "Showing 1-1 of 1")
+        self.assertContains(response, "Medication incident")
+        self.assertNotContains(response, "Road safety issue")
+
+    @patch("wb_collections.services.load_reports")
+    def test_collection_reports_panel_uses_selected_collection_slug(self, mock_load_reports):
+        mock_load_reports.return_value = self.dataset.copy()
+        response = self.client.get(
+            reverse("collection-reports-panel", kwargs={"collection_slug": "nhs"}),
+            {"show_reports": "1"},
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "Showing 1-1 of 1")
+        self.assertContains(response, "NHS Trust")
+        self.assertNotContains(response, "Local Authority")
+
+    @patch("wb_collections.services.load_reports")
+    def test_collection_export_csv_uses_collection_scope(self, mock_load_reports):
+        mock_load_reports.return_value = self.dataset.copy()
+        response = self.client.get(
+            reverse("collection-export-csv", kwargs={"collection_slug": "custom-search"}),
+            {"q": "medication"},
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response["Content-Type"], "text/csv; charset=utf-8")
+        csv_text = response.content.decode("utf-8")
+        self.assertIn("Medication incident", csv_text)
+        self.assertNotIn("Road safety issue", csv_text)
 
     @patch("wb_collections.services.load_reports")
     def test_collection_copy_creates_workspace_and_investigation_scope(self, mock_load_reports):
@@ -102,27 +142,38 @@ class CollectionViewTests(TestCase):
         self.assertEqual(len(allowlist), 1)
 
     @patch("wb_collections.services.load_reports")
-    def test_collection_list_includes_theme_collection_cards(self, mock_load_reports):
+    def test_collection_list_hides_only_all_and_custom_cards(self, mock_load_reports):
         mock_load_reports.return_value = self.dataset.copy()
         refresh_collection_cards_snapshot(force_refresh_dataset=False)
         response = self.client.get(reverse("collection-list"))
         self.assertEqual(response.status_code, 200)
[…]
