+++ django_workbench_v02/wb_collections/views.py
@@ -1,15 +1,28 @@
 from __future__ import annotations
 
+from io import StringIO
 import logging
+import hashlib
+from math import ceil
 
+import pandas as pd
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import get_object_or_404, redirect, render
+from django.http import HttpResponse, HttpResponseBadRequest
+from django.shortcuts import redirect, render
+from django.urls import reverse
+from django.utils import timezone
 from django.views.decorators.http import require_GET, require_POST
 
-from wb_workspaces.models import Workspace
-from wb_workspaces.permissions import can_edit_workspace
-
+from accounts.views import (
+    EXPLORE_REPORTS_PAGE_SIZE,
+    _build_explore_payload,
+    _coerce_positive_int,
+    _explore_column_label,
+    _explore_ordered_columns,
+    _explore_report_rows,
+    _explore_shared_querydict,
+)
 from .services import (
     collection_cards,
     copy_collection_to_workbook,
@@ -20,6 +33,216 @@ from .services import (
 )
 
 logger = logging.getLogger(__name__)
+
+_COLLECTION_BADGE_BY_SLUG: dict[str, str] = {
+    "custom": "Complete archive",
+    "custom-search": "Custom",
+    "wales": "Geography",
+    "local_gov": "Geography",
+    "nhs": "Health",
+    "health_regulators": "Health",
+    "gov_department": "Government",
+    "prisons": "Justice",
+}
+_COLLECTION_SPARK_COLOR_BY_BADGE: dict[str, str] = {
+    "Complete archive": "var(--accent)",
+    "Custom": "var(--accent-2)",
+    "Geography": "oklch(0.70 0.12 150)",
+    "Health": "oklch(0.65 0.15 200)",
+    "Government": "oklch(0.70 0.16 55)",
+    "Justice": "oklch(0.68 0.14 30)",
+    "Theme": "oklch(0.65 0.18 300)",
+}
+_COLLECTION_CARD_ORDER: tuple[str, ...] = (
+    "wales",
+    "nhs",
+    "gov_department",
+    "prisons",
+    "health_regulators",
+    "local_gov",
+)
+_COLLECTION_CARD_COPY: dict[str, dict[str, object]] = {
+    "custom": {
+        "title": "All reports",
+        "description": "The full PFD archive. Start here for a broad pattern search across every report, coroner, and receiver since 2013.",
+        "chips": [{"key": "", "value": "2013 → 2026"}, {"key": "", "value": "All coroners"}],
+        "footer": "up",
+    },
+    "custom-search": {
+        "title": "Custom collection · Active AI filter",
+        "description": "Your active AI filter. Reports matching your current Explore scope and query.",
+        "chips": [{"key": "filter", "value": "AI screened"}, {"key": "status", "value": "live"}],
+        "footer": "saved",
+    },
+    "wales": {
+        "title": "Wales",
+        "description": "Reports where the coroner area is a Welsh jurisdiction, including legacy area names recoded to current Welsh areas.",
+        "chips": [{"key": "areas", "value": "Welsh scope"}, {"key": "status", "value": "live"}],
+        "footer": "updated",
+    },
+    "nhs": {
+        "title": "NHS Bodies",
+        "description": "Reports addressed to NHS organisations including trusts, foundation trusts, integrated care boards, and health boards.",
+        "chips": [{"key": "scope", "value": "Receiver-linked"}, {"key": "status", "value": "live"}],
+        "footer": "up",
+    },
+    "gov_department": {
+        "title": "Government departments",
+        "description": "Reports addressed to UK government departments and central offices, including Home Office, Cabinet Office, and ministerial departments.",
[…]
