+++ django_workbench_v02/wb_investigations/tests.py
@@ -13,18 +13,28 @@ from wb_workspaces.models import (
     WorkspaceReportExclusion,
     WorkspaceVisibility,
 )
-from wb_workspaces.services import create_workspace_for_user, upsert_workspace_credential
+from wb_workspaces.services import (
+    create_workspace_for_user,
+    has_workspace_credential,
+    upsert_workspace_credential,
+)
 
 from .forms import (
     InvestigationWizardExtractConfigForm,
     InvestigationWizardMethodForm,
+    InvestigationWizardScopeForm,
     InvestigationWizardState,
     InvestigationWizardThemesConfigForm,
     TemporalScopeOption,
     temporal_scope_parameters,
 )
 from .models import Investigation, InvestigationStatus
-from .services import InvestigationServiceError, create_investigation, update_investigation
+from .services import (
+    InvestigationServiceError,
+    create_investigation,
+    evaluate_investigation_launch_readiness,
+    update_investigation,
+)
 
 
 User = get_user_model()
@@ -130,6 +140,93 @@ class InvestigationServiceTests(TestCase):
                 status=InvestigationStatus.DRAFT,
             )
 
+    def test_launch_readiness_blocks_unknown_provider(self):
+        inv = create_investigation(
+            actor=self.owner,
+            workspace=self.workspace,
+            title="Provider Guardrail",
+            question_text="Q",
+            scope_json={},
+            method_json={},
+            status=InvestigationStatus.DRAFT,
+        )
+        state = InvestigationWizardState(
+            stage="review",
+            title=inv.title,
+            question_text=inv.question_text,
+            run_filter=True,
+            review_config={"provider": "bogus-provider", "model_name": "gpt-4.1-mini"},
+        )
+        readiness = evaluate_investigation_launch_readiness(
+            actor=self.owner,
+            investigation=inv,
+            wizard_state=state,
+        )
+        self.assertFalse(readiness["can_launch"])
+        provider_model_check = next(
+            check for check in readiness["checks"] if check["key"] == "provider_model"
+        )
+        self.assertFalse(provider_model_check["ready"])
+        self.assertIn("not supported", provider_model_check["message"])
+
+    def test_launch_readiness_blocks_empty_model_name(self):
+        inv = create_investigation(
+            actor=self.owner,
+            workspace=self.workspace,
+            title="Model Guardrail",
+            question_text="Q",
+            scope_json={},
+            method_json={},
+            status=InvestigationStatus.DRAFT,
+        )
+        state = InvestigationWizardState(
+            stage="review",
+            title=inv.title,
+            question_text=inv.question_text,
+            run_filter=True,
+            review_config={"provider": "openai", "model_name": ""},
+        )
+        readiness = evaluate_investigation_launch_readiness(
+            actor=self.owner,
+            investigation=inv,
+            wizard_state=state,
+        )
+        self.assertFalse(readiness["can_launch"])
+        provider_model_check = next(
+            check for check in readiness["checks"] if check["key"] == "provider_model"
+        )
+        self.assertFalse(provider_model_check["ready"])
+        self.assertIn("required", provider_model_check["message"])
+
+    def test_launch_readiness_blocks_model_not_allowed_for_provider(self):
+        inv = create_investigation(
+            actor=self.owner,
+            workspace=self.workspace,
+            title="Allowlist Guardrail",
+            question_text="Q",
+            scope_json={},
+            method_json={},
+            status=InvestigationStatus.DRAFT,
+        )
+        state = InvestigationWizardState(
+            stage="review",
+            title=inv.title,
+            question_text=inv.question_text,
+            run_filter=True,
+            review_config={"provider": "openrouter", "model_name": "gpt-4.1-mini"},
[…]
