+++ django_workbench_v02/wb_workspaces/context_processors.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from .permissions import can_edit_workspace
-from .services import get_active_workspace_for_user
+from .services import get_active_workspace_for_user, get_workspace_llm_setting
 
 
 def active_workbook(request):
@@ -10,10 +10,13 @@ def active_workbook(request):
         return {
             "active_workbook": None,
             "can_start_over_active_workbook": False,
+            "active_llm_config": None,
         }
 
     workbook = get_active_workspace_for_user(user=user)
+    llm_config = get_workspace_llm_setting(user=user, workspace=workbook) if workbook else None
     return {
         "active_workbook": workbook,
         "can_start_over_active_workbook": bool(workbook and can_edit_workspace(user, workbook)),
+        "active_llm_config": llm_config,
     }

[…]
