### 2026-04-22 fc63cf9
Lock workspace contract and launch payload integrity
