+++ django_workbench_v02/wb_runs/migrations/0003_runworkerheartbeat.py
@@ -1,0 +1,62 @@
+# Generated by Django 5.2.12 on 2026-04-22 11:45
+
+import django.db.models.deletion
+import django.utils.timezone
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("wb_runs", "0002_investigationrun_idx_run_stat_queue_cr_and_more"),
+    ]
+
+    operations = [
+        migrations.CreateModel(
+            name="RunWorkerHeartbeat",
+            fields=[
+                (
+                    "id",
+                    models.BigAutoField(
+                        auto_created=True,
+                        primary_key=True,
+                        serialize=False,
+                        verbose_name="ID",
+                    ),
+                ),
+                ("worker_id", models.CharField(max_length=255, unique=True)),
+                ("state", models.CharField(blank=True, max_length=32)),
+                ("last_run_status", models.CharField(blank=True, max_length=16)),
+                ("last_error", models.TextField(blank=True)),
+                (
+                    "last_seen_at",
+                    models.DateTimeField(default=django.utils.timezone.now),
+                ),
+                (
+                    "created_at",
+                    models.DateTimeField(
+                        default=django.utils.timezone.now, editable=False
+                    ),
+                ),
+                ("updated_at", models.DateTimeField(auto_now=True)),
+                (
+                    "last_run",
+                    models.ForeignKey(
+                        blank=True,
+                        null=True,
+                        on_delete=django.db.models.deletion.SET_NULL,
+                        related_name="+",
+                        to="wb_runs.investigationrun",
+                    ),
+                ),
+            ],
+            options={
+                "ordering": ["-last_seen_at"],
+                "indexes": [
+                    models.Index(
+                        fields=["last_seen_at"], name="idx_worker_hb_last_seen"
+                    )
+                ],
+            },
+        ),
+    ]
