### 2026-04-22 e33b4bf
Investigation now leads to custom dashboard
