+++ django_workbench_v02/wb_workspaces/views.py
@@ -1,18 +1,52 @@
+from datetime import timedelta
+import csv
+import json
+from io import StringIO, TextIOWrapper
+
+import pandas as pd
+from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import PermissionDenied, ValidationError
+from django.http import HttpResponse, HttpResponseBadRequest
 from django.shortcuts import get_object_or_404, redirect, render
+from django.urls import reverse
 from django.utils import timezone
 from django.views.decorators.http import require_GET, require_http_methods, require_POST
 
+from accounts.views import (
+    EXPLORE_REPORTS_PAGE_SIZE,
+    _coerce_positive_int,
+    _explore_column_label,
+    _explore_ordered_columns,
+    _explore_report_rows,
+    _explore_shared_querydict,
+)
 from wb_investigations.models import Investigation
-from wb_runs.models import InvestigationRun, RunStatus
+from wb_collections.services import (
+    _network_truthy,
+    _split_receivers,
+    _theme_collection_map_from_reports,
+    build_explore_metrics,
+    reports_for_collection,
+)
+from wb_runs.artifact_storage import ArtifactStorageError, open_artifact_for_download
+from wb_runs.models import (
+    ArtifactStatus,
+    ArtifactType,
+    InvestigationRun,
+    RunArtifact,
+    RunStatus,
+    RunType,
+    RunWorkerHeartbeat,
+)
 from wb_sharing.forms import ShareLinkCreateForm
 from wb_sharing.models import ShareMode, WorkspaceShareLink
 
 from .activity import is_human_view_request, should_update_last_viewed
 from .forms import (
+    ActiveLLMCredentialDeleteForm,
     ActiveLLMConfigForm,
     WorkspaceCreateForm,
     WorkspaceCredentialDeleteForm,
@@ -39,6 +73,7 @@ from .services import (
     WorkspaceReportExclusionError,
     add_workspace_member,
     create_workspace_for_user,
+    delete_user_llm_credential,
     delete_workspace_credential,
     delete_workspace_immediately,
     get_active_workspace_for_user,
@@ -95,6 +130,232 @@ def _pipeline_status_for_run(run: InvestigationRun | None) -> str:
     if run.status == RunStatus.SUCCEEDED:
         return "complete"
     return ""
+
+
+def _worker_health() -> tuple[bool, str]:
+    stale_seconds = max(1, int(getattr(settings, "WORKER_HEARTBEAT_STALE_SECONDS", 120)))
+    threshold = timezone.now() - timedelta(seconds=stale_seconds)
+    latest = RunWorkerHeartbeat.objects.order_by("-last_seen_at").first()
+    if latest is None:
+        return (
+            False,
+            "Worker offline: runs will stay queued until `run_runs_worker` is running.",
+        )
+    if latest.last_seen_at < threshold:
+        return (
+            False,
+            "Worker heartbeat is stale: restart `run_runs_worker` to process queued runs.",
+        )
+    return (True, f"Worker online ({latest.worker_id}).")
+
+
+def _artifact_preview(artifact: RunArtifact | None, *, max_rows: int = 10) -> dict:
+    empty = {"columns": [], "rows": [], "error": "", "row_count": 0}
+    if artifact is None:
+        return empty
+    try:
+        file_obj, _ = open_artifact_for_download(artifact)
+    except ArtifactStorageError as exc:
+        return {**empty, "error": str(exc)}
+
+    columns: list[str] = []
+    rows: list[list[str]] = []
+    row_count = 0
+    try:
[…]
