+++ django_workbench_v02/pfd_workbench_v02/settings.py
@@ -321,7 +321,7 @@ RUN_RETRY_MAX_ATTEMPTS = _env_int("RUN_RETRY_MAX_ATTEMPTS", 3)
 RUN_RETRY_BACKOFF_SECONDS = _env_int_list("RUN_RETRY_BACKOFF_SECONDS", (30, 120, 600))
 RUN_RETRY_JITTER_PCT = _env_int("RUN_RETRY_JITTER_PCT", 20)
 RUN_STAGE_TIMEOUT_SECONDS = _env_int("RUN_STAGE_TIMEOUT_SECONDS", 1800)
-RUN_TOTAL_TIMEOUT_SECONDS = _env_int("RUN_TOTAL_TIMEOUT_SECONDS", 7200)
+RUN_TOTAL_TIMEOUT_SECONDS = _env_int("RUN_TOTAL_TIMEOUT_SECONDS", 28800)
 RUN_STUCK_THRESHOLD_SECONDS = _env_int("RUN_STUCK_THRESHOLD_SECONDS", 1800)
 WORKER_HEARTBEAT_STALE_SECONDS = _env_int("WORKER_HEARTBEAT_STALE_SECONDS", 120)
 
