### 2026-04-22 51c51d6
Update UI for workspaces
