### 2026-04-22 58101be
Add share function for investigations
