+++ django_workbench_v02/wb_sharing/urls.py
@@ -36,5 +36,7 @@ urlpatterns = [
         views.revoke_workspace_share,
         name="workspace-share-revoke",
     ),
+    path("s/<uuid:share_id>/reports/panel/", views.share_reports_panel, name="share-link-reports-panel"),
+    path("s/<uuid:share_id>/export.csv", views.export_share_link_csv, name="share-link-export-csv"),
     path("s/<uuid:share_id>/", views.view_share_link, name="share-link-detail"),
 ]
