### 2026-04-23 43efa53
Add global user notice feature with modal and homepage banner
