### 2026-04-23 f421974
Modify screening prompt to reduce focus on the deceased
