### 2026-04-23 2e555dd
Introduce 2 more worker replicas
