### 2026-04-23 244d6d9
Fix bug where live csv wasn't getting refreshed
