### 2026-04-23 872b3c5
User friendliness tweaks
