+++ django_workbench_v02/wb_workspaces/views.py
@@ -14,6 +14,7 @@ from django.http import HttpResponse, HttpResponseBadRequest
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 from django.utils import timezone
+from django.utils.http import url_has_allowed_host_and_scheme
 from django.views.decorators.http import require_GET, require_http_methods, require_POST
 
 from accounts.views import (
@@ -441,7 +442,6 @@ def dashboard(request):
     )
     memberships_list = list(memberships)
     dashboard_rows: list[dict] = []
-    can_hard_delete = bool(request.user.is_superuser)
     for membership in memberships_list:
         workspace = membership.workspace
         can_manage_lifecycle = bool(request.user.is_superuser) or (
@@ -456,7 +456,7 @@ def dashboard(request):
                 "is_archived": workspace.archived_at is not None,
                 "can_restore": can_manage_lifecycle,
                 "can_archive": can_manage_lifecycle and workspace.archived_at is None,
-                "can_hard_delete": can_hard_delete,
+                "can_hard_delete": can_manage_lifecycle,
                 "can_cancel_running_run": False,
                 "running_run": None,
                 "pending_stage_label": "",
@@ -1155,8 +1155,6 @@ def archive_workspace_view(request, workbook_id):
         archive_workspace(actor=request.user, workspace=workspace, request=request)
     except (PermissionDenied, ValidationError, WorkspaceLifecycleError) as exc:
         messages.error(request, str(exc))
-    else:
-        messages.success(request, "Workbook archived. It will be auto-deleted after 60 days.")
     return redirect("workbook-dashboard")
 
 
@@ -1164,13 +1162,20 @@ def archive_workspace_view(request, workbook_id):
 @require_POST
 def restore_workspace_view(request, workbook_id):
     workspace = get_object_or_404(Workspace, id=workbook_id)
+    next_url = str(request.POST.get("next") or "").strip()
+    if next_url and not url_has_allowed_host_and_scheme(
+        url=next_url,
+        allowed_hosts={request.get_host()},
+        require_https=request.is_secure(),
+    ):
+        next_url = ""
     try:
         restore_workspace(actor=request.user, workspace=workspace, request=request)
     except (PermissionDenied, ValidationError, WorkspaceLifecycleError) as exc:
         messages.error(request, str(exc))
-    else:
-        messages.success(request, "Workbook restored.")
-    return redirect("workbook-detail", workbook_id=workspace.id)
+    if next_url:
+        return redirect(next_url)
+    return redirect("workbook-open", workbook_id=workspace.id)
 
 
 @login_required
@@ -1187,8 +1192,7 @@ def delete_workspace_view(request, workbook_id):
         )
     except (PermissionDenied, ValidationError) as exc:
         messages.error(request, str(exc))
-        return redirect("workbook-detail", workbook_id=workspace.id)
-    messages.success(request, "Workbook permanently deleted.")
+        return redirect("workbook-open", workbook_id=workspace.id)
     return redirect("workbook-dashboard")
 
 
@@ -1197,6 +1201,8 @@ def workspace_detail(request, workbook_id):
     workspace = get_object_or_404(Workspace, id=workbook_id)
     if not can_view_workspace(request.user, workspace):
         return redirect("accounts-login")
+    # Retired user-facing workspace detail page: canonical user view is workbook-open.
+    return redirect("workbook-open", workbook_id=workspace.id)
     if request.user.is_authenticated and workspace.archived_at is None:
         set_active_workspace_for_user(user=request.user, workspace=workspace, request=request)
 
[…]
