### 2026-04-24 657eb74
Update how themes renders in investigation dashboard
