+++ django_workbench_v02/wb_workspaces/views.py
@@ -667,6 +667,33 @@ def _pipeline_theme_label_map(*, workspace: Workspace) -> dict[str, str]:
     return label_map
 
 
+def _pipeline_theme_metadata_map(*, workspace: Workspace) -> dict[str, dict[str, str]]:
+    theme_summary_artifact = _latest_workspace_artifact(
+        workspace=workspace,
+        artifact_type=ArtifactType.THEME_SUMMARY,
+    )
+    summary_df = _artifact_dataframe(theme_summary_artifact)
+    if summary_df.empty or "theme" not in summary_df.columns:
+        return {}
+    metadata_map: dict[str, dict[str, str]] = {}
+    for _, row in summary_df.iterrows():
+        cleaned = str(row.get("theme") or "").strip()
+        if not cleaned:
+            continue
+        description = str(row.get("description") or "").strip()
+        normalized = cleaned.casefold()
+        prefixed = cleaned if cleaned.startswith("theme_") else f"theme_{cleaned}"
+        payload = {
+            "label": cleaned,
+            "description": description,
+        }
+        metadata_map[cleaned] = payload
+        metadata_map[normalized] = payload
+        metadata_map[prefixed] = payload
+        metadata_map[prefixed.casefold()] = payload
+    return metadata_map
+
+
 def _prune_non_pipeline_theme_columns(*, workspace: Workspace, reports_df: pd.DataFrame) -> pd.DataFrame:
     if reports_df.empty:
         return reports_df
@@ -687,11 +714,12 @@ def _prune_non_pipeline_theme_columns(*, workspace: Workspace, reports_df: pd.Da
     return reports_df.drop(columns=drop_columns, errors="ignore")
 
 
-def _theme_rows_from_pipeline_scope(*, workspace: Workspace, scoped_reports: pd.DataFrame) -> list[tuple[str, int]]:
+def _theme_rows_from_pipeline_scope(*, workspace: Workspace, scoped_reports: pd.DataFrame) -> list[dict[str, object]]:
     if scoped_reports.empty:
         return []
     label_map = _pipeline_theme_label_map(workspace=workspace)
-    rows: list[tuple[str, int]] = []
+    metadata_map = _pipeline_theme_metadata_map(workspace=workspace)
+    rows: list[dict[str, object]] = []
     for column in scoped_reports.columns:
         column_name = str(column or "").strip()
         if not column_name.startswith("theme_"):
@@ -707,15 +735,28 @@ def _theme_rows_from_pipeline_scope(*, workspace: Workspace, scoped_reports: pd.
             or label_map.get(column_name.replace("theme_", "").casefold())
             or default_label
         )
-        rows.append((str(label), count))
-    rows.sort(key=lambda item: (-int(item[1]), str(item[0]).casefold()))
+        metadata = (
+            metadata_map.get(column_name)
+            or metadata_map.get(column_name.casefold())
+            or metadata_map.get(column_name.replace("theme_", ""))
+            or metadata_map.get(column_name.replace("theme_", "").casefold())
+            or {}
+        )
+        rows.append(
+            {
+                "label": str(label),
+                "count": count,
+                "description": str(metadata.get("description") or "").strip(),
+            }
+        )
+    rows.sort(key=lambda item: (-int(item["count"]), str(item["label"]).casefold()))
     return rows
 
 
-def _theme_rows_from_generic_scope(*, reports_df: pd.DataFrame, scoped_reports: pd.DataFrame) -> list[tuple[str, int]]:
+def _theme_rows_from_generic_scope(*, reports_df: pd.DataFrame, scoped_reports: pd.DataFrame) -> list[dict[str, object]]:
     if reports_df.empty or scoped_reports.empty:
[…]

[…]
