### 2026-04-24 df74cb9
Fix temporal graph smoothing between week, month and year view
