+++ django_workbench_v02/wb_collections/services.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 import json
+import math
 from collections import Counter
 from pathlib import Path
 from threading import Lock
@@ -559,11 +560,35 @@ def build_explore_metrics(
         "year": {"points": [], "line_path": "", "area_path": "", "axis_labels": [], "start_label": "", "end_label": "", "latest_label": "", "latest_count": 0},
     }
 
+    def _complete_period_counts(*, source_dates: pd.Series, freq: str) -> pd.Series:
+        period_series = source_dates.dt.to_period(freq)
+        if period_series.empty:
+            return pd.Series(dtype="int64")
+        counts = period_series.value_counts().sort_index()
+        full_index = pd.period_range(
+            start=period_series.min(),
+            end=period_series.max(),
+            freq=freq,
+        )
+        return counts.reindex(full_index, fill_value=0).astype("int64")
+
+    def _year_axis_labels(*, start_year: int, end_year: int, max_labels: int = 6) -> list[str]:
+        if end_year <= start_year:
+            return [str(start_year), "Now"]
+        span_years = max(1, end_year - start_year)
+        step = max(1, int(math.ceil(span_years / max(1, max_labels - 2))))
+        labels = [str(year) for year in range(start_year, end_year + 1, step)]
+        if labels[-1] != str(end_year):
+            labels.append(str(end_year))
+        if labels[-1] != "Now":
+            labels.append("Now")
+        return labels
+
     def _build_temporal_series(
         counts_by_period: pd.Series,
         *,
+        mode: str,
         max_points: int | None,
-        smooth_window: int,
     ) -> dict[str, Any]:
         result: dict[str, Any] = {
             "points": [],
@@ -578,13 +603,20 @@ def build_explore_metrics(
         if counts_by_period.empty:
             return result
 
+        period_values = [period for period in counts_by_period.index.tolist()]
         counts = [int(value) for value in counts_by_period.tolist()]
-        labels = [str(period) for period in counts_by_period.index.tolist()]
-        count_series = pd.Series(counts, dtype="float64")
-        smoothed_series = count_series.rolling(window=max(1, smooth_window), min_periods=1).mean()
-        smoothed_counts = [float(value) for value in smoothed_series.tolist()]
+        labels: list[str] = []
+        for period in period_values:
+            if mode == "week":
+                labels.append(
+                    f"{period.start_time.date().isoformat()} to {period.end_time.date().isoformat()}"
+                )
+            elif mode == "month":
+                labels.append(period.strftime("%Y-%m"))
+            else:
+                labels.append(str(period.year))
 
-        point_pairs = list(zip(labels, counts, smoothed_counts))
+        point_pairs = list(zip(period_values, labels, counts))
         if max_points is not None and len(point_pairs) > max_points:
             sampled_indexes = sorted(
                 {
@@ -594,12 +626,12 @@ def build_explore_metrics(
             )
             point_pairs = [point_pairs[index] for index in sampled_indexes]
 
-        smoothed_only = [float(smoothed) for _, _, smoothed in point_pairs]
-        clip_low = float(pd.Series(smoothed_only).quantile(0.05))
-        clip_high = float(pd.Series(smoothed_only).quantile(0.95))
-        if clip_high <= clip_low:
-            clip_high = clip_low + 1.0
-        span = clip_high - clip_low
+        plotted_counts = [float(count) for _, _, count in point_pairs]
+        count_low = min(plotted_counts)
+        count_high = max(plotted_counts)
+        if count_high <= count_low:
+            count_high = count_low + 1.0
+        span = count_high - count_low
 
[…]
