### 2026-04-24 b43c2be
Add copy into new investigation feature
