### 2026-04-24 a957115
More dynamic loaders for ongoing workspace
