### 2026-04-24 976ee6d
Temporal graph now defaults to either weeks/months/years display given deterministic rules
