+++ django_workbench_v02/wb_workspaces/views.py
@@ -458,6 +458,7 @@ def _pipeline_dashboard_context(*, workspace: Workspace, request) -> dict:
             "temporal_axis_labels": [],
             "temporal_start_label": "",
             "temporal_end_label": "",
+            "temporal_mode": "month",
             "temporal_series": {"month": {"points": []}},
             "unique_receiver_count": 0,
             "most_common_theme": "No theme signals",
@@ -980,6 +981,7 @@ def _workspace_build_explore_payload(*, workspace: Workspace, request):
             "temporal_axis_labels": [],
             "temporal_start_label": "",
             "temporal_end_label": "",
+            "temporal_mode": "month",
             "temporal_series": {"month": {"points": []}},
             "unique_receiver_count": 0,
             "themes_from_investigation": False,
