+++ django_workbench_v02/wb_workspaces/context_processors.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from django.conf import settings
 
-from .permissions import can_edit_workspace
 from .services import get_active_workspace_for_user, get_workspace_llm_setting
 
 
@@ -19,7 +18,6 @@ def active_workbook(request):
     if not user or not getattr(user, "is_authenticated", False):
         return {
             "active_workbook": None,
-            "can_start_over_active_workbook": False,
             "active_llm_config": None,
             "user_notice": notice,
         }
@@ -28,7 +26,6 @@ def active_workbook(request):
     llm_config = get_workspace_llm_setting(user=user, workspace=workbook) if workbook else None
     return {
         "active_workbook": workbook,
-        "can_start_over_active_workbook": bool(workbook and can_edit_workspace(user, workbook)),
         "active_llm_config": llm_config,
         "user_notice": notice,
     }
