### 2026-04-26 eb43e15
Create admin interface
