+++ django_workbench_v02/wb_ops/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class WbOpsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "wb_ops"
