### 2026-04-27 92e83b6
Workspace list/dashboard now counts from the canonical dataset artifact
