### 2026-04-27 56ac383
Cancelling run no longer sends user to empty dashboard
