### 2026-04-29 15fe01d
Upgrade gpt 4.1 to 5.4 & add compatibility guard for inconsistent openai params: temperature, reasoning effort
