### 2026-05-02 f6cca16
Add study protocol and other info for local LLM experiment
