+++ open_llm_ons_experiment/ons_replication/scripts/kappa_results.R
@@ -1,0 +1,181 @@
+# install.packages(c("readxl","dplyr","tidyr","purrr","irr"))
+library(readxl)
+library(dplyr)
+library(tidyr)
+library(purrr)
+library(irr)
+
+# irr doesn’t return a confidence interval for k.
+# It gives k and its z-statistic (z = k / SE), but not SE or variance.
+# We therefore recover SE as 'k / z' and build a 95% Wald CI: k +/- 1.96*SE.
+
+
+# Read data
+# ...data is an Excel spreadsheet with multiple tabs. Resolve it using
+# relative paths so this script can be run from either the repo root or
+# this scripts directory.
+candidate_paths <- c(
+  "ons_replication/PFD Toolkit--Consensus Comparison.xlsx",
+  "../PFD Toolkit--Consensus Comparison.xlsx",
+  "PFD Toolkit--Consensus Comparison.xlsx"
+)
+existing_paths <- candidate_paths[file.exists(candidate_paths)]
+if(length(existing_paths) == 0){
+  stop("Could not find replication spreadsheet. Tried:\n", paste(candidate_paths, collapse = "\n"))
+}
+
+file_path  <- existing_paths[[1]]
+sheet_name <- "Consensus annotations"
+
+raw <- read_excel(file_path, sheet = sheet_name)
+
+# Keep only the annotation cols + Toolkit output and put them in a tidier form 
+
+annotation_cols <- c("Dan: child & suicide", 
+                     "Arpan: child & suicide", 
+                     "Sahil: child & suicide")
+
+df <- raw %>% 
+  select(`PFD Toolkit: child suicide?`,
+         `Post-consensus verdict: Is this a child suicide case? (Yes or No)`,
+         all_of(annotation_cols)) %>% 
+  # trim leading/trailing spaces in all selected columns (handles stray spaces)
+  mutate(across(everything(), ~ trimws(as.character(.x)))) %>% 
+  mutate(across(everything(),
+                ~factor(.x, levels = c("No","Yes")))) %>% 
+  # short, convenient aliases for analysis while preserving original names above
+  rename(Toolkit  = `PFD Toolkit: child suicide?`,
+         Consensus = `Post-consensus verdict: Is this a child suicide case? (Yes or No)`)
+
+
+# Calculations!
+
+## For each clin, raw agreement & Cohen's k vs. Toolkit 
+
+### Documentation for kappa2(): https://www.rdocumentation.org/packages/irr/versions/0.84.1/topics/kappa2
+
+agreement_vs_toolkit <- map_dfr(
+  annotation_cols,
+  function(cln){
+    rater_col <- cln
+    k_out     <- kappa2(df[, c(rater_col, "Toolkit")], weight = "unweighted")
+    se        <- abs(k_out$value / k_out$statistic)
+    ci        <- k_out$value + c(-1.96, 1.96) * se
+    tibble(
+      Clinician       = cln,
+      Raw_Agreement   = mean(df[[rater_col]] == df$Toolkit, na.rm = TRUE),
+      Cohen_Kappa     = k_out$value,
+      CI_lower        = ci[1],
+      CI_upper        = ci[2]
+    )
+  }
+)
+
+## Consensus vs. Toolkit: raw agreement & Cohen's kappa
+
+### Documentation for kappa2(): https://www.rdocumentation.org/packages/irr/versions/0.84.1/topics/kappa2
+
+consensus_out <- kappa2(df[, c("Consensus", "Toolkit")], weight = "unweighted")
[…]

[…]
