### 2026-05-02 b207d29
Add variant selection policy
