+++ open_llm_ons_experiment/scripts/apply_manual_exclusions.py
@@ -1,0 +1,209 @@
+#!/usr/bin/env python3
+"""
+Apply manual exclusions to discovery outputs.
+
+Reads:
+- eligible_tags_latest.csv
+- excluded_tags_latest.csv
+- manual_exclusions.csv
+
+Writes:
+- eligible_tags_latest.csv (updated)
+- excluded_tags_latest.csv (updated)
+- exclusion_reason_counts_latest.csv (recomputed)
+"""
+
+from __future__ import annotations
+
+import argparse
+import csv
+from dataclasses import dataclass
+from pathlib import Path
+
+import pandas as pd
+
+
+@dataclass
+class ManualExclusion:
+    family: str
+    tag: str
+    stage: str
+    reason_code: str
+    rationale: str
+
+    def matches(self, row: pd.Series) -> bool:
+        row_family = str(row.get("family", ""))
+        row_tag = str(row.get("tag", ""))
+        family = self.family.strip()
+        tag = self.tag.strip()
+
+        if family and tag:
+            return row_family == family and row_tag == tag
+        if family:
+            return row_family == family
+        if tag:
+            return row_tag == tag
+        return False
+
+
+def parse_args() -> argparse.Namespace:
+    p = argparse.ArgumentParser(description=__doc__)
+    p.add_argument(
+        "--eligible",
+        default="open_llm_ons_experiment/artifacts/discovery/eligible_tags_latest.csv",
+    )
+    p.add_argument(
+        "--excluded",
+        default="open_llm_ons_experiment/artifacts/discovery/excluded_tags_latest.csv",
+    )
+    p.add_argument(
+        "--manual-exclusions",
+        default="open_llm_ons_experiment/config/manual_exclusions.csv",
+    )
+    p.add_argument(
+        "--counts-out",
+        default="open_llm_ons_experiment/artifacts/discovery/exclusion_reason_counts_latest.csv",
+    )
+    return p.parse_args()
+
+
+def load_manual_exclusions(path: Path) -> list[ManualExclusion]:
+    rows: list[ManualExclusion] = []
+    if not path.exists():
+        return rows
+    with path.open(newline="", encoding="utf-8") as f:
+        reader = csv.DictReader(f)
+        for r in reader:
+            family = (r.get("family") or "").strip()
+            tag = (r.get("tag") or "").strip()
+            stage = (r.get("stage") or "manual_exclusion").strip()
+            reason_code = (r.get("reason_code") or "").strip()
+            rationale = (r.get("rationale") or "").strip()
+            if (not family and not tag) or family.startswith("#"):
+                continue
+            if stage not in {"preprocessing", "manual_exclusion"}:
+                raise ValueError(
+                    f"Invalid stage '{stage}' for manual exclusion row family='{family}' tag='{tag}'. "
+                    "Stage must be 'preprocessing' or 'manual_exclusion'."
+                )
+            rows.append(
+                ManualExclusion(
+                    family=family,
+                    tag=tag,
+                    stage=stage,
+                    reason_code=reason_code,
+                    rationale=rationale,
+                )
+            )
+    return rows
+
+
+def append_reason(existing: str, new_reason: str) -> str:
+    if existing is None or (isinstance(existing, float) and pd.isna(existing)) or str(existing).strip().lower() == "nan":
+        existing_text = ""
+    else:
+        existing_text = str(existing)
+    parts = [p.strip() for p in existing_text.split(";") if p.strip() and p.strip().lower() != "nan"]
+    if new_reason not in parts:
+        parts.append(new_reason)
+    return ";".join(parts)
+
+
+def _reason_stage(reason: str) -> str:
[…]
