+++ open_llm_ons_experiment/scripts/discover_eligible_ollama_models.py
@@ -1,0 +1,596 @@
+#!/usr/bin/env python3
+"""
+Discover and filter eligible Ollama library models for the open LLM experiment.
+
+This script:
+1) Crawls family pages from https://ollama.com/library
+2) Expands to tag-level variants
+3) Applies protocol filters
+4) Applies manual exclusions from a CSV template
+5) Writes discovered / eligible / excluded CSV outputs
+"""
+
+from __future__ import annotations
+
+import argparse
+import csv
+import re
+from dataclasses import dataclass
+from datetime import date, datetime, timedelta
+from pathlib import Path
+from typing import Iterable
+from urllib.parse import quote, unquote
+
+import pandas as pd
+import requests
+from bs4 import BeautifulSoup
+
+
+LIBRARY_URL = "https://ollama.com/library"
+
+# Protocol defaults
+DEFAULT_START_DATE = "2024-05-01"
+DEFAULT_END_DATE = "2026-05-01"
+DEFAULT_MIN_PARAMS_B = 5.0
+DEFAULT_DENSE_MAX_PARAMS_B = 120.0
+DEFAULT_MOE_MAX_ACTIVE_PARAMS_B = 40.0
+DEFAULT_MOE_MAX_TOTAL_PARAMS_B = 400.0
+DEFAULT_MAX_MODEL_SIZE_GB = 100.0
+
+# Family/tag keyword heuristics for specialised model exclusions.
+# Keep this conservative to reduce false positives.
+SPECIALISED_PATTERNS = [
+    re.compile(p, flags=re.IGNORECASE)
+    for p in [
+        r"\bcoder\b",
+        r"\bcodeqwen\b",
+        r"\bcodellama\b",
+        r"\bcodestral\b",
+        r"\bcodegemma\b",
+        r"\bdeepseek-coder\b",
+        r"\bgranite-code\b",
+        r"\bstarcoder\b",
+        r"\bwizardcoder\b",
+        r"\byi-coder\b",
+        r"\bdolphincoder\b",
+        r"\bsqlcoder\b",
+        r"\bqwen2-math\b",
+        r"\bwizard-math\b",
+        r"\bllama-guard\b",
+        r"\bembedding\b",
+        r"\bembed\b",
+        r"\btranslategemma\b",
+    ]
+]
+
+# VLM-specialised families commonly marketed as vision-first endpoints.
+VISION_SPECIALISED_FAMILIES = {
+    "llava",
+    "llava-llama3",
+    "minicpm-v",
+    "moondream",
+    "qwen2.5vl",
+    "qwen3-vl",
+    "llama3.2-vision",
+    "granite3.2-vision",
+}
+
+
+@dataclass
+class ManualExclusion:
+    family: str
+    tag: str
+    stage: str
+    reason_code: str
+    rationale: str
+
+    def matches(self, row: pd.Series) -> bool:
+        row_tag = str(row.get("tag", ""))
+        row_family = str(row.get("family", ""))
+        family = self.family.strip()
+        tag = self.tag.strip()
+
+        if family and tag:
+            return row_family == family and row_tag == tag
+        if family:
+            return row_family == family
+        if tag:
+            return row_tag == tag
+        return False
+
+
+def parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument("--start-date", default=DEFAULT_START_DATE, help="YYYY-MM-DD")
+    parser.add_argument("--end-date", default=DEFAULT_END_DATE, help="YYYY-MM-DD")
+    parser.add_argument("--min-params-b", type=float, default=DEFAULT_MIN_PARAMS_B)
[…]
