### 2026-05-05 a953a18
Update protocol and re-run partial
