### 2026-05-05 4e29779
Fix graph temporal malalignment and hover tooltip
