+++ django_workbench_v02/wb_collections/services.py
@@ -553,11 +553,12 @@ def build_explore_metrics(
     temporal_start_label = ""
     temporal_end_label = ""
     temporal_axis_labels: list[str] = []
+    temporal_axis_ticks: list[dict[str, Any]] = []
     temporal_mode = "month"
     temporal_series: dict[str, dict[str, Any]] = {
-        "week": {"points": [], "line_path": "", "area_path": "", "axis_labels": [], "start_label": "", "end_label": "", "latest_label": "", "latest_count": 0},
-        "month": {"points": [], "line_path": "", "area_path": "", "axis_labels": [], "start_label": "", "end_label": "", "latest_label": "", "latest_count": 0},
-        "year": {"points": [], "line_path": "", "area_path": "", "axis_labels": [], "start_label": "", "end_label": "", "latest_label": "", "latest_count": 0},
+        "week": {"points": [], "line_path": "", "area_path": "", "axis_labels": [], "axis_ticks": [], "start_label": "", "end_label": "", "latest_label": "", "latest_count": 0},
+        "month": {"points": [], "line_path": "", "area_path": "", "axis_labels": [], "axis_ticks": [], "start_label": "", "end_label": "", "latest_label": "", "latest_count": 0},
+        "year": {"points": [], "line_path": "", "area_path": "", "axis_labels": [], "axis_ticks": [], "start_label": "", "end_label": "", "latest_label": "", "latest_count": 0},
     }
 
     def _complete_period_counts(*, source_dates: pd.Series, freq: str) -> pd.Series:
@@ -586,15 +587,28 @@ def build_explore_metrics(
 
     def _year_axis_labels(*, start_year: int, end_year: int, max_labels: int = 6) -> list[str]:
         if end_year <= start_year:
-            return [str(start_year), "Now"]
+            return [str(start_year)]
         span_years = max(1, end_year - start_year)
         step = max(1, int(math.ceil(span_years / max(1, max_labels - 2))))
         labels = [str(year) for year in range(start_year, end_year + 1, step)]
         if labels[-1] != str(end_year):
             labels.append(str(end_year))
-        if labels[-1] != "Now":
-            labels.append("Now")
         return labels
+
+    def _latest_period_is_current(*, latest_period, mode: str) -> bool:
+        today = timezone.now().date()
+        if mode == "year":
+            return int(latest_period.year) == int(today.year)
+        if mode == "month":
+            return (
+                int(latest_period.year) == int(today.year)
+                and int(latest_period.month) == int(today.month)
+            )
+        if mode == "week":
+            start = latest_period.start_time.date()
+            end = latest_period.end_time.date()
+            return start <= today <= end
+        return False
 
     def _build_temporal_series(
         counts_by_period: pd.Series,
@@ -607,6 +621,7 @@ def build_explore_metrics(
             "line_path": "",
             "area_path": "",
             "axis_labels": [],
+            "axis_ticks": [],
             "start_label": "",
             "end_label": "",
             "latest_label": "",
@@ -675,7 +690,45 @@ def build_explore_metrics(
 
         start_year = int(point_pairs[0][0].year)
         end_year = int(point_pairs[-1][0].year)
-        result["axis_labels"] = _year_axis_labels(start_year=start_year, end_year=end_year)
+        raw_axis_labels = _year_axis_labels(start_year=start_year, end_year=end_year)
+        year_to_x: dict[int, float] = {}
+        for (period, _, _), (x, _) in zip(point_pairs, raw_points):
+            year_to_x.setdefault(int(period.year), float(x))
+
+        ticks: list[dict[str, Any]] = []
+        for axis_label in raw_axis_labels:
+            label_text = str(axis_label)
+            try:
+                label_year = int(label_text)
[…]
