+++ django_workbench_v02/wb_workspaces/views.py
@@ -486,10 +486,11 @@ def _pipeline_dashboard_context(*, workspace: Workspace, request) -> dict:
             "temporal_line_path": "",
             "temporal_area_path": "",
             "temporal_axis_labels": [],
+            "temporal_axis_ticks": [],
             "temporal_start_label": "",
             "temporal_end_label": "",
             "temporal_mode": "month",
-            "temporal_series": {"month": {"points": []}},
+            "temporal_series": {"month": {"points": [], "axis_ticks": []}},
             "unique_receiver_count": 0,
             "most_common_theme": "No theme signals",
         }
@@ -1023,10 +1024,11 @@ def _workspace_build_explore_payload(*, workspace: Workspace, request):
             "temporal_line_path": "",
             "temporal_area_path": "",
             "temporal_axis_labels": [],
+            "temporal_axis_ticks": [],
             "temporal_start_label": "",
             "temporal_end_label": "",
             "temporal_mode": "month",
-            "temporal_series": {"month": {"points": []}},
+            "temporal_series": {"month": {"points": [], "axis_ticks": []}},
             "unique_receiver_count": 0,
             "themes_from_investigation": False,
             "dataset_min_date": "",

[…]
