### 2026-05-05 a667707
Add planning files for issues tracking
