### 2026-05-06 271ba5d
Add text normalisation pipeline component
