+++ pyproject.toml
@@ -48,6 +48,10 @@ dependencies = [
     "tqdm>=4.67.1",
     "whitenoise>=6.11.0",
     "boto3>=1.42.91",
+    "scikit-learn>=1.7.2",
+    "umap-learn>=0.5.12",
+    "sentence-transformers>=5.4.1",
+    "ollama>=0.6.2",
 ]
 
 [project.optional-dependencies]
