### 2026-05-06 10cec50
Create local dgx spark workflow route
