+++ django_workbench_v02/wb_workspaces/views.py
@@ -63,6 +63,7 @@ from .models import (
     MembershipRole,
     Workspace,
     WorkspaceCredential,
+    WorkspaceLLMProvider,
     WorkspaceMembership,
     WorkspaceReportExclusion,
     WorkspaceVisibility,
@@ -256,8 +257,12 @@ def _wizard_copy_prefill(
         "force_assign": bool(config.get("force_assign", False)),
         "skip_if_present": bool(config.get("skip_if_present", True)),
         "extract_include_supporting_quotes": bool(config.get("produce_spans", False)),
-        "provider": str(config.get("provider") or "openai").strip().lower(),
-        "model_name": str(config.get("model_name") or "gpt-4.1-mini").strip(),
+        "provider": (
+            WorkspaceLLMProvider.OPENAI
+            if str(config.get("provider") or "").strip().lower() == WorkspaceLLMProvider.OPENROUTER
+            else str(config.get("provider") or WorkspaceLLMProvider.LOCAL_OLLAMA).strip().lower()
+        ),
+        "model_name": str(config.get("model_name") or "gemma4:27b").strip(),
         "max_parallel_workers": max_parallel_workers,
         "request_completion_email": True,
     }
@@ -1713,8 +1718,8 @@ def save_active_llm_config(request):
         messages.error(request, "Invalid LLM configuration submission.")
         return redirect("llm-config")
 
-    provider = str(form.cleaned_data.get("provider") or "openai").strip().lower()
-    model_name = str(form.cleaned_data.get("model_name") or "gpt-4.1-mini").strip()
+    provider = str(form.cleaned_data.get("provider") or WorkspaceLLMProvider.LOCAL_OLLAMA).strip().lower()
+    model_name = str(form.cleaned_data.get("model_name") or "gemma4:27b").strip()
     max_parallel_workers = int(form.cleaned_data.get("max_parallel_workers") or 1)
     api_key = str(form.cleaned_data.get("api_key") or "").strip()
     base_url = str(form.cleaned_data.get("base_url") or "").strip()
@@ -1728,7 +1733,7 @@ def save_active_llm_config(request):
             max_parallel_workers=max_parallel_workers,
             request=request,
         )
-        if api_key:
+        if api_key and provider == WorkspaceLLMProvider.OPENAI:
             upsert_user_llm_credential(
                 actor=request.user,
                 provider=provider,
