### 2026-05-06 4b5d935
Failsafe to prevent server workers from claiming DGX runs
