### 2026-05-07 0012254
Normalise yes no values
