### 2026-05-07 d7b4980
Fix fallback error path
