### 2026-05-07 2ba5aca
Fix fallback_exec bug
