+++ django_workbench_v02/wb_workspaces/migrations/0010_alter_userllmsetting_model_name_and_more.py
@@ -1,0 +1,23 @@
+# Generated by Django 5.2.12 on 2026-05-07 11:20
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('wb_workspaces', '0009_alter_userllmcredential_provider_and_more'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='userllmsetting',
+            name='model_name',
+            field=models.CharField(default='gemma4:26b', max_length=255),
+        ),
+        migrations.AlterField(
+            model_name='workspacellmsetting',
+            name='model_name',
+            field=models.CharField(default='gemma4:26b', max_length=255),
+        ),
+    ]
