### 2026-05-07 be4d8f1
Add ops dashboard refresh
