+++ django_workbench_v02/wb_ops/views.py
@@ -5,6 +5,7 @@ from collections import defaultdict
 from datetime import timedelta
 from dataclasses import dataclass
 from decimal import Decimal, InvalidOperation
+from itertools import zip_longest
 
 import pandas as pd
 from django.conf import settings
@@ -109,6 +110,11 @@ class TypedRunReviewPayload:
     model_name: str
     max_parallel_workers: int
     request_completion_email: bool
+    feature_fields: list[dict]
+    allow_multiple: bool
+    force_assign: bool
+    skip_if_present: bool
+    extract_include_supporting_quotes: bool
 
 
 def _coerce_lookback_days(raw_value) -> int:
@@ -213,6 +219,29 @@ def _typed_payload_from_run(run: InvestigationRun) -> TypedRunReviewPayload:
     model_name = str(config.get("model_name") or "gemma4:26b").strip()
     if not model_name:
         model_name = "gemma4:26b"
+    feature_rows = config.get("feature_fields") if isinstance(config.get("feature_fields"), list) else []
+    feature_fields: list[dict] = []
+    for row in feature_rows:
+        if not isinstance(row, dict):
+            continue
+        name = str(row.get("name") or row.get("field_name") or "").strip()
+        description = str(row.get("description") or "").strip()
+        field_type = str(row.get("type") or "text").strip().lower() or "text"
+        if field_type == "number":
+            field_type = "decimal"
+        if field_type not in {"text", "decimal", "boolean"}:
+            field_type = "text"
+        if not name:
+            continue
+        feature_fields.append(
+            {
+                "name": name,
+                "description": description,
+                "type": field_type,
+            }
+        )
+    if not feature_fields:
+        feature_fields = [{"name": "", "description": "", "type": "text"}]
 
     return TypedRunReviewPayload(
         title=str(investigation.title or "").strip(),
@@ -234,7 +263,39 @@ def _typed_payload_from_run(run: InvestigationRun) -> TypedRunReviewPayload:
         model_name=model_name,
         max_parallel_workers=_parse_int(config.get("max_parallel_workers"), default=1, minimum=1, maximum=32) or 1,
         request_completion_email=True,
+        feature_fields=feature_fields,
+        allow_multiple=bool(config.get("allow_multiple", False)),
+        force_assign=bool(config.get("force_assign", False)),
+        skip_if_present=bool(config.get("skip_if_present", True)),
+        extract_include_supporting_quotes=bool(config.get("produce_spans", False)),
     )
+
+
+def _feature_fields_from_post(request) -> list[dict]:
+    names = request.POST.getlist("feature_field_name")
+    descriptions = request.POST.getlist("feature_field_description")
+    types = request.POST.getlist("feature_field_type")
+    rows: list[dict] = []
+    for raw_name, raw_description, raw_type in zip_longest(names, descriptions, types, fillvalue=""):
+        name = str(raw_name or "").strip()
+        description = str(raw_description or "").strip()
+        field_type = str(raw_type or "text").strip().lower() or "text"
+        if field_type == "number":
+            field_type = "decimal"
+        if field_type not in {"text", "decimal", "boolean"}:
+            field_type = "text"
+        if not name and not description:
+            continue
+        if not name:
+            continue
+        rows.append(
+            {
+                "name": name,
+                "description": description,
+                "type": field_type,
+            }
+        )
+    return rows
 
 
 def _typed_payload_from_post(request) -> TypedRunReviewPayload:
@@ -253,6 +314,8 @@ def _typed_payload_from_post(request) -> TypedRunReviewPayload:
[…]
