+++ pyproject.toml
@@ -52,6 +52,8 @@ dependencies = [
     "umap-learn>=0.5.12",
     "sentence-transformers>=5.4.1",
     "ollama>=0.6.2",
+    "plotly>=6.7.0",
+    "kaleido>=1.3.0",
 ]
 
 [project.optional-dependencies]
