### 2026-06-15 ac9b4e7
Restore workers
