### 2026-06-15 e59deeb
Make server settings persist on switch
