### 2026-06-15 f407691
Force LLM config to remember user's last saved option
