### 2026-06-15 d11b67c
Investigation LLM server setting now inherits from config
