### 2026-06-17 5715654
Fix pre-mature email notification; add sticky table scroller
