### 2026-06-17 b958466
prorates the in-progress current year
