### 2026-06-22 502a431
Allow API to be assigned to user in Ops
