+++ django_workbench_v02/wb_ops/views.py
@@ -14,6 +14,7 @@ from django.db.models import Count, Q
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 from django.utils import timezone
+from django.utils.http import url_has_allowed_host_and_scheme
 from django.views.decorators.http import require_GET, require_POST
 
 from wb_auditlog.models import AuditEvent
@@ -27,8 +28,9 @@ from wb_runs.models import (
     RunStatus,
     RunWorkerHeartbeat,
 )
+from wb_runs.services import RunServiceError, configure_pending_run_for_ops
 from wb_sharing.models import WorkspaceShareLink
-from wb_workspaces.models import Workspace, WorkspaceMembership, WorkspaceReportExclusion
+from wb_workspaces.models import Workspace, WorkspaceLLMProvider, WorkspaceMembership, WorkspaceReportExclusion
 from wb_workspaces.permissions import can_edit_workspace
 from wb_workspaces.report_identity import REPORT_IDENTITY_COLUMN, with_report_identities
 from wb_workspaces.services import (
@@ -36,6 +38,8 @@ from wb_workspaces.services import (
     restore_workspace_report_exclusion,
     upsert_workspace_report_exclusion,
 )
+
+from .forms import OpsPendingRunConfigForm
 
 
 User = get_user_model()
@@ -76,6 +80,41 @@ def _safe_text(value) -> str:
     if text.casefold() in {"nan", "nat", "none", "null"}:
         return ""
     return text
+
+
+def _resolve_next_url(request, *, fallback: str) -> str:
+    next_url = str(request.POST.get("next_url") or "").strip()
+    if not next_url:
+        next_url = str(request.META.get("HTTP_REFERER") or "").strip()
+    if next_url and url_has_allowed_host_and_scheme(
+        url=next_url,
+        allowed_hosts={request.get_host()},
+        require_https=request.is_secure(),
+    ):
+        return next_url
+    return fallback
+
+
+def _ops_run_provider(run: InvestigationRun) -> str:
+    config = run.input_config_json if isinstance(run.input_config_json, dict) else {}
+    provider = str(config.get("provider") or WorkspaceLLMProvider.OPENAI).strip().lower()
+    if provider not in {WorkspaceLLMProvider.OPENAI, WorkspaceLLMProvider.OPENROUTER}:
+        return WorkspaceLLMProvider.OPENAI
+    return provider
+
+
+def _ops_can_reconfigure_run(run: InvestigationRun) -> bool:
+    return run.status == RunStatus.QUEUED and not str(run.worker_id or "").strip()
+
+
+def _ops_run_row(run: InvestigationRun) -> dict:
+    return {
+        "run": run,
+        "input_config_json": json.dumps(run.input_config_json or {}, indent=2, sort_keys=True),
+        "current_provider": _ops_run_provider(run),
+        "can_reconfigure": _ops_can_reconfigure_run(run),
+        "has_ops_override": bool(str(run.ops_override_key_last4 or "").strip()),
+    }
 
 
 def _worker_snapshot(*, now):
@@ -361,11 +400,12 @@ def dashboard(request):
     concurrency = _concurrency_window_metrics(start=start, end=now)
     visitors = _human_visitor_metrics(start=start, end=now)
 
-    recent_runs = list(
-        InvestigationRun.objects.select_related("workspace", "requested_by", "investigation").order_by("-created_at")[
-            :RECENT_RUN_LIMIT
-        ]
-    )
+    recent_runs = [
+        _ops_run_row(run)
+        for run in InvestigationRun.objects.select_related("workspace", "requested_by", "investigation").order_by(
+            "-created_at"
+        )[:RECENT_RUN_LIMIT]
+    ]
     workspaces = list(
         Workspace.objects.select_related("created_by")
         .annotate(
[…]
