+++ django_workbench_v02/wb_runs/admin.py
@@ -46,7 +46,17 @@ class InvestigationRunAdmin(admin.ModelAdmin):
         "requested_by",
         "cancel_requested_by",
     ]
-    readonly_fields = ["created_at", "updated_at", "queued_at", "started_at", "finished_at"]
+    exclude = ["ops_override_encrypted_api_key"]
+    readonly_fields = [
+        "created_at",
+        "updated_at",
+        "queued_at",
+        "started_at",
+        "finished_at",
+        "ops_override_provider",
+        "ops_override_key_last4",
+        "ops_override_base_url",
+    ]
     inlines = [RunEventInline, RunArtifactInline]
 
 

[…]
