### 2026-06-22 5e71f84
Revert "Allow API to be assigned to user in Ops"
This reverts commit 502a43166a2ff4256e4e3749eaf2697fd59b28a6.
