+++ django_workbench_v02/wb_runs/services.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from copy import deepcopy
 from datetime import datetime, time as dt_time
 
 from django.conf import settings
@@ -10,15 +9,8 @@ from django.db import transaction
 from django.utils import timezone
 
 from wb_auditlog.services import log_action_cache_event, log_audit_event
-from wb_workspaces.credentials import WorkspaceCredentialError, encrypt_secret
 from wb_workspaces.activity import is_human_view_request, should_update_last_viewed
-from wb_workspaces.models import WorkspaceLLMProvider
 from wb_workspaces.permissions import can_run_workflows, can_view_workspace
-from wb_workspaces.services import (
-    WorkspaceCredentialValidationError,
-    has_workspace_credential,
-    validate_provider_api_key,
-)
 
 from .models import InvestigationRun, RunArtifact, RunEvent, RunEventType, RunStatus
 from .scope import resolve_run_scope_config
@@ -190,137 +182,6 @@ def _validate_status_transition(current_status: str, next_status: str) -> None:
     allowed = ALLOWED_STATUS_TRANSITIONS.get(current_status, set())
     if next_status not in allowed:
         raise RunServiceError(f"Invalid run transition: {current_status} -> {next_status}.")
-
-
-@transaction.atomic
-def configure_pending_run_for_ops(
-    *,
-    actor,
-    run,
-    provider: str,
-    api_key: str = "",
-    request=None,
-) -> InvestigationRun:
-    if not actor or not getattr(actor, "is_staff", False):
-        raise PermissionDenied("Only staff users can reconfigure pending runs from ops.")
-
-    current_run = (
-        InvestigationRun.objects.select_related("workspace", "investigation", "requested_by")
-        .select_for_update()
-        .get(id=run.id)
-    )
-    if current_run.status != RunStatus.QUEUED:
-        raise RunServiceError("Only queued runs can be reconfigured from the ops dashboard.")
-    if str(current_run.worker_id or "").strip():
-        raise RunServiceError(
-            f"Run has already been claimed by worker {current_run.worker_id} and can no longer be reconfigured."
-        )
-
-    resolved_provider = str(provider or "").strip().lower()
-    if resolved_provider not in {WorkspaceLLMProvider.OPENAI, WorkspaceLLMProvider.OPENROUTER}:
-        raise RunServiceError(f"Unsupported provider '{provider}'.")
-
-    compact_api_key = str(api_key or "").strip()
-    if resolved_provider != WorkspaceLLMProvider.OPENAI and compact_api_key:
-        raise RunServiceError("One-time ops credential overrides are currently supported only for OpenAI runs.")
-
-    if not compact_api_key:
-        if not has_workspace_credential(
-            user=current_run.requested_by,
-            workspace=current_run.workspace,
-            provider=resolved_provider,
-        ):
-            if resolved_provider == WorkspaceLLMProvider.OPENAI:
-                raise RunServiceError(
-                    "Provide a one-time OpenAI API key for this queued run, or save an OpenAI credential for the requester first."
-                )
-            raise RunServiceError(
-                f"The requester does not have a saved {resolved_provider} credential for this queued run."
-            )
-    try:
-        encrypted_api_key = ""
-        key_last4 = ""
-        if compact_api_key:
-            validated_api_key = validate_provider_api_key(
-                provider=resolved_provider,
-                api_key=compact_api_key,
-            )
-            encrypted_api_key = encrypt_secret(validated_api_key)
-            key_last4 = validated_api_key[-4:]
-    except (WorkspaceCredentialValidationError, WorkspaceCredentialError) as exc:
-        raise RunServiceError(str(exc)) from exc
-
[…]
