### 2026-06-22 f9993f4
Allow API key to be assigned one-off in Ops
