### 2026-06-22 4150849
Fix bug where OpenAI API tries to run Gemma
