+++ django_workbench_v02/wb_ops/forms.py
@@ -3,6 +3,12 @@ from django import forms
 from wb_workspaces.models import WorkspaceLLMProvider
 
 
+OPS_OPENAI_MODEL_CHOICES = (
+    ("gpt-4.1-mini", "gpt-4.1-mini"),
+    ("gpt-4.1", "gpt-4.1"),
+)
+
+
 class OpsPendingRunConfigForm(forms.Form):
     provider = forms.ChoiceField(
         choices=(
@@ -15,6 +21,11 @@ class OpsPendingRunConfigForm(forms.Form):
         required=False,
         widget=forms.PasswordInput(render_value=False),
     )
+    model_name = forms.ChoiceField(
+        choices=OPS_OPENAI_MODEL_CHOICES,
+        required=False,
+        initial="gpt-4.1-mini",
+    )
     next_url = forms.CharField(required=False)
 
     def clean_provider(self):
@@ -25,3 +36,12 @@ class OpsPendingRunConfigForm(forms.Form):
 
     def clean_api_key(self):
         return str(self.cleaned_data.get("api_key") or "").strip()
+
+    def clean_model_name(self):
+        provider = str(self.cleaned_data.get("provider") or WorkspaceLLMProvider.LOCAL_OLLAMA).strip().lower()
+        model_name = str(self.cleaned_data.get("model_name") or "").strip()
+        if provider != WorkspaceLLMProvider.OPENAI:
+            return ""
+        if model_name not in dict(OPS_OPENAI_MODEL_CHOICES):
+            return "gpt-4.1-mini"
+        return model_name

[…]
