### 2026-06-22 15b31ea
Fix UI bug where openai models were being shown in Ops for local runs
