### 2026-06-22 ae0e1bb
Fix OpenAI run approval and worker routing bugs in Ops
