### 2026-06-22 445ca11
Fix OpenAI run approval and worker routing bugs in Ops
has_workspace_credential() always returned None because its return
logic was accidentally placed inside validate_provider_api_key() after
a return statement, becoming unreachable dead code. This caused:
- Workbench OpenAI run submissions to fail with "No saved API key" even
  when a saved credential existed, forcing users to fall back to
  local_ollama (which triggers the approval queue)
- Ops approval of OpenAI runs to fail with "Provide a one-time API key"
  even when the requester had a saved credential

Worker local mode silently aband
[…]
