### 2026-07-07 4c4e8ba
Cancelling now requires UI confirmation
