+++ scripts/issue_tracker_mvp/normalize_issue_occurrences.py
@@ -1,0 +1,546 @@
+#!/usr/bin/env python3
+"""Add actor- and object-explicit canonical issues without re-extracting reports."""
+
+from __future__ import annotations
+
+import argparse
+import hashlib
+import json
+import os
+from concurrent.futures import ThreadPoolExecutor, as_completed
+from contextlib import contextmanager
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import Any, Iterable
+
+import pandas as pd
+from tqdm import tqdm
+
+import build_issue_index as pipeline
+
+
+NORMALIZATION_VERSION = "issue-normalization-v1"
+DEFAULT_SCHEMA_PATH = Path(__file__).with_name("issue_normalization_schema_v1.json")
+
+SYSTEM_PROMPT = """You normalize already-extracted safety issues from Prevention of Future Death reports.
+Return only JSON matching the requested schema. Use British English.
+
+This is a representational normalization task, not a new extraction task. Preserve the meaning,
+failure direction, responsible party, and level of certainty in the supplied evidence. Do not add a
+new concern, causal claim, actor, or fact. Do not medicalise neutral or non-medical concerns.
+"""
+
+USER_PROMPT = """Normalize every supplied issue. Return exactly one output for every issue_id, in the
+same order, with no additions or omissions.
+
+Derive fields in this order before writing canonical_issue:
+
+1. responsible_actor_role: a short, generic role responsible for the failed action, condition, duty,
+   system, or decision. Prefer roles such as "mental health service", "prison healthcare team",
+   "local authority", "highway authority", "employer", "manufacturer", or "regulator". Remove names
+   of people and organisations. Treat a non-"not_stated" input responsible_actor_type as supported
+   schema information: if no narrower generic role is supported, use its readable generic label
+   (for example "provider organisation", "government or public authority", or "individual
+   practitioner"). Use "not stated" only when responsible_actor_type is "not_stated" and no actor is
+   supported by responsible_actor_text or evidence_quote. The person exposed to harm is not
+   automatically the responsible actor.
+2. issue_object: a neutral, reusable noun phrase naming what failed or was affected, normally 2-12
+   words. It may be an action, information, service, policy, system, equipment, environment, decision,
+   duty, or other object. Examples include "risk assessment", "discharge information", "staffing
+   levels", "bridge barrier design", "product safety warning", and "custody observation policy".
+   Do not call every object a safeguard and do not assume a clinical setting.
+3. canonical_issue: at most {canonical_max_words} words, stating who failed, what failed, and how.
+   Include process, communication direction, recipient, or setting only where it changes the issue
+   type. When responsible_actor_role is not "not stated", canonical_issue MUST contain that exact
+   generic role phrase (normally as its grammatical subject). If the actor is "not stated", use a
+   clear passive formulation rather than inventing one.
+
+Direction and agency rules:
+- Never use "disengagement from services" to describe a service failing to contact, follow up,
+  escalate, offer reasonable adjustments, or respond to help-seeking.
+- Distinguish a person's voluntary refusal or non-attendance from a service's omission. Attribute
+  the action to the person only where the evidence explicitly does so.
+- For communication failures, identify the sending/receiving sides or recipient when supported.
+- Do not turn uncertainty about responsibility into an accusation.
+- State the problem, not a recommendation or desired remedy.
+
+The input facets may be imperfect. Resolve them conservatively against evidence_quote; do not copy a
[…]
