+++ scripts/issue_tracker_mvp/assign_issue_prototypes.py
@@ -1,0 +1,376 @@
+#!/usr/bin/env python3
+"""Conservatively assign ungrouped issue occurrences to reviewed prototypes."""
+
+from __future__ import annotations
+
+import argparse
+import json
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+import pandas as pd
+
+import build_issue_index as pipeline
+
+
+@dataclass
+class Prototype:
+    prototype_id: str
+    source: str
+    status: str
+    subject_domain: str
+    issue_ids: list[str]
+    report_ids: set[str]
+    centroid: np.ndarray
+
+
+def parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description="Assign ungrouped occurrences to recurring and emerging prototypes."
+    )
+    parser.add_argument("--run-dir", type=Path, required=True)
+    parser.add_argument("--tuning-dir", type=Path, default=None)
+    parser.add_argument("--repair-dir", type=Path, default=None)
+    parser.add_argument("--output-dir", type=Path, default=None)
+    parser.add_argument(
+        "--similarity",
+        type=float,
+        default=0.845,
+        help="Minimum cosine similarity; the reviewed default achieved 87.5%% precision.",
+    )
+    parser.add_argument(
+        "--margin",
+        type=float,
+        default=0.03,
+        help="Minimum lead over the second-best eligible prototype.",
+    )
+    parser.add_argument(
+        "--domain-match",
+        action=argparse.BooleanOptionalAction,
+        default=True,
+        help="Require occurrence and prototype subject_domain to match.",
+    )
+    return parser.parse_args()
+
+
+def _prototype(
+    prototype_id: str,
+    source: str,
+    status: str,
+    subject_domain: str,
+    issue_ids: list[str],
+    occurrences: pd.DataFrame,
+    embeddings: np.ndarray,
+    issue_index: dict[str, int],
+) -> Prototype:
+    indices = [issue_index[issue_id] for issue_id in issue_ids]
+    centroid = embeddings[indices].mean(axis=0)
+    centroid /= max(float(np.linalg.norm(centroid)), 1e-12)
+    return Prototype(
+        prototype_id=prototype_id,
+        source=source,
+        status=status,
+        subject_domain=subject_domain,
+        issue_ids=issue_ids,
+        report_ids=set(occurrences.iloc[indices]["report_key"]),
+        centroid=centroid,
+    )
+
+
+def build_prototypes(
+    occurrences: pd.DataFrame,
+    embeddings: np.ndarray,
+    candidate_groups: pd.DataFrame,
+    candidate_assignments: pd.DataFrame,
+    repaired_groups: pd.DataFrame,
+    repaired_assignments: pd.DataFrame,
+) -> list[Prototype]:
+    issue_index = {
+        issue_id: position for position, issue_id in enumerate(occurrences["issue_id"])
+    }
+    prototypes: list[Prototype] = []
+    repaired = repaired_groups[
+        repaired_groups["recurrence_status"].isin(["recurring", "emerging"])
+    ]
+    for row in repaired.itertuples(index=False):
+        issue_ids = repaired_assignments.loc[
+            repaired_assignments["repaired_group_id"] == row.repaired_group_id, "issue_id"
+        ].tolist()
+        prototypes.append(
+            _prototype(
+                row.repaired_group_id,
+                "review_repair",
+                row.recurrence_status,
+                row.subject_domain,
+                issue_ids,
+                occurrences,
+                embeddings,
+                issue_index,
+            )
+        )
+    emerging = candidate_groups[candidate_groups["recurrence_status"] == "emerging"]
+    for row in emerging.itertuples(index=False):
+        issue_ids = candidate_assignments.loc[
[…]

[…]
