### 2026-07-23 f743583
Create adjudication component
