+++ tests/test_issue_index_pipeline.py
@@ -17,12 +17,362 @@ SCRIPT_DIR = Path(__file__).resolve().parents[1] / "scripts" / "issue_tracker_mv
 sys.path.insert(0, str(SCRIPT_DIR))
 
 import build_issue_index as pipeline  # noqa: E402
+import adjudicate_recurring_issue_groups as adjudication  # noqa: E402
 import assign_issue_prototypes as prototype_assignment  # noqa: E402
+import build_issue_registry as issue_registry  # noqa: E402
 import build_recurring_issue_audit as quality_audit  # noqa: E402
 import normalize_issue_occurrences as normalization  # noqa: E402
 import repair_reviewed_issue_groups as repair  # noqa: E402
+import run_full_issue_tracker as full_workflow  # noqa: E402
 import score_recurring_issue_audit as audit_scoring  # noqa: E402
 import tune_issue_index as tuning  # noqa: E402
+
+
+def test_targeted_risk_gate_flags_directional_and_action_conflicts():
+    groups = pd.DataFrame(
+        [
+            {
+                "subissue_id": "sub_test",
+                "recurrence_status": "recurring",
+                "report_count": 4,
+                "issue_count": 4,
+                "median_centroid_similarity": 0.97,
+            }
+        ]
+    )
+    indexed = pd.DataFrame(
+        [
+            {
+                "subissue_id": "sub_test",
+                "canonical_issue": "The service failed to escalate a referral",
+                "issue_object": "referral escalation",
+                "failure_state": "omitted",
+                "process_stage": "referral",
+                "communication_direction": "professional_to_professional",
+                "responsible_actor_role": "provider organisation",
+            },
+            {
+                "subissue_id": "sub_test",
+                "canonical_issue": "The receiving team failed to process the referral",
+                "issue_object": "referral processing",
+                "failure_state": "delayed",
+                "process_stage": "follow_up",
+                "communication_direction": "between_organisations",
+                "responsible_actor_role": "team",
+            },
+        ]
+    )
+
+    result = adjudication.build_risk_gate(
+        groups, indexed, risk_threshold=3, low_cohesion=0.955
+    ).iloc[0]
+
+    assert bool(result["flagged_for_adjudication"])
+    assert bool(result["direction_conflict"])
+    assert bool(result["lifecycle_conflict"])
+    assert bool(result["semantic_action_conflict"])
+
+
+def test_recurrence_boundary_alone_does_not_trigger_adjudication():
+    groups = pd.DataFrame(
+        [
+            {
+                "subissue_id": "sub_tight",
+                "recurrence_status": "recurring",
+                "report_count": 3,
+                "issue_count": 3,
+                "median_centroid_similarity": 0.98,
+            }
+        ]
+    )
+    indexed = pd.DataFrame(
+        [
+            {
+                "subissue_id": "sub_tight",
+                "canonical_issue": "The employer omitted an automated external defibrillator",
+                "issue_object": "automated external defibrillator",
+                "failure_state": "omitted",
+                "process_stage": "equipment_availability",
+                "communication_direction": "not_applicable",
+                "responsible_actor_role": "employer",
+            }
+            for _ in range(3)
+        ]
+    )
+
+    result = adjudication.build_risk_gate(
+        groups, indexed, risk_threshold=3, low_cohesion=0.955
+    ).iloc[0]
+
+    assert result["risk_reasons"] == "recurrence_boundary"
[…]
