+++ scripts/issue_tracker_mvp/adjudicate_recurring_issue_groups.py
@@ -1,0 +1,969 @@
+#!/usr/bin/env python3
+"""Risk-gate, adjudicate, repair, and label recurring issue groups."""
+
+from __future__ import annotations
+
+import argparse
+import json
+import re
+from collections import Counter
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+import pandas as pd
+from tqdm import tqdm
+
+import build_issue_index as pipeline
+
+
+DECISIONS = {"accept", "exclude", "split", "reject"}
+RISK_GATE_VERSION = "targeted-risk-v2"
+GENERIC_OBJECT_TERMS = {
+    "assessment",
+    "communication",
+    "documentation",
+    "information",
+    "records",
+    "record keeping",
+    "referral",
+    "resources",
+    "response",
+    "staff training",
+    "training",
+}
+UNINFORMATIVE_VALUES = {"", "not stated", "not_stated", "not applicable", "not_applicable"}
+OBJECT_STOPWORDS = {
+    "and",
+    "care",
+    "clinical",
+    "communication",
+    "documentation",
+    "for",
+    "information",
+    "management",
+    "of",
+    "patient",
+    "process",
+    "provider",
+    "record",
+    "recording",
+    "records",
+    "response",
+    "service",
+    "staff",
+    "system",
+    "the",
+    "to",
+}
+DOCUMENTATION_STAGES = {"information_record_management"}
+ACTION_STAGE_FAMILIES = {
+    "documentation": DOCUMENTATION_STAGES,
+    "escalation": {"escalation"},
+    "referral": {"referral"},
+    "follow_up": {"follow_up"},
+    "training": {"training_delivery"},
+    "performance": {
+        "assessment",
+        "emergency_response",
+        "medication_administration",
+        "monitoring_observation",
+        "professional_practice",
+        "treatment_care_delivery",
+    },
+}
+ACTION_PATTERNS = {
+    "documentation": re.compile(r"\b(record|recorded|recording|document|documented|notes?|charts?)\b"),
+    "performance": re.compile(
+        r"\b(perform|performed|carry out|carried out|conduct|conducted|undertake|undertaken|continue|continued)\b"
+    ),
+    "initiate": re.compile(r"\b(refer\w*|referral|escalat\w*|request(?:ed)? review)\b"),
+    "respond": re.compile(r"\b(respond|response|act(?:ed)? on|chase|process(?:ed|ing)?|attend)\b"),
+    "obtain": re.compile(r"\b(obtain|seek|consult|liaise)\b"),
+    "communicate": re.compile(
+        r"\b(communicat\w*|inform\w*|notif\w*|hand(?: |-)?over|bring to .*attention)\b"
+    ),
+}
+CONFLICTING_ACTION_PAIRS = {
+    frozenset(("documentation", "performance")),
+    frozenset(("initiate", "respond")),
+    frozenset(("obtain", "communicate")),
+}
+
+ADJUDICATION_SYSTEM_PROMPT = """You adjudicate proposed recurring issue groups from UK Prevention of Future Deaths reports.
+
+A recurring issue must describe the same directional failure in at least three distinct reports. Shared vocabulary or subject matter alone is insufficient. Treat these as materially different unless the concrete obligation is the same:
+- performing an action versus documenting it;
+- initiating escalation/referral versus responding to it;
+- obtaining information/advice versus communicating existing information/advice;
+- communication to different recipients where responsibility changes;
+- generic training, records, resources, response, referral, or communication with different objects.
+
+Choose exactly one decision:
+- accept: every member supports one precise recurring issue;
+- exclude: one coherent core remains after removing explicit outliers;
+- split: members form two or more directionally distinct groups;
+- reject: no precise three-report recurring core is established.
+
[…]

[…]
