### 2026-07-30 e36ff13
add relational issue linkage and PFD feasibility workflows
