+++ tests/test_issue_index_pipeline.py
@@ -23,6 +23,8 @@ import build_issue_registry as issue_registry  # noqa: E402
 import build_recurring_issue_audit as quality_audit  # noqa: E402
 import normalize_issue_occurrences as normalization  # noqa: E402
 import repair_reviewed_issue_groups as repair  # noqa: E402
+import prepare_relational_linkage_pilot as relational_pilot  # noqa: E402
+import run_relational_linkage_experiment as relational_linkage  # noqa: E402
 import run_full_issue_tracker as full_workflow  # noqa: E402
 import score_recurring_issue_audit as audit_scoring  # noqa: E402
 import tune_issue_index as tuning  # noqa: E402
@@ -1505,13 +1507,17 @@ def test_normalization_preserves_requested_order_and_trims_fields():
         {
             "issue_id": "i2",
             "responsible_actor_role": "the local highway authority",
+            "failed_action": "inspect and maintain",
             "issue_object": "road barrier inspection and maintenance arrangements",
+            "counterparty_role": "road users",
             "canonical_issue": "The local highway authority did not adequately inspect and maintain road barriers before foreseeable vehicle impacts occurred",
         },
         {
             "issue_id": "i1",
             "responsible_actor_role": "community mental health service",
+            "failed_action": "contact",
             "issue_object": "follow-up contact",
+            "counterparty_role": "patient",
             "canonical_issue": "The community mental health service did not make follow-up contact after the patient sought help",
         },
     ]
@@ -1528,7 +1534,9 @@ def test_normalization_rejects_missing_or_unknown_issue_ids():
         {
             "issue_id": "unknown",
             "responsible_actor_role": "employer",
+            "failed_action": "assess",
             "issue_object": "workplace risk assessment",
+            "counterparty_role": "worker",
             "canonical_issue": "The employer omitted a workplace risk assessment",
         }
     ]
@@ -1543,7 +1551,9 @@ def test_normalization_recovers_unique_near_match_issue_id():
         {
             "issue_id": "iss_3c64aff0b03b479",
             "responsible_actor_role": "provider organisation",
+            "failed_action": "maintain",
             "issue_object": "clinical records",
+            "counterparty_role": "not stated",
             "canonical_issue": "Provider organisation failed to maintain clinical records",
         }
     ]
@@ -1566,7 +1576,9 @@ def test_normalization_requires_stated_actor_in_canonical_issue():
         {
             "issue_id": "i1",
             "responsible_actor_role": "mental health service",
+            "failed_action": "contact",
             "issue_object": "follow-up contact",
+            "counterparty_role": "patient",
             "canonical_issue": "Follow-up contact was not attempted",
         }
     ]
@@ -1618,7 +1630,9 @@ def test_normalized_output_preserves_original_issue_and_evidence():
                 {
                     "issue_id": "i1",
                     "responsible_actor_role": "mental health team",
+                    "failed_action": "contact",
                     "issue_object": "follow-up contact",
+                    "counterparty_role": "patient",
                     "canonical_issue": "The mental health team omitted follow-up contact",
                 }
             ],
@@ -1658,7 +1672,9 @@ def test_normalized_output_preserves_results_from_partial_batch():
                 {
                     "issue_id": "i1",
                     "responsible_actor_role": "provider organisation",
+                    "failed_action": "maintain",
                     "issue_object": "clinical records",
+                    "counterparty_role": "not stated",
                     "canonical_issue": "Provider organisation maintained incomplete clinical records",
                     "normalization_warnings": [],
                 }
[…]
