### 2026-08-03 4ec8f12
Add topic taxonomy loader and relational case tracing scripts
