+++ scripts/issue_tracker_mvp/research/automatic_parent_families/build_issue_family_recall_pilot.py
@@ -1,0 +1,372 @@
+#!/usr/bin/env python3
+"""Archived: build candidate samples for automatic issue-family auditing."""
+
+from __future__ import annotations
+
+import argparse
+import json
+import re
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+import pandas as pd
+
+
+PILOT_VERSION = "issue-family-recall-pilot-v1"
+
+
+@dataclass(frozen=True)
+class Family:
+    family_id: str
+    label: str
+    definition: str
+    inclusions: str
+    exclusions: str
+    strict_pattern: str
+    stages: tuple[str, ...] = ()
+    themes: tuple[str, ...] = ()
+
+
+FAMILIES = (
+    Family(
+        "continuity_coordination",
+        "Failures in continuity and coordination of care",
+        (
+            "Service-side failures that interrupt, fragment, or leave unclear "
+            "responsibility for care across staff, teams, organisations, settings, "
+            "transfers, or time."
+        ),
+        (
+            "Discontinuity after transfer or discharge; poorly coordinated or "
+            "fragmented care; unclear ownership; absent joined-up working."
+        ),
+        (
+            "Staff-retention continuity alone; generic poor treatment with no "
+            "continuity or coordination element; record defects unless they caused "
+            "fragmented care."
+        ),
+        r"\bcontinuity\b|joined[- ]up care|fragmented care|"
+        r"(?:fail\w*|poor\w*|inadequate\w*|lack\w*)\W+(?:to )?coordinat\w*|"
+        r"unclear responsibility for (?:ongoing )?care",
+        stages=("care_coordination",),
+        themes=("communication_handover",),
+    ),
+    Family(
+        "records_information",
+        "Medical and care-record failures",
+        (
+            "Failures to create, complete, maintain, integrate, retain, access, "
+            "review, or use medical, clinical, nursing, care, or custody health records."
+        ),
+        (
+            "Incomplete, inaccurate, delayed, inaccessible, fragmented, unavailable, "
+            "or unread records and record systems."
+        ),
+        (
+            "General verbal communication with no record component; unrelated "
+            "commercial or administrative records; a clinical action merely mentioned "
+            "in a record when the record itself did not fail."
+        ),
+        r"\b(?:medical|clinical|nursing|patient|care|health|custody) records?\b|"
+        r"\brecord[- ]keeping\b|\bclinical documentation\b|\bcare notes?\b",
+        stages=("information_record_management",),
+        themes=("records_information",),
+    ),
+    Family(
+        "staffing_capacity",
+        "Insufficient staffing capacity and workforce availability",
+        (
+            "Insufficient numbers, availability, deployment, skill mix, recruitment, "
+            "or retention of staff to operate a service safely."
+        ),
+        (
+            "Understaffing, vacancies, unsafe rotas, workload pressure, unavailable "
+            "specialist staff, recruitment and retention failures."
+        ),
+        (
+            "Training or competence alone; individual misconduct; equipment or bed "
+            "capacity without a workforce shortage."
+        ),
+        r"\bunderstaff\w*|\bstaff(?:ing)? (?:levels?|shortage|capacity|numbers?|"
+        r"availability|vacanc\w*|rota)|\bworkforce (?:shortage|capacity|vacanc\w*|"
+        r"recruit\w*|retention)|\binsufficient (?:numbers? of )?staff\b",
+        stages=("workforce_management",),
+        themes=("staffing_capacity",),
+    ),
+    Family(
+        "discharge_transitions",
[…]
