+++ scripts/issue_tracker_mvp/research/automatic_parent_families/expand_issue_families.py
@@ -1,0 +1,525 @@
+#!/usr/bin/env python3
+"""Archived: expand tight automatic issue-family cores with auditable overlap.
+
+Precise recurring groups remain unchanged. Families are seeded from a strict
+average-linkage cut, while child groups may receive a limited number of
+secondary family memberships when their centroid is close to the alternative
+core and the attachment has independent multi-resolution or facet evidence.
+"""
+
+from __future__ import annotations
+
+import argparse
+import json
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+import pandas as pd
+
+import discover_issue_families as discovery
+
+
+EXPANSION_VERSION = "overlapping-issue-family-expansion-v1"
+
+
+def parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument("--occurrences-csv", required=True, type=Path)
+    parser.add_argument("--embeddings-npy", required=True, type=Path)
+    parser.add_argument("--assignments-csv", required=True, type=Path)
+    parser.add_argument("--groups-csv", required=True, type=Path)
+    parser.add_argument("--output-dir", required=True, type=Path)
+    parser.add_argument("--core-similarity", type=float, default=0.84)
+    parser.add_argument("--broad-similarity", type=float, default=0.82)
+    parser.add_argument("--attachment-similarity", type=float, default=0.90)
+    parser.add_argument("--maximum-primary-gap", type=float, default=0.04)
+    parser.add_argument("--maximum-secondary-families", type=int, default=2)
+    parser.add_argument("--minimum-core-children", type=int, default=2)
+    parser.add_argument("--review-family-limit", type=int, default=40)
+    return parser.parse_args()
+
+
+def values(value: Any) -> set[str]:
+    return set(discovery.split_pipe_values([value]))
+
+
+def build_cores(
+    child_features: pd.DataFrame,
+    child_centroids: np.ndarray,
+    strict_labels: np.ndarray,
+    broad_labels: np.ndarray,
+    *,
+    minimum_core_children: int,
+) -> tuple[list[dict[str, Any]], dict[int, int]]:
+    cores: list[dict[str, Any]] = []
+    child_owner: dict[int, int] = {}
+    for strict_label in sorted(set(strict_labels)):
+        positions = np.where(strict_labels == strict_label)[0]
+        if len(positions) < minimum_core_children:
+            continue
+        centroid = discovery.unit_normalize(child_centroids[positions].mean(axis=0))
+        similarities = child_centroids[positions] @ centroid
+        medoid_position = positions[int(np.argmax(similarities))]
+        family_id = discovery.stable_family_id(
+            child_features.iloc[positions]["child_group_id"].astype(str)
+        )
+        broad_counts = pd.Series(broad_labels[positions]).value_counts()
+        broad_label = int(broad_counts.index[0])
+        core_index = len(cores)
+        for position in positions:
+            child_owner[int(position)] = core_index
+        cores.append(
+            {
+                "family_id": family_id,
+                "strict_label": int(strict_label),
+                "broad_label": broad_label,
+                "positions": positions,
+                "centroid": centroid,
+                "medoid_position": int(medoid_position),
+                "themes": set().union(
+                    *(
+                        values(value)
+                        for value in child_features.iloc[positions]["dominant_themes"]
+                    )
+                ),
+                "stages": set().union(
+                    *(
+                        values(value)
+                        for value in child_features.iloc[positions][
+                            "dominant_process_stages"
+                        ]
+                    )
+                ),
+            }
+        )
+    return cores, child_owner
+
+
+def attachment_rows(
[…]

[…]
